% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{angelier1990}
\alias{angelier1990}
\title{Fault-Slip Data}
\format{
a list containing four objects, each of class \code{Fault}
}
\usage{
data('angelier1990')
}
\description{
List of four different fault-slip data from Angelier (1990):
}
\details{
\code{"AVB"} (Agia Varvara, Crete, Greece), \code{"KAM"} (Kamogawa, Boso Peninsula, Central Japan),
\code{"MD1"} (Arroyo Montado, Baja California, Mexico), and \code{"TYM"} (Tymbaki, Crete, Greece).
}
\examples{
data("angelier1990")
head(angelier1990)
}
\references{
Angelier, J. (1990). Inversion of field data in fault tectonics to obtain the
regional stress-III. A new rapid direct inversion method by analytical means.
Geophysical Journal International, 103(2), 363–376. \doi{10.1111/j.1365-246X.1990.tb01777.x}
}
\seealso{
Other datasets:
\code{\link{example_lines}},
\code{\link{example_lines_df}},
\code{\link{example_planes}},
\code{\link{example_planes_df}},
\code{\link{gray_example}},
\code{\link{holst}},
\code{\link{hossack1968}},
\code{\link{osmundsen2010}},
\code{\link{ramsay}},
\code{\link{shebandowan}},
\code{\link{simongomez}},
\code{\link{strabo_prj}}
}
\concept{datasets}
\keyword{datasets}
