% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_math.R
\name{antipode}
\alias{antipode}
\alias{antipode.default}
\alias{antipode.Vec3}
\alias{antipode.Line}
\alias{antipode.Ray}
\alias{antipode.Plane}
\alias{antipode.Pair}
\alias{antipode.Fault}
\title{Antipode vector}
\usage{
antipode(x, ...)

\method{antipode}{default}(x, ...)

\method{antipode}{Vec3}(x, ...)

\method{antipode}{Line}(x, ...)

\method{antipode}{Ray}(x, ...)

\method{antipode}{Plane}(x, ...)

\method{antipode}{Pair}(x, ...)

\method{antipode}{Fault}(x, ...)
}
\arguments{
\item{x}{a spherical object}

\item{...}{arguments passed to function call}
}
\description{
Returns the opposite vector
}
\examples{
Line(120, 55) |> antipode()
Ray(120, 55) |> antipode()
}
