% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faults.R
\name{apparent_dip_direction}
\alias{apparent_dip_direction}
\title{Apparent dip direction}
\usage{
apparent_dip_direction(p, apparent_dip)
}
\arguments{
\item{p}{object of class \code{"Plane"}}

\item{apparent_dip}{angle in degrees}
}
\value{
Azimuth in degrees
}
\description{
Apparent dip direction
}
\examples{
apparent_dip_direction(Plane(63, 45), 10)
}
