% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assign-color}
\alias{assign-color}
\alias{assign_col_d}
\alias{assign_col}
\alias{assign_col_binned}
\alias{legend_col}
\alias{legend_col_d}
\title{Helper functions to assign plotting colors to a vector}
\usage{
assign_col_d(
  x,
  pal = colorblind_pal,
  na.translate = TRUE,
  na.values = "grey",
  ...
)

assign_col(
  x,
  n = length(x),
  pal = viridis::viridis,
  na.translate = TRUE,
  na.values = "grey",
  ...
)

assign_col_binned(x, breaks = 5L, pal = viridis::viridis, ...)

legend_col(breaks, title = NULL, pal = viridis::viridis, cex = 1, ...)

legend_col_d(fill, legend = names(fill), position = "topright", ...)
}
\arguments{
\item{x}{vector to colorize}

\item{pal}{color function; Default is \code{\link[viridis:viridis]{viridis::viridis()}}}

\item{na.translate}{Unlike continuous scales, discrete scales can easily show
missing values, and do so by default. If you want to remove missing values
from a discrete scale, specify \code{na.translate = FALSE}.}

\item{na.values}{If \code{na.translate = TRUE}, what color should the
missing values be displayed as?}

\item{...}{arguments passed to color function}

\item{n}{integer. The number of colors (\eqn{\ge}1) to be in the palette.}

\item{breaks}{integer giving the desired number of intervals. Non-integer values are rounded down.}

\item{title}{character. Legend title}

\item{cex}{character expansion factor relative to current par("cex").
Used for text in legend.}

\item{fill}{color vector}

\item{legend}{character vector. Names of discrete colors.
Can be ignored when \code{cols} is a named vector.}

\item{position}{Legend position. Either a two-column vector of the x and y coordinates, or a
keyword from the list \code{"bottomright"}, \code{"bottom"}, \code{"bottomleft"}, \code{"left"},
\code{"topleft"}, \code{"top"}, \code{"topright"}, \code{"right"} and \code{"center"}.}
}
\value{
character vector of colors in hexadecimal code
}
\description{
Helper functions to assign plotting colors to a vector
}
\examples{
set.seed(20250411)

# example for discrete colors
x <- rvmf(5, mu = Line(120, 50), k = 5)
key <- letters[round(runif(5, 1, 26))]
key[2] <- NA
plot(x, col = assign_col_d(key), grid.params = list(guides = FALSE))
legend_col_d(assign_col_d(key))

# example for continuous colors:
x <- rvmf(100, mu = Line(120, 50), k = 5)
cols <- runif(100)
cols[sample.int(length(cols), size = 10)] <- NA
plot(x, col = assign_col(cols), grid.params = list(guides = FALSE))
legend_col(seq(0, 1, .1), title = "test")
}
\seealso{
\code{\link[PlotTools:SpectrumLegend]{PlotTools::SpectrumLegend()}} - an alternative tool to generate a
color-bar in base-R plots

Other assign:
\code{\link{assign-cex}},
\code{\link{assign-pch}}
}
\concept{assign}
