% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assign-pch}
\alias{assign-pch}
\alias{assign_pch}
\alias{assign_pch_binned}
\alias{legend_pch}
\title{Assigns plotting characters (pch values) to a vector}
\usage{
assign_pch(x, solid = TRUE)

assign_pch_binned(x, solid = TRUE, breaks = 6)

legend_pch(x, solid = TRUE, breaks = NULL, position = "topright", ...)
}
\arguments{
\item{x}{vector.}

\item{solid}{Should the plotting character be solid, \code{TRUE} (the default), or hollow, \code{FALSE}?}

\item{breaks}{integer giving the desired number of intervals. Non-integer values are rounded down.}

\item{position}{Legend position. Either a two-column vector of the x and y coordinates, or a
keyword from the list \code{"bottomright"}, \code{"bottom"}, \code{"bottomleft"}, \code{"left"},
\code{"topleft"}, \code{"top"}, \code{"topright"}, \code{"right"} and \code{"center"}.}

\item{...}{arguments passed to \code{graphics::legend()}}
}
\value{
named integer vector
}
\description{
\code{assign_pch()} maps discrete variables to six easily discernible shapes. If you have more
than six levels, you will get a warning message, and the seventh and
subsequent levels will not appear on the plot.
You can not map a continuous variable to shape unless \code{assign_pch_binned()} is
used.
}
\examples{
set.seed(20250411)

# example for discrete colors
x <- rvmf(5, mu = Line(120, 50), k = 5)
key <- sample(letters, 5, replace = TRUE)
plot(x, pch = assign_pch(key), grid.params = list(guides = FALSE))
legend_pch(key)
}
\seealso{
Other assign:
\code{\link{assign-cex}},
\code{\link{assign-color}}
}
\concept{assign}
