% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{azimuth_to_cardinal}
\alias{azimuth_to_cardinal}
\title{Converts azimuth angles into Cardinal directions}
\usage{
azimuth_to_cardinal(x, n_directions = 8L)
}
\arguments{
\item{x}{angles in degree.}

\item{n_directions}{either 4, for 4-points (N, E, S, W), 8 for 8-point (N, NE, E, …) or 6 for 16-point (N, NNE, NE, …) cardinal version.}
}
\value{
character vector
}
\description{
Converts azimuth angles into Cardinal directions
}
\examples{
azimuth_to_cardinal(c(0, 23, 45, 100, 190, 270, 350), 4) # 8-point compass
azimuth_to_cardinal(c(0, 23, 45, 100, 190, 270, 350), 8) # 8-point compass
azimuth_to_cardinal(c(0, 23, 45, 100, 190, 270, 350), 16) # 16-point compass
}
\seealso{
Other parse-orientations:
\code{\link[=Fault_from_rake_quadrant]{Fault_from_rake_quadrant()}},
\code{\link[=quadrant2dd]{quadrant2dd()}},
\code{\link[=sense_from_strikeslip]{sense_from_strikeslip()}},
\code{\link{split}},
\code{\link[=strikeslip_kinematics]{strikeslip_kinematics()}}
}
\concept{parse-orientations}
