% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_plots.R
\name{balebrun_plot}
\alias{balebrun_plot}
\title{Dip-Pitch-Plunge Triangular Diagram}
\usage{
balebrun_plot(
  x,
  labels = NULL,
  main = "Ternary Fabric Orientation",
  extra_labels = TRUE,
  add = FALSE,
  ...
)
}
\arguments{
\item{x}{object of class \code{"Pair"}}

\item{labels}{character. Text labels}

\item{main}{character. The title of the plot.}

\item{extra_labels}{logical}

\item{...}{optional plotting parameters passed to \code{\link[graphics:points]{graphics::points()}}}
}
\value{
matrix of plotting parameters
}
\description{
Ternary fabric orientation diagram after Balé and Brun (1989) showing the
pitch and plunge of stretching lineation X) and the dip of the foliation plane (XY).
}
\examples{
data(angelier1990)
balebrun_plot(angelier1990$TYM)

lapply(seq(0, 90, 10), function(i) {
  p.i <- Pair_from_pitch(Plane(120, 0), i)
  invisible(balebrun_plot(p.i, as.character(i), col = i / 9, add = i != 0))
})

balebrun_plot(Pair_from_pitch(Plane(0, 80), 80), "vertical flow", col = 1, add = F)
balebrun_plot(Pair_from_pitch(Plane(0, 80), 10), "strike-slip", col = 2, add = T)
balebrun_plot(Pair_from_pitch(Plane(0, 10), 10), "horizontal flow", col = 2, add = T)
}
\references{
Balé, P., & Brun, J.-P. (1989). Late Precambrian thrust and wrench zones in
northern Brittany (France). Journal of Structural Geology, 11(4), 391–405.
https://doi.org/10.1016/0191-8141(89)90017-5
}
