% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_pole.R
\name{best_fit}
\alias{best_fit}
\alias{regression_greatcircle}
\alias{regression_smallcircle}
\title{Least-Squares Geodesic Regression of Small and Great Circles for Spherical Data}
\usage{
regression_greatcircle(
  x,
  val = seq_len(nrow(x)),
  iterations = 1000L,
  n_points = 100L
)

regression_smallcircle(
  x,
  val = seq_len(nrow(x)),
  num_seeds = 5L,
  iterations = 1000L,
  n_points = 100L
)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}, where the
rows are the observations and the columns are the coordinates.}

\item{val}{A vector of real numbers. Values of independent scalar variables. Same length as \code{x}.}

\item{iterations}{A real number (positive integer). A bound on the number of numerical optimization iterations allowed.}

\item{n_points}{A real number (positive integer). The number of points along the regression curve requested.}

\item{num_seeds}{A real number (positive integer)}
}
\value{
list. \describe{
\item{\code{vec}}{axis of best-fit great- or small-circle. Class as \code{x}}
\item{\code{cone}}{half-apical angle of the best-fit small-circle (in radians if \code{x} is of class \code{"Vec3"}, degrees otherwise)}
\item{\code{convergence}}{An integer code. \code{0} indicates successful completion.}
\item{\code{min_eigenvalue}}{Positive if successful completion}
\item{\code{r_squared}}{\eqn{R^2} value of the regression}
\item{\code{points}}{A set of Vec3 points for the regression line}
\item{\code{range}}{Range of the regression line (in radians if \code{x} is of class \code{"Vec3"}, degrees otherwise)}
}
}
\description{
Fits a great circle arc to a set of lines, based on an independent scalar variable.
}
\examples{
set.seed(20250411)
data("gray_example")
bestgc_clea <- regression_greatcircle(gray_example[1:8, ])
bestgc_bedd <- regression_greatcircle(gray_example[9:16, ])

bestsc_clea <- regression_smallcircle(gray_example[1:8, ])
bestsc_bedd <- regression_smallcircle(gray_example[9:16, ])
bestsc_all <- regression_smallcircle(gray_example)

stereoplot()
points(gray_example[1:8, ], col = "blue")
points(gray_example[9:16, ], col = "red", pch = "x")

# best for cleavage
lines(bestsc_clea$vec, bestsc_clea$cone, col = "lightblue")
points(bestsc_clea$vec, col = "lightblue", pch = 16)
lines(bestgc_clea$vec, lty = 2, col = "blue")
points(bestgc_clea$vec, col = "blue", pch = 17)

# best for bedding
lines(bestsc_bedd$vec, bestsc_bedd$cone, col = "sienna")
points(bestsc_bedd$vec, col = "sienna", pch = 16)
lines(bestgc_bedd$vec, lty = 2, col = "red")
points(bestgc_bedd$vec, col = "red", pch = 17)
}
\references{
Davis, J.R. and Titus, S.J. (2017). Modern methods of analysis
for three-dimensional orientational data. Journal of Structural Geology, 96,
65–89. \doi{10.1016/j.jsg.2017.01.002}
}
\seealso{
Other geodesic-regression:
\code{\link{perm-rsq}}
}
\concept{geodesic-regression}
