% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{bingham_statistics}
\alias{bingham_statistics}
\title{Elliptical concentration and confidence cone estimation}
\source{
Borradaile, G. (2003). Spherical-Orientation Data. In: Statistics of
Earth Science Data. Springer, Berlin, Heidelberg. https://doi.org/10.1007/978-3-662-05223-5_10
}
\usage{
bingham_statistics(x, w = NULL)
}
\arguments{
\item{x}{numeric. Can be three element vector, three column array, or an
object of class \code{"line"} or \code{"plane"}}

\item{w}{numeric. Weights}
}
\value{
list
\describe{
\item{\code{k}}{two-column vector containing the estimates for the minimum and maximum concentration (\eqn{\kappa}).}
\item{\code{a95}}{two-column vector containing the estimates for the minimum and maximum 95\% confidence cone.}
}
}
\description{
Elliptical concentration and confidence cone estimation
}
\examples{
x <- rfb(100, mu = Line(120, 50), k = 5, A = diag(c(-1, 0, 1)))
bingham_statistics(x)
}
