% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortensor.R
\name{center}
\alias{center}
\title{Centering vectors}
\usage{
center(x, max_vertical = FALSE)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"},
where the rows are the observations and the columns are the coordinates.}

\item{max_vertical}{Whether the maximum of the von Mises-Fisher distribution
is already vertical or not.}
}
\value{
Object of class of \code{x}
}
\description{
Rotate vector object to position that ortensor eigenvectors are parallel to
axes of coordinate system: E3||X (north-south), E2||X(east-west),
E1||X(vertical).
Useful when one wants to inspect the distribution of vectors, especially when vectors plot
near the perimeter of the stereonet.
}
\examples{
set.seed(1)
mu <- Line(120, 10)
x <- rkent(100, mu = mu, k = 20, b = 5)
x_centered <- center(x)

# plot results
plot(x, col = "grey")
points(x_centered, col = "#B63679", pch = 16)
legend("topright", legend = c("original", "centered"), col = c("grey", "#B63679"), pch = 16)
}
\seealso{
\code{\link[=ot_eigen]{ot_eigen()}}
}
