% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_mean.R
\name{class-rot}
\alias{class-rot}
\alias{pair2rot}
\alias{is.Rotation}
\alias{as.Rotation}
\alias{rot2pair}
\title{Orientation matrix from fault planes and slip directions}
\usage{
pair2rot(p)

is.Rotation(x)

as.Rotation(x)

rot2pair(x, fault = FALSE)
}
\arguments{
\item{p}{object of class \code{"Pair"} or \code{"Fault"}}

\item{x}{object of class \code{"Rotation"}, a 3x3 matrix}

\item{fault}{logical. Whether to coerces to a fault or a pair object.}
}
\value{
list of rotation matrices
}
\description{
Converts a set of planes and lines into a list of rotation matrices in SO(3).
}
\examples{
my_fault <- Fault(
  c("a" = 120, "b" = 120, "c" = 100),
  c(60, 60, 50),
  c(110, 25, 30),
  c(58, 9, 23),
  c(1, -1, 1)
)
pair2rot(my_fault)
}
