% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{colorblind_pal}
\alias{colorblind_pal}
\title{Colorblind Color Palette (Discrete) and Scales}
\usage{
colorblind_pal(n)
}
\arguments{
\item{n}{number of colors to extract. Gives a warning when \code{n} exceeds 8.}
}
\value{
character vector with color values
}
\description{
Colorblind Color Palette (Discrete) and Scales
}
\examples{
colorblind_pal(3)
}
