% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{colorize}
\alias{colorize}
\alias{assign_col_d}
\alias{assign_col}
\alias{assign_col_binned}
\alias{legend_c}
\alias{legend_d}
\title{Helper functions for color assignment and legends}
\usage{
assign_col_d(x, pal = viridis::viridis, ...)

assign_col(x, n = length(x), pal = viridis::viridis, ...)

assign_col_binned(x, breaks, pal = viridis::viridis, ...)

legend_c(breaks, title = NULL, pal = viridis::viridis)

legend_d(fill, labels = names(fill), position = "topright", ...)
}
\arguments{
\item{x}{vector to colorize}

\item{pal}{color function; Default is \code{\link[viridis:reexports]{viridis::viridis()}}}

\item{...}{arguments passed to color function}

\item{n}{integer.}

\item{breaks}{numeric.}

\item{title}{character. Legend title}

\item{fill}{color vector}

\item{labels}{character.vector. Names of discrete colors.
Can be ignored when \code{cols} is a named vector.}

\item{position}{Legend position}
}
\value{
.
}
\description{
Helper functions for color assignment and legends
}
\examples{
set.seed(1234)

# example for discrete colors
x <- rvmf(5, mu = Line(120, 50), k = 5)
key <- letters[round(runif(5, 1, 26))]
stereoplot(guides = FALSE)
stereo_point(x, col = assign_col_d(key))
legend_d(assign_col_d(key))

# example for continuous colors:
x <- rvmf(100, mu = Line(120, 50), k = 5)
stereoplot(guides = FALSE)
stereo_point(x, col = assign_col(runif(100)))
legend_c(seq(0, 1, .1), title = "test")
}
