% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion_wissi.R
\name{combine_weights}
\alias{combine_weights}
\title{Combine multiple weight vectors multiplicatively and normalise.}
\usage{
combine_weights(...)
}
\description{
Each element of ... is a numeric vector of length N.
Returns a normalised weight vector (mean = 1).
}
