% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coordinates}
\alias{coordinates}
\alias{line2vec}
\alias{plane2vec}
\alias{vec2line}
\alias{vec2plane}
\title{Coordinate conversion}
\usage{
line2vec(l)

plane2vec(p)

vec2line(v)

vec2plane(v)
}
\arguments{
\item{l, p}{two column matrix, or two element vector. Spherical coordinates
(in degrees) of lines (azimuth, plunge) or plane (dir direction, dip).}

\item{v}{three column matrix, or three-element vector. Cartesian coordinates}
}
\description{
Coordinate conversion
}
\examples{
cbind(c(90, 89), c(45, 46)) |>
  lin2vec() |>
  vec2lin()
}
