% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faults.R
\name{correct_pair}
\alias{correct_pair}
\title{Orthogonalization of plane and line measurement}
\usage{
correct_pair(x)
}
\arguments{
\item{x}{Pair or Fault}
}
\value{
list with the orthogonalized plane and line measurements
(as 3d vectors) and the misfit angles (in radians)
}
\description{
Orthogonalization of plane and line measurement
}
\examples{
p <- Pair(120, 60, 110, 58)
correct_pair(p)
}
