% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deftensor.R
\name{defgrad}
\alias{defgrad}
\alias{defgrad_from_ratio}
\alias{defgrad_from_pair}
\alias{defgrad_from_vectors}
\alias{defgrad_from_axisangle}
\alias{defgrad_from_comp}
\title{Deformation Gradient Tensor}
\usage{
defgrad_from_ratio(Rxy = 1, Ryz = 1)

defgrad_from_pair(p)

defgrad_from_vectors(v1, v2)

defgrad_from_axisangle(axis, angle)

defgrad_from_comp(
  xx = 1,
  xy = 0,
  xz = 0,
  yx = 0,
  yy = 1,
  yz = 0,
  zx = 0,
  zy = 0,
  zz = 1
)
}
\arguments{
\item{Rxy, Ryz}{numeric. the XY and YZ strain ratio to create a strain tensor
with axial stretches.Values must be greater than or equal to 1.}

\item{p}{object of class \code{pair}}

\item{v1, v2}{objects of class \code{"spherical"} or three-element vector.
Deformation gradient results from the rotation around axis perpendicular to
both vectors to rotate \code{v1} to \code{v2}.}

\item{axis, angle}{rotation axis and angle, axis can be an object of class
\code{"spherical"} (incl. \code{"line"} and \code{"plane"}) or a three-element vector. Angle in
degrees when axis is a object of class \code{"spherical"}, and radians otherwise.}

\item{xx, xy, xz, yx, yy, yz, zx, zy, zz}{numeric. Directly specify components of
the tensor. Identity matrix by default.}
}
\value{
3x3 matrix.
}
\description{
Deformation Gradient Tensor
}
\examples{
defgrad_from_ratio(2, 3)
defgrad_from_axisangle(Line(120, 30), 45)
defgrad_from_vectors(Line(120, 30), Line(210, 60))
defgrad_from_pair(Pair(40, 20, 75, 16))
}
