% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contouring.R
\name{density_grid}
\alias{density_grid}
\title{Estimates point density}
\usage{
density_grid(x, FUN = exponential_kamb, sigma = 3, n = 100, weights = NULL)
}
\arguments{
\item{x}{spherical object}

\item{n}{(optional) numeric. The size of the grid that the density is estimated on. Defaults to \code{100}.}

\item{arguments}{passed to \code{\link[=count_points]{count_points()}}}
}
\value{
list with the cartesian coordinates of the grid and density values of the regularly gridded density estimates.
}
\description{
Estimates point density of the given linear orientation measurements
Returns a regular (in lat-long space) grid of density estimates over a
hemispherical surface.
}
\examples{
x <- Line(c(120, 315, 86), c(22, 85, 31))
density_grid(x, n = 100)
}
