\name{dict_repl}
\alias{dict_repl}
\title{Function for representing hashed objects as strings}
\usage{
  dict_repl(object, obj_name)
}
\arguments{
  \item{object}{an arbitrary thing}

  \item{obj_name}{name of the variable as defined by the
  user (not currently being used)}
}
\description{
  Purely visual.
}

