% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stats.R
\name{dist-sphere}
\alias{dist-sphere}
\alias{dist}
\title{Angular distance matrix for orientation vectors}
\usage{
dist(x, ...)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, or \code{"Plane"}.}

\item{...}{optional parameters passed to \code{\link[=stats:as.dist]{stats:as.dist()}}}
}
\value{
distance matrix
}
\description{
This function computes and returns the distance matrix computed by using the
Cosine similarity to compute the distances between the rows of a data matrix.
}
\examples{
set.seed(20250411)
dist(rvmf(100, mu = Line(90, 0), k = 20))
}
