% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortensor.R
\name{.or_eigen_helper}
\alias{.or_eigen_helper}
\title{Helper function for Eigenvalues and Eigenvectors of a Set of Vectors}
\usage{
.or_eigen_helper(x, scaled = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}, where the
rows are the observations and the columns are the coordinates.}

\item{scaled}{logical. Whether the Eigenvectors should be scaled by the
Eigenvalues (only effective if \code{x} is in Cartesian coordinates).}

\item{...}{additional arguments passed to \code{\link[=ortensor]{ortensor()}}.}
}
\description{
Helper function for Eigenvalues and Eigenvectors of a Set of Vectors
}
\keyword{internal}
