% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabeta.R
\name{drillcore}
\alias{drillcore}
\alias{drill_plane}
\alias{drill_line}
\title{Orientation of structures from drill core orientation angles}
\usage{
drill_plane(alpha, beta, azi, inc)

drill_line(alpha, beta, gamma, azi, inc)
}
\arguments{
\item{alpha}{numeric vector. Alpha angle in degrees}

\item{beta}{numeric vector. Beta angle in degrees}

\item{azi}{numeric. Azimuth of drill core axis orientation in degrees (measured from North).}

\item{inc}{numeric. Inclination of drill core axis in degrees (negative values for downward!).}

\item{gamma}{numeric. Gamma angle in degrees}
}
\value{
numeric two-column matrix with Dip direction and dip angle of plane or
azimuth and plunge of line (in degrees).
}
\description{
Calculates the orientation of a plane or line from alpha, beta and gamma
angles in oriented drill cores
}
\examples{
drill_plane(45, 220, 225, -45)
drill_plane(60, 320, 225, -45)

# multiple alpha-beta measurements
drill_plane(alpha = c(45, 60), beta = c(220, 320), azi = 225, inc = -45)
}
