% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabeta.R
\name{drillcore_orientation}
\alias{drillcore_orientation}
\title{Orientation of structures from drill core orientation angles}
\usage{
drillcore_orientation(azi, inc, alpha, beta, gamma = NULL)
}
\arguments{
\item{azi}{numeric. Azimuth of drill core axis orientation in degrees
(measured clockwise from North).}

\item{inc}{numeric. Inclination of drill core axis in degrees
(Note: \strong{negative values for downward}).}

\item{alpha}{numeric vector. Alpha angle in degrees}

\item{beta}{numeric vector. Beta angle in degrees}

\item{gamma}{numeric. (optional). Gamma angle in degrees}
}
\value{
object of class \code{"plane"}. If gamma is specified, \code{"line"} object is
returned.
}
\description{
Calculates the orientation of a plane or line from internal core angles (
alpha, beta, and gamma) of oriented drill cores
}
\examples{
azi <- 225
inc <- -45
drillcore_orientation(azi, inc, 60, 320)
drillcore_orientation(azi, inc, 45, 220)

# multiple alpha-beta measurements
stereoplot()
stereo_point(Line(azi, -inc), lab = "CA")
drillcore_orientation(azi, inc, alpha = c(60, 45), beta = c(320, 220)) |>
  stereo_point(lab = c("A", "B"))
}
