% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid.R
\name{ellipsoid-class}
\alias{ellipsoid-class}
\alias{is.ellipsoid}
\alias{as.ellipsoid}
\alias{ellipsoid}
\alias{ellipsoid.default}
\alias{ellipsoid.defgrad}
\title{Ellipsoid class}
\usage{
is.ellipsoid(x)

as.ellipsoid(x)

ellipsoid(x, left, ...)

\method{ellipsoid}{default}(x, left = NULL, ...)

\method{ellipsoid}{defgrad}(x, left = TRUE, ...)
}
\arguments{
\item{x}{Either a matrix or a \code{"defgrad"} object}

\item{left}{logical. \code{TRUE}  for left Cauchy–Green deformation tensor (or
Finger deformation tensor, the default), \code{FALSE} for for right Cauchy–Green deformation tensor (or Green’s deformation tensor)}

\item{...}{optional parameters passed to function call.}
}
\value{
\code{is.ellipsoid} returns \code{TRUE} if \code{x} is an \code{"ellipsoid"} object, and \code{FALSE} otherwise.

\code{as.ellipsoid} coerces a 3x3 matrix into an \code{"ellipsoid"} object.
}
\description{
In deformation analysis, the quadratic forms of the three-dimensional stretches
are represented by the \code{ellipsoid} class.
It can be used to represents either ellipsoid objects or finite strain ellipsoids.
}
\details{
The eigenvalues \eqn{\lambda} of the deformation matrix are the quadratic
forms of the principal stretches \eqn{s} (\eqn{s = 1 + \epsilon = l/l_0}).
}
\examples{
test <- as.ellipsoid(diag(3))
is.ellipsoid(test)

R <- defgrad_from_ratio(2, 3)
ellipsoid(R)
}
\seealso{
\code{\link[=ortensor]{ortensor()}}

Other ellipsoid:
\code{\link{ellipsoid-params}},
\code{\link[=ellipsoid_from_stretch]{ellipsoid_from_stretch()}}
}
\concept{ellipsoid}
