% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid.R
\name{ellipsoid-params}
\alias{ellipsoid-params}
\alias{volume}
\alias{volume.default}
\alias{volume.ellipsoid}
\alias{volume.ortensor}
\alias{lode}
\alias{lode.default}
\alias{lode.ellipsoid}
\alias{lode.ortensor}
\alias{nadai}
\alias{nadai.default}
\alias{nadai.ellipsoid}
\alias{nadai.ortensor}
\alias{jelinek}
\alias{jelinek.default}
\alias{jelinek.ellipsoid}
\alias{jelinek.ortensor}
\alias{flinn}
\alias{flinn.default}
\alias{flinn.ortensor}
\alias{flinn.ellipsoid}
\alias{size_invariant}
\alias{size_invariant.default}
\alias{size_invariant.ortensor}
\alias{size_invariant.ellipsoid}
\alias{strain_invariant}
\alias{strain_invariant.default}
\alias{strain_invariant.ortensor}
\alias{strain_invariant.ellipsoid}
\alias{shape_invariant}
\alias{shape_invariant.default}
\alias{shape_invariant.ortensor}
\alias{shape_invariant.ellipsoid}
\alias{kind}
\alias{kind.default}
\alias{kind.ortensor}
\alias{kind.ellipsoid}
\title{Ellipsoid shape parameters}
\usage{
volume(x, ...)

\method{volume}{default}(x, ...)

\method{volume}{ellipsoid}(x, ...)

\method{volume}{ortensor}(x, ...)

lode(x, ...)

\method{lode}{default}(x, ...)

\method{lode}{ellipsoid}(x, ...)

\method{lode}{ortensor}(x, ...)

nadai(x, ...)

\method{nadai}{default}(x, ...)

\method{nadai}{ellipsoid}(x, ...)

\method{nadai}{ortensor}(x, ...)

jelinek(x, ...)

\method{jelinek}{default}(x, ...)

\method{jelinek}{ellipsoid}(x, ...)

\method{jelinek}{ortensor}(x, ...)

flinn(x, ...)

\method{flinn}{default}(x, ...)

\method{flinn}{ortensor}(x, ...)

\method{flinn}{ellipsoid}(x, ...)

size_invariant(x, ...)

\method{size_invariant}{default}(x, ...)

\method{size_invariant}{ortensor}(x, ...)

\method{size_invariant}{ellipsoid}(x, ...)

strain_invariant(x, ...)

\method{strain_invariant}{default}(x, ...)

\method{strain_invariant}{ortensor}(x, ...)

\method{strain_invariant}{ellipsoid}(x, ...)

shape_invariant(x, ...)

\method{shape_invariant}{default}(x, ...)

\method{shape_invariant}{ortensor}(x, ...)

\method{shape_invariant}{ellipsoid}(x, ...)

kind(x, ...)

\method{kind}{default}(x, ...)

\method{kind}{ortensor}(x, ...)

\method{kind}{ellipsoid}(x, ...)
}
\arguments{
\item{x}{numeric. either a 3-element vector giving the ellipsoid's semi-axis
lengths (in any order), an object of class \code{"ellipsoid"}, or an object of class \code{"ortensor"}.}

\item{...}{optional arguments passed to \code{\link[=ortensor]{ortensor()}}}
}
\value{
positive numeric
}
\description{
Ellipsoid shape parameters
}
\details{
The natural strain is
\deqn{\bar{\epsilon}_i = \log s_i = \log(1 + \epsilon_i)} with \eqn{s_1 \geq s_2 \geq s_3} the semi-axis lengths of the ellipsoid, or
\eqn{\epsilon_i} the strains (elongation) given by \eqn{\epsilon = \frac{l-l_0}{l_0}} (hence \eqn{s = \frac{l}{l_0}}).

Lode's parameter for \strong{strain symmetry}:
\deqn{\nu = \frac{2 \bar{\epsilon}_2 - \bar{\epsilon}_1 - \bar{\epsilon}_3}{\bar{\epsilon}_1-\bar{\epsilon}_3}}
with \eqn{\bar{\epsilon}_1 \geq \bar{\epsilon}_2 \geq \bar{\epsilon}_3}. Note that \eqn{\nu} is undefined for spheres,
but we arbitrarily declare \eqn{\nu=0} for them (plane strain).
Otherwise \eqn{-1 \geq \nu \geq 1}. \eqn{\nu=-1} for prolate spheroids (constriction) and \eqn{\nu=1} for oblate spheroids (flattening).

\strong{Octahedral shear strain} \eqn{\bar{\epsilon}_s} (Nádai 1963):
\deqn{\bar{\epsilon}_s = \sqrt{\frac{(\bar{\epsilon}_1 - \bar{\epsilon}_2)^2 + (\bar{\epsilon}_2 - \bar{\epsilon}_3)^2 + (\bar{\epsilon}_1 - \bar{\epsilon}_3)^2 }{3}}}
This is the amount of strain assuming coaxial deformation (pure-shear).

\strong{Strain symmetry} (Flinn 1963):
\deqn{k = \frac{s_1/s_2 - 1}{s_2/s_3 - 1}}
The value ranges from 0 to \eqn{\infty}, and is 0 for oblate ellipsoids
(flattening), 1 for plane strain and  \eqn{\infty} for prolate ellipsoids (constriction).

and \strong{strain intensity} (Flinn 1963):
\deqn{d = \sqrt{(s_1/s_2 - 1)^2 + (s_2/s_3 - 1)^2}}
This is analogous to Nadai's strain parameter.

Jelinek (1981)'s \eqn{P_j} parameter:
\deqn{P_j = \bar{\epsilon}^{\sqrt{2 \vec{v}\cdot \vec{v}}}}
with \eqn{\vec{v} = \bar{\epsilon}_i - \frac{\sum \bar{\epsilon}_i}{3}}
}
\examples{
# Generate some random orientation data
set.seed(20250411)
dat <- rvmf(100, k = 20)
s <- principal_stretch(dat)

# Volume of ellipsoid
volume(s)

#  Size-related tensor invariant of ellipsoids
size_invariant(s)

# Strain-related tensor invariant of ellipsoids
strain_invariant(s)

# Shape-related tensor invariant of ellipsoids
shape_invariant(s)

# Lode's shape parameter for the strain symmetry ratio
lode(s)

# Nadai's octahedral shear strain
nadai(s)

# Jelinek Pj parameter
jelinek(s)

# Flinn's intensity and symmetry parameters
flinn(s)

kind(s)

# Ellipsoid data
hossack_ell <- lapply(seq.int(nrow(hossack1968)), function(i) {
  ellipsoid_from_stretch(hossack1968[i, 3], hossack1968[i, 2], hossack1968[i, 1])
})
sapply(hossack_ell, nadai)
}
\references{
Flinn, Derek.(1963): "On the statistical analysis of fabric diagrams."
Geological Journal 3.2: 247-253.

Lode, Walter (1926): "Versuche über den Einfluß der mittleren Hauptspannung
auf das Fließen der Metalle Eisen, Kupfer und Nickel“
(\emph{"Experiments on the influence of the mean principal stress on the flow of
the metals iron, copper and nickel"}], Zeitschrift für Physik, vol. 36 (November),
pp. 913–939, \doi{10.1007/BF01400222}

Nádai, A., and Hodge, P. G., Jr. (1963): "Theory of Flow and Fracture of Solids,
vol. II." ASME. J. Appl. Mech. December 1963; 30(4): 640. \doi{10.1115/1.3636654}

Jelinek, Vit. "Characterization of the magnetic fabric of rocks."
Tectonophysics 79.3-4 (1981): T63-T67.
}
\seealso{
\code{\link[=shape_params]{shape_params()}}, \code{\link[=ot_eigen]{ot_eigen()}}, \code{\link[=hsu_plot]{hsu_plot()}}, \code{\link[=flinn_plot]{flinn_plot()}}

Other ellipsoid:
\code{\link{ellipsoid-class}},
\code{\link[=ellipsoid_from_stretch]{ellipsoid_from_stretch()}}
}
\concept{ellipsoid}
