\name{encapsulate}
\alias{encapsulate}
\title{Helper function for making character vectors have quotes around each item when
printed to the console.}
\usage{
  encapsulate(values)
}
\arguments{
  \item{values}{a vector of values}
}
\description{
  Helper function for making character vectors have quotes
  around each item when printed to the console.
}

