% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohR.R
\name{failure_criterion}
\alias{failure_criterion}
\title{Failure Criterion}
\usage{
failure_criterion(
  sigma_n = seq(-100, 100, 0.01),
  cohesion = 1,
  friction = 0.6,
  ...
)
}
\arguments{
\item{cohesion}{numeric. Cohesion}

\item{friction}{numeric. Coefficient of friction}

\item{...}{optional plotting arguments passed to \code{\link[graphics:lines]{graphics::lines()}}}
}
\value{
a matrix of the normal and shear stresses.
}
\description{
Adds the Griffith-Coulomb-fracture criterion to a plot.
}
\examples{
Mohr_plot(sigma1 = 66, sigma2 = 30, sigma3 = 20, xlim = c(-50, 125), full.circle = TRUE)
failure_criterion(col = "red")
}
