% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{fault-plot}
\alias{fault-plot}
\alias{fault_plot}
\alias{hoeppener}
\alias{angelier}
\title{Plot Fault Data in a StereoPlot}
\usage{
fault_plot(x, type = c("angelier", "hoeppener"), ...)

hoeppener(x, pch = 1, col = "black", cex = 1, bg = NULL, points = TRUE, ...)

angelier(
  x,
  pch = 1,
  lwd = 1,
  lty = 1,
  col = "black",
  cex = 1,
  points = TRUE,
  bg = NULL,
  ...
)
}
\arguments{
\item{x}{\code{"Fault"} object where the rows are the observations, and the
columns the coordinates. Object must be complete, i.e. no \code{NA} values.
For Michael's, Angelier's, and Yamaji-Sato's methods, at least 4 rows
of fault measurements are required, while Hansen's method requires at least 7.}

\item{type}{character. One of \code{"angelier"} (for "\emph{Angelier plot}") or
\code{"hoeppener"} (for "\emph{Hoeppener plot}"). See details.}

\item{...}{arguments passed to \code{\link[=stereo_arrows]{stereo_arrows()}}}

\item{points}{logical. Whether the lineation points (\emph{Angelier plot}) or poles
(\emph{Hoeppener plot}) should be added to the plot}

\item{lty, lwd, cex, pch, col, bg}{plotting parameters}
}
\description{
Adds fault plane normals and slip directions to an existing a stereographic/equal-area
projection plot.
}
\details{
\strong{Angelier plot} shows all planes as \emph{great circles} and lineations as points.
Fault striae are plotted as vectors on top of the lineation pointing in the
movement direction of the hangingwall. Easy to read in case of homogeneous or
small datasets.

\strong{Hoeppener plot} shows all planes as \emph{poles} while lineations are not shown.
Fault striae are plotted as vectors on top of poles pointing in the movement
direction of the hangingwall. Useful in case of large or heterogeneous datasets.
Arrows point towards \eqn{\sigma_1}
}
\examples{
f <- Fault(
  c("a" = 120, "b" = 125, "c" = 100),
  c(60, 62, 50),
  c(110, 25, 30),
  c(58, 9, 23),
  c(1, -1, 1)
)

stereoplot(title = "Angelier plot")
angelier(f, col = 1:nrow(f), pch = 16, scale = 0.1)

stereoplot(title = "Hoeppener plot")
hoeppener(f, col = 1:nrow(f), cex = 1, scale = 0.1, points = FALSE)

# or
stereoplot()
fault_plot(f, type = "hoeppener", col = 1:nrow(f), cex = 1, scale = 0.1, points = FALSE)
}
\references{
Angelier, J. Tectonic analysis of fault slip data sets, J. Geophys. Res. 89 (B7), 5835-5848 (1984)

Hoeppener, R. Tektonik im Schiefergebirge. Geol Rundsch 44, 26-58 (1955). https://doi.org/10.1007/BF01802903
}
\seealso{
Other stereo-plot:
\code{\link{arrows}},
\code{\link{lines}},
\code{\link{plot-spherical}},
\code{\link[=points.spherical]{points.spherical()}},
\code{\link{stereo_cones}},
\code{\link[=stereo_confidence]{stereo_confidence()}},
\code{\link{stereo_contour}},
\code{\link[=stereo_lines]{stereo_lines()}},
\code{\link[=stereo_point]{stereo_point()}},
\code{\link[=stereo_segment]{stereo_segment()}},
\code{\link[=stereoplot]{stereoplot()}},
\code{\link[=stereoplot_guides]{stereoplot_guides()}},
\code{\link[=stereoplot_ticks]{stereoplot_ticks()}},
\code{\link[=text.spherical]{text.spherical()}}
}
\concept{stereo-plot}
