% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion.R
\name{fault_instability_criterion}
\alias{fault_instability_criterion}
\title{Fault Instability Criterion}
\usage{
fault_instability_criterion(fault, R, friction = 0.6)
}
\arguments{
\item{fault}{\code{"Fault"} object where the rows are the observations, and the columns the coordinates.}

\item{R}{numeric. Stress ratio after Gephart and Forsyth (1984): \eqn{(\sigma_1 - \sigma_2)/(\sigma_1 - \sigma_3)}}

\item{friction}{numeric. Coefficient of friction (0.6 by default)}
}
\value{
numeric. Instability ranges from 0 (most stable) to 1 (most unstable).
The most unstable fault is the optimally oriented fault for shear faulting.
}
\description{
Calculates the instability criterion \eqn{I} after Vavrycuk (2013, Eq. 3).
Instability ranges from 0 (most stable) to 1 (most unstable)
The most unstable fault is the optimally oriented fault for shear faulting.
}
\details{
\deqn{I = \frac{\tau - \mu(\sigma - \sigma_1)}{\tau_c - \mu(\sigma_c - \sigma_1)}}

where \eqn{\tau_c} and \eqn{\sigma_c} are the shear traction and effective
normal traction along the optimally oriented fault, and \eqn{\tau} and \eqn{\sigma}
are the shear traction and effective normal traction along the analysed fault plane.
}
\examples{
f <- angelier1990$TYM
s <- reduced_stress(f)
r <- stress_shape(s)

fault_instability_criterion(f, r$R)
}
\references{
Vavrycuk, V., Bouchaala, F. & Fischer, T., 2013. High-resolution
fault image from accurate locations and focal mechanisms of the 2008 swarm
earthquakes in West Bohemia, Czech Republic, Tectonophysics, 590, 189–195.
}
\seealso{
Other stress-tensor:
\code{\link[=reduced_stress]{reduced_stress()}},
\code{\link{sigma-comp}},
\code{\link[=sigma2rup]{sigma2rup()}},
\code{\link[=sigma2stress]{sigma2stress()}},
\code{\link[=stress_shape]{stress_shape()}}
}
\concept{stress-tensor}
