% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faults.R
\name{fault_rake}
\alias{fault_rake}
\title{Rake of a fault}
\usage{
fault_rake(x)
}
\arguments{
\item{x}{fault object}
}
\value{
numeric. Angle in degrees
}
\description{
Angle between fault slip vector and fault strike.
}
\examples{
f <- Fault(c(120, 120, 100), c(60, 60, 50), c(110, 25, 30), c(58, 9, 23), c(1, -1, 1))
fault_rake(f)

f2 <- Pair(c(120, 120, 100), c(60, 60, 50), c(110, 25, 30), c(58, 9, 23))
fault_rake(f2)
}
