% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{fisher_ftest}
\alias{fisher_ftest}
\title{Test of mean orientations}
\usage{
fisher_ftest(x, y, alpha = 0.05)
}
\arguments{
\item{x, y}{Can be three element vectors, three column arrays, or objects of class \code{"line"} or \code{"plane"}}

\item{alpha}{Significance level}
}
\value{
list indicating the F-statistic and the p-value.
}
\description{
Test against the null-hypothesis that the samples are drawn from the same Fisher population.
}
\examples{
set.seed(1234)
x <- rvmf(100, mu = Line(120, 50), k = 20)
y <- rvmf(100, mu = Line(180, 45), k = 20)

ggstereo() +
  geom_point(data = gg(x), aes(x, y, color = "x")) +
  geom_point(data = gg(y), aes(x, y, color = "y"))

fisher_ftest(x, y)
}
