\docType{package}
\name{foldr}
\alias{foldr}
\alias{foldr-package}
\title{dicts and lists in R}
\description{
  \code{foldr} provides functionality similar to dicts
  and lists in Python and Ruby. Everything in \code{foldr} is
  object oriented and copies much of the API from Python and Ruby. 
}
\details{

  Creating a list

  blah blah blah

  Merging lists

  use the + operator

  Creating a dict

  blah blah blah

  Iteration

  blah bla blah; pitfalls, shortcomings, features...

  When not to use them

  Don't use them with large sets of data...

}
\references{
  blog post by Yhat
  \url{http://blog.yhathq.com/a-blog-post/}.
}