% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_mean.R
\name{geodesic-line}
\alias{geodesic-line}
\alias{geodesic_mean_line}
\alias{geodesic_var_line}
\alias{geodesic_meanvariance_line}
\title{The Frechet (geodesic \eqn{L^2}) mean of a set of lines}
\source{
lineMeanVariance from geologyGeometry (J.R. Davis)
}
\usage{
geodesic_mean_line(x, ...)

geodesic_var_line(x, ...)

geodesic_meanvariance_line(x, seeds = 5L, steps = 100L)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, or \code{"Plane"}}

\item{...}{parameters passed to \code{\link[=geodesic_meanvariance_line]{geodesic_meanvariance_line()}}}

\item{seeds}{positive integer. How many \code{x} to try as seeds}

\item{steps}{positive integer. Bound on how many iterations to use.}
}
\value{
\code{geodesic_meanvariance_line} returns a \code{list} consisting of
\verb{$variance} (numeric), \verb{$mean} (a line),
\verb{$error} (an integer) and \verb{$min.eigenvalue} (numeric).
\code{geodesic_mean_line} and \code{geodesic_var_line} are convenience wrapper and only
return the mean and the variance, respectively.
}
\description{
An iterative algorithm for computing the Frechet mean — the line that
minimizes the Frechet variance. The iterations continue until error squared of
epsilon is achieved or \code{steps} iterations have been used. Try multiple
seeds, to improve your chances of finding the global optimum.
}
\details{
Error should be \code{0} and \code{min.eigenvalue} should be positive.
Otherwise there was some problem in the optimization. If error is non-zero, then try increasing \code{steps}.
}
\examples{
geodesic_meanvariance_line(example_lines)
geodesic_mean_line(example_lines)
geodesic_var_line(example_lines)
}
\references{
Davis, J. R., & Titus, S. J. (2017). Modern methods of analysis
for three-dimensional orientational data. Journal of Structural Geology,
96, 65–89. https://doi.org/10.1016/j.jsg.2017.01.002
}
\seealso{
\code{\link[=sph_mean]{sph_mean()}} for arithmetic mean and \code{\link[=geodesic_mean_pair]{geodesic_mean_pair()}} for geodesic mean of pairs.
}
