% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_mean.R
\name{geodesic-var}
\alias{geodesic-var}
\alias{geodesic_var}
\alias{geodesic_var.Vec3}
\alias{geodesic_var.Line}
\alias{geodesic_var.Plane}
\alias{geodesic_var.Ray}
\alias{geodesic_var.Pair}
\title{The Fréchet (geodesic \eqn{L^2}) variance}
\source{
geologyGeometry (J.R. Davis)
}
\usage{
geodesic_var(x, ...)

\method{geodesic_var}{Vec3}(x, y = NULL, ...)

\method{geodesic_var}{Line}(x, y = NULL, ...)

\method{geodesic_var}{Plane}(x, y = NULL, ...)

\method{geodesic_var}{Ray}(x, y = NULL, ...)

\method{geodesic_var}{Pair}(x, y = NULL, group = NULL, ...)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"}.}

\item{...}{parameters passed to \code{\link[=geodesic_meanvariance_ray]{geodesic_meanvariance_ray()}} (if \code{x} is a Ray), \code{\link[=geodesic_meanvariance_line]{geodesic_meanvariance_line()}} (if \code{x} is a Vec3, Line or Plane)
or \code{\link[=geodesic_mean_pair]{geodesic_mean_pair()}} (if \code{x} is a Pair or a Fault).}

\item{y}{Only for variance. object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"} about which the Fréchet variance should be calculated for.
If \code{NULL} (the default), Fréchet variance about the Fréchet mean.}

\item{group}{character. Symmetry group of \code{x}. One of
\code{"orthorhombic"} (line-in-plane symmetry, e.g. foliation-lineations, cylindrical fold
orientations, triaxial ellipsoid orientations, and earthquake
focal mechanisms, and olivine),
\code{"triclinic"} (ray-in-plane symmetry, e.g. faults with slip directions),
\code{"trigonal"} (e.g. alpha-quartz),
\code{"hexagonal"} (e.g. beta-quartz), or
\code{"trivial"} (rotations).
If \code{NULL}, the group will be
automatically picked based on the class of \code{x}.}
}
\value{
the Fréchet variance as a numeric number. Because distances in SO(3) never exceed \eqn{\pi}, the maximum possible variance
is \eqn{\frac{\pi^2}{2} \approx 4.93}.
}
\description{
Dispersion  measured using the Fréchet variance, i.e the sum of the squared
geodesic distances between all vectors and a specified vector.
}
\details{
The variance of a dataset \eqn{{x_1, \ldots, x_n}} about a vector \eqn{y} is defined as
\deqn{ \Psi(x) = \frac{1}{2n} \sum_{i=1}^n d_G(y, x_i)^2}
where \eqn{d_G(x, y)} is the geodesic distance between vectors \eqn{x} and \eqn{y} (see \code{\link[=angle]{angle()}}).
}
\examples{
set.seed(20250411)
geodesic_var(example_planes, example_planes[1, ])
geodesic_var(example_planes)
}
\references{
Davis, J. R., & Titus, S. J. (2017). Modern methods of analysis
for three-dimensional orientational data. Journal of Structural Geology,
96, 65–89. https://doi.org/10.1016/j.jsg.2017.01.002
}
\seealso{
\code{\link[=geodesic_mean]{geodesic_mean()}} for the Fréchet mean, \code{\link[=sph_mean]{sph_mean()}} for the arithmetic mean, \code{\link[=projected_mean]{projected_mean()}} for projected mean
}
