% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_gg.R
\name{ggframe}
\alias{ggframe}
\title{Stereoplot Perimeter}
\usage{
ggframe(n = 10000, color = "black", fill = NA, lwd = 1, ...)
}
\arguments{
\item{n}{integer. Resolution of line.}

\item{color, fill, lwd}{Graphical parameters}

\item{...}{optional graphical parameters passed to \code{\link[ggplot2:geom_polygon]{ggplot2::geom_polygon()}}}
}
\description{
Adds a frame to the stereographic projection
}
