% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_stereonet.R
\name{ggstereo}
\alias{ggstereo}
\title{Stereonet using ggplot}
\usage{
ggstereo(
  earea = TRUE,
  centercross = TRUE,
  grid = FALSE,
  grid.spacing = 10,
  grid.rot = 0,
  ...
)
}
\arguments{
\item{earea}{logical. Whether the projection is equal-area ("Schmidt net")
(\code{TRUE}, the default), or equal-angle ("Wulff net") (\code{FALSE}).}

\item{grid.spacing}{numeric. Grid spacing in degree}

\item{grid.rot}{numeric. Angle (in degrees) to rotate the grid.}

\item{...}{argument passed to \code{\link[ggplot2:geom_polygon]{ggplot2::geom_polygon()}}}

\item{logical.}{Whether a centercross should be added to the plot.}

\item{grid.}{Whether a gird should be drawn.}
}
\value{
ggplot
}
\description{
Stereonet using ggplot
}
\examples{
test_data <- rbind(rvmf(100, mu = Line(90, 45), k = 10), rvmf(50, mu = Line(0, 0), k = 20)) |> as.line()
ggstereo(grid = TRUE) +
  geom_point(data = gg(test_data), aes(x = x, y = y))

ggstereo(earea = FALSE, centercross = TRUE) +
  geom_point(data = gg(test_data), aes(x = x, y = y))
}
