% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defgrad.R
\name{gradient}
\alias{gradient}
\alias{velgrad}
\alias{defgrad}
\title{Velocity gradient and Deformation gradient tensors}
\usage{
velgrad(R, time = 1)

defgrad(V, time = 1, steps = 1)
}
\arguments{
\item{R}{deformation gradient tensor. object of class \code{matrix} or \code{defgrad}}

\item{time}{numeric. total time (default is 1)}

\item{V}{velocity gradient tensor. object of class \code{matrix} or \code{velgrad}}

\item{steps}{numeric. time increments (default is 1)}
}
\value{
object of class \code{velgrad} or \code{defgrad}. If steps is > 1, then a list of \code{defgrad} objects is returned.
}
\description{
Calculates the velocity gradient tensor as the matrix logarithm  of the deformation gradient tensor divided by given time, and
the deformation gradient tensor accumulated after some time.
}
\examples{
D <- defgrad_from_comp(xx = 2, xy = 1, zz = 0.5)
L <- velgrad(D, time = 10)
L
defgrad(L, time = 10, steps = 2)
}
