% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-S3.R
\name{head-sphere}
\alias{head-sphere}
\alias{head.spherical}
\alias{tail.spherical}
\title{Return the First or Last Parts of an Object}
\usage{
\method{head}{spherical}(x, n = 6L, ...)

\method{tail}{spherical}(x, n = 6L, ...)
}
\arguments{
\item{x}{objects of class \code{"Vec3"}, \code{"Line"}, \code{"Plane"}, \code{"Pair"}, or \verb{"Fault}}

\item{n}{an integer vector of length up to \code{dim(x)} (or 1,
    for non-dimensioned objects). Values specify the indices to be
    selected in the corresponding dimension (or along the length) of the
    object. A positive value of \code{n[i]} includes the first/last
    \code{n[i]} indices in that dimension, while a negative value
    excludes the last/first \code{abs(n[i])}, including all remaining
    indices. \code{NA} or non-specified values (when \code{length(n) <
      length(dim(x))}) select all indices in that dimension. Must
    contain at least one non-missing value.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
Returns the first or last parts of a vector.
}
\examples{
x <- rvmf(n = 10)
head(x)
tail(x)
}
