% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holst}
\alias{holst}
\title{3D Strain Data I}
\format{
An object of class \code{matrix}
}
\usage{
data('holst')
}
\description{
Example data from Holst and Fossen (1987) containing 3D strain data from 17 localities in the West Norwegian Caledonides.
The data gives the  difference of the  \eqn{\epsilon_1} and \eqn{\epsilon_2} (\code{e1e2}) and \eqn{\epsilon_2} and \eqn{\epsilon_3} (\code{e2e3})
}
\examples{
data("holst")
head(holst)
}
\seealso{
Other datasets:
\code{\link{angelier1990}},
\code{\link{example_lines}},
\code{\link{example_lines_df}},
\code{\link{example_planes}},
\code{\link{example_planes_df}},
\code{\link{gray_example}},
\code{\link{hossack1968}},
\code{\link{osmundsen2010}},
\code{\link{ramsay}},
\code{\link{shebandowan}},
\code{\link{simongomez}},
\code{\link{strabo_prj}}
}
\concept{datasets}
\keyword{datasets}
