% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_plots.R
\name{hsu_fabric_plot}
\alias{hsu_fabric_plot}
\title{Fabric plot of Hsü (1965)}
\usage{
hsu_fabric_plot(x, labels = NULL, add = FALSE, es.max = 3, ...)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}.}

\item{labels}{character. text labels}

\item{add}{logical. Should data be plotted to an existing plot?}

\item{es.max}{maximum strain for scaling.}

\item{...}{optional parameters passed to \code{\link[=hsu_plot]{hsu_plot()}}}
}
\value{
plot and when stored as object, a list containing the Lode parameter \code{lode} and the natural octahedral strain \code{es}.
}
\description{
Fabric plot of Hsü (1965)
}
\examples{
set.seed(20250411)
mu <- Line(120, 50)
x <- rvmf(100, mu = mu, k = 1)
hsu_fabric_plot(x, labels = "x")

set.seed(20250411)
y <- rvmf(100, mu = mu, k = 20)
hsu_fabric_plot(y, labels = "y", col = "red", add = TRUE)
}
