% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_analysis.R
\name{hypercontour}
\alias{hypercontour}
\title{Calculates densities for fabric and strain data}
\usage{
hypercontour(
  rphi,
  angfmt = c("deg", "rad", "grd"),
  proj = c("eqd", "eqa", "stg", "ort", "gno", "lin", "rdl", "rfp"),
  normalize = TRUE,
  rmax = NULL,
  kappa = 40,
  nnodes = 50L
)
}
\arguments{
\item{rphi}{matrix with two  named columns representing the \code{R} and \code{phi} values}

\item{angfmt}{character. Angle format, \code{'deg'} for degree, \code{'rad'} for radians, and \code{'grd'} for gradians.}

\item{proj}{character. Projection,  \code{'eqd'} for equidistant (Elliot plot),
\code{'eqa'} for equal-area, \code{'stg'} for stereographic, \code{'ort'} for orthographic,
\code{'gno'} for gnomonic, \code{'lin'} for exponential (linear R), \code{'rdl'} for radial,
\code{'rfp'} for Rf/phi (cylindrical instead of polar).}

\item{normalize}{logical.}

\item{rmax}{maximum R value (if \code{NULL}, computed automatically)}

\item{kappa}{smoothing parameter}

\item{nnodes}{grid resolution. higher is more accurate but slower, 30 is good, but 50 is recommended for final plots, default = 50.}
}
\value{
list with:
\itemize{
\item x, y: vectors of the projection coordinates for the density grid
\item z: density matrix
\item points: projected data points
\item frame: plot frame (circle or square)
}
}
\description{
Densities in hyperbaloidal projections of geological fabric and
finite strain data density calculations done on the unit hyperbaloid
(Vollmer, 2018). Options are given for equidistant (Elliott), equal-area,
stereographic, orthographic, exponential, and radial projections, as polar
azimuthal or cylindrical (cartesian, RfPhi-type) plots.
}
\details{
The data must be in a comma delimited csv text file with one (R, phi) pair
per line, where R = strain ratio (max/min), phi = orientation of long (max)
axis from x. Contours are equally spaced over the probability density
distribution.
}
\examples{
data(ramsay)
out <- hypercontour(ramsay, angfmt = "deg", proj = 'rfp')

image(out$x, out$y, out$z, asp = ifelse(out$proj != 'rfp', 1, NA),
 ylim = c(1, out$rmax),
 xlab = bquote(varphi~'('*degree*')'), ylab = 'R',
 main = out$proj, 
col = viridis::viridis(100)
)
contour(out$x, out$y, out$z, add = TRUE, lwd = .5)
points(out$points)


out <- hypercontour(ramsay, angfmt = "deg", proj = 'eqa')
levels <- pretty(range(out$z, na.rm = TRUE), 100)
cols <- viridis::viridis(n = length(levels) - 1)
plot(out$x, y = out$y, 'n', xlab = NULL, ylab = NULL, asp = 1, axes = FALSE, ann = FALSE)
graphics::.filled.contour(out$x, out$y, out$z, levels = levels, col = cols)
points(out$points)
lines( out$frame, lwd = 2)
title(main = out$proj)
}
\references{
Vollmer, F.W., 2018. Automatic contouring of geological fabric and finite
strain data on the unit hyperboloid. Computers & Geosciences,
https://doi.org/10.1016/j.cageo.2018.03.006
}
