% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortensor.R
\name{inertia}
\alias{inertia}
\alias{inertia_tensor.spherical}
\alias{inertia_tensor}
\title{Inertia tensor}
\usage{
\method{inertia_tensor}{spherical}(x, w = NULL)

inertia_tensor(x, w = NULL)
}
\arguments{
\item{x}{Object of class \code{"Vec3"}, \code{"Line"} or \code{"Plane"}}

\item{w}{numeric. weightings}
}
\value{
3 x 3 matrix
}
\description{
Inertia tensor
}
\details{
\deqn{D = n - (x_i, y_i, z_i) (x_i, y_i, z_i)^T}
}
\examples{
set.seed(20250411)
x <- rfb(100, mu = Line(120, 50), k = 1, A = diag(c(10, 0, 0)))
inertia_tensor(x)
}
\seealso{
\code{\link[=ortensor]{ortensor()}}
}
