% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{inertia_tensor}
\alias{inertia_tensor}
\title{Inertia tensor}
\usage{
inertia_tensor(x, w = NULL)
}
\arguments{
\item{x}{numeric. Can be three element vector, three column array, or an
object of class \code{"line"} or \code{"plane"}}

\item{w}{numeric. weightings}
}
\value{
3 x 3 matrix
}
\description{
Inertia tensor
}
\details{
\deqn{D = n - (x_i, y_i, z_i) (x_i, y_i, z_i)^T}
}
\examples{
set.seed(1)
x <- rfb(100, mu = Line(120, 50), k = 1, A = diag(c(10, 0, 0)))
inertia_tensor(x)
}
\seealso{
\code{\link[=ortensor]{ortensor()}}
}
