% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{lines.spherical}
\alias{lines.spherical}
\title{Add Lines to a Plot}
\usage{
\method{lines}{spherical}(x, ...)
}
\arguments{
\item{x}{objects of class \code{"Vec3"}, \code{"Line"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"}.}

\item{...}{arguments passed to \code{\link[graphics:lines]{graphics::lines()}}}
}
\description{
Add Lines to a Plot
}
\examples{
stereoplot()
lines(rvmf(n = 5), d = runif(5, 0, 90), col = 1:5)
}
