% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_mean.R
\name{mean-pair}
\alias{mean-pair}
\alias{geodesic_mean_pair}
\alias{geodesic_var_pair}
\title{Mean orientation of a set of pairs or faults}
\source{
oriMeanVariance from geologyGeometry (J.R. Davis)
}
\usage{
geodesic_mean_pair(x, group = NULL)

geodesic_var_pair(x, group = NULL)
}
\arguments{
\item{x}{object of class \code{"Pair"} or \code{"Fault"}}

\item{group}{Symmetry group of \code{x}. See \code{\link[=symmetry_group]{symmetry_group()}} for details
If \code{NULL}, the group will be
automatically picked based on the class of \code{x}.}
}
\value{
object of class \code{"Pair"} or \code{"Fault"}, respectively
}
\description{
The Frechet (geodesic \eqn{L^2}) mean and variance of a pair of foliations
and lineations
}
\examples{
my_fault <- Fault(
  c("a" = 120, "b" = 120, "c" = 100),
  c(60, 60, 50),
  c(110, 25, 30),
  c(58, 9, 23),
  c(1, -1, 1)
)
geodesic_mean_pair(my_fault)
geodesic_var_pair(my_fault)
}
\references{
Davis, J. R., & Titus, S. J. (2017). Modern methods of analysis
for three-dimensional orientational data. Journal of Structural Geology,
96, 65–89. https://doi.org/10.1016/j.jsg.2017.01.002
}
\seealso{
\code{\link[=sph_mean]{sph_mean()}} for arithmetic mean, and \code{\link[=geodesic_mean_line]{geodesic_mean_line()}} for geodesic mean of lines.
}
