% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_analysis.R
\name{mean_strain_ellipse}
\alias{mean_strain_ellipse}
\title{Mean strain ellipse}
\usage{
mean_strain_ellipse(r, phi, boot = TRUE, n_iter = 1000, boot.values = FALSE)
}
\arguments{
\item{r}{numeric. Aspect ratio of deformed object (long axis / short axis)}

\item{phi}{numeric. Orientation of long axis of deformed object (in degrees)}

\item{boot}{logical. Whether a 95\% confidence interval from on bootstrapping should be calculated. \code{TRUE} by default.}

\item{n_iter}{integer. Number of bootstrap resamples (\code{1000} by default). Ignored when \code{boot = FALSE}.}

\item{boot.values}{logical. Whether the bootstrapped R and phi values should be added to the output. \code{FALSE} by default.}
}
\value{
list. \code{R} gives the mean aspect ratio of the strain ellipse,
and \code{phi} gives the orientation of its long axis. If \code{boot=TRUE}, then
the bootstrapped 95\% confidence interval for the mean aspect ratio (\code{R_CI})
and for its orientation (\code{phi_CI}) are added. If \code{boot.values=TRUE}, a matrix
containing the bootstrapped R and phi values are added.
}
\description{
determines the shape and orientation of the strain ellipse by using
deformed elliptical objects as strain markers. The algorithm is based on the
mean shape matrix and its eigenvalues (Shimamoto and Ikeda, 1976).
}
\examples{
set.seed(20250411)
data(ramsay)
mean_strain_ellipse(ramsay[, "R"], ramsay[, "phi"])
}
\references{
Shimamoto, T., Ikeda, Y., 1976. A simple algebraic method for
strain estimation from ellipsoidal objects. Tectonophysics 36, 315–337.
\doi{10.1016/0040-1951(76)90107-4}
}
\seealso{
\code{\link[=mean_strain_ellipse_ramsay]{mean_strain_ellipse_ramsay()}}
}
