% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_analysis.R
\name{mean_strain_ellipse_ramsay}
\alias{mean_strain_ellipse_ramsay}
\title{Mean strain ellipse after Ramsay (1967)}
\usage{
mean_strain_ellipse_ramsay(r, phi = NULL, alpha = 0.05)
}
\arguments{
\item{r}{numeric. Aspect ratio of deformed object (long axis / short axis)}

\item{phi}{numeric. Orientation of long axis of deformed object (in degrees)}

\item{alpha}{numeric. alpha value to remove outliers}
}
\value{
list. \code{Rs} aspect ratio of strian ellipse, \code{Ri} initial aspect ratio,
\code{Fluctuation} angle (Ramsay 1967), and \code{phi} the mean orientation of the
strain ellipse long axis.
}
\description{
Mean strain ellipse after Ramsay (1967)
}
\examples{
data(ramsay)
mean_strain_ellipse_ramsay(ramsay[, "R"], ramsay[, "phi"])
}
