% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohR.R
\name{MohrCircle.calc}
\alias{MohrCircle.calc}
\title{MOHR CIRCLE:  CALCULATE}
\usage{
MohrCircle.calc(
  sigmaX = NA,
  sigmaZ = NA,
  tauXZ = NA,
  sigma1 = NA,
  sigma3 = NA,
  theta = seq(from = 0, to = 180, by = 1)
)
}
\arguments{
\item{sigmaX}{normal stress acting in the horizontal direction}

\item{sigmaZ}{normal stress acting in the vertical direction}

\item{tauXZ}{shear stress acting on the same plane as sigmaX}

\item{sigma1}{major principal stress}

\item{sigma3}{minor principal stress}

\item{theta}{vector of angles (degrees); defaults to 0-180 in increments of 1}
}
\value{
A five-element list containing:
           C = center of Mohr circle
           R = radius of Mohr circle
           sigma = vector of normal stresses for Mohr circle t
           au = vector of shear stresses for Mohr circle
           theta = vector of angles (deg)
}
\description{
MOHR CIRCLE:  CALCULATE
}
\note{
o  One of the following two sets of data must be entered:
               1.  sigmaX, sigmaZ, tauXZ
               2.  sigma1, sigma3
      o  If theta is entered in conjunction with sigmaX, sigmaZ, and tauXZ, it is interpreted
         as the angle of inclination above the horizontal.  If theta is entered in conjunction
         with the principal stresses, then it is interpreted as the angle of inclination above the
         major principal plane.
}
\examples{
MohrCircle.calc(sigmaX = 80, sigmaZ = 120, tauXZ = 20)
}
\author{
Kyle Elmy and Jim Kaklamanos
}
