% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohR.R
\name{MohrCircle_calc}
\alias{MohrCircle_calc}
\title{Mohr circle parameters}
\usage{
MohrCircle_calc(
  sigmaX = NA,
  sigmaZ = NA,
  tauXZ = NA,
  sigma1 = NA,
  sigma3 = NA,
  theta = seq(from = 0, to = 180, by = 1)
)
}
\arguments{
\item{sigmaX}{normal stress acting in the horizontal direction}

\item{sigmaZ}{normal stress acting in the vertical direction}

\item{tauXZ}{shear stress acting on the same plane as \code{"sigmaX"}}

\item{sigma1}{major principal stress}

\item{sigma3}{minor principal stress}

\item{theta}{vector of angles (degrees); defaults to 0-180 in increments of 1}
}
\value{
A five-element list containing
\describe{
\item{\code{"C"}}{center of Mohr circle}
\item{\code{"R"}}{radius of Mohr circle}
\item{\code{"sigma"}}{vector of normal stresses for Mohr circle}
\item{\code{"tau"}}{vector of shear stresses for Mohr circle}
\item{\code{"theta"}}{vector of angles (degrees)}
}
}
\description{
calculates parameters of the Mohr circle
}
\details{
If theta is entered in conjunction with \code{"sigmaX"}, \code{"sigmaZ"}, and \code{"tauXZ"}, it is interpreted
as the angle of inclination above the horizontal.  If \code{"theta"} is entered in conjunction
with the principal stresses, then it is interpreted as the angle of inclination above the
major principal plane.
}
\note{
One of the following two sets of data must be entered
\enumerate{
\item{\code{"sigmaX"}, \code{"sigmaZ"}, \code{"tauXZ"}}
\item{\code{"sigma1"}, \code{"sigma3"}}
}
}
\examples{
MohrCircle_calc(sigmaX = 80, sigmaZ = 120, tauXZ = 20)
}
\seealso{
\code{\link[=MohrCircle_plot]{MohrCircle_plot()}}
}
\author{
Kyle Elmy and Jim Kaklamanos
}
