% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohR.R
\name{MohrCircle_plot}
\alias{MohrCircle_plot}
\title{Mohr Circle plot}
\usage{
MohrCircle_plot(
  sigmaX = NA,
  sigmaZ = NA,
  tauXZ = NA,
  sigma1 = NA,
  sigma3 = NA,
  metric = TRUE
)
}
\arguments{
\item{sigmaX}{normal stress acting in the horizontal direction}

\item{sigmaZ}{normal stress acting in the vertical direction}

\item{tauXZ}{shear stress acting on the same plane as \code{"sigmaX"}}

\item{sigma1}{major principal stress}

\item{sigma3}{minor principal stress}

\item{metric}{logical variable: \code{TRUE} for metric units (kPa), and \code{FALSE} for English units}
}
\description{
plots the Mohr Circle
}
\note{
One of the following two sets of data must be entered
\enumerate{
\item{\code{"sigmaX"}, \code{"sigmaZ"}, \code{"tauXZ"}}
\item{\code{"sigma1"}, \code{"sigma3"}}
}
}
\examples{
MohrCircle_plot(sigmaX = 80, sigmaZ = 120, tauXZ = 20, metric = FALSE)
}
\seealso{
\code{\link[=MohrCircle_calc]{MohrCircle_calc()}}, \code{\link[=ggMohr]{ggMohr()}}
}
\author{
Kyle Elmy and Jim Kaklamanos
}
