% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-S3.R
\name{na.omit.spherical}
\alias{na.omit.spherical}
\title{Handle Missing Values in Spherical Objects}
\usage{
\method{na.omit}{spherical}(object, ...)
}
\arguments{
\item{object}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"}.}

\item{...}{further arguments special methods could require.}
}
\description{
Handle Missing Values in Spherical Objects
}
\examples{
x <- Line(c(120, NA, 100), c(50, 60, 70))
na.omit(x)
}
\keyword{internal}
