% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{operations}
\alias{operations}
\alias{vlength}
\alias{vnorm}
\alias{vcross}
\alias{vdot}
\alias{vrotate}
\alias{vangle}
\alias{vproject}
\alias{vreject}
\title{Vector operations}
\usage{
vlength(x)

vnorm(x)

vcross(x, y)

vdot(x, y)

vrotate(x, rotaxis, rotangle)

vangle(x, y)

vproject(x, y)

vreject(x, y)
}
\arguments{
\item{x, y, rotaxis}{numeric vector or matrix}

\item{rotangle}{numeirc. Angle in radians}
}
\description{
Vector operations
}
\examples{
vec1 <- cbind(1, 0, 0)
vec2 <- cbind(0, 0, 1)

vlength(vec1)
vcross(vec1, vec2)
vdot(vec1, vec2)
vrotate(vec1, vec2, pi / 2)
vangle(vec1, vec2)
vproject(vec1, vec2)
vreject(vec1, vec2)
}
