% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{or_eigen}
\alias{or_eigen}
\title{Eigenvalues and eigenvectors of a set of vectors}
\usage{
or_eigen(x, scaled = FALSE)
}
\arguments{
\item{x}{numeric. Can be three element vector, three column array, or an
object of class \code{"line"} or \code{"plane"}}

\item{scaled}{logical. Whether the eigenvectors should be scaled by the
eigenvalues (only effective if \code{x} is in Cartesian coordinates)}
}
\value{
list containing
\describe{
\item{\code{values}}{Eigenvalues}
\item{\code{vectors}}{Eigenvectors in coordinates system of \code{x}}
}
}
\description{
Decomposition of Orientation tensor
}
\examples{
set.seed(1)
mu <- Line(120, 50)
x <- rvmf(100, mu = mu, k = 20)
x_eigen <- or_eigen(x)
x_eigen
stereoplot()
stereo_point(x, col = "grey")
stereo_point(mu, lab = "mu", col = 4)
stereo_point(x_eigen$vectors, col = c(1, 2, 3), lab = c("E1", "E2", "E3"))
}
\seealso{
\code{\link[=ortensor]{ortensor()}}
}
