% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO3.R
\name{oriHexagonalTrapezohedralGroup}
\alias{oriHexagonalTrapezohedralGroup}
\title{Hexagonal trapezohedral is the point group of beta-quartz.}
\usage{
oriHexagonalTrapezohedralGroup()
}
\description{
Hexagonal trapezohedral is the point group of beta-quartz.
}
