% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO3.R
\name{oriLineInPlaneGroup}
\alias{oriLineInPlaneGroup}
\title{Line-in-plane symmetry is applicable to foliation-lineations, cylindrical fold orientations, triaxial ellipsoid orientations, and earthquake focal mechanisms (and olivine).}
\usage{
oriLineInPlaneGroup()
}
\description{
Line-in-plane symmetry is applicable to foliation-lineations, cylindrical fold orientations, triaxial ellipsoid orientations, and earthquake focal mechanisms (and olivine).
}
