% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO3.R
\name{oriRayInPlaneGroup}
\alias{oriRayInPlaneGroup}
\title{Ray-in-plane symmetry is applicable to faults-with-slip-directions, such as slickensides (and certain minerals).}
\usage{
oriRayInPlaneGroup()
}
\description{
Ray-in-plane symmetry is applicable to faults-with-slip-directions, such as slickensides (and certain minerals).
}
