% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO3.R
\name{oriTrigonalTrapezohedralGroup}
\alias{oriTrigonalTrapezohedralGroup}
\title{Trigonal trapezohedral is the point group of alpha-quartz.}
\usage{
oriTrigonalTrapezohedralGroup()
}
\description{
Trigonal trapezohedral is the point group of alpha-quartz.
}
