% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO3.R
\name{oriTrivialGroup}
\alias{oriTrivialGroup}
\title{When trivial symmetry is used, the orientations are simply rotations. This case is so important that we have separate code, in rot.R, for doing it. But let's include the trivial group, for completeness.}
\usage{
oriTrivialGroup()
}
\description{
When trivial symmetry is used, the orientations are simply rotations. This case is so important that we have separate code, in rot.R, for doing it. But let's include the trivial group, for completeness.
}
