% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortensor.R
\name{ot_eigen}
\alias{ot_eigen}
\alias{ot_eigen.spherical}
\alias{ot_eigen.ortensor}
\title{Eigenvalues and Eigenvectors of a Set of Vectors}
\usage{
ot_eigen(x, scaled = FALSE, ...)

\method{ot_eigen}{spherical}(x, scaled = FALSE, ...)

\method{ot_eigen}{ortensor}(x, scaled = FALSE, ...)
}
\arguments{
\item{x}{either an object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"},
\code{"Pair"}, or \code{"Fault"} where the
rows are the observations and the columns are the coordinates, or an \code{"ortensor"} object.}

\item{scaled}{logical. Whether the Eigenvectors should be scaled by the
Eigenvalues (only effective if \code{x} is in Cartesian coordinates).}

\item{...}{additional arguments passed to \code{\link[=ortensor]{ortensor()}} (ignored if \code{x} is \code{"ortensor"} object).}
}
\value{
list containing
\describe{
\item{\code{values}}{Eigenvalues}
\item{\code{vectors}}{Eigenvectors in coordinate system of \code{x}}
}
}
\description{
Decomposition of Orientation Tensor Eigenvectors and Eigenvalues
}
\examples{
set.seed(20250411)
mu <- rvmf(n = 1)
x <- rfb(100, mu = mu, k = 1, A = diag(c(10, 0, 0)))
x_eigen <- ot_eigen(x)
x_eigen
plot(x, col = "grey")
points(mu, col = 4)
text(mu, labels = "Mean", col = 4, pos = 4)
points(x_eigen$vectors, col = c(1, 2, 3))
text(x_eigen$vectors, col = c(1, 2, 3), labels = c("E1", "E2", "E3"), pos = 4)
}
\seealso{
Other ortensor:
\code{\link[=ortensor]{ortensor()}},
\code{\link{strain_shape}}
}
\concept{ortensor}
