% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO3.R
\name{pair2rot}
\alias{pair2rot}
\title{Orientation matrix from fault planes and slip directions}
\usage{
pair2rot(x)
}
\arguments{
\item{x}{object of class \code{"Pair"} or \code{"Fault"}}
}
\value{
list of rotation matrices
}
\description{
Converts a set of planes and lines into a list of rotation matrices.
}
\examples{
my_fault <- Fault(c("a" = 120, "b" = 120, "c" = 100), c(60, 60, 50), c(110, 25, 30), c(58, 9, 23), c(1, -1, 1))
pair2rot(my_fault)
}
