% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faults.R
\name{pair_correct}
\alias{pair_correct}
\alias{misfit_pair}
\alias{correct_pair}
\title{Orthogonalization of plane and line measurement}
\usage{
misfit_pair(x)

correct_pair(x)
}
\arguments{
\item{x}{Pair or Fault}
}
\value{
\code{misfit_pair} returns a list with the orthogonalized plane and line measurements
(as 3d vectors) and the misfit angles (in radians). \code{correct_pair} returns the orthogonalized vectors.
}
\description{
Both the line and the plane  are rotated in opposite directions by half the angle between
the slip and the plane normal vector about the vector that is pperpendicular to both.
}
\examples{
p <- Pair(120, 60, 110, 58, correction = FALSE)
misfit_pair(p)

correct_pair(p)
}
