% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faults.R
\name{Pair_pitch}
\alias{Pair_pitch}
\title{Pitch of a pair}
\usage{
Pair_pitch(x)
}
\arguments{
\item{x}{object of class "\code{Pair}"}
}
\value{
numeric. Angle in degrees
}
\description{
Pitch of a pair
}
\examples{
Pair_pitch(Fault_from_rake(Plane(120, 50), 80))
}
