% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{parse_json_impl}
\alias{parse_json_impl}
\title{Parses a JSON string according to a provided structure definition using streaming.}
\usage{
parse_json_impl(json_string, structure)
}
\arguments{
\item{json_string}{A character string containing the JSON data.}

\item{structure}{An R list object representing the structure, created by \code{build_structure()}.}
}
\value{
An R object (list, vector, atomic) representing the parsed JSON if validation succeeds.
}
\description{
Parses a JSON string according to a provided structure definition using streaming.
}
