% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{parse_quadrant_measurement}
\alias{parse_quadrant_measurement}
\title{Title}
\usage{
parse_quadrant_measurement(x)
}
\arguments{
\item{x}{}
}
\value{
numeric
}
\description{
Title
}
\examples{
parse_quadrant_measurement(c("E30N", "W10S"))
}
