% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faults.R
\name{Plane_from_apparent_dips}
\alias{Plane_from_apparent_dips}
\title{Apparent dip direction}
\usage{
Plane_from_apparent_dips(a1, a2)
}
\arguments{
\item{a1, a2}{\code{"Line"} objects containing the apparent dips and dip directions}
}
\value{
\code{"Plane"} object
}
\description{
Apparent dip direction
}
\examples{
a1 <- Line(45, 22)
a2 <- Line(352, 10)
res <- Plane_from_apparent_dips(a1, a2)

stereoplot()
points(rbind(a1, a2))
lines(res, lty = 2)
}
