% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faults.R
\name{Plane_pitch}
\alias{Plane_pitch}
\title{Pitch of a line of a plane}
\usage{
Plane_pitch(p, l)
}
\arguments{
\item{p}{object of class \code{"Plane"}}

\item{l}{object of class \code{"Line"}}
}
\value{
Pitch angle in degrees
}
\description{
Pitch of a line of a plane
}
\examples{
Plane_pitch(Plane(90, 70), Line(175, 13)) # 13.9
}
