% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{plot-spherical}
\alias{plot-spherical}
\alias{plot.Line}
\alias{plot.Vec3}
\alias{plot.Ray}
\alias{plot.Plane}
\alias{plot.Pair}
\alias{plot.Fault}
\title{Plot Spherical Objects}
\usage{
\method{plot}{Line}(x, upper.hem = NULL, earea = NULL, grid.params = list(), ...)

\method{plot}{Vec3}(x, upper.hem = NULL, earea = NULL, grid.params = list(), ...)

\method{plot}{Ray}(x, upper.hem = NULL, earea = NULL, grid.params = list(), pch = NULL, ...)

\method{plot}{Plane}(x, upper.hem = NULL, earea = NULL, grid.params = list(), ...)

\method{plot}{Pair}(x, upper.hem = NULL, earea = NULL, grid.params = list(), ...)

\method{plot}{Fault}(x, upper.hem = NULL, earea = NULL, grid.params = list(), ...)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"},
where the rows are the observations and the columns are the coordinates.}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}). Defaults to
\code{getOption("structr.upper.hem")}.}

\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection, or \code{FALSE} for meridional stereographic projection.
Defaults to \code{getOption("structr.earea")}.}

\item{grid.params}{list.}

\item{...}{parameters passed to \code{\link[=stereo_point]{stereo_point()}}, \code{\link[=stereo_smallcircle]{stereo_smallcircle()}}, \code{\link[=stereo_greatcircle]{stereo_greatcircle()}}, or \code{\link[=fault_plot]{fault_plot()}}}

\item{pch}{plotting character}
}
\description{
Plot Spherical Objects
}
\details{
If \code{x} is a Ray and \code{pch} is \code{NULL}, solid symbols show rays pointing in the lower hemisphere,
while open symbols point into the upper hemisphere.
}
\examples{
plot(rvmf(10, mu = Vec3(1, 0, 0))) # Vec
plot(Line(c(90, 80), c(10, 75)), lab = c("L1", "L2"))
plot(Ray(c(90, 80), c(10, 75), sense = c(1, -1)), lab = c("L1", "L2"))
plot(Plane(120, 30), col = "red")
plot(Pair(120, 50, 36, 8))
plot(Fault(120, 50, 36, 8, -1))
}
\seealso{
Other stereo-plot:
\code{\link{arrows}},
\code{\link{fault-plot}},
\code{\link{lines}},
\code{\link[=points.spherical]{points.spherical()}},
\code{\link{stereo_cones}},
\code{\link[=stereo_confidence]{stereo_confidence()}},
\code{\link{stereo_contour}},
\code{\link[=stereo_lines]{stereo_lines()}},
\code{\link[=stereo_point]{stereo_point()}},
\code{\link[=stereo_segment]{stereo_segment()}},
\code{\link[=stereoplot]{stereoplot()}},
\code{\link[=stereoplot_guides]{stereoplot_guides()}},
\code{\link[=stereoplot_ticks]{stereoplot_ticks()}},
\code{\link[=text.spherical]{text.spherical()}}
}
\concept{stereo-plot}
