% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{plot.spherical}
\alias{plot.spherical}
\title{Plot spherical objects}
\usage{
\method{plot}{spherical}(x, upper.hem = FALSE, earea = TRUE, grid.params = list(), ...)
}
\arguments{
\item{x}{objects of class \code{"Vec3"}, \code{"Line"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"}.}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}, the default).}

\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection (the default), or \code{FALSE} for meridional stereographic projection.}

\item{grid.params}{list.}

\item{...}{parameters passed to \code{\link[=stereo_point]{stereo_point()}}, \code{\link[=stereo_smallcircle]{stereo_smallcircle()}}, \code{\link[=stereo_greatcircle]{stereo_greatcircle()}}, or \code{\link[=fault_plot]{fault_plot()}}}
}
\description{
Plot spherical objects
}
\examples{
plot(Line(c(90, 80), c(10, 75)), lab = c("L1", "L2"))
plot(Plane(120, 30), col = "red")
plot(Pair(120, 50, 36, 8))
plot(Fault(120, 50, 36, 8, -1))
}
