% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{points.spherical}
\alias{points.spherical}
\title{Add Points to a Plot}
\usage{
\method{points}{spherical}(x, upper.hem = NULL, earea = NULL, ...)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"},
where the rows are the observations and the columns are the coordinates.}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}). Defaults to
\code{getOption("structr.upper.hem")}.}

\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection, or \code{FALSE} for meridional stereographic projection.
Defaults to \code{getOption("structr.earea")}.}

\item{...}{arguments passed to \code{\link[graphics:points]{graphics::points()}}}
}
\description{
Add Points to a Plot
}
\examples{
stereoplot()
points(rvmf(n = 100))
points(Plane(120, 30), col = "red", pch = 19)
}
\seealso{
Other stereo-plot:
\code{\link{arrows}},
\code{\link{fault-plot}},
\code{\link{lines}},
\code{\link{plot-spherical}},
\code{\link{stereo_cones}},
\code{\link[=stereo_confidence]{stereo_confidence()}},
\code{\link{stereo_contour}},
\code{\link[=stereo_lines]{stereo_lines()}},
\code{\link[=stereo_point]{stereo_point()}},
\code{\link[=stereo_segment]{stereo_segment()}},
\code{\link[=stereoplot]{stereoplot()}},
\code{\link[=stereoplot_guides]{stereoplot_guides()}},
\code{\link[=stereoplot_ticks]{stereoplot_ticks()}},
\code{\link[=text.spherical]{text.spherical()}}
}
\concept{stereo-plot}
