% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohR.R
\name{PR_stress}
\alias{PR_stress}
\title{Principal stresses from 2D stress components}
\usage{
PR_stress(sigma_x, sigma_y, tau_xy)
}
\arguments{
\item{sigma_x}{numeric. Magnitude of normal stress acting in the horizontal direction}

\item{sigma_y}{numeric. Magnitude of normal stress acting on plane facing in Y direction
(\deqn{\sigma_y}).}

\item{tau_xy}{numeric. Magnitude of shear stress acting on planes facing X and Y
(\deqn{\tau_{xy}}).}
}
\value{
angle in degrees
}
\description{
Determines the principal stresses and their orientations from the stress
components \deqn{sigma_x}, \deqn{\sigma_y}, \deqn{\tau_{xy}}.
}
\examples{
PR_stress(sigma_x = 80, sigma_y = 120, tau_xy = 20)
}
\references{
Richard J. Lisle (1999)
}
