% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_stereonet.R
\name{prepare-ggplot}
\alias{prepare-ggplot}
\alias{gg}
\alias{ggl}
\title{Prepare points and lines for ggplot}
\usage{
gg(x, ...)

ggl(x, ..., d = 90, n = 1000)
}
\arguments{
\item{x}{Spherical or vector orientation data}

\item{...}{\code{\link[rlang:args_data_masking]{<data-masking>}} Name-value pairs. The name gives the name of the column in the output.
The value can be:
\itemize{
\item A vector of length 1, which will be recycled to the correct length.
\item A vector the same length as the current group (or the whole data frame if ungrouped).
\item \code{NULL}, to remove the column.
\item A data frame or tibble, to create multiple columns in the output.
}}

\item{d}{numeric. Cone angle (small circle radius) in degrees. \code{90} (the default) produces great circles.}

\item{n}{integer. Resolution of line.}
}
\value{
data.frame
}
\description{
Prepare points and lines for ggplot
}
\examples{
x <- Plane(120, 85)
ggstereo() +
  geom_point(data = gg(x), aes(x, y), color = "red") +
  geom_path(data = ggl(x), aes(x, y), color = "red")

x <- Line(120, 5)
ggstereo() +
  geom_point(data = gg(x), aes(x, y), color = "darkgreen") +
  geom_path(data = ggl(x, d = 8), aes(x, y, group = group), color = "darkgreen")
}
