% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion.R
\name{principal_fault}
\alias{principal_fault}
\title{Calculate Principal Fault planes from stress vectors and friction}
\usage{
principal_fault(s1, s3, friction = 0.6)
}
\arguments{
\item{s1, s3}{Principal stress vectors as spherical objects}

\item{friction}{numeric. Coefficient of friction (0.6 by default)}
}
\value{
\code{"Fault"} object
}
\description{
Calculate Principal Fault planes from stress vectors and friction
}
\examples{
s <- reduced_stress(angelier1990$TYM)
stress <- sigma2stress(s)
pr_TYM <- principal_fault(stress$axes[1, ], stress$axes[3, ])

stereoplot()
fault_plot(angelier1990$TYM, col = "grey")
fault_plot(pr_TYM, col = "red")
points(stress$axes, pch = 16)

# Focal mechanism plot
stereoplot()
stereo_fms(s)
fault_plot(angelier1990$TYM, col = "grey30", type = "hoeppener")
points(stress$axes, col = 2:4, pch = 16)
}
\seealso{
\code{\link[=stereo_fms]{stereo_fms()}} to plot the focal mechanism solution
}
