% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortensor.R
\name{projected_mean}
\alias{projected_mean}
\title{Projected Mean}
\usage{
projected_mean(x)
}
\arguments{
\item{x}{Object of class \code{"Vec3"}, \code{"Line"} or \code{"Plane"}}
}
\value{
Vector in coordinate system of \code{x}
}
\description{
Eigenvector with largest eigenvalue of the orientation tensor
}
\examples{
projected_mean(example_lines)
}
\seealso{
\code{\link[=ot_eigen]{ot_eigen()}} for eigenvalues of orientation tensor,
\code{\link[=sph_mean]{sph_mean()}} for arithmetic mean, \code{\link[=geodesic_mean]{geodesic_mean()}} for geodesic mean.
}
