% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{quadrant2dd}
\alias{quadrant2dd}
\title{Converts strike and dip quadrant notation into into dip direction}
\usage{
quadrant2dd(strike, dip_quadrant, n_directions = c(4L, 8L, 16L))
}
\arguments{
\item{strike}{numeric. Strike in degree (left- or right-hand rule)}

\item{dip_quadrant}{character. Quadrant of dip direction}

\item{n_directions}{integer.}
}
\value{
Dip direction in degrees
}
\description{
Converts strike and dip quadrant notation into into dip direction
}
\examples{
s <- c(270, 315, 0, 45, 90, 135, 180, 225, 270) # strike in left-hand-rule
q <- c("N", "E", "E", "S", "S", "W", "W", "N", "N") # dip quadrant
quadrant2dd(s, q)
}
\seealso{
Other parse-orientations:
\code{\link[=Fault_from_rake_quadrant]{Fault_from_rake_quadrant()}},
\code{\link[=azimuth_to_cardinal]{azimuth_to_cardinal()}},
\code{\link[=sense_from_strikeslip]{sense_from_strikeslip()}},
\code{\link{split}},
\code{\link[=strikeslip_kinematics]{strikeslip_kinematics()}}
}
\concept{parse-orientations}
