% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_pole.R
\name{ramsay_coords}
\alias{ramsay_coords}
\alias{cartesian_to_acoscartesian}
\alias{acoscartesian_to_cartesian}
\title{Spherical coordinate conversion}
\source{
Ramsay, 1967, p. 15-16
}
\usage{
cartesian_to_acoscartesian(x)

acoscartesian_to_cartesian(x)
}
\arguments{
\item{x}{matrix. Cartesian coordinates of points}
}
\description{
Spherical coordinate conversion
}
\details{
acoscartesian coordinates are given as
\describe{
\item{\code{a}}{deviation from x-axis E-W horizontal (angle in degrees)}
\item{\code{b}}{deviation from y-axis N-S horizontal (angle in degrees)}
\item{\code{c}}{deviation from z-axis vertical (angle in degrees)}
}
}
\examples{
Stereographic coordinates (angle notation):
x <- rbind(
  c(58, 78, 35),
  c(47, 72, 47),
  c(57, 68, 41),
  c(56, 64, 45),
  c(52, 65, 47),
  c(61, 62, 51),
  c(55, 60, 49),
  c(61, 62, 42),
  c(68, 52, 46),
  c(42, 58, 65)
)
acoscartesian_to_cartesian(x)
acoscartesian_to_cartesian(x) |> to_spherical()
acoscartesian_to_cartesian(x) |> cartesian_to_acoscartesian()
}
