% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-S3.R
\name{rbind.spherical}
\alias{rbind.spherical}
\title{Combine R Objects by Rows or Columns}
\usage{
\method{rbind}{spherical}(..., .class = NULL)
}
\arguments{
\item{...}{objects  to bind; note that all objects have to be class \code{"Vec3"}, \code{"Line"}, or \code{"Plane"}.}

\item{.class}{character. The output class. If \code{NULL}, all combined objects
will be coerced to the first element's class}
}
\value{
combined objects as class defined in \code{.class}
}
\description{
Combine R Objects by Rows or Columns
}
\examples{
set.seed(20250411)
rbind.spherical(
  rvmf(n = 5, mu = Line(90, 45)),
  runif.spherical('Vec3', n = 5),
  rkent(n = 5, mu = Plane(0, 10), b = 1)
)
}
