% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_strabo}
\alias{read_strabo}
\title{Import data from StraboSpot export}
\usage{
read_strabo(file, tag_cols = FALSE, sf = TRUE)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.}

\item{tag_cols}{logical. Whether the Tag columns should be summarized in a
single column (may lead to duplicate rows).}

\item{sf}{logical. Whether the output should be a spatial \code{"sf"} object
using the Longitude and Latitude columns.}
}
\value{
\code{list} containing the following objects:
\describe{
\item{\code{data}}{\code{"tbl_df"} object. \code{"sf"} if \code{sf == TRUE}.}
\item{\code{planar}}{Plane elements. Same row ids as in \code{data}.}
\item{\code{linear}}{Line elements. Same row ids as in \code{data}.}
}
}
\description{
Import data from StraboSpot export
}
\examples{
\dontrun{
file = "E:/Lakehead/Field Work/StraboSpot_Output_11_01_2022_TS.xlsx"
dt <- read_strabo(file)
stereoplot(); 
stereo_greatcircle(dt$planar, col = "lightgrey") 
stereo_point(dt$linear)
}
}
