% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion.R
\name{reduced_stress}
\alias{reduced_stress}
\title{Reduced stress tensor}
\usage{
reduced_stress(fault, method = c("michael", "angelier"))
}
\arguments{
\item{fault}{\code{"Fault"} object where the rows are the observations, and the columns the coordinates.}

\item{method}{character. The inversion algorithm, one of \code{"michael"} (the default)
for a bootstrapped linear inversion after Micheal (1984), \code{"angelier"}
for an iterative direct inversion after Angelier (1990) and Mostafa (2005),
\code{"yamaji"} for direct inversion using the 5d parameter space after Yamaji and Sato (2006),
\code{"hansen"} for direct inversion using the 9d parameter space after Hansen (2013), and
\code{"wissi"} for the Weighted Iterative Sigma-Space Inversion (WISSI).}
}
\value{
\code{"ellipsoid"} object. 3x3 matrix (i.e. 2nd rank tensor).
}
\description{
Calculates the reduced stress tensor (without optimization).
}
\examples{
f <- angelier1990$TYM
reduced_stress(f)
}
\seealso{
\code{\link[=slip_inversion]{slip_inversion()}}

Other stress-tensor:
\code{\link[=fault_instability_criterion]{fault_instability_criterion()}},
\code{\link{sigma-comp}},
\code{\link[=sigma2rup]{sigma2rup()}},
\code{\link[=sigma2stress]{sigma2stress()}},
\code{\link[=stress_shape]{stress_shape()}}
}
\concept{stress-tensor}
