% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sph_distributions.R
\name{rfb}
\alias{rfb}
\title{Spherical Fisher-Bingham Distribution}
\source{
Adapted from \code{\link[Directional:rfb]{Directional::rfb()}}
}
\usage{
rfb(n = 100, mu = Vec3(1, 0, 0), k = 5, A)
}
\arguments{
\item{n}{integer. number of random samples to be generated}

\item{mu}{Mean vector. object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}, where the
rows are the observations and the columns are the coordinates.}

\item{k}{numeric. The concentration parameter (\eqn{\kappa}) of the von
Mises-Fisher distribution}

\item{A}{symmetric matrix}
}
\description{
Simulation of random values from a spherical Fisher-Bingham distribution.
}
\examples{
set.seed(20250411)
x <- rfb(100, mu = Line(120, 50), k = 5, A = diag(c(-1, 0, 1)))

contour(x)
points(x)
}
\seealso{
Other random:
\code{\link{rbing}},
\code{\link[=rkent]{rkent()}},
\code{\link[=rrot]{rrot()}},
\code{\link[=runif.spherical]{runif.spherical()}},
\code{\link[=rwatson]{rwatson()}},
\code{\link{vonmises-fisher}}
}
\concept{random}
