% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_analysis.R
\name{RGN_plot}
\alias{RGN_plot}
\title{Rigid Grain Net (RGN)}
\usage{
RGN_plot(
  r,
  theta,
  angle_error = 3,
  n_iter = 100L,
  probs = 0.975,
  grid = 0.05,
  main = "Rigid-Grain-Net",
  guides = NULL,
  ...
)
}
\arguments{
\item{r}{numeric. The porphyroclast aspect ratio (long axis/short axis)}

\item{theta}{numeric. Angle between long axis and foliation (in degrees)}

\item{angle_error}{numeric. Uncertainty of angle measurement. \code{3} by default.}

\item{n_iter}{integer. Number of bootstrap resamples}

\item{probs}{integer. Probability with values in \eqn{[0, 1]} to estimate
critical shape factor, i.e. the largest shape factor of measurements outside
of critical hyperbole.}

\item{grid}{numeric. Spacing of hyperboles.}

\item{main}{character. The main title (on top).}

\item{guides}{logical. Whether guides of the projection grid should be added
to the plot. Defaults to \code{getOption("structr.guides")}.}

\item{...}{plotting arguments passed to \code{\link[graphics:points]{graphics::points()}}}
}
\value{
a plot or a list of the calculated \code{B} (shape factor) and \code{theta} values,
and the bootstrapped confidence interval of the critical B value (\code{Rc_CI}).
}
\description{
The rigid grain net after (Jessup et al. 2007) plots the distribution the finite
strain ratio (\code{R}) and orientation (\code{phi}) of
porphyroclast over the theoretical distribution of tailless clasts. The plot estimates
the critical shape factor \code{Rc} marking the transition between the stable-sink
position and infinitely rotating porphyroclasts.
This critical shape factor can be interpreted as the the \strong{mean kinematic vorticity number}.
Here the \code{Rc} is estimated using the bootstrap method described in Stephan et al. (2025).
}
\examples{
data(shebandowan)
set.seed(20250411)
RGN_plot(shebandowan$r, shebandowan$phi, col = "darkred")
}
\references{
Jessup, Micah J., Richard D. Law, and Chiara Frassi.
"The rigid grain net (RGN): an alternative method for estimating mean
kinematic vorticity number (Wm)." Journal of Structural Geology 29.3 (2007):
411-421. \doi{10.1016/j.jsg.2006.11.003}

Stephan, T., Phillips, N., Tiitto, H., Perez, A., Nwakanma, M.,
Creaser, R., & Hollings, P. (2025). Going with the flow - Changes of
vorticity control gold enrichment in Archean shear zones
(Shebandowan Greenstone Belt, Superior Province, Canada).
Journal of Structural Geology, 201, 105542.
\doi{10.1016/j.jsg.2025.105542}
}
