% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rhr}
\alias{rhr}
\alias{rhr2dd}
\alias{dd2rhr}
\title{Converts strike into dip direction using right-hand rule}
\usage{
rhr2dd(strike)

dd2rhr(dipdirection)
}
\arguments{
\item{strike}{strike direction in degrees}

\item{dipdirection}{strike direction in degrees}
}
\description{
Converts strike into dip direction using right-hand rule
}
