% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{rotate_stereogrid}
\alias{rotate_stereogrid}
\title{Center grid lines on a given vector}
\usage{
rotate_stereogrid(
  x,
  d = 10,
  col = "gray90",
  lwd = 0.5,
  equator = TRUE,
  equator_lwd = 1.5 * lwd,
  lty = 1,
  ...
)
}
\arguments{
\item{x}{center position of grid lines.}

\item{d}{angle spacing between grid lines of the projection}

\item{col}{Color of guide lines. Defaults to \code{getOption("structr.col")}.}

\item{lwd}{Width of guide lines. Defaults to \code{getOption("structr.lwd")}.}

\item{equator}{logical. Whether the grid equator should be shown no matter how the grid is constructed via \code{d}. \code{TRUE} by default.}

\item{equator_lwd}{numeric. The line width of the drawn equator grid line. By default, the equator is \code{1.5} times thicker then \code{lwd}.}

\item{lty}{Type of guide lines. Defaults to \code{getOption("structr.lty")}.}

\item{...}{arguments passed to \code{\link[graphics:lines]{graphics::lines()}}}
}
\description{
Center grid lines on a given vector
}
\examples{
stereoplot(guide = FALSE)
rotate_stereogrid(Plane(120, 50), earea = FALSE)
}
