% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geologyGeometry_functions.R
\name{rotLeftCovariance}
\alias{rotLeftCovariance}
\title{Sample covariance matrix, approximated in the tangent space at a given rotation.}
\usage{
rotLeftCovariance(rs, center)
}
\arguments{
\item{rs}{A list of rotation matrices.}

\item{center}{A rotation matrix. Typically the geodesic mean of the \code{rs}.}
}
\value{
A 3x3 real matrix (symmetric, non-negative eigenvalues).
}
\description{
Appropriate only if the sample is tightly concentrated near the center.
}
