% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_analysis.R
\name{Rphi}
\alias{Rphi}
\title{\deqn{R_f/\phi}}
\usage{
Rphi(Rs, Ri = 1, theta)
}
\arguments{
\item{Rs}{numeric.}

\item{Ri}{numeric. Initial or pre-deformation aspect ratio of an elliptical object}

\item{theta}{numeric. Fluctuation angle of undeformed marker w.r.t. computed principal direction of strain (in degrees)}
}
\value{
list. \code{phi}: the final angle (in degrees) between long axis and principal strain direction, \code{Rf}: final axial ratio
}
\description{
\deqn{R_f/\phi}
}
\examples{
#' data(ramsay)
Rphi(ramsay[, 'R'], theta = ramsay[, 'phi'])
}
