% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_analysis.R
\name{Rphi_polar_plot}
\alias{Rphi_polar_plot}
\title{Polar R/phi plot}
\usage{
Rphi_polar_plot(
  r,
  phi,
  proj = c("eqd", "eqa", "stg", "ort", "gno", "lin", "rdl"),
  contour = TRUE,
  image = FALSE,
  n = 10L,
  contour.lines = TRUE,
  contour.lines.params = list(lwd = 0.5, col = par("col")),
  contour.col = viridis,
  contour.col.params = list(),
  mean.ellipse = TRUE,
  mean.ellipse.params = list(col = "red", lwd = 2),
  point.params = list(col = "grey", pch = 16, cex = 0.5),
  main = "Polar Rf/phi plot",
  ...
)
}
\arguments{
\item{r}{numeric. Aspect ratio of deformed object (long axis / short axis)}

\item{phi}{numeric. Orientation of long axis of deformed object (in degrees)}

\item{proj}{character.}

\item{contour}{logical. Whether a grid of densities should be drawn in the background.}

\item{image}{logical. Whether a raster image or filled contour lines should be
drawn as density grid. Ignored when \code{contour=FALSE}}

\item{n}{integer. Grid resolution or number of fille dcontours}

\item{contour.lines}{logical. Whether contour lines should be added.}

\item{contour.lines.params}{list of plotting arguments passed to \code{\link[graphics:contour]{graphics::contour()}}}

\item{contour.col}{function to produce color palette used for contouring}

\item{contour.col.params}{list of plotting arguments passed \code{col.palette}}

\item{mean.ellipse}{logical. Whether the mean ellipse should be plotted}

\item{mean.ellipse.params}{list of plotting arguments passed to \code{\link[=ellipse]{ellipse()}}}

\item{point.params}{list of plotting arguments passed to \code{\link[graphics:points]{graphics::points()}}}

\item{main}{character. The main title (on top).}

\item{...}{optional arguments passed to \code{\link[=hypercontour]{hypercontour()}}}
}
\value{
plot
}
\description{
Plots R/phi fabric plot in polar coordinates
}
\examples{
Rphi_polar_plot(ramsay[, 1], ramsay[, 2], proj = "eqa")
}
\seealso{
\code{\link[=hypercontour]{hypercontour()}}, \code{\link[=Rphi_plot]{Rphi_plot()}}
}
