% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sph_distributions.R
\name{rrot}
\alias{rrot}
\title{Random Rotation Matrices}
\usage{
rrot(n)
}
\arguments{
\item{n}{integer. number of random samples to be generated}
}
\value{
list of rotation matrices
}
\description{
Random sample of matrices in SO(3).
}
\examples{
set.seed(20250411)
# Generate 10 random SO(3) rotation matrices
r <- rrot(10)

# convert SO(3) matrices to "Pair"
rp <- lapply(r, rot2pair) |> lapply(unclass)
rp <- do.call(rbind, args = rp) |>
  as.Pair()
  
# plot pairs
plot(rp)
}
\seealso{
Other random:
\code{\link{rbing}},
\code{\link[=rfb]{rfb()}},
\code{\link[=rkent]{rkent()}},
\code{\link[=runif.spherical]{runif.spherical()}},
\code{\link[=rwatson]{rwatson()}},
\code{\link{vonmises-fisher}}
}
\concept{random}
