% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-S3.R
\name{sample_spherical}
\alias{sample_spherical}
\title{Random Samples and Permutations of Spherical Objects}
\usage{
sample_spherical(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"}.}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement?}

\item{prob}{a vector of probability weights for obtaining the elements
    of the vector being sampled.}
}
\value{
A spherical object of length \code{size} with elements drawn from \code{x}
}
\description{
\code{sample_spherical} takes a sample of the specified size from the elements of
the spherical object \code{x} using either with or without replacement.
}
\examples{
set.seed(20250411)
x <- rvmf(n = 100, mu = Line(90, 45))
sample_spherical(x, size = 5)
}
