% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faults.R
\name{sense_from_strikeslip}
\alias{sense_from_strikeslip}
\title{Dip-slip Kinematics from Strike-Slip Faults}
\usage{
sense_from_strikeslip(x, left)
}
\arguments{
\item{x}{\code{"Pair"} object(s) of the plane(s) and line(s) with unknown dip-slip offset}

\item{left}{logical. \code{TRUE} if \code{x} is a left-lateral (sinistral) fault, and
\code{FALSE} if \code{x} is a right-lateral (dextral) fault.
Must have same length as rows in \code{x}}
}
\value{
numeric. 1 if normal, -1 if reverse offset
}
\description{
Dip-slip sense (1 for normal, -1 for reverse) when only strike-slip kinematics are known
}
\examples{
# Sinistral fault
sense_from_strikeslip(Pair(120, 89, 30, 5), left = TRUE) # 1: normal offset

# Dextral fault
sense_from_strikeslip(Pair(120, 89, 30, 5), left = FALSE) # -1: reverse offset
}
\seealso{
\code{\link[=strikeslip_kinematics]{strikeslip_kinematics()}}

Other parse-orientations:
\code{\link[=Fault_from_rake_quadrant]{Fault_from_rake_quadrant()}},
\code{\link[=azimuth_to_cardinal]{azimuth_to_cardinal()}},
\code{\link[=quadrant2dd]{quadrant2dd()}},
\code{\link{split}},
\code{\link[=strikeslip_kinematics]{strikeslip_kinematics()}}
}
\concept{parse-orientations}
