% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lund.R
\name{SH}
\alias{SH}
\title{Direction of maximum horizontal stress from the stress tensor}
\usage{
SH(S1, S2, S3, R, tol = NULL, ortho.tol = 0.005)
}
\arguments{
\item{S1, S2, S3}{The principal stress orientations.
The variables hold the coordinates in the North, East and Down geographical
coordinate system, e.g. \code{S1 = c(s1N,s1E,s1D)}. Given as object of class \code{"Vec3"} or \code{"Line"}}

\item{R}{numeric. Relative magnitude of \code{S2} with respect to \code{S1} and \code{S3}:
\eqn{R = \frac{S1 - S2}{S1 - S3}}. Values ranging from 0 to 1, with 0 being
\code{S1==S2} and 1 being \code{S2==S3}. Equivalent to the stress shape ratio of Gephart & Forsyth (1984).}

\item{tol}{Tolerance of comparison. Defaults to \code{getOption("structr.tol")}.}

\item{ortho.tol}{tolerance angle (in degree) for orthogonality check of the
three principal stress vectors.}
}
\value{
The direction of SH from North as numeric angle in degrees (radians if all principal stress axes were given as \code{"Vec3"} objects).
}
\description{
Calculates the direction of maximum horizontal stress using only the
directions of the principal stress and \eqn{R = \frac{S1 - S2}{S1 - S3}}.
This function implements equations 11 and 10 from Lund and Townend (2007).
}
\examples{
# first example from https://www.snsn.se/SH/SHcode/benchmark.out
S1 <- Line(250.89, 70.07)
S3 <- Line(103.01, 17.07)
S2 <- crossprod(S3, S1)
SH(S1, S2, S3, R = 1) #  70.89

R <- seq(0, 1, .05)
cbind(R, SH = SH(S1, S2, S3, R = R))
}
\references{
Lund and Townend, (2007). Calculating horizontal stress
orientations with full or partial knowledge of the tectonic stress tensor,
Geophys. J. Int., doi:\doi{10.1111/j.1365-246X.2007.03468.x}.
}
\seealso{
\code{\link[=slip_inversion]{slip_inversion()}} for stress inversion of fault slip data.

Other SH-from-tensor:
\code{\link[=SH_from_tensor]{SH_from_tensor()}}
}
\concept{SH-from-tensor}
