% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohR.R
\name{sigma13}
\alias{sigma13}
\title{Principle stresses}
\usage{
sigma13(sigma_x, sigma_z, tau_xz)
}
\arguments{
\item{sigma_x}{numeric. Magnitude of normal stress acting in the horizontal direction}

\item{sigma_z}{numeric. Magnitude of normal stress acting in the vertical direction}

\item{tau_xz}{numeric. Magnitude of shear stress acting on the same plane as \code{"sigma_x"}}
}
\value{
A four-element list containing
\describe{
\item{\code{"sigma1"}}{magnitude of major principal stress}
\item{\code{"sigma3"}}{magnitude of minor principal stress}
\item{\code{"theta1"}}{direction of major principal stress (degrees)}
\item{\code{"theta3"}}{direction of minor principal stress (degrees)}
}
}
\description{
calculates the magnitudes and directions of the principal stresses \eqn{\sigma_1} and \eqn{\sigma_3}
}
\examples{
sigma13(sigma_x = 80, sigma_z = 120, tau_xz = 20)
}
\author{
Kyle Elmy and Jim Kaklamanos
}
