% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion.R
\name{sigma2rup}
\alias{sigma2rup}
\title{Angelier's Ratio Upsilon (RUP)}
\usage{
sigma2rup(sigma, fault, lambda = sqrt(3)/2)
}
\arguments{
\item{sigma}{symmetric 3x3 matrix. The (reduced) stress tensor.}

\item{fault}{\code{"Fault"} object where the rows are the observations, and the columns the coordinates.}

\item{lambda}{numeric. The maximum shear stress. \eqn{\sqrt(3)/2} by default.}
}
\value{
numeric. The per-fault "ratio upsilon" (RUP) parameter in percent.
}
\description{
The per-fault "ratio upsilon" (RUP) parameter after Angelier (1990) as
an estimator for the quality-of-fit of fault-slip inversions.
}
\details{
The RUP estimator varies from 0\%, indicating maximum shear stress parallel to fault
slip and with same sense (i.e. perfect fit), to 200\%, corresponding to maximum
shear stress parallel to fault slip but with opposite sense (i.e. largest misfit).
The quality of the fit is good if RUP \eqn{\le} 50\%, (potentially) acceptable
if 50\%<RUP\eqn{\le} 75\%, and poor otherwise.

Angelier (1990) introduced the \emph{ratio upsilon}, which is the ratio of \eqn{\upsilon_i} to the largest shear stress on a fault \eqn{i}
expressed in percentage:
\deqn{\text{RUP} = \upsilon_i / \lambda \times 100}
where \eqn{\lambda = \frac{\sqrt{3}}{2}} is the maximum shear stress.
}
\examples{
f <- angelier1990$TYM
s <- reduced_stress(f)
sigma2rup(s, f)
}
\references{
Angelier, J. (1990). Inversion of field data in fault tectonics to obtain the
regional stress—III. A new rapid direct inversion method by analytical means.
Geophys. J. Int, 103, 363–376. \doi{10.1111/j.1365-246X.1990.tb01777.x}
}
\seealso{
\code{\link[=slip_inversion]{slip_inversion()}}

Other stress-tensor:
\code{\link[=fault_instability_criterion]{fault_instability_criterion()}},
\code{\link[=reduced_stress]{reduced_stress()}},
\code{\link{sigma-comp}},
\code{\link[=sigma2stress]{sigma2stress()}},
\code{\link[=stress_shape]{stress_shape()}}
}
\concept{stress-tensor}
