% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohR.R
\name{sigmaTrans}
\alias{sigmaTrans}
\title{Stress transformation}
\usage{
sigmaTrans(
  theta,
  sigmaX = NA,
  sigmaZ = NA,
  tauXZ = NA,
  sigma1 = NA,
  sigma3 = NA
)
}
\arguments{
\item{theta}{angle of inclination (degrees)}

\item{sigmaX}{normal stress acting in the horizontal direction}

\item{sigmaZ}{normal stress acting in the vertical direction}

\item{tauXZ}{shear stress acting on the same plane as \code{"sigmaX"}}

\item{sigma1}{major principal stress}

\item{sigma3}{minor principal stress}
}
\value{
A two-element list containing
\describe{
\item{\code{"sigma"}}{normal stress on an inclined plane}
\item{\code{"tau"}}{shear stress on an inclined plane}
}
}
\description{
calculates the magnitudes of the normal stress and the shear stress
}
\note{
In addition to theta, One of the following two sets of data must be entered:
\enumerate{
\item{\code{"sigmaX"}, \code{"sigmaZ"}, \code{"tauXZ"}}
\item{\code{"sigma1"}, \code{"sigma3"}}
}

If theta is entered in conjunction with \code{"sigmaX"}, \code{"sigmaZ"}, and \code{"tauXZ"},
it is interpreted as the angle of inclination above the horizontal.  If
theta is entered in conjunction with the principal stresses, then it is
interpreted as the angle of inclination above the major principal plane.
}
\examples{
sigmaTrans(sigmaX = 80, sigmaZ = 120, tauXZ = 20, theta = 78)
}
\author{
Kyle Elmy and Jim Kaklamanos
}
