% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_math.R
\name{slerp}
\alias{slerp}
\title{Spherical Linear Interpolation (Slerp)}
\usage{
slerp(x0, x1, t)
}
\arguments{
\item{x0, x1}{objects of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"} of the first
and the last points of the to be interpolated arc.}

\item{t}{numeric. Interpolation factor(s) (\verb{t = [0, 1]}).}
}
\description{
Returns the spherical linear interpolation of points between two vectors
}
\details{
A Slerp path is the spherical geometry equivalent of a path along a line
segment in the plane; a great circle is a spherical geodesic.
}
\note{
For non-unit vectors the interpolation is not uniform.
}
\examples{
x0 <- Line(120, 7)
x1 <- Line(10, 13)
t <- seq(0, 1, .05)
xslerp <- slerp(x0, x1, t)

plot(xslerp, col = assign_col(t))
points(rbind(x0, x1))
}
\seealso{
\code{\link[=stereo_segment]{stereo_segment()}} for plotting the Slerp path as a line
}
