% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion.R
\name{slip_inversion}
\alias{slip_inversion}
\title{Stress Inversion for Fault-Slip Data}
\usage{
slip_inversion(
  x,
  method = c("michael", "angelier", "hansen", "yamaji", "wissi"),
  ...
)
}
\arguments{
\item{x}{\code{"Fault"} object where the rows are the observations, and the
columns the coordinates. Object must be complete, i.e. no \code{NA} values.
For Michael's, Angelier's, and Yamaji-Sato's methods, at least 4 rows
of fault measurements are required, while Hansen's method requires at least 7.}

\item{method}{character. The inversion algorithm, one of \code{"michael"} (the default)
for a bootstrapped linear inversion after Micheal (1984), \code{"angelier"}
for an iterative direct inversion after Angelier (1990) and Mostafa (2005),
\code{"yamaji"} for direct inversion using the 5d parameter space after Yamaji and Sato (2006),
\code{"hansen"} for direct inversion using the 9d parameter space after Hansen (2013), and
\code{"wissi"} for the Weighted Iterative Sigma-Space Inversion (WISSI).}

\item{...}{arguments passed to \code{\link[=slip_inversion_angelier]{slip_inversion_angelier()}},  \code{\link[=slip_inversion_michael]{slip_inversion_michael()}}, \code{\link[=slip_inversion_yamaji_sato]{slip_inversion_yamaji_sato()}},
\code{\link[=slip_inversion_hansen]{slip_inversion_hansen()}}, or \code{\link[=slip_inversion_wissi]{slip_inversion_wissi()}} depending on \code{method}.}
}
\value{
a named list with the following components:
\describe{
\item{\code{sigma}}{\code{"ellipsoid"} object. Best-fit devitoric stress tensor in input coordinate frame}
\item{\code{principal_axes}}{\code{"Line"} objects. Orientation of the principal stress axes as unit vectors (max to min)}
\item{\code{tensor_params}}{the four tensor parameters (Eq. 4.87)}
\item{\code{principal_vals}}{eigenvalues of the stress tensor (\eqn{\sigma_1 >= \sigma_2 >= \sigma_3})}
\item{\code{stress_shape}}{list Stress shape ratio. See \code{\link[=stress_shape]{stress_shape()}}.}
\item{\code{misfit}}{list. Misfit parameters. See \code{\link[=slip_inversion_misfit]{slip_inversion_misfit()}}.}
\item{\code{SHmax}}{numeric. Direction of maximum horizontal stress (in degrees)}
\item{\code{tau_mean}}{numeric. Average resolved shear stress on each plane. Should be close to 1.}
\item{\code{stress_components}}{matrix. The resolved shear and normal stresses, the slip and dilation tendency on each plane. See \code{\link[=sigma2shearnorm]{sigma2shearnorm()}} and \code{\link[=sigma2tendency]{sigma2tendency()}}.}
\item{\code{n_iter}}{number of Mostafa iterations performed}
\item{\code{method}}{character. The inversion method used, equal to \code{method} argument.}
}
}
\description{
Convenience function for direction inversion of fault-slip data to
derive the reduced stress tensor.
}
\examples{
set.seed(20250411)

# Use Angelier 1990 examples
nx <- length(angelier1990)

par(mfrow = c(nx/2, nx/2), mar = c(4, 1, 4, 1))
invisible(lapply(seq_len(nx), function(i) {
  x <- angelier1990[[i]]

  # Inversion after Michael (1984)
  res_michael <- slip_inversion(x, method = "michael", n_iter = 100, n = 100, res = 100)

  # Inversion after Angelier (1990)
  res_angelier <- slip_inversion(x, method = "angelier")
  
  res_yamaji <- slip_inversion(x, method = "yamaji")

  res_hansen <- slip_inversion(x, method = "hansen", type = "6d")
  
  res_wissi <- slip_inversion(x, method = 'wissi')

  stereoplot(title = names(angelier1990)[i], guides = FALSE)
  fault_plot(x, col = "gray80")
  points(res_michael$principal_axes, pch = 1:3, col = 2)
  points(res_angelier$principal_axes, pch = 1:3, col = 3)
  points(res_yamaji$principal_axes, pch = 1:3, col = 4)
  points(res_hansen$principal_axes, pch = 1:3, col = 5)
  points(res_wissi$principal_axes, pch = 1:3, col = 6)
  legend("topleft",
    pch = 1,
    legend = c("Michael (1984)", "Angelier (1990)", "Yamaji & Sato (2006)", 
      "Hansen (2013)", "WISSI"),
    col = 2:6, cex = 0.8
  )
  legend("bottom",
    pch = 1:3,
    legend = c("S1", "S2", "S3"),
    horiz = TRUE,
    cex = 0.8
  )
}))
dev.off()
}
\references{
Angelier, J. (1990). Inversion of field data in fault tectonics to obtain
the regional stress—III. A new rapid direct inversion method by analytical
means. Geophys. J. Int, 103, 363–376. \doi{10.1111/j.1365-246X.1990.tb01777.x}

Hansen, J. A. (2013). Direct inversion of stress, strain or strain rate
including vorticity: A linear method of homogenous fault-slip data inversion
independent of adopted hypothesis. Journal of Structural Geology, 51, 3–13.
\doi{10.1016/j.jsg.2013.03.014}

Michael, A. J. (1984). Determination of stress from slip data: Faults and
folds. Journal of Geophysical Research: Solid Earth, 89(B13), 11517–11526.
\doi{10.1029/JB089iB13p11517}

Yamaji, A., & Sato, K. (2006). Distances for the solutions of stress tensor
inversion in relation to misfit angles that accompany the solutions.
Geophysical Journal International, 167(2), 933–942.
\doi{10.1111/j.1365-246X.2006.03188.x}
}
\seealso{
Other stress-inversion:
\code{\link[=Fault_PT]{Fault_PT()}},
\code{\link[=slip_inversion_angelier]{slip_inversion_angelier()}},
\code{\link[=slip_inversion_hansen]{slip_inversion_hansen()}},
\code{\link[=slip_inversion_michael]{slip_inversion_michael()}},
\code{\link[=slip_inversion_simple]{slip_inversion_simple()}},
\code{\link[=slip_inversion_wissi]{slip_inversion_wissi()}},
\code{\link[=slip_inversion_yamaji_sato]{slip_inversion_yamaji_sato()}}
}
\concept{stress-inversion}
