% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{sort_by.strabo}
\alias{sort_by.strabo}
\title{Sorting StraboSpot Projects}
\usage{
\method{sort_by}{strabo}(x, y, ...)
}
\arguments{
\item{x}{object of class \code{"strabo"}}

\item{y}{Variables to sort by. Typically a column in \code{x$data} data.frame. To
sort by more than one column, take the form \code{list(g1, g2, ...)}.}

\item{...}{Additional arguments, typically passed on to \code{\link[=order]{order()}}.}
}
\value{
A sorted version of \code{x}.
}
\description{
Sorts \code{strabo} objects in the order determined by one or more other objects, typically columns in the
\code{data} data.frame. This will also sort the \code{planar} and \code{linear} objects.
}
\examples{
sort_by(strabo_prj, strabo_prj$data$dip_direction)
}
