% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcomp.R
\name{sph-prcomp}
\alias{sph-prcomp}
\alias{prcomp.Vec3}
\alias{prcomp.Ray}
\alias{prcomp.Line}
\alias{prcomp.Plane}
\alias{prcomp.Pair}
\title{Principal component (geodesic) analysis in the tangent space.}
\usage{
\method{prcomp}{Vec3}(x, center = geodesic_mean(x), n = 0L)

\method{prcomp}{Ray}(x, center = geodesic_mean(x), n = 0L)

\method{prcomp}{Line}(x, center = geodesic_mean(x), n = 0L)

\method{prcomp}{Plane}(x, center = geodesic_mean(x), n = 0L)

\method{prcomp}{Pair}(x, center = geodesic_mean(x), n = 0L, group = NULL)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"},
where the rows are the observations and the columns are the coordinates.}

\item{center}{A spherical object. Typically the geodesic mean of the x.}

\item{n}{real number (integer, 0 or >= 3). The number of points to
return on each of the four geodesics through the center.}

\item{group}{Symmetry group of \code{x}. See \code{\link[=symmetry_group]{symmetry_group()}} for details
If \code{NULL}, the group will be
automatically picked based on the class of \code{x}.}
}
\value{
A list consisting of
\describe{
\item{\code{rotation}}{3x3 rotation matrix}
\item{\code{magnitudes}}{2D real vector, non-negative. Magnitudes are
analogous to sample standard deviations. They are in  decreasing order and
quantify how dispersed the data are in those two directions.}
\item{\code{directions}}{2x2 real matrix, whose columns are unit-length vectors.
The corresponding directions to the magnitudes}
\item{\code{pcsFromRay}}{function to convert rays to 2-dimensional vectors}
\item{\code{rayFromPCs}}{function to convert 2-dimensional vectors to rays}
\item{\code{curves}}{list of two lists of (2 \code{n} + 1) rays, only if \code{n} >= 1}
\item{\code{tangents}}{Tangents from \code{directions} and \code{rotation}}
}

If \code{x} is a \code{"Pair"} object the list only contains \code{magnitudes}, \code{directions}
and \code{curves} (if \code{n}>=1).
}
\description{
Principal component (geodesic) analysis in the tangent space.
}
\details{
Appropriate only if the data set is tightly concentrated about the center.
}
\examples{
res <- prcomp(example_lines, n = 10)

stereoplot(sub = paste("SD1:", round(res$magnitudes[1], 2), "| SD2:", round(res$magnitudes[2], 2)))
points(example_lines, pch = 16, cex = .8)
invisible(lapply(res$curves, stereo_lines, col = "red", lwd = 1.5))
}
