% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stats.R
\name{sph_cluster}
\alias{sph_cluster}
\title{Cluster spherical data}
\usage{
sph_cluster(
  x,
  k,
  method = c("hclust", "kmeans", "diana", "agnes", "pam", "clara", "fanny"),
  ...
)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"},
where the rows are the observations and the columns are the coordinates.}

\item{k}{integer. Number of desired clusters.}

\item{method}{character. Clustering method to be applied. Currently implemented are
\describe{
\item{\code{"hclust"}}{Hierarchical Clustering using \code{\link[stats:hclust]{stats::hclust()}}, the default)}
\item{\code{"kmeans"}}{K-Means Clustering using \code{\link[stats:kmeans]{stats::kmeans()}})}
\item{\code{"pam"}}{Partitioning Around Medoids using \code{\link[cluster:pam]{cluster::pam()}}}
\item{\code{"agnes"}}{Agglomerative hierarchical clustering using \code{\link[cluster:agnes]{cluster::agnes()}}}
\item{\code{"diana"}}{Divisive hierarchical clustering using \code{\link[cluster:diana]{cluster::diana()}}}
\item{\code{"clara"}}{Clustering Large Applications using \code{\link[cluster:clara]{cluster::clara()}}}
\item{\code{"fanny"}}{Fuzzy Analysis Clustering using \code{\link[cluster:fanny]{cluster::fanny()}}}
}}

\item{...}{optional arguments passed to cluster algorithm.}
}
\value{
output of applied cluster function
}
\description{
Finds k groups of clusters using the angular distance matrix
}
\examples{
set.seed(20250411)
x1 <- rvmf(100, mu = Line(90, 0), k = 20)
x2 <- rvmf(100, mu = Line(0, 0), k = 20)
x3 <- rvmf(100, mu = Line(0, 90), k = 20)
x123 <- rbind(x1, x2, x3)
cl <- sph_cluster(x123, k = 3)
plot(x123, col = cl$cluster)

# sph_cluster(simongomez, k = 3)
}
\seealso{
\code{\link[=dist]{dist()}}
}
