% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{split_trailing_letters}
\alias{split_trailing_letters}
\title{Parse measurement and direction strings}
\usage{
split_trailing_letters(x)
}
\arguments{
\item{x}{character or number}
}
\value{
list
}
\description{
Parse measurement and direction strings
}
\examples{
test <- c("45NW", "4SE")
split_trailing_letters(test)
}
