% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stats.R
\name{stats}
\alias{stats}
\alias{sph_mean}
\alias{sph_sd}
\alias{sph_var}
\alias{sph_confidence_angle}
\alias{rdegree}
\alias{sd_error}
\alias{delta}
\alias{estimate_k}
\title{Statistical estimators of the distribution of a set of vectors}
\usage{
sph_mean(x, na.rm = TRUE, ...)

sph_sd(x, ...)

sph_var(x, ...)

sph_confidence_angle(x, w = NULL, alpha = 0.05, na.rm = TRUE)

rdegree(x, w = NULL, na.rm = FALSE)

sd_error(x, w = NULL, na.rm = FALSE)

delta(x, w = NULL, na.rm = TRUE)

estimate_k(x, w = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}, where the
rows are the observations and the columns are the coordinates.}

\item{na.rm}{logical. Whether \code{NA} values should be removed before the computation proceeds.}

\item{...}{arguments passed to function call}

\item{w}{numeric. Optional weights for each observation.}

\item{alpha}{numeric. Significance level for the confidence angle (default is 0.05 for a 95\% confidence angle).}
}
\description{
Statistical estimators of the distribution of a set of vectors
}
\details{
These statistical estimators are based on the resultant vector of a set of \eqn{n} vectors \eqn{x_1, \ldots, x_n} (Mardia 1972). The resultant vector is given by
\deqn{\bar{\mathbf{x}} = \sum_{i=1}^{n} \mathbf{x}_i}

The mean resultant is defined as
\deqn{\bar{\mathbf{R}} = ||\bar{\mathbf{x}}|| = \sqrt{x_x^2 + x_y^2 + x_z^2}}

\code{sph_mean} returns the spherical mean of a set of vectors
(object of class of \code{x}), that is the maximum likelihood estimate of the mean direction given by the arithmetic mean of the vector components normalized by the mean resultant vector:
\deqn{\mu = \frac{\bar{\mathbf{x}}}{\bar{\mathbf{R}}}}

\code{sph_var} returns the spherical variance (numeric).
\deqn{S = 1 - \bar{\mathbf{R}}}

\code{sph_sd} returns the spherical standard deviation (numeric) given as the half
apical angle of a cone about the mean vector. In degrees if \code{x} is a
\code{"Plane"} or \code{"Line"}, or in radians if otherwise.
\deqn{s = \sqrt{\log(1 / \bar{\mathbf{R}}^2))}}

\code{delta} returns the half apical angle of the cone containing ~63\% of the data
(in degrees if \code{x} is a \code{"Plane"} or \code{"Line"}, or in radians
if otherwise). For enough large sample it approaches the angular standard
deviation (\code{"csd"}) of the Fisher statistics.
\deqn{\delta = \arccos(\bar{\mathbf{R}})}

\code{rdegree} returns the degree of preferred orientation of vectors, range: (0, 1).
\deqn{r = \frac{2 \bar{\mathbf{R}} - n}{n}}

\code{sd_error} returns the spherical standard error (numeric). If the number of
data is less than 25, if will print a additional message, that the output
value might not be a good estimator.
\deqn{\text{SDE} = \sqrt{\frac{1 - \frac{1}{n} \sum_{i=1}^{n} (\mu \cdot x_i)^2}{n \bar{\mathbf{R}}^2}}}

\code{sph_confidence_angle} returns the half-apical angle \eqn{q} of a cone about the
mean \eqn{\mu} (in degrees if \code{x} is a \code{"Plane"} or \code{"Line"}, or in radians
if otherwise). The \eqn{100(1-\alpha)\%} confidence interval is than given by \eqn{\mu \pm q}.
\deqn{q = \arcsin(\sqrt{-\log(\alpha)} \cdot \text{SDE})}

\code{estimate_k} returns the estimated concentration of the von Mises-Fisher distribution \eqn{\hat{\kappa}} (after Sra, 2011).
\deqn{\hat{\kappa} = \frac{\bar{R}(p - \bar{R}^2)}{1 - \bar{R}^2}}
where \eqn{p} is the dimension of the data (3 for spherical data).
}
\examples{
set.seed(20250411)
x <- rvmf(100, mu = Line(120, 50), k = 5)
sph_mean(x)
sph_sd(x)
sph_var(x)
delta(x)
rdegree(x)
sd_error(x)
sph_confidence_angle(x)
estimate_k(x)

#' weights:
x2 <- Line(c(0, 0), c(0, 90))
sph_mean(x2)
sph_mean(x2, w = c(1, 2))
sph_var(x2)
sph_var(x2, w = c(1, 2))
}
\references{
Mardia, Kanti; Jupp, P. E. (1999). Directional Statistics. John Wiley & Sons Ltd. ISBN 978-0-471-95333-3.

Sra, S. A short note on parameter approximation for von Mises-Fisher distributions: and a fast implementation of I s (x). Comput Stat 27, 177–190 (2012). https://doi.org/10.1007/s00180-011-0232-x
}
\seealso{
\code{\link[=projected_mean]{projected_mean()}} for projected mean, \code{\link[=geodesic_mean_line]{geodesic_mean_line()}} and \code{\link[=geodesic_mean_pair]{geodesic_mean_pair()}} for the geodesic mean of lines and pairs, respectively.
or \code{\link[=geodesic_mean]{geodesic_mean()}} and \code{\link[=geodesic_var]{geodesic_var()}} as a convenience wrapper for all spherical data types.
}
