% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereo-pair}
\alias{stereo-pair}
\alias{stereo_pair}
\title{Stereographic projection of pairs}
\usage{
stereo_pair(
  x,
  pch = 16,
  col = 1,
  lwd = 1,
  lty = 1,
  lab = NULL,
  cex = 1,
  greatcircles = TRUE,
  upper.hem = FALSE,
  earea = TRUE
)
}
\arguments{
\item{x}{Object of class \code{"Fault"}}

\item{pch, col, lwd, lty}{plotting parameters for planes and lines}

\item{lab}{character. text labels}

\item{cex}{character expansion of labels}

\item{greatcircles}{logical. Whether greatcircles are displayed (\code{TRUE}, the default) or poles to planes (\code{FALSE})}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}, the default).}

\item{earea}{logical \code{TRUE} for Lambert equal-area projection (also "Schmidt net"; the default), or
\code{FALSE} for meridional stereographic projection (also "Wulff net" or "Stereonet").}
}
\description{
Visualization of pairs (planes and lines) in a stereographic projection.
}
\note{
\code{"Plane"} objects will be displayed as pole to the plane.
}
\examples{
faults <- Fault(
  c(0, 90, 180, 270),
  c(80, 45, 80, 45),
  c(0, 170, 180, 315),
  c(80, 10, 80, 36),
  c(1, -1, 0, 1)
)
# stereoplot()
# stereo_fault(faults, col = 1:4)
# stereo_fault(faults, col =1:4, hoeppner = TRUE)
# legend("bottomright", c("normal", "thrust", "unknown", "normal"), fill = 1:4)

stereoplot()
stereo_pair(faults, col = 1:4)
}
