% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereo_cones}
\alias{stereo_cones}
\alias{stereo_smallcircle}
\alias{stereo_greatcircle}
\title{Spherical Projection of Cones}
\usage{
stereo_smallcircle(
  x,
  d = 90,
  col = par("col"),
  N = 1000,
  upper.hem = NULL,
  earea = NULL,
  lty = par("lty"),
  lwd = par("lwd"),
  fill = FALSE,
  border = NA,
  radius = NULL,
  ...
)

stereo_greatcircle(x, ...)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"},
where the rows are the observations and the columns are the coordinates.}

\item{d}{numeric. conical angle in degrees.}

\item{col, lty, lwd}{color, line type, and line width parameters}

\item{N}{integer. number of points to calculate}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}). Defaults to
\code{getOption("structr.upper.hem")}.}

\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection, or \code{FALSE} for meridional stereographic projection.
Defaults to \code{getOption("structr.earea")}.}

\item{fill}{logical. Whether to fill the inner part of the small-circle? \code{FALSE} by default.}

\item{border}{Color of the filled small-circle's outline (ignored if \code{fill=FALSE})}

\item{radius}{numeric. Radius of circle. Defaults to \code{getOption("structr.radius")}.}

\item{...}{optional graphical parameters passed to \code{\link[graphics:lines]{graphics::lines()}} and (if \code{fill=TRUE}) \code{\link[graphics:polygon]{graphics::polygon()}}}
}
\description{
Visualization of small-circles and great-circles in a stereographic or equal-area projection.
}
\examples{
stereoplot()
stereo_point(Line(90, 5), lab = "L")
stereo_smallcircle(Line(90, 5), d = 10)
stereo_point(Plane(120, 30), lab = "P", col = "red")
stereo_greatcircle(Plane(120, 30), col = "red")

stereoplot()
stereo_point(Line(c(129, 90), c(30, 5)), lab = c("L1", "L2"))
stereo_smallcircle(Line(c(129, 90), c(30, 5)), d = c(10, 5), col = 1:2, lty = 1:2, lwd = 1:2)

# Filled cones:
stereoplot()
stereo_smallcircle(Line(c(90, 120), c(5, 5)), d = c(5, 20), col = c('grey60', 'grey40'), border = c('red', 'blue'), fill = TRUE)
stereo_point(Line(c(90, 120), c(5, 5)), col = c('red', 'blue'))
}
\seealso{
\code{\link[=lines.spherical]{lines.spherical()}}, \code{\link[=stereo_segment]{stereo_segment()}}, \code{\link[=stereo_lines]{stereo_lines()}}

Other stereo-plot: 
\code{\link{arrows}()},
\code{\link{fault-plot}},
\code{\link{lines}()},
\code{\link{plot-spherical}},
\code{\link{points.spherical}()},
\code{\link{stereo_confidence}()},
\code{\link{stereo_contour}},
\code{\link{stereo_lines}()},
\code{\link{stereo_point}()},
\code{\link{stereo_segment}()},
\code{\link{stereoplot}()},
\code{\link{stereoplot_guides}()},
\code{\link{stereoplot_ticks}()},
\code{\link{text.spherical}()}
}
\concept{stereo-plot}
