% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_contours.R
\name{stereo_contour}
\alias{stereo_contour}
\alias{contour}
\alias{image}
\alias{density-plot}
\alias{density_plot}
\alias{contour.spherical}
\alias{contour.sph_density}
\alias{contourf}
\alias{image.spherical}
\alias{image.sph_density}
\title{Plot Spherical Densities in a Spherical Projection}
\usage{
\method{contour}{spherical}(
  x,
  add = FALSE,
  nlevels = 10L,
  density.params = list(),
  col.palette = viridis::viridis,
  col = NULL,
  col.params = list(),
  upper.hem = NULL,
  radius = NULL,
  ...
)

\method{contour}{sph_density}(
  x,
  add = FALSE,
  nlevels = 10L,
  density.params = list(),
  col.palette = viridis::viridis,
  col = NULL,
  col.params = list(),
  upper.hem = NULL,
  radius = NULL,
  ...
)

contourf(
  x,
  add = FALSE,
  nlevels = 10L,
  density.params = list(),
  col.palette = viridis::viridis,
  col.params = list(),
  upper.hem = NULL,
  radius = NULL,
  ...
)

\method{image}{spherical}(
  x,
  add = FALSE,
  nlevels = 10L,
  density.params = list(),
  col.palette = viridis::viridis,
  col.params = list(),
  upper.hem = NULL,
  radius = NULL,
  ...
)

\method{image}{sph_density}(
  x,
  add = FALSE,
  nlevels = 10L,
  density.params = list(),
  col.palette = viridis::viridis,
  col.params = list(),
  upper.hem = NULL,
  radius = NULL,
  ...
)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"} or \code{'sph_density'} (for plotting only).}

\item{add}{logical. Whether the contours should be added to an existing plot.}

\item{nlevels}{integer. Number of contour levels for plotting}

\item{density.params}{list. Optional arguments passed to \code{\link[=density.spherical]{density.spherical()}},
such as \code{FUN}, \code{n}, \code{sigma}, and \code{weights}}

\item{col.palette}{a color palette function to be used to assign colors in the plot.}

\item{col}{colour(s) for the contour lines drawn. If \code{NULL}, lines are color based on \code{col.palette}.}

\item{col.params}{list. Arguments passed to \code{col.palette}}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}). Defaults to
\code{getOption("structr.upper.hem")}.}

\item{radius}{numeric. Radius of circle. Defaults to \code{getOption("structr.radius")}.}

\item{...}{optional parameters passed to \code{\link[graphics:image]{graphics::image()}} or \code{\link[graphics:contour]{graphics::contour()}}.}
}
\value{
list containing the stereographic coordinates of the grid,
the counts, and the density.
}
\description{
Kamb counts and densities on the sphere. \code{contour} plots the contour lines,
\code{contourf} displays a contour plot with the areas between the contours filled,
and \code{image} creates a dense grid of colored rectangles.
}
\note{
Densities are plotted in the equal-area projection!
}
\examples{
set.seed(20250411)
x <- rfb(100, mu = Line(120, 10), k = 5, A = diag(c(-1, 0, 1)))
contour(x)

x_density <- density(x)
contourf(x_density,
  col.params = list(direction = -1, begin = .05, end = .95, alpha = .75)
)
stereo_point(x, col = "black", pch = 21)

image(x)
stereo_point(x, col = "lightgrey", pch = 21)


# complete example:
par(mfrow = c(1, 2))
wp <- 6 / ifelse(is.na(example_planes_df$quality), 6, example_planes_df$quality)
my_planes <- Plane(example_planes_df$dipdir, example_planes_df$dip)
fabric_p <- vollmer(my_planes)["D"]
my_planes_eig <- ot_eigen(my_planes)

stereoplot(guides = TRUE, col = "grey96")
points(my_planes, col = "grey", pch = 16, cex = .5)
contour(my_planes, add = TRUE, weights = wp)
points(my_planes_eig$vectors[3, ], col = "black", pch = 16)
lines(my_planes_eig$vectors[3, ], ang = 90, col = "black", pch = 16)
title(
  main = "Planes",
  sub = paste0(
    "N: ", nrow(my_planes), " | Fabric strength: ", round(fabric_p, 2),
    "\nLambert equal area, lower hemisphere projection"
  )
)

my_lines <- Line(example_lines_df$trend, example_lines_df$plunge)
wl <- 6 / ifelse(is.na(example_lines_df$quality), 6, example_lines_df$quality)
fabric_l <- vollmer(my_lines)["D"]

stereoplot(guides = TRUE, col = "grey96")
points(my_lines, col = "grey", pch = 16, cex = .5)
contour(my_lines, add = TRUE, weights = wl)
points(sph_mean(my_lines, w = wl), col = "black", pch = 16)
title(
  main = "Lines",
  sub = paste0(
    "N: ", nrow(my_lines), " | Fabric strength: ", round(fabric_l, 2),
    "\nLambert equal area, lower hemisphere projection"
  )
)
}
\seealso{
Other stereo-plot:
\code{\link{arrows}},
\code{\link{fault-plot}},
\code{\link{lines}},
\code{\link{plot-spherical}},
\code{\link[=points.spherical]{points.spherical()}},
\code{\link{stereo_cones}},
\code{\link[=stereo_confidence]{stereo_confidence()}},
\code{\link[=stereo_lines]{stereo_lines()}},
\code{\link[=stereo_point]{stereo_point()}},
\code{\link[=stereo_segment]{stereo_segment()}},
\code{\link[=stereoplot]{stereoplot()}},
\code{\link[=stereoplot_guides]{stereoplot_guides()}},
\code{\link[=stereoplot_ticks]{stereoplot_ticks()}},
\code{\link[=text.spherical]{text.spherical()}}

Other density:
\code{\link{density}},
\code{\link{density-funs}}
}
\concept{density}
\concept{stereo-plot}
