% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereo_coords}
\alias{stereo_coords}
\title{Stereographic and Equal-Area Projection}
\usage{
stereo_coords(az, inc, upper.hem = NULL, earea = NULL, radius = 1, fix = TRUE)
}
\arguments{
\item{az, inc}{numeric vectors. Azimuth and Inclination in degrees.}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}). Defaults to
\code{getOption("structr.upper.hem")}.}

\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection, or \code{FALSE} for meridional stereographic projection.
Defaults to \code{getOption("structr.earea")}.}

\item{radius}{numeric. Radius of circle. Defaults to \code{getOption("structr.radius")}.}

\item{fix}{logical. Whether coordinates should always plot in one hemisphere?}
}
\value{
two-column vector with the transformed coordinates
}
\description{
Transformation of spherical coordinates into the stereographic or equal-area projection
}
\examples{
stereo_coords(90, 10)
stereo_coords(90, 10, earea = TRUE, upper.hem = TRUE)
}
\seealso{
\link{structr-options}
}
