% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contouring.R
\name{stereo_density_grid}
\alias{stereo_density_grid}
\title{Plot density grid in a stereonet}
\usage{
stereo_density_grid(x, ..., binned = FALSE, pal = viridis::viridis, breaks = 5)
}
\arguments{
\item{x}{spherical data}

\item{...}{arguments passed top \code{\link[=blank_grid]{blank_grid()}}}

\item{binned}{logical. Whether the density colors should be binned}

\item{pal}{color function}

\item{breaks}{for binning}
}
\value{
plot
}
\description{
Plot density grid in a stereonet
}
\examples{
test <- Line(c(90, 85, 105), c(45, 50, 40))
stereoplot(guides = F, centercross = FALSE)
stereo_density_grid(test, binned = FALSE, n = 10000)
stereo_point(test, cex = 2, col = "red")
}
