% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereo_fms}
\alias{stereo_fms}
\title{Plot Focal Mechanism Solution in a Spherical Plot}
\usage{
stereo_fms(
  sigma,
  friction = 0.6,
  fill = TRUE,
  col = "#BEBEBE80",
  border = "black",
  ...
)
}
\arguments{
\item{sigma}{symmetric 3x3 matrix. The (reduced) stress tensor.}

\item{friction}{numeric. Coefficient of friction (0.6 by default)}

\item{fill}{logical. Whether to fill the inner part of the small-circle? \code{FALSE} by default.}

\item{col}{fill color of quadrant}

\item{border}{Color of the filled small-circle's outline (ignored if \code{fill=FALSE})}

\item{...}{optional plotting parameters passed to \code{\link[=stereo_greatcircle]{stereo_greatcircle()}}}
}
\value{
\code{"Fault"} object
}
\description{
Shows the principal faults associated with a stress tensor and a given
friction coefficient using the focal mechanism / beach ball style.
}
\examples{
f <- angelier1990$TYM
sig <- reduced_stress(f)

stereoplot()
stereo_fms(sig)
}
\seealso{
\code{\link[=stereo_fms]{stereo_fms()}} and \code{\link[=principal_fault]{principal_fault()}}
}
