% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereo_path}
\alias{stereo_path}
\title{Deformation Paths in Stereoplot}
\usage{
stereo_path(
  x,
  type = c("l", "p", "b"),
  add = TRUE,
  n = 5L,
  upper.hem = NULL,
  earea = NULL,
  ...
)
}
\arguments{
\item{x}{list of spherical objects.}

\item{type}{character. Type of plotting the path. \code{"l"}if only the path line
should be shown, \code{"p"} if only the vectors should be shown, and \code{"b"} if both
should be plotted.}

\item{add}{logical. Whether the paths should be added to a existing plot or a
new one should be created?}

\item{n}{integer. Number of points to be interpolated}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}). Defaults to
\code{getOption("structr.upper.hem")}.}

\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection, or \code{FALSE} for meridional stereographic projection.
Defaults to \code{getOption("structr.earea")}.}

\item{...}{parameters passed to \code{\link[=stereo_point]{stereo_point()}}, \code{\link[=stereo_smallcircle]{stereo_smallcircle()}}, \code{\link[=stereo_greatcircle]{stereo_greatcircle()}}, or \code{\link[=fault_plot]{fault_plot()}}}
}
\value{
\code{NULL}
}
\description{
Plot the paths of deformed vectors in a equal-area or stereographic projection
}
\examples{
d <- defgrad_from_generalshear(k = 2.5, gamma = 0.9)
v <- velgrad(d, time = 10)
d_steps <- defgrad(v, time = 10, steps = 2)

# apply on orientation data
set.seed(20250411)
l <- rvmf(100, mu = Line(0, 90), k = 100)
l_trans <- lapply(d_steps, function(i) {
  transform_linear(l, i)
})

# plot in stereonet
axes <- Vec3(c(1, 0, 0), c(0, 1, 0), c(0, 0, 1))
stereo_path(l_trans, type = "l", add = FALSE)
stereo_path(l_trans, type = "p", col = assign_col(seq_along(l_trans)), pch = 16, cex = .4)
}
\seealso{
\code{\link[=defgrad]{defgrad()}} and \code{\link[=velgrad]{velgrad()}}
}
