% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereo_segment}
\alias{stereo_segment}
\title{Great-circle Segment Between Two Vectors}
\usage{
stereo_segment(
  x,
  y,
  upper.hem = NULL,
  earea = NULL,
  n = 100L,
  radius = NULL,
  ...
)
}
\arguments{
\item{x, y}{objects of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}). Defaults to
\code{getOption("structr.upper.hem")}.}

\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection, or \code{FALSE} for meridional stereographic projection.
Defaults to \code{getOption("structr.earea")}.}

\item{n}{integer. number of points along great-circle (100 by default)}

\item{radius}{numeric. Radius of circle. Defaults to \code{getOption("structr.radius")}.}

\item{...}{optional graphical parameters passed to \code{\link[graphics:lines]{graphics::lines()}} and (if \code{fill=TRUE}) \code{\link[graphics:polygon]{graphics::polygon()}}}
}
\description{
Plots the great-circle segment between two vectors
}
\examples{
x <- Line(120, 7)
y <- Line(10, 13)
plot(rbind(x, y))
stereo_segment(x, y, col = "red")

# For multiple segments use lapply():
set.seed(20250411)
mu <- Line(45, 10)
x <- rvmf(100, mu = mu) |> Line()
plot(x)
invisible(lapply(seq_len(nrow(x)), FUN = function(i) {
  stereo_segment(x[i, ], mu, col = i)
}))
points(mu, pch = 16, col = "white")
}
\seealso{
\code{\link[=slerp]{slerp()}}, \link{stereo_greatcircle}, \code{\link[=stereo_lines]{stereo_lines()}}

Other stereo-plot: 
\code{\link{arrows}()},
\code{\link{fault-plot}},
\code{\link{lines}()},
\code{\link{plot-spherical}},
\code{\link{points.spherical}()},
\code{\link{stereo_cones}},
\code{\link{stereo_confidence}()},
\code{\link{stereo_contour}},
\code{\link{stereo_lines}()},
\code{\link{stereo_point}()},
\code{\link{stereoplot}()},
\code{\link{stereoplot_guides}()},
\code{\link{stereoplot_ticks}()},
\code{\link{text.spherical}()}
}
\concept{stereo-plot}
