% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereo_shmax}
\alias{stereo_shmax}
\title{Horizontal directions}
\usage{
stereo_shmax(
  azi,
  ...,
  type = c("arrows", "line"),
  shmin = FALSE,
  radius = NULL,
  arrow.offset = 0.02,
  arrow.length = 0.1,
  arrow.head = arrow.length
)
}
\arguments{
\item{azi}{numeric. Angle of maximum horizontal stress in degrees}

\item{...}{... arguments passed to \code{\link[graphics:segments]{graphics::segments()}} or \code{\link[graphics:arrows]{graphics::arrows()}}}

\item{type}{character. Either \code{'line'} or \code{'arrows'} for a straight line or arrows at the perimeter of the plot.}

\item{shmin}{logical. Whether the minimum horizontal stress should be indicated too?}

\item{radius}{numeric. Radius of circle. Defaults to \code{getOption("structr.radius")}.}

\item{arrow.offset}{numeric. Offset of the arrows from the perimeter.}

\item{arrow.length}{numeric. Length of the arrows.}

\item{arrow.head}{numeric. Length of the arrow head.}
}
\description{
Horizontal directions
}
\examples{
stereoplot()
stereo_shmax(30, shmin = TRUE)
}
