% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion.R
\name{stereo_stress_grid}
\alias{stereo_stress_grid}
\title{Stress Grid}
\usage{
stereo_stress_grid(sigma, n = 15, radius = NULL, ...)
}
\arguments{
\item{sigma}{symmetric 3x3 matrix. The (reduced) stress tensor.}

\item{n}{integer. The size of the grid of fault planes. \code{20} by default.}

\item{radius}{numeric. Radius of circle. Defaults to \code{getOption("structr.radius")}.}

\item{...}{optional arguments passed to \code{\link[=stereo_hoeppener]{stereo_hoeppener()}}}
}
\value{
object of class \code{"Fault"}
}
\description{
Plots the slip directions for a grid of fault plane normals and a given stress
tensor in a stereographic/equal-area projection
}
\examples{
sigma <- reduced_stress(angelier1990$AVB)

dev.off()
stereoplot(guides = FALSE)
stereo_stress_grid(sigma, type = 'hoeppener')

pr <- sigma2stress(sigma)
points(pr$axes, col = 2:4, pch = 16)
text(pr$axes, label = rownames(pr$axes), col = 2:4, adj = -.25)
}
\seealso{
\code{\link[=slip_inversion]{slip_inversion()}}; \code{\link[=sigma2slip]{sigma2slip()}}
}
