% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereoplot}
\alias{stereoplot}
\title{Stereographic and Equal Area Projection}
\usage{
stereoplot(
  earea = NULL,
  guides = NULL,
  d = NULL,
  col = NULL,
  lwd = NULL,
  lty = NULL,
  border.col = NULL,
  title = NULL,
  sub = NULL,
  origin.text = NULL,
  labels = NULL,
  ladj = NULL,
  centercross = NULL,
  ticks = NULL,
  radius = NULL,
  center = NULL
)
}
\arguments{
\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection, or \code{FALSE} for meridional stereographic projection.
Defaults to \code{getOption("structr.earea")}.}

\item{guides}{logical. Whether guides of the projection grid should be added
to the plot. Defaults to \code{getOption("structr.guides")}.}

\item{d}{integer. Angle distance between guides. Defaults to \code{getOption("structr.d")}.}

\item{col}{Color of guide lines. Defaults to \code{getOption("structr.col")}.}

\item{lwd}{Width of guide lines. Defaults to \code{getOption("structr.lwd")}.}

\item{lty}{Type of guide lines. Defaults to \code{getOption("structr.lty")}.}

\item{border.col}{color of primitive circle (frame), center-cross and ticks of the stereo plot.
Defaults to \code{getOption("structr.border.col")}.}

\item{title, sub}{character. Title and subtitle of plot}

\item{origin.text}{character. Text at origin of plot. Defaults to \code{getOption("structr.origin.text")}.}

\item{labels}{this can either be a logical value specifying whether (numerical)
annotations are to be made next to the tick marks, or a character or expression
vector of labels to be placed next to the tick points. Defaults to \code{getOption("structr.labels")}.}

\item{ladj}{adjustment for all labels away from origin of projection circle.
This essentially an amount that is added to \code{radius} and the length of the ticks. Defaults to \code{getOption("structr.ladj")}.}

\item{centercross}{logical. Whether a center cross should be added (\code{TRUE} by default).
Defaults to \code{getOption("structr.centercross")}.}

\item{ticks}{integer. Angle between ticks. if \code{NULL} (the default), no ticks are drawn.
Defaults to \code{getOption("structr.ticks")}.}

\item{radius}{numeric. Radius of circle. Defaults to \code{getOption("structr.radius")}.}

\item{center}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}
specifying the center of the projection grid. If \code{NULL} (the default), the center
is at the origin of the plot. Overwrites \code{guides}.}
}
\description{
Initialize the plot for equal-angle stereographic projections (Wulff) or Lambert
equal-area projections (Schmidt).
}
\examples{
stereoplot(guides = TRUE)

stereoplot(ticks = 30, title = "title", sub = "subtitle", border.col = "purple", labels = TRUE)

stereoplot(center = Line(120, 50))
}
\seealso{
\link{structr-options}

Other stereo-plot: 
\code{\link{arrows}()},
\code{\link{fault-plot}},
\code{\link{lines}()},
\code{\link{plot-spherical}},
\code{\link{points.spherical}()},
\code{\link{stereo_cones}},
\code{\link{stereo_confidence}()},
\code{\link{stereo_contour}},
\code{\link{stereo_lines}()},
\code{\link{stereo_point}()},
\code{\link{stereo_segment}()},
\code{\link{stereoplot_guides}()},
\code{\link{stereoplot_ticks}()},
\code{\link{text.spherical}()}
}
\concept{stereo-plot}
