% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereoplot_frame}
\alias{stereoplot_frame}
\title{Stereoplot frame}
\usage{
stereoplot_frame(ndiv = 144, radius = 1, ...)
}
\arguments{
\item{ndiv}{integer. Resolution of circle's line}

\item{radius}{numeric. Radius of circle}

\item{...}{optional arguments passed to \code{\link[graphics:lines]{graphics::lines()}}}
}
\description{
Adds a (primitive) circle with given radius to an existing plot
}
\examples{
plot(c(-1, 1), c(-1, 1), type = "n", asp = 1)
stereoplot_frame(col = "red", lwd = 3)
}
\seealso{
\code{\link[=stereoplot]{stereoplot()}}, \code{\link[=stereoplot_ticks]{stereoplot_ticks()}}, \code{\link[=stereoplot_guides]{stereoplot_guides()}}
}
