% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereoplot_guides}
\alias{stereoplot_guides}
\title{Stereoplot Gridlines}
\usage{
stereoplot_guides(
  d = 10,
  earea = NULL,
  radius = NULL,
  center = NULL,
  equator = TRUE,
  lwd = 0.5,
  equator_lwd = lwd * 1.5,
  ...
)
}
\arguments{
\item{d}{angle spacing between grid lines of the projection}

\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection, or \code{FALSE} for meridional stereographic projection.
Defaults to \code{getOption("structr.earea")}.}

\item{radius}{numeric. Radius of circle. Defaults to \code{getOption("structr.radius")}.}

\item{center}{Center position of grid lines. If \code{NULL} (the default),
grid lines are centered on the origin of the stereoplot. Otherwise, this
should be a spherical object.}

\item{equator}{logical. Whether the grid equator should be shown no matter how the grid is constructed via \code{d}. \code{TRUE} by default.}

\item{lwd}{Width of guide lines. Defaults to \code{getOption("structr.lwd")}.}

\item{equator_lwd}{numeric. The line width of the drawn equator grid line. By default, the equator is \code{1.5} times thicker then \code{lwd}.}

\item{...}{optional arguments passed to \code{\link[graphics:lines]{graphics::lines()}}}
}
\description{
Adds equal-area or equal-angle projection grid to an existing stereoplot
}
\examples{
plot(c(-1, 1), c(-1, 1), type = "n", asp = 1)
stereoplot_guides(d = 5, earea = FALSE, col = "green", rotation = 20)

plot(c(-1, 1), c(-1, 1), type = "n", asp = 1)
stereoplot_guides(d = 15, earea = TRUE, col = "orange", rotation = 90)

plot(c(-1, 1), c(-1, 1), type = "n", asp = 1)
stereoplot_guides(d = 15, earea = FALSE, center = Line(120, 50), col = 'red')
}
\seealso{
Other stereo-plot: 
\code{\link{arrows}()},
\code{\link{fault-plot}},
\code{\link{lines}()},
\code{\link{plot-spherical}},
\code{\link{points.spherical}()},
\code{\link{stereo_cones}},
\code{\link{stereo_confidence}()},
\code{\link{stereo_contour}},
\code{\link{stereo_lines}()},
\code{\link{stereo_point}()},
\code{\link{stereo_segment}()},
\code{\link{stereoplot}()},
\code{\link{stereoplot_ticks}()},
\code{\link{text.spherical}()}
}
\concept{stereo-plot}
