% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_quantitites.R
\name{strain}
\alias{strain}
\alias{longitudinal_strain}
\alias{angular_strain}
\alias{quadratic_elongation}
\alias{stretch}
\title{Strain quantities}
\usage{
longitudinal_strain(l, l0)

angular_strain(psi, degree = TRUE)

quadratic_elongation(e = NULL, l = NULL, l0 = NULL)

stretch(e = NULL, l = NULL, l0 = NULL, lambda = NULL)
}
\arguments{
\item{l, l0}{numeric. Final and original length, respectively}

\item{psi}{numeric. Angle}

\item{degree}{logical. Whether \code{psi} is given in degree (the default) or radians}

\item{e}{numeric. Elongation}

\item{lambda}{numeric. Quadratic elongation}
}
\description{
Converts different quantifications of strain.
}
\details{
\emph{Longitudianal strain} (also \strong{elongation}) is the change in length divided by the original length:
\deqn{e = (l-l_o)/l_o}

\emph{Angular strain} is the change in angle between two lines that were initially perpendicular"
\deqn{\gamma = \tan \psi}

\emph{Stretch} is the ratio of the final length and the original length:
\deqn{s = \sqrt{\lambda} = l/l_o = 1 + e}

\emph{Quadratic elongation} is the quadratic stretch:
\deqn{\lambda = s^2 = (l/l_o)^2 = (1+e)^2}
}
\examples{
longitudinal_strain(l=10,l0=5) 
angular_strain(25)
quadratic_elongation(l=10,l0=5)
stretch(l=10,l0=5)
}
