% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohR.R
\name{stress-comp}
\alias{stress-comp}
\alias{diff_stress}
\alias{mean_stress}
\alias{shear_stress}
\alias{normal_stress}
\alias{fracture_angle}
\title{Stress components}
\usage{
diff_stress(s1, s3)

mean_stress(s1, s3)

shear_stress(s1, s3, theta)

normal_stress(s1, s3, theta)

fracture_angle(mu)
}
\arguments{
\item{s1, s3}{numeric. Magnitudes of maximum and minimum principal stress (\eqn{\sigma_1} and \eqn{\sigma_3}), respectively.}

\item{theta}{numeric. Angle \eqn{\theta} between fracture and \eqn{\sigma_1}.}

\item{mu}{numeric. Coefficient of internal friction \eqn{\mu}.}
}
\value{
numeric
}
\description{
Calculates some stress components
}
\examples{
s1 <- 1025
s3 <- 250

diff_stress(s1, s3)
mean_stress(s1, s3)
shear_stress(s1, s3, theta = 35)
normal_stress(s1, s3, theta = 35)
fracture_angle(mu = 0.6)
}
