% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohR.R
\name{stress_transformation}
\alias{stress_transformation}
\title{Stress transformation}
\usage{
stress_transformation(
  theta,
  sigma_x = NA,
  sigma_z = NA,
  tau_xz = NA,
  sigma1 = NA,
  sigma3 = NA
)
}
\arguments{
\item{theta}{numeric. Angles (degrees); defaults to 0-180 in increments of 1}

\item{sigma_x}{numeric. Magnitude of normal stress acting in the horizontal direction}

\item{sigma_z}{numeric. Magnitude of normal stress acting in the vertical direction}

\item{tau_xz}{numeric. Magnitude of shear stress acting on the same plane as \code{"sigma_x"}}

\item{sigma1}{numeric. Magnitude of major principal stress}

\item{sigma3}{numeric. Magnitude of minor principal stress}
}
\value{
A two-element list containing
\describe{
\item{\code{"normal"}}{normal stress on an inclined plane}
\item{\code{"shear"}}{shear stress on an inclined plane}
}
}
\description{
calculates the magnitudes of the normal stress and the shear stress
}
\note{
In addition to theta, One of the following two sets of data must be entered:
\enumerate{
\item{\code{"sigma_x"}, \code{"sigma_z"}, \code{"tau_xz"}}
\item{\code{"sigma1"}, \code{"sigma3"}}
}

If theta is entered in conjunction with \code{"sigma_x"}, \code{"sigma_z"}, and \code{"tau_xz"},
it is interpreted as the angle of inclination above the horizontal.  If
theta is entered in conjunction with the principal stresses, then it is
interpreted as the angle of inclination above the major principal plane.
}
\examples{
stress_transformation(sigma_x = 80, sigma_z = 120, tau_xz = 20, theta = 78)
}
\author{
Kyle Elmy and Jim Kaklamanos
}
