% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faults.R
\name{strikeslip_kinematics}
\alias{strikeslip_kinematics}
\title{Strike-slip Kinematics}
\usage{
strikeslip_kinematics(x)
}
\arguments{
\item{x}{\code{"Fault"} object(s)}
}
\value{
character. \code{"left"} - left-lateral (sinistral) offset, \code{"right"} - right-lateral (dextral) offset
}
\description{
Returns the strike-slip kinematics of a fault
}
\examples{
strikeslip_kinematics(Fault(120, 50, 104, 49, sense = -1))
}
\seealso{
\code{\link[=sense_from_strikeslip]{sense_from_strikeslip()}}

Other parse-orientations:
\code{\link[=Fault_from_rake_quadrant]{Fault_from_rake_quadrant()}},
\code{\link[=azimuth_to_cardinal]{azimuth_to_cardinal()}},
\code{\link[=quadrant2dd]{quadrant2dd()}},
\code{\link[=sense_from_strikeslip]{sense_from_strikeslip()}},
\code{\link{split}}
}
\concept{parse-orientations}
