% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structr-package.R
\docType{package}
\name{structr-package}
\alias{structr-package}
\title{structr: Structural Geology in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Free and open-source R package for structural geology to analyze and visualize orientation, strain and stress data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tobiste.github.io/structr/}
  \item Report bugs at \url{https://github.com/tobiste/structr/issues}
}

}
\author{
\strong{Maintainer}: Tobias Stephan \email{tobias.stephan1@yahoo.com} (\href{https://orcid.org/0000-0002-9290-014X}{ORCID})

}
\keyword{internal}
