% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structr-package.R
\docType{package}
\name{structr}
\alias{structr}
\alias{structr-package}
\title{library(structr)}
\description{
Free and open-source R package for analyzing and visualizing
orientation and stress data for structural geology.
}
\details{
A list of documented functions may be viewed by typing
\code{help(package='structr')}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tobiste.github.io/structr/}
  \item Report bugs at \url{https://github.com/tobiste/structr/issues}
}

}
\author{
\strong{Maintainer}: Tobias Stephan \email{tobias.stephan1@yahoo.com} (\href{https://orcid.org/0000-0002-9290-014X}{ORCID})

Authors:
\itemize{
  \item Tobias Stephan \email{tobias.stephan1@yahoo.com} (\href{https://orcid.org/0000-0002-9290-014X}{ORCID})
}

}
