% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{subset.strabo}
\alias{subset.strabo}
\title{Subsetting StraboSpot Projects}
\usage{
\method{subset}{strabo}(x, ..., ds = c("data", "spots", "tags"))
}
\arguments{
\item{x}{object of class \code{"strabo"}}

\item{...}{arguments to be passed to \code{\link[=subset]{subset()}}}

\item{ds}{character. Dataset that should be the base for subsetting.
One of \code{"data"} (the default), \code{"spots"}, or \code{"tags"}.}
}
\value{
An object of class \code{"strabo"} containing just the selected rows and
columns of \code{"x"}.
}
\description{
Returns a subset from \code{"strabo"} objects which meet conditions.
}
\examples{
subset(strabo_prj, strabo_prj$data$quality > 3)
}
\seealso{
\code{\link[=subset]{subset()}}

Other strabo:
\code{\link{strabo}}
}
\concept{strabo}
