% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stats.R
\name{summary.spherical}
\alias{summary.spherical}
\title{Summary statistics}
\usage{
\method{summary}{spherical}(x, ...)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, or \code{"Plane"}.}

\item{...}{parameters passed to \code{\link[=mean]{mean()}}, \code{\link[=sd]{sd()}},}
}
\value{
named vector
}
\description{
Calculates the mean, variance, 68\% cone, and the confidence cone around the mean.
}
\examples{
set.seed(20250411)
summary.spherical(rvmf(100, mu = Line(90, 20), k = 20))
}
