% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_mean.R
\name{symmetry_group}
\alias{symmetry_group}
\title{Symmetry groups}
\usage{
symmetry_group(
  group = c("triclinic", "ray_in_plane", "line_in_plane", "monoclinic",
    "trigonal-trapezohedral", "hexagonal-trapezohedral", "trivial")
)
}
\arguments{
\item{group}{character (symmetry class) or integer number of the enantiomorphic point group. See table below for details.
Also accepts \verb{""ray_in_plane"} (equivalent to triclinic symmetry) and \code{"line_in_plane"} (monoclinic).}
}
\value{
list of rotation parameters
}
\description{
Frequently orientations are subject to some symmetry group \eqn{\mathbb{G}}, which is a
finite set of rotations (satisfying certain properties). For any rotation \eqn{Q}
in \eqn{\mathbb{G}}, the rotation matrix \eqn{Q \cdot R} represents the same orientation as \eqn{R} does.
}
\details{
\tabular{lcl}{
Symmetry \tab Enantiomorphic Point Group \tab Example \cr
\code{triclinic} 	\tab 1 \tab Ray in plane, plagioclase \cr
\code{monoclinic} \tab 2 \tab Line in plane, orthoclase, gypsum, muscovite, clinopyroxene, clinoamphibole \cr
\code{orthorhombic} \tab 222 \tab olivine, aragonite, marcasite, orthopyroxenes \cr
\code{tetragonal} \tab 	4 \tab Pyramidal: zircon \cr
\tab 422 \tab Trapezohedral\cr
\code{trigonal} \tab 	3 \tab Pyramidal, Rhombohedral \cr
\tab	32 \tab Trapezohedral: \eqn{\alpha}-Quartz \cr
\code{hexagonal} \tab	6 \tab Pyramidal, Rhombohedral\cr
\tab	622 \tab Trapezohedral: \eqn{\beta}-Quartz \cr
\code{cubic} \tab	 23 \tab Tetartoidal \cr
\tab	432 \tab Hexoctahedral: galena, pyrite, fluorite \cr
}
}
\examples{
symmetry_group("triclinic")

symmetry_group(2)
}
