% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion.R
\name{tau-comp}
\alias{tau-comp}
\alias{tau2shearnorm}
\alias{tau2tendency}
\title{Resolved Shear and Normal Stress}
\usage{
tau2shearnorm(tau, fault, friction = 0.6)

tau2tendency(tau, fault, friction = 0.6)
}
\arguments{
\item{tau}{symmetric 3x3 matrix. The (reduced) stress tensor.}

\item{fault}{\code{"Fault"} object where the rows are the observations, and the columns the coordinates.}

\item{friction}{numeric. Coefficient of friction (0.6 by default)}
}
\value{
2-column numeric array giving the relative normal and shear stress components
for each fault in \code{fault}.
}
\description{
\code{tau2stress()} calculate normal and shear stress components, while
\code{tau2tendency()} computes the tendency for slip and dilatency for a given set
of faults and a given strerss tensor.
}
\examples{
f <- angelier1990$TYM
tau <- reduced_stress(f)
tau2shearnorm(tau, f)  

tau2tendency(tau, f)  
}
\seealso{
\code{\link[=slip_inversion]{slip_inversion()}}

Other stress-tensor: 
\code{\link{fault_instability_criterion}()},
\code{\link{reduced_stress}()},
\code{\link{stress_shape}()},
\code{\link{tau2rup}()},
\code{\link{tau2stress}()}
}
\concept{stress-tensor}
