% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion.R
\name{tau2stress}
\alias{tau2stress}
\title{Principal Stresses from Stress Tensor}
\usage{
tau2stress(tau)
}
\arguments{
\item{tau}{symmetric 3x3 matrix. The (reduced) stress tensor.}
}
\value{
list with the following components
\describe{
\item{\code{"sigma_vals"}}{numeric. The relative magnitudes of the principal stress axes.}
\item{\code{"principal_axes"}}{The principal stress axes as \code{Line} objects.}
}
}
\description{
The eigenvector and eigenvalues of a stress tensor give the orientations and
relative magnitudes of the principal stress axes.
}
\examples{
f <- angelier1990$TYM
tau <- reduced_stress(f)
tau2stress(tau)  
}
\seealso{
\code{\link[=slip_inversion]{slip_inversion()}}

Other stress-tensor: 
\code{\link{fault_instability_criterion}()},
\code{\link{reduced_stress}()},
\code{\link{stress_shape}()},
\code{\link{tau-comp}},
\code{\link{tau2rup}()}
}
\concept{stress-tensor}
