% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohR.R
\name{tau_max}
\alias{tau_max}
\title{Maximum in-plane shear stress}
\usage{
tau_max(sigma_x, sigma_z, tau_xz)
}
\arguments{
\item{sigma_x}{numeric. Magnitude of normal stress acting in the horizontal direction}

\item{sigma_z}{numeric. Magnitude of normal stress acting in the vertical direction}

\item{tau_xz}{numeric. Magnitude of shear stress acting on the same plane as \code{"sigma_x"}}
}
\value{
A two-element list containing
\describe{
\item{\code{"tauMax"}}{maximum in-plane shear stress}
\item{\code{"theta"}}{angle of maximum in-plane shear stress (in degrees)}
}
}
\description{
calculates the magnitude and direction of the maximum in-plane shear stress
}
\examples{
tau_max(sigma_x = 80, sigma_z = 120, tau_xz = 20)
}
\author{
Kyle Elmy and Jim Kaklamanos
}
