% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohR.R
\name{tauMax}
\alias{tauMax}
\title{MAXIMUM IN-PLANE SHEAR STRESS}
\usage{
tauMax(sigmaX, sigmaZ, tauXZ)
}
\arguments{
\item{sigmaX}{normal stress acting in the horizontal direction}

\item{sigmaZ}{normal stress acting in the vertical direction}

\item{tauXZ}{shear stress acting on the same plane as sigmaX}
}
\value{
A two-element list containing:
tauMax = maximum in-plane shear stress
theta = angle of maximum in-plane shear stress
}
\description{
MAXIMUM IN-PLANE SHEAR STRESS
}
\examples{
tauMax(sigmaX = 80, sigmaZ = 120, tauXZ = 20)
}
\author{
Kyle Elmy and Jim Kaklamanos
}
