% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{text.spherical}
\alias{text.spherical}
\title{Add Points to a Plot}
\usage{
\method{text}{spherical}(x, labels = seq_along(x[, 1]), upper.hem = NULL, earea = NULL, ...)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"},
where the rows are the observations and the columns are the coordinates.}

\item{labels}{a character vector or \link{expression} specifying
    the \emph{text} to be written.  An attempt is made to coerce other
    language objects (names and calls) to expressions, and vectors and
    other classed objects to character vectors by \code{\link{as.character}}.
    If \code{labels} is longer than \code{x} and
    \code{y}, \code{labels} is truncated to \code{max(length(x), length(y))}.}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}). Defaults to
\code{getOption("structr.upper.hem")}.}

\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection, or \code{FALSE} for meridional stereographic projection.
Defaults to \code{getOption("structr.earea")}.}

\item{...}{arguments passed to \code{\link[graphics:text]{graphics::text()}}}
}
\description{
Add Points to a Plot
}
\examples{
stereoplot()
points(Line(c(90, 80), c(10, 75)), col = 1:2)
text(Line(c(90, 80), c(10, 75)), labels = c("L1", "L2"), col = 1:2, pos = 3)
}
\seealso{
Other stereo-plot:
\code{\link{arrows}},
\code{\link{fault-plot}},
\code{\link{lines}},
\code{\link{plot-spherical}},
\code{\link[=points.spherical]{points.spherical()}},
\code{\link{stereo_cones}},
\code{\link[=stereo_confidence]{stereo_confidence()}},
\code{\link{stereo_contour}},
\code{\link[=stereo_lines]{stereo_lines()}},
\code{\link[=stereo_point]{stereo_point()}},
\code{\link[=stereo_segment]{stereo_segment()}},
\code{\link[=stereoplot]{stereoplot()}},
\code{\link[=stereoplot_guides]{stereoplot_guides()}},
\code{\link[=stereoplot_ticks]{stereoplot_ticks()}}
}
\concept{stereo-plot}
