% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{v_delta}
\alias{v_delta}
\title{Cone angle containing ~63\% of the data in degrees.}
\usage{
v_delta(x)
}
\arguments{
\item{x}{numeric. Can be three element vector, three column matrix, or an object of class \code{"line"} or \code{"plane"}}
}
\description{
Cone angle containing ~63\% of the data in degrees.
}
\note{
For enough large sample it approachs the angular standard deviation (csd) of the Fisher statistics
}
\examples{
x <- rvmf(100, mu = Line(120, 50), k = 5)
v_delta(x)
}
