% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_math.R
\name{v_lower}
\alias{v_lower}
\title{Antipodal map sending upper-hemisphere rays to the lower hemisphere.}
\usage{
v_lower(v)
}
\arguments{
\item{v}{object of class \code{'Vec3'}}
}
\value{
object of class \code{'Vec3'}
}
\description{
Negates any vector with positive \code{z}-component. Used in lower-hemisphere plots.
}
\examples{
v <- rbind(Vec3(1, 0, -1), Vec3(1, 0, 1))
v_lower(v)
}
\keyword{internal}
