% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{v_orthogonalize}
\alias{v_orthogonalize}
\title{Orthogonalize two vectors}
\source{
https://www.snsn.se/SH/SHcode/Python/BL.py
}
\usage{
v_orthogonalize(x, y)
}
\arguments{
\item{x, y}{Coordinates of the two vectors. Any coordinates system}
}
\value{
Orthogonalized vectors. In the same coordinate system
}
\description{
Make two vectors orthogonal by moving them in opposite directions in
the plane they span. \code{vdot(x,y) < 1.0e-4} are considered orthogonal.
}
\examples{
v_orthogonalize(c(1, 0, 0), c(.1, 1, 0))
}
