% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{v_rdegree}
\alias{v_rdegree}
\title{Degree of preferred orientation of vectors.}
\usage{
v_rdegree(x)
}
\arguments{
\item{x}{numeric. Can be three element vector, three column array, or an
object of class \code{"line"} or \code{"plane"}}
}
\description{
Degree of preferred orientation of vectors.
}
\examples{
x <- rvmf(100, mu = Line(120, 50), k = 5)
v_rdegree(x)
}
