% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sph_distribution.R
\name{v_unif}
\alias{v_unif}
\title{Uniformly distributed vectors}
\usage{
v_unif(class = NULL, n = 100, method = c("gss", "sfs", "rotasym"))
}
\arguments{
\item{class}{character. Coordinate class of the output vectors.}

\item{n}{number of observations}

\item{method}{character. The algorithm for generating uniformly distributed
vectors. Either \code{"sfs"} for the "Spherical Fibonacci Spiral points on a sphere",
\code{"gss"} for "Golden Section Spiral points on a sphere", or the algorithm
\code{\link[rotasym:unif]{rotasym::r_unif_sphere()}} from the rotasym package.}
}
\value{
object of class specified by \code{"class"} argument
}
\description{
Create uniformly distributed vectors using the algorithm
\emph{Spherical Fibonacci Spiral points on a sphere} algorithm (John Burkardt) or
\emph{Golden Section Spiral points on a sphere}.
}
\details{
\code{"sfs"} algorithm is from on John Burkardt (http://people.sc.fsu.edu/~jburkardt/),
\verb{"gss} is from  http://www.softimageblog.com/archives/115
}
\examples{
set.seed(20250411)
v_unif("line", n = 100, method = "sfs") |>
  ortensor() |>
  or_eigen()
v_unif("line", n = 100, method = "gss") |>
  ortensor() |>
  or_eigen()
v_unif("line", n = 100, method = "rotasym") |>
  ortensor() |>
  or_eigen()
}
\seealso{
\code{\link[=rvmf]{rvmf()}} to draw samples from the von Mises Fisher distribution
around a specified mean vector.
}
