% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{v_var}
\alias{v_var}
\title{Spherical variance based on resultant length (Mardia 1972).}
\usage{
v_var(x)
}
\arguments{
\item{x}{numeric. Can be three element vector, three column matrix, or an object of class \code{"line"} or \code{"plane"}}
}
\description{
Spherical variance based on resultant length (Mardia 1972).
}
\examples{
x <- rvmf(100, mu = Line(120, 50), k = 5)
v_var(x)
}
