% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{variance_plot}
\alias{variance_plot}
\title{Variance visualization}
\usage{
variance_plot(
  x,
  y = NULL,
  .mean = c("geodesic", "arithmetic", "projected"),
  segments = TRUE,
  upper.hem = FALSE,
  earea = TRUE,
  ...
)
}
\arguments{
\item{x}{set of vectors. Object of class \code{"Vec3"}, \code{"Line"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"}.}

\item{y}{The vector from which the variance should be visualized (only one vector allowed).
When \code{NULL}, then the mean vector of \code{x} is used (the default).}

\item{.mean}{character. The type of mean to be used if \code{y} is \code{NULL}.
One of \code{"geodesic"} (the default), \code{"arithmetic"} or \code{"projected"}.}

\item{segments}{logical. Whether the segments should be shown or only the points?}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}, the default).}

\item{earea}{logical \code{TRUE} for Lambert equal-area projection (also "Schmidt net"; the default), or
\code{FALSE} for meridional stereographic projection (also "Wulff net" or "Stereonet").}

\item{...}{optional arguments passed to \code{\link[=assign_col]{assign_col()}}}
}
\value{
angles between all vectors in \code{x} and \code{y}.
}
\description{
Shows the greatcircle of the shortest distance between a set of vectors to a
specified vector in a stereoplot.
The greatcircles are color-coded by the angular distance.
}
\examples{
variance_plot(example_lines)
variance_plot(example_planes, example_planes[1, ], segments = FALSE)
}
\seealso{
\code{\link[=stereo_segment]{stereo_segment()}}, \code{\link[=sph_mean]{sph_mean()}}, \code{\link[=geodesic_mean]{geodesic_mean()}}, \code{\link[=projected_mean]{projected_mean()}}
}
