% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vec-list}
\alias{vec-list}
\alias{vec_list}
\alias{list_vec}
\title{List of vectors}
\usage{
vec_list(x)

list_vec(ls)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}, where the
rows are the observations and the columns are the coordinates.}

\item{ls}{list of 3-element vectors (Cartesian coordinates x, y, z)}
}
\value{
\code{vec_list} returns a list of 3-element vectors (Cartesian coordinates x, y, z).
\code{list_vec} returns a \code{"Vec3"} object
}
\description{
Creates a list of Cartesian vectors  from an spherical objects. This is a convenience
function to link with the package \code{geologyGeometry} by J. R. Davis
}
\examples{
ls <- vec_list(example_lines[1:5, ])
print(ls)

list_vec(ls)
}
