% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deftensor.R
\name{vel_rate}
\alias{vel_rate}
\alias{velgrad_rate}
\alias{velgrad_spin}
\title{Rate and spin of velocity gradient tensor}
\source{
\code{apsg} by O. Lexa
}
\usage{
velgrad_rate(x)

velgrad_spin(x)
}
\arguments{
\item{x}{3x3 matrix. Velocity gradient tensor.}
}
\value{
3x3 matrix
}
\description{
The velocity gradient tensor \strong{L} can be decomposed into a symmetric matrix \strong{S}
(the rate or stretching tensor) and the skew-symmetric matrix \strong{W} (the spin or
vorticity tensor).
}
\details{
The velocity gradient tensor\eqn{\mathbf{L}} can be decomposed into the
sum of a symmetric matrix \eqn{\mathbf{\dot{S}}} and a skew-symmetric matrix \eqn{\mathbf{W}}
\deqn{\mathbf{L} = \mathbf{\dot{S}} + \mathbf{W}}

where
\eqn{\mathbf{\dot{S}}} is the stretching tensor (or strain-rate tensor) that describes
the portion of the deformation that over time produces strain. \eqn{\mathbf{W}}
is the vorticity or spin tensor and describes the
internal rotation (rate) during the deformation.
}
\examples{
R <- defgrad_from_comp(xx = 2, xy = 1, zz = 0.5)
L <- velgrad(R, time = 10)
velgrad_rate(L)
velgrad_spin(L)
}
\seealso{
\code{\link[=velgrad]{velgrad()}}
}
