% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_plots.R
\name{vollmer-plot}
\alias{vollmer-plot}
\alias{vollmer_plot}
\alias{vollmer_plot.default}
\alias{vollmer_plot.spherical}
\alias{vollmer_plot.list}
\title{Fabric Plot of Vollmer (1990)}
\usage{
vollmer_plot(x, ...)

\method{vollmer_plot}{default}(
  x,
  labels = NULL,
  add = FALSE,
  ngrid = c(5, 5, 5),
  main = "Vollmer diagram",
  guides = NULL,
  ...
)

\method{vollmer_plot}{spherical}(
  x,
  labels = NULL,
  add = FALSE,
  ngrid = c(5, 5, 5),
  main = "Vollmer diagram",
  weights = NULL,
  guides = NULL,
  ...
)

\method{vollmer_plot}{list}(
  x,
  labels = NULL,
  add = FALSE,
  ngrid = c(5, 5, 5),
  main = "Vollmer diagram",
  guides = NULL,
  ...
)
}
\arguments{
\item{x}{spherical object or a three-column matrix, where the first column is
P, the second is G, and the third one is R of the Vollmer parameters.}

\item{...}{optional graphical parameters}

\item{labels}{character. text labels}

\item{add}{logical. Should data be plotted to an existing plot?}

\item{ngrid}{integer or 3-element vector specifying the amount of gridlines
for the P, G, and G axes. Constant grid spacing when only one integer is given.
\code{NULL} when no grid.}

\item{main}{character. The main title for the plot.}

\item{guides}{logical. Whether guides of the projection grid should be added
to the plot. Defaults to \code{getOption("structr.guides")}.}

\item{weights}{numeric. Weightings}
}
\value{
plot and when stored as an object, the \code{P}, \code{G}, and \code{R} values as a numeric vector.
}
\description{
Creates a fabric plot using the eigenvalue method
}
\examples{
# Orientation data
set.seed(20250411)
mu <- Line(120, 50)
a <- rvmf(10, mu = mu, k = 10)
vollmer_plot(a, labels = "VMF")

b <- rfb(100, mu = mu, k = 1, A = diag(c(10, 0, 0)))
vollmer_plot(b, labels = "FB", add = TRUE, col = "red")

c <- runif.spherical(n = 100, "Line", method = "rotasym")
vollmer_plot(c, labels = "UNIF", add = TRUE, col = "green")

d <- rkent(100, mu = mu, k = 10, b = 4)
vollmer_plot(d, labels = "KENT", add = TRUE, col = "blue")
}
\references{
Vollmer, F. W. (1990). An application of eigenvalue methods to
structural domain analysis. Geological Society of America Bulletin, 102, 786<U+2013>791.
}
\seealso{
\code{\link[=vollmer]{vollmer()}}, \code{\link[=ortensor]{ortensor()}}

Other fabric-plot: 
\code{\link{flinn_plot}()},
\code{\link{hsu_plot}()},
\code{\link{woodcock_plot}()}
}
\concept{fabric-plot}
