% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_plots.R
\name{vollmer}
\alias{vollmer}
\title{Fabric Intensity and Shape of Orientation Tensor}
\usage{
vollmer(x, ...)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"},
where the rows are the observations and the columns are the coordinates.}

\item{...}{parameters passed to \code{\link[=ortensor]{ortensor()}}}
}
\value{
numeric vector containing the fabric shape and intensity indices:
\describe{
\item{\code{P}}{Point (Vollmer 1990). Range: (0, 1)}
\item{\code{G}}{Girdle (Vollmer 1990). Range: (0, 1)}
\item{\code{R}}{Random (Vollmer 1990). Range: (0, 1)}
\item{\code{B}}{Cylindricity (Vollmer 1990). Range: (0, 1)}
\item{\code{C}}{Cylindricity or Fabric strength (Woodcock 1977). Range: (0, \code{Inf})}
\item{\code{I}}{Cylindricity or Fabric intensity (Lisle 1985). Range: (0, 5)}
\item{\code{D}}{"Distance" from uniformity, linear from R to P, and R to G
(Vollmer 2020). Range: (0, 1). End members are: uniform D = 0, girdle D = 0.5,
cluster D = 1. The 99\% level for a test against uniformity for a sample size
of 300 is D = 0.1.}
\item{\code{U}}{Uniformity statistic of Mardia (1972)}
}
}
\description{
Fabric intensity and shape parameters of the orientation tensor based on Vollmer (1990)
}
\examples{
set.seed(20250411)
mu <- Line(120, 50)
x <- rvmf(100, mu = mu, k = 1)
vollmer(x)

# Pair objects:
vollmer(simongomez)
}
\references{
Lisle, Richard J.  (1985): "The use of the orientation tensor for the description and statistical testing of fabrics." Journal of Structural Geology 7.1: 115-117.

Mardia, Kantilal Varichand (1975): "Statistics of directional data." Journal of the Royal Statistical Society Series B: Statistical Methodology 37.3: 349-371.

Vollmer, Frederick W. (1990): "An application of eigenvalue methods to structural domain analysis." Geological Society of America Bulletin 102.6: 786-791.

Vollmer, Frederick W. (2020): "Representing Progressive Fabric Paths on a Triangular Plot Using a Fabric Density Index and Crystal Axes Eigenvector Barycenters." Geological Society of America Abstracts. Vol. 52.

Woodcock, Nigel H.  (1977): "Specification of fabric shapes using an eigenvalue method." Geological Society of America Bulletin 88.9: 1231-1236.
}
\seealso{
\code{\link[=shape_params]{shape_params()}}, \code{\link[=ortensor]{ortensor()}}, \code{\link[=vollmer_plot]{vollmer_plot()}}
}
