% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_plots.R
\name{vollmer_plot}
\alias{vollmer_plot}
\title{Fabric plot of Vollmer (1990)}
\usage{
vollmer_plot(x, labels = NULL, add = FALSE, ngrid = c(5, 5, 5), ...)
}
\arguments{
\item{x}{either an object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}, where the
rows are the observations and the columns are the coordinates, or an \code{"ortensor"} object.}

\item{labels}{character. text labels}

\item{add}{logical. Should data be plotted to an existing plot?}

\item{ngrid}{integer or 3-element vector specifying the amount of gridlines
for the P, G, and G axes. Constant grid spacing when only one integer is given.
\code{NULL} when no grid.}

\item{...}{optional graphical parameters}
}
\value{
plot and when stored as an object, the \code{P}, \code{G}, and \code{R} values as a numeric vector.
}
\description{
Creates a fabric plot using the eigenvalue method
}
\examples{
set.seed(20250411)
mu <- Line(120, 50)
a <- rvmf(100, mu = mu, k = 10)
vollmer_plot(a, lab = "VMF")

set.seed(20250411)
b <- rfb(100, mu = mu, k = 1, A = diag(c(10, 0, 0)))
vollmer_plot(b, lab = "FB", add = TRUE, col = "red")

set.seed(20250411)
c <- runif.spherical(n = 100, "Line", method = "rotasym")
vollmer_plot(c, lab = "UNIF", add = TRUE, col = "green")

set.seed(20250411)
d <- rkent(100, mu = mu, k = 10, b = 4)
vollmer_plot(d, lab = "KENT", add = TRUE, col = "blue")
title("Fabric plot of Vollmer (1990)")
}
\references{
Vollmer, F. W. (1990). An application of eigenvalue methods to
structural domain analysis. Geological Society of America Bulletin, 102, 786<U+2013>791.
}
\seealso{
\code{\link[=fabric_indexes]{fabric_indexes()}}

Other fabric-plot: 
\code{\link{flinn_plot}()},
\code{\link{hsu_plot}()},
\code{\link{woodcock_plot}()}
}
\concept{fabric-plot}
