% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_plots.R
\name{VollmerPlot}
\alias{VollmerPlot}
\title{Fabric plot of Vollmer (1990)}
\usage{
VollmerPlot(x, labels = NULL, add = FALSE, ngrid = c(5, 5, 5), ...)
}
\arguments{
\item{x}{numeric. Can be three element vector, three column array, or an object of class \code{"line"} or \code{"plane"}}

\item{labels}{character. text labels}

\item{add}{logical. Should data be plotted to an existing plot?}

\item{ngrid}{integer or 3-element vector specifying the amount of gridlines
for the P, G, and G axes. Constant grid spacing when only one integer is given.
\code{NULL} when no grid.}

\item{...}{optional graphical parameters}
}
\description{
Creates a fabric plot using the eigenvalue method
}
\examples{
set.seed(1)
mu <- Line(120, 50)
x <- rvmf(100, mu = mu, k = 1)
VollmerPlot(x, lab = "x")
y <- rvmf(100, mu = mu, k = 20)
VollmerPlot(y, lab = "y", add = TRUE, col = "red")
title("Fabric plot of Vollmer (1990)")
}
\references{
Vollmer, F. W. (1990). An application of eigenvalue methods to structural domain analysis. Geological Society of America Bulletin, 102, 786–791.
}
\seealso{
\code{\link[=WoodcockPlot]{WoodcockPlot()}}, \code{\link[=fabric_indexes]{fabric_indexes()}}
}
