% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deftensor.R
\name{vorticity}
\alias{vorticity}
\alias{flow_apophyses}
\alias{vorticity_axis}
\alias{kinematic_vorticity_from_velgrad}
\alias{instantaneous_stretching}
\alias{instantaneous_stretching_axes}
\title{Flow Apophyses, Vorticity, and Instantaneous Stretching Axes}
\usage{
flow_apophyses(x)

vorticity_axis(x)

kinematic_vorticity_from_velgrad(x)

instantaneous_stretching(x)

instantaneous_stretching_axes(x)
}
\arguments{
\item{x}{object of class \code{"velgrad"}}
}
\value{
\code{vorticity_axis} and \code{instantaneous_stretching_axes} return \code{"Vec3"} object;
\code{"flow_apophyses"} returns \code{"Plane"} object; \code{kinematic_vorticity_from_velgrad} and
\code{instantaneous_stretching} return numeric.
}
\description{
Computes flow apophyses, vorticity vectors, kinematic vorticity numbers, and
instantaneous stretching axes from eigenvalues and eigenvectors of velocity
gradient tensor
}
\examples{
d <- defgrad_from_generalshear(k = 2.5, gamma = 0.9)
l <- velgrad(d, time = 10)

flow_apophyses(l)
vorticity_axis(l)
kinematic_vorticity_from_velgrad(l)
instantaneous_stretching_axes(l)
instantaneous_stretching(l)
}
