% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{vresultant}
\alias{vresultant}
\title{Mean resultant of a set of vectors}
\usage{
vresultant(x, w = NULL, mean = FALSE)
}
\arguments{
\item{x}{numeric. Can be three element vector or a three column array}

\item{w}{numerical vector of weights the same length as \code{x} giving the
weights to use for elements of \code{x}.}

\item{mean}{logical. Whether the mean resultant (\code{TRUE}) or resultant
(\code{FALSE}, the default) is returned.}
}
\value{
if \code{mean==TRUE}, mean resultant is returned
and numeric otherwise.
}
\description{
Mean resultant of a set of vectors
}
\examples{
x <- rvmf(100, mu = Line(120, 50), k = 5) |> to_vec()
vresultant(x, mean = FALSE)
vresultant(x, mean = TRUE)
}
