% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{vslerp}
\alias{vslerp}
\title{Return a spherical linear interpolation between self and other vector}
\usage{
vslerp(x, y, t)
}
\arguments{
\item{x, y}{numeric. Can be three element vector, three column array, or an
object of class \code{"line"} or \code{"plane"}}

\item{t}{description}
}
\description{
Return a spherical linear interpolation between self and other vector
}
\note{
For non-unit vectors the interpolation is not uniform
}
