% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{vtransform}
\alias{vtransform}
\title{Affine transformation of vector by matrix}
\usage{
vtransform(x, A, norm = FALSE)
}
\arguments{
\item{x}{numeric vector, array, or object of class \code{"line"} or \code{"plane"}}

\item{A}{3x3 matrix}

\item{norm}{logical. Whether the transformed vector should be normalized
(\code{TRUE}) or not (\code{FALSE}, the default).}
}
\description{
Affine transformation of vector by matrix
}
\examples{
mat <- cbind(V1 = c(1, 0, 0), V2 = c(0, 1, 0), V3 = c(0, 0, -1))
vec <- c(1, 1, 1)
vtransform(vec, mat)
}
