% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellner.R
\name{wellner_exact_inference_ray}
\alias{wellner_exact_inference_ray}
\title{Two-sample test, based on permutations and Wellner's Rayleigh-style T-statistic (Wellner, 1979, Example 1a).}
\usage{
wellner_exact_inference_ray(xs, ys)
}
\arguments{
\item{xs}{A list of rays.}

\item{ys}{A list of rays.}
}
\value{
A real number, between 0 and 1 inclusive. The fraction of tests in which T exceeds the original T for the data. You can interpret this as a p-value for the null hypothesis that the two populations are identical (not just that their means are identical). In other words, small values of p indicate that the distinction between the two populations is meaningful.
}
\description{
Assumes small sample sizes. Deterministically generates all choose(nx + ny, nx) reassignments of the data to the two groups. For large sample sizes, see rayWellnerInference.
}
