% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellner.R
\name{Wellner_inference_ray}
\alias{Wellner_inference_ray}
\title{Two-sample test, based on permutations and Wellner's Rayleigh-style T-statistic (Wellner, 1979, Example 1a).}
\source{
\code{geologyGeometry} (J.R. Davis): http://www.joshuadavis.us/software/
}
\usage{
Wellner_inference_ray(xs, ys, numPerms)
}
\arguments{
\item{xs}{A list of rays.}

\item{ys}{A list of rays.}

\item{numPerms}{A real number (positive integer). The number of permutations, say 1,000 or 10,000.}
}
\value{
A real number, between 0 and 1 inclusive. The fraction of tests in which T exceeds the original T for the data. You can interpret this as a p-value for the null hypothesis that the two populations are identical (not just that their means are identical). In other words, small values of p indicate that the distinction between the two populations is meaningful.
}
\description{
Assumes large sample sizes, specifically that choose(nx + ny, nx) >> numPerms. For small sample sizes, see rayWellnerExactInference.
}
