% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_plots.R
\name{WoodcockPlot}
\alias{WoodcockPlot}
\title{Fabric plot of Woodcock (1977)}
\usage{
WoodcockPlot(x, labels = NULL, add = FALSE, ...)
}
\arguments{
\item{x}{numeric. Can be three element vector, three column array, or an object of class \code{"line"} or \code{"plane"}}

\item{labels}{character. text labels}

\item{add}{logical. Should data be plotted to an existing plot?}

\item{...}{optional graphical parameters}
}
\description{
Creates a fabric plot using the eigenvalue method
}
\examples{
set.seed(1)
mu <- Line(120, 50)
x <- rvmf(100, mu = mu, k = 1)
WoodcockPlot(x, lab = "x", main = "Fabric plot of Woodcock (1977)")
y <- rvmf(100, mu = mu, k = 20)
WoodcockPlot(y, lab = "y", add = TRUE, col = "red")
}
\references{
Woodcock, N. H. (1977). Specification of fabric shapes using an eigenvalue method. Geological Society of America Bulletin88, 1231<U+2013>1236. http://pubs.geoscienceworld.org/gsa/gsabulletin/article-pdf/88/9/1231/3418366/i0016-7606-88-9-1231.pdf
}
\seealso{
\code{\link[=VollmerPlot]{VollmerPlot()}}, \code{\link[=fabric_indexes]{fabric_indexes()}}
}
