% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{angelier1990}
\alias{angelier1990}
\title{Example fault slip data}
\format{
An object of class \code{data.frame}
}
\usage{
data('angelier1990')
}
\description{
List of different fault-slip data from Angelier (1990)
}
\examples{
data("angelier1990")
head(angelier1990)
}
\references{
Angelier, J. (1990). Inversion of field data in fault tectonics to obtain the
regional stress-III. A new rapid direct inversion method by analytical means.
Geophysical Journal International, 103(2), 363–376. https://doi.org/10.1111/j.1365-246X.1990.tb01777.x
}
\keyword{datasets}
