% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{angelier1990}
\alias{angelier1990}
\title{Fault-Slip Data}
\format{
An object of class \code{data.frame}
}
\usage{
data('angelier1990')
}
\description{
List of different fault-slip data from Angelier (1990)
}
\examples{
data("angelier1990")
head(angelier1990)
}
\references{
Angelier, J. (1990). Inversion of field data in fault tectonics to obtain the
regional stress-III. A new rapid direct inversion method by analytical means.
Geophysical Journal International, 103(2), 363–376. \doi{10.1111/j.1365-246X.1990.tb01777.x}
}
\seealso{
Other datasets: 
\code{\link{example_lines}},
\code{\link{example_lines_df}},
\code{\link{example_planes}},
\code{\link{example_planes_df}},
\code{\link{gray_example}},
\code{\link{holst}},
\code{\link{hossack1968}},
\code{\link{ramsay}},
\code{\link{shebandowan}},
\code{\link{simongomez}}
}
\concept{datasets}
\keyword{datasets}
