% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{arrows}
\alias{arrows}
\alias{stereo_arrows}
\alias{stereo_hoeppener}
\alias{arrows.spherical}
\title{Add Arrows to a Stereoplot}
\usage{
stereo_arrows(
  x,
  sense,
  scale = 0.1,
  angle = 10,
  length = 0.1,
  upper.hem = NULL,
  earea = NULL,
  ...
)

stereo_hoeppener(
  pole,
  slip,
  scale = 0.1,
  angle = 10,
  length = 0.1,
  upper.hem = NULL,
  earea = NULL,
  ...
)

\method{arrows}{spherical}(
  x,
  sense,
  scale = 0.1,
  angle = 10,
  length = 0.1,
  upper.hem = NULL,
  earea = NULL,
  ...
)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}, where the
rows are the observations and the columns are the coordinates.}

\item{sense}{numeric. Sense of the line on a fault plane. Either
\code{1}or \code{-1} for normal or thrust offset, respectively. The \code{"sense"} is the sign
of the fault's rake (see \code{\link[=Fault_from_rake]{Fault_from_rake()}} for details).}

\item{scale}{numeric. Scales the length of the vector. \code{0.1} by default}

\item{angle}{numeric. Angle from the shaft of the arrow to the edge of the arrow head.}

\item{length}{numeric. Length of the edges of the arrow head (in inches).}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}). Defaults to
\code{getOption("structr.upper.hem")}.}

\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection, or \code{FALSE} for meridional stereographic projection.
Defaults to \code{getOption("structr.earea")}.}

\item{...}{arguments passed to \code{\link[graphics:arrows]{graphics::arrows()}}}

\item{pole, slip}{spherical object. The plane normal and a slip vector}
}
\description{
A quiver plot displays displacement vectors into pointing into the direction of movement.
}
\examples{
set.seed(20250411)
stereoplot()
p <- rvmf(n = 100)
stereo_arrows(p, sense = 1, col = "red")
}
\seealso{
\code{\link[=hoeppener]{hoeppener()}}, \code{\link[=angelier]{angelier()}}

Other stereo-plot: 
\code{\link{fault-plot}},
\code{\link{lines}()},
\code{\link{plot-spherical}},
\code{\link{points.spherical}()},
\code{\link{stereo_cones}},
\code{\link{stereo_confidence}()},
\code{\link{stereo_contour}},
\code{\link{stereo_lines}()},
\code{\link{stereo_point}()},
\code{\link{stereo_segment}()},
\code{\link{stereoplot}()},
\code{\link{stereoplot_guides}()},
\code{\link{stereoplot_ticks}()},
\code{\link{text.spherical}()}
}
\concept{stereo-plot}
