% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assign-cex}
\alias{assign-cex}
\alias{assign_cex}
\alias{assign_cex_binned}
\alias{assign_cex_d}
\alias{legend_cex}
\title{Assign plotting size (cex values) to a vector}
\usage{
assign_cex(x, range = c(0.25, 2), area = FALSE)

assign_cex_binned(x, range = c(0.25, 2), breaks = 5, area = FALSE)

assign_cex_d(x, values = NULL, range = c(0.25, 2))

legend_cex(
  x,
  range = c(0.25, 2),
  breaks = 5,
  values = NULL,
  area = FALSE,
  position = "topright",
  pch = 16,
  ...
)
}
\arguments{
\item{x}{vector}

\item{range}{numeric 2-element vector. Output range of \code{cex} values. Minimum
value must be greater than 0.}

\item{area}{logical. Whether \code{cex} should be proportional to the area (\code{TRUE})
or the radius (\code{FALSE}, the default) of the plotting character.}

\item{breaks}{integer giving the desired number of intervals. Non-integer values are rounded down.}

\item{values}{numeric. \code{cex} values to manually assign to \code{x}. Must be at least
the number of unique values in \code{x}.}

\item{position}{Legend position. Either a two-column vector of the x and y coordinates, or a
keyword from the list \code{"bottomright"}, \code{"bottom"}, \code{"bottomleft"}, \code{"left"},
\code{"topleft"}, \code{"top"}, \code{"topright"}, \code{"right"} and \code{"center"}.}

\item{pch}{plotting character to be used in legend.}

\item{...}{arguments passed to \code{\link[graphics:legend]{graphics::legend()}}}
}
\value{
numeric vector
}
\description{
\code{assign_cex()} maps the character expansion (size).
The cex is most commonly used for points and text, and humans perceive the
area of points (not their radius), so this provides for optimal perception.
The argument \code{area} ensures that a value of \code{0} is mapped to a size of \code{0};
\code{assign_cex_binned()} is a binned version, and \code{assign_cex_d()} assigns
cex values to discrete values.
}
\details{
The character expansion \code{cex} is a number indicating the amount by
which plotting text and symbols
should be scaled relative to the default. \code{1}=default, \code{1.5} is 50\% larger,
\code{0.5} is 50\% smaller, etc.
}
\examples{
set.seed(20250411)

# example for continuous colors:
x <- rvmf(100, mu = Line(120, 50), k = 5)
key <- runif(100)
plot(x, cex = assign_cex(key), grid.params = list(guides = FALSE))
legend_cex(key, position = "topright", area = TRUE)
}
\seealso{
\code{\link[PlotTools:SpectrumLegend]{PlotTools::SizeLegend()}} - alternative tool to generate nice
looking legends for cex values

Other assign: 
\code{\link{assign-color}},
\code{\link{assign-pch}}
}
\concept{assign}
