% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{azimuth_to_cardinal}
\alias{azimuth_to_cardinal}
\title{Converts azimuth angles into Cardinal directions}
\usage{
azimuth_to_cardinal(x, n_directions = 8)
}
\arguments{
\item{x}{angles in degree.}

\item{n_direction}{either 8 for 8-point (N, NE, E, …) or 6 for 16-point (N, NNE, NE, …) cardinal version.}
}
\value{
character vector
}
\description{
Converts azimuth angles into Cardinal directions
}
\examples{
azimuth_to_cardinal(c(0, 23, 45, 100, 190, 270, 350))          # 8-point compass
azimuth_to_cardinal(c(0, 23, 45, 100, 190, 270, 350), 16)     # 16-point compass
}
