% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_pole.R
\name{best_fit_plane}
\alias{best_fit_plane}
\title{Least-square fit of small and great circles to spherically projected data}
\usage{
best_fit_plane(x)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, or \code{"Plane"}.}
}
\value{
list
}
\description{
Finds the best small and great circles using the algorithm by Gray et al. (1980)
}
\examples{
data("gray_example")
gray_example1 <- gray_example |>
  dplyr::mutate(
    dipdir = Strike + 90,
    dip = Dip,
    id = seq_along(dip)
  )
gray_cleavage <- dplyr::filter(gray_example1, Type == "Cleavage")
gray_bedding <- dplyr::filter(gray_example1, Type == "Bedding")
test_clea <- Plane(gray_cleavage$dipdir, gray_cleavage$dip)
test_bedd <- Plane(gray_bedding$dipdir, gray_bedding$dip)
best_clea <- best_fit_plane(test_clea)
best_bedd <- best_fit_plane(test_bedd)

stereoplot()
points(test_clea, col = "blue")
points(test_bedd, col = "red")
lines(best_clea$axis_c, best_clea$cone_angle, col = "lightblue")
lines(best_clea$axis_g, 90, lty = 2, col = "blue")
lines(best_bedd$axis_c, best_bedd$cone_angle, col = "sienna")
lines(best_bedd$axis_g, 90, lty = 2, col = "red")
}
\references{
Gray, N.H., Geiser, P.A., Geiser, J.R. (1980). On the
least-square fit of small and great circles to spherically projected data.
Mathematical Geology, Vol. 12, No. 3, 1980.
}
