% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_pole_ramsay.R
\name{best_pole}
\alias{best_pole}
\alias{best_cone_ramsay}
\alias{best_cone_ramsay2}
\title{The cone or plane best fit of conically or cylindrical disposed s-plane poles}
\usage{
best_cone_ramsay(x)

best_cone_ramsay2(x)
}
\arguments{
\item{x}{matrix. Cartesian coordinates of points}
}
\value{
numeric vector with
\describe{
\item{\code{x},\code{y},\code{z}}{Cartesian coordinates of best fit pole of plane or cone axis,}
\item{\code{e}}{residual of the sum of square of the deviations of the observed poles to the planes from the best fit pole, and}
\item{\code{K}}{(only for cones) half apical angle of best fit cone (in radians).}
}
}
\description{
Finding the best fit pole of rotation for a given set of points that are
assumed to lie on a mutual small or great circle circle
}
\examples{
\dontrun{
# example from Ramsay, 1967, p. 20
x <- rbind(
  c(-67, -31, -71),
  c(-62, -53, -50),
  c(-62, -75, -34),
  c(-58, 85, -34),
  c(-79, 40, -52),
  c(90, 14, -75),
  c(80, 10, 90)
) |> acoscartesian_to_cartesian()
best_cone_ramsay(x) # expect: c(0.856, -0.157, -0.492, NA, 1.56207)
best_plane_ramsay(x) # expect: c(0.852, -0.154, -0.502, 1-1.002)
}
}
\references{
Ramsay, 1967, p. 18-21
}
