% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{bingham-inference}
\alias{bingham-inference}
\alias{bingham_inference}
\alias{bingham_inference.Vec3}
\alias{bingham_inference.Line}
\alias{bingham_inference.Plane}
\title{Confidence region for the mean of the Bingham distribution.}
\usage{
bingham_inference(x, n_points)

\method{bingham_inference}{Vec3}(x, n_points = 0L)

\method{bingham_inference}{Line}(x, n_points = 0)

\method{bingham_inference}{Plane}(x, n_points = 0)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, or \code{"Plane"}, where the rows are
the observations and the columns are the coordinates.}

\item{n_points}{A real number (non-negative integer). If \code{n_points} > 0, then
this function constructs a curve for the boundary of the 95\% confidence region.}
}
\value{
A list with members \code{directions}, \code{scatter}, and \code{angles}. The first two
members are identical to \code{vectors} and \code{values} in \code{\link[=ot_eigen]{ot_eigen()}}. \code{angles}
is a pair of real numbers. They describe the 95\% confidence region, as two
distances from the mean toward the two other principal dispersions, measured
in radians along the unit sphere's surface. If the inference fails, then the
angles are \code{NA}. If \code{n_points} > 0, then there is also a member \code{points},
which is a list of \code{n_points} + 1 lines delineating the confidence region.
}
\description{
Confidence region for the mean of a Bingham distribution for anisotropic axial vectors.
This function sometimes fails, if the data set is too concentrated,
dispersed, or small? This is not a particularly high-quality implementation
of the technique.
}
\examples{
r <- bingham_inference(example_planes, n_points = 1e3)
r$directions

stereoplot()
points(example_planes, cex = 0.7)
points(r$directions, col = 1:3, pch = 16, cex = 1.5)
stereo_lines(r$points, col = 'red')
}
\seealso{
\code{\link[=rbingham]{rbingham()}} for simulating a Bingham distribution, and \code{\link[=bingham_MLE]{bingham_MLE()}} to
estimate distribution parameters.

Other distribution-inference: 
\code{\link{fisher-inference}},
\code{\link{watson-inference}}
}
\concept{distribution-inference}
