% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{bingham-mle}
\alias{bingham-mle}
\alias{bingham_MLE}
\alias{bingham_MLE.Vec3}
\alias{bingham_MLE.Line}
\alias{bingham_MLE.Plane}
\title{Maximum likelihood estimation of the Bingham distribution parameters.}
\usage{
bingham_MLE(x, w, n_nonadapt, n_steps)

\method{bingham_MLE}{Vec3}(x, w = NULL, n_nonadapt = 5L, n_steps = 1000L)

\method{bingham_MLE}{Line}(x, w = NULL, n_nonadapt = 5L, n_steps = 1000L)

\method{bingham_MLE}{Plane}(x, w = NULL, n_nonadapt = 5L, n_steps = 1000L)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, or \code{"Plane"}, where the rows are
the observations and the columns are the coordinates.}

\item{w}{optional weights for each observation in \code{x}. A vector of real
numbers (non-negative), of length equal to \code{x}. They need not sum to 1;
the function automatically normalizes them to do so.}

\item{n_nonadapt}{A real number (non-negative integer). The number of
refinements to use in the numerical integration. Note that each increment of
\code{n_nonadapt} increases time and memory requirements by a factor of 4!}

\item{n_steps}{A real number (positive integer). The number of steps to
use in the numerical optimization.}
}
\value{
A list with members
\describe{
\item{\code{A}}{symmetric 3x3 real matrix A,}
\item{\code{values}}{a real 3D vector; the eigenvalues of A; sum to zero,}
\item{\code{vectors}}{a rotation matrix; the eigenvectors of A are the columns,}
\item{\code{error}}{integer; increase \code{n_steps} if \code{error} != 0, and}
\item{\code{minEigenvalue}}{the minimum eigenvalue of the Hessian at the putative optimum; worry if this is not positive.}
}
}
\description{
MLE parameters of a Bingham distribution for anisotropic axial vectors.
Uses a numerical integration (to compute the normalization constant) inside
an numerical optimization (to maximize the likelihood). The Bingham
probability density is proportional to \eqn{\exp{(-x^T A x)}}, not \eqn{\exp{(x^T A x)}}.
}
\examples{
set.seed(2025041)
r <- bingham_MLE(example_planes)
print(r)

stereoplot()
points(example_planes, cex = 0.7)
points(r$vectors, col = 'red', pch = 16, cex = 1.5)

rnd <- rbingham(100, r$A, "Plane")
points(rnd, col = 'grey', cex = 0.7, pch = 16)
}
\seealso{
Other distribution-MLE: 
\code{\link{dist.mle}},
\code{\link{fisher-mle}},
\code{\link{watson-mle}}
}
\concept{distribution-MLE}
