% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{bingham_statistics}
\alias{bingham_statistics}
\title{Elliptical concentration and confidence cone estimation}
\source{
Borradaile, G. (2003). Spherical-Orientation Data. In: Statistics of
Earth Science Data. Springer, Berlin, Heidelberg. https://doi.org/10.1007/978-3-662-05223-5_10
}
\usage{
bingham_statistics(x, w = NULL)
}
\arguments{
\item{x}{numeric. Can be three element vector, three column array, or an
object of class \code{"line"} or \code{"plane"}}

\item{w}{numeric. Weights}
}
\value{
list
\describe{
\item{\code{k}}{two-column vector containing the estimates for the minimum (\eqn{k_\text{min}}) and maximum concentration (\eqn{k_\text{max}}).}
\item{\code{a95}}{two-column vector containing the estimates for the minimum and maximum 95\% confidence cone.}
\item{\code{beta}}{The shape factor of the distribution given by the ratio \eqn{\frac{k_\text{min}}{k_\text{max}}}.}
}
}
\description{
Elliptical concentration and confidence cone estimation
}
\examples{
set.seed(1234)
x <- rfb(100, mu = Line(120, 50), k = 15, A = diag(c(-5, 0, 5)))

ggstereo() +
  geom_point(data = gg(x), aes(x, y))

bingham_statistics(x)
}
\seealso{
\code{\link[=inertia_tensor]{inertia_tensor()}}
}
