% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortensor.R
\name{center}
\alias{center}
\title{Centering vectors}
\usage{
center(x, max_vertical = FALSE)
}
\arguments{
\item{x}{Object of class \code{"Vec3"}, \code{"Line"} or \code{"Plane"}}

\item{max_vertical}{Whether the maximum of the von Mises-Fisher distribution
is already vertical or not.}
}
\value{
Object of class of \code{x}
}
\description{
Rotate vector object to position that eigenvectors are parallel to
axes of coordinate system: E3||X (north-south), E2||X(east-west),
E1||X(vertical)
}
\examples{
set.seed(1)
mu <- Line(120, 50)
x <- rvmf(100, mu = mu, k = 20)
x_centered <- center(x)
plot(x, col = "grey")
# points(mu, col = "red")
points(x_centered, col = "black")
points(Line(c(0, 90, 180), c(0, 0, 90)), col = 2:4)
text(Line(c(0, 90, 180), c(0, 0, 90)), col = 2:4, labels = c("E3", "E2", "E1"), pos = 3)
}
\seealso{
\code{\link[=eigen.spherical]{eigen.spherical()}}
}
