% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{center}
\alias{center}
\title{Centering vectors}
\usage{
center(x, max_vertical = FALSE)
}
\arguments{
\item{x}{numeric. Can be three element vector, three column matrix, or an object of class \code{"line"} or \code{"plane"}}

\item{max_vertical}{Whether the maximum of the von Mises-Fisher distribution is already vertical or not.}
}
\value{
Object of class of \code{x}
}
\description{
Rotate vector object to position that eigenvectors are parallel to
axes of coordinate system: E1||X (north-south), E2||X(east-west),
E3||X(vertical)
}
\examples{
x <- rvmf(100, mu = Line(120, 50), k = 5)
x_centered <- center(x)
stereoplot()
stereo_point(x, col = "grey")
stereo_point(x_centered, col = "black")
}
\seealso{
\code{\link[=or_eigen]{or_eigen()}}
}
