% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interpolation.R
\name{compact_grid}
\alias{compact_grid}
\title{Compact grid}
\usage{
compact_grid(grid)
}
\arguments{
\item{grid}{output of \code{\link[=stress2grid]{stress2grid()}}, \code{\link[=PoR_stress2grid]{PoR_stress2grid()}}, or \code{\link[=kernel_dispersion]{kernel_dispersion()}}}
}
\value{
\code{sf} object
}
\description{
Filter spatial interpolation containing a range of search radii or kernel
half widths to find smallest wavelength (R) with the least spherical sd.
}
\seealso{
\code{\link[=spatial_interpolation]{spatial_interpolation()}}, \code{\link[=v_mean]{v_mean()}}, \code{\link[=v_delta]{v_delta()}}
}
