% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{confidence_ellipse}
\alias{confidence_ellipse}
\title{Confidence ellipse}
\source{
\code{geologyGeometry} (J.R. Davis): http://www.joshuadavis.us/software/
}
\usage{
confidence_ellipse(
  x,
  n_iter = 10000L,
  alpha = 0.05,
  res = 512L,
  isotropic = FALSE
)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}, where the
rows are the observations and the columns are the coordinates.}

\item{n_iter}{integer (10000 by default).}

\item{alpha}{numeric (0.05 by default).}

\item{res}{integer. The resolution with which to sample the boundary curve of the (1 - \code{alpha}) * 100\% percentile region.}

\item{isotropic}{logical. If \code{TRUE}, forces the inverse covariance to the identity matrix and hence the region to be circular.}
}
\value{
list.
\describe{
\item{\code{center}}{Projected mean of \code{x}}
\item{\code{cov}}{Inverse covariance matrix in the tangent space, which is the identity if \code{isotropic} is \code{TRUE}}
\item{\code{rotation}}{Rotation matrix used}
\item{\code{quantiles}}{Quantiles of Mahalanobis norm}
\item{\code{pvalue.ray}}{For rays: p-value for each ray in \code{x}, i.e. the fraction of \code{x} that are farther from \code{center} than the given ray.}
\item{\code{pvalue.line}}{For lines: p-value for each line in \code{x}, i.e. the fraction of \code{x} that are farther from \code{center} than the given line}
\item{\code{pvalue.line.FUN},\code{pvalue.ray.FUN}}{The function to calculate the p-value for a given vector}
\item{\code{angles}}{Angles of the semi-axis of the confidence ellipse (in radians if \code{x} is an \code{"Vec3"} object, in degrees if otherwise.)}
\item{\code{ellipse}}{Confidence ellipse given as \code{"Vec3"} object with \code{res} vectors}
}
}
\description{
Bootstrapped projected mean with percentile confidence region and hypothesis tests
}
\note{
The inference is based on percentiles of Mahalanobis distance in the tangent
space at the mean of the bootstrapped means. The user should check that the
bootstrapped means form a tight ellipsoidal cluster, before taking such a
region seriously.
}
\examples{
set.seed(20250411)
ce <- confidence_ellipse(example_lines, n_iter = 1000, res = 10)
# print(ce)

# Check how many vectors lie outside quantiles:
stats::quantile(ce$pvalue, probs = c(0.00, 0.05, 0.25, 0.50, 0.75, 1.00))

# Hypothesis testing (reject if p-value < alpha):
ce$pvalue.FUN((Line(90, 0)))
}
\references{
Davis, J. R., & Titus, S. J. (2017). Modern methods of analysis
for three-dimensional orientational data. Journal of Structural Geology,
96, 65–89. \doi{10.1016/j.jsg.2017.01.002}
}
