% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{count_points}
\alias{count_points}
\title{Point density}
\usage{
count_points(azi, inc, FUN, sigma, ngrid, weights, r)
}
\arguments{
\item{azi, inc}{degrees.}

\item{FUN}{The method of density estimation to use. Defaults to \code{\link[=exponential_kamb]{exponential_kamb()}}.}

\item{sigma}{(optional) numeric. The number of standard deviations defining the expected number of
standard deviations by which a random sample from a uniform
distribution of points would be expected to vary from being evenly
distributed across the hemisphere.  This controls the size of the
counting circle, and therefore the degree of smoothing.  Higher \code{sigma}s
will lead to more smoothing of the resulting density distribution. This
parameter only applies to Kamb-based methods.  Defaults to \code{3}.}

\item{ngrid}{numeric. The size of the grid that the density is estimated on.}

\item{weights}{(optional) numeric vector of length of \code{azi}.
The relative weight to be applied to each input measurement. The array
will be normalized to sum to 1, so absolute value of the \code{weights} do not
affect the result. Defaults to \code{NULL}}

\item{r}{numeric. radius of stereonet circle}
}
\value{
list
}
\description{
This function calculates the point density of the input
spherical points at a series of counter stations. Creates
\code{gridsize} regular grid of counter stations, calculates the
distance to all input points at each counter station, and then calculates
the density using \code{FUN}.  Each input point is weighted by the corresponding
item of \code{weights}. The weights are normalized to 1 before calculation.
}
