% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deftensor.R
\name{defgrad}
\alias{defgrad}
\alias{is.defgrad}
\alias{as.defgrad}
\alias{defgrad_from_ratio}
\alias{defgrad.Pair}
\alias{defgrad_from_vectors}
\alias{defgrad_from_axisangle}
\alias{defgrad_from_comp}
\alias{defgrad.default}
\alias{defgrad.velgrad}
\title{Deformation Gradient Tensor}
\usage{
is.defgrad(x)

as.defgrad(object)

defgrad(x, ...)

defgrad_from_ratio(Rxy = 1, Ryz = 1)

\method{defgrad}{Pair}(x, ...)

defgrad_from_vectors(v1, v2)

defgrad_from_axisangle(axis, angle)

defgrad_from_comp(
  xx = 1,
  xy = 0,
  xz = 0,
  yx = 0,
  yy = 1,
  yz = 0,
  zx = 0,
  zy = 0,
  zz = 1
)

\method{defgrad}{default}(x, ...)

\method{defgrad}{velgrad}(x, time, steps)
}
\arguments{
\item{x}{object of class \code{"Pair"}, \code{"velgrad"} or 3x3 \code{"matrix"}}

\item{object}{3x3 \code{"matrix"}}

\item{...}{parameters passed to function call}

\item{Rxy, Ryz}{numeric. the XY and YZ strain ratio to create a strain tensor
with axial stretches.Values must be greater than or equal to 1.}

\item{v1, v2}{spherical objects.
Deformation gradient results from the rotation around axis perpendicular to
both vectors to rotate \code{v1} to \code{v2}.}

\item{axis, angle}{rotation axis and angle, axis can be an object of class \code{"Vec3"},
\code{"Line"}, \code{"Ray"}, or \code{"Plane"}, or a three-element vector. Angle in
degrees when axis is a object of class \code{"Line"}, \code{"Ray"}, or \code{"Plane"}, and radians otherwise.}

\item{xx, xy, xz, yx, yy, yz, zx, zy, zz}{numeric. Directly specify components of
the tensor. Identity matrix by default.}

\item{time}{numeric. Total time (default is 1)}

\item{steps}{numeric. Time increments (default is 1)}
}
\value{
3x3 matrix.
}
\description{
Deformation Gradient Tensor
}
\examples{
defgrad_from_ratio(2, 3)
defgrad_from_axisangle(Line(120, 50), 60)
defgrad_from_vectors(Line(120, 50), Line(270, 80))
defgrad(Pair(40, 20, 75, 16))
}
