% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{density}
\alias{density}
\alias{density.spherical}
\alias{density_spherical}
\alias{density_calc}
\title{Spherical density estimation}
\usage{
\method{density}{spherical}(x, ...)

density_calc(
  x,
  FUN = exponential_kamb,
  n = 128L,
  sigma = 3,
  weights = NULL,
  upper.hem = FALSE,
  r = 1
)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}, where the
rows are the observations and the columns are the coordinates.}

\item{...}{arguments passed to \code{\link[=density_calc]{density_calc()}}}

\item{FUN}{density estimation function if \code{kamb=TRUE}; one of \code{\link[=exponential_kamb]{exponential_kamb()}} (the default),
\link{kamb_count}, and \code{\link[=schmidt_count]{schmidt_count()}}.}

\item{n}{integer. Grid size. \code{128} by default.}

\item{sigma}{numeric. Radius for Kamb circle used for counting. 3 by default.}

\item{weights}{(optional) numeric vector of length of \code{nrow(x)}.
The relative weight to be applied to each input measurement. The array
will be normalized to sum to 1, so absolute value of the \code{weights} do not
affect the result. Defaults to \code{NULL}}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}, the default).}

\item{r}{numeric. radius of stereonet circle}
}
\description{
Spherical density estimation
}
\examples{
set.seed(20250411)
test <- rfb(100, mu = Line(120, 10), k = 5, A = diag(c(-1, 0, 1)))
density(x = test, n = 100, sigma = 3, weights = runif(100))
}
