% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{dist.mle}
\alias{dist.mle}
\alias{kent_mle}
\alias{vmf_mle}
\title{Maximum likelihood estimation of Spherical Rotational Symmetric Distributions}
\source{
Adapted from \code{\link[Directional:kent.mle]{Directional::kent.mle()}} and \code{\link[Directional:vmf.mle]{Directional::vmf.mle()}}
}
\usage{
kent_mle(x)

vmf_mle(x)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"},
where the rows are the observations and the columns are the coordinates.}
}
\description{
Estimates the parameters of a von Mises-Fisher or Kent distribution.
}
\examples{
x <- rkent(100, mu = Line(120, 50), k = 5, b = 1)
kent_mle(x)
vmf_mle(x)
}
\seealso{
Other distribution-MLE: 
\code{\link{bingham-mle}},
\code{\link{fisher-mle}},
\code{\link{watson-mle}}
}
\concept{distribution-MLE}
