% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sph_distributions.R
\name{dist.mle}
\alias{dist.mle}
\alias{kent_mle}
\alias{vmf_mle}
\title{MLE of spherical rotational symmetric distributions}
\source{
Adapted from \code{\link[Directional:kent.mle]{Directional::kent.mle()}} and \code{\link[Directional:vmf.mle]{Directional::vmf.mle()}}
}
\usage{
kent_mle(x)

vmf_mle(x)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}, where the
rows are the observations and the columns are the coordinates.}
}
\description{
Estimates the parameters of a von Mises-Fisher or Kent distribution.
}
\examples{
x <- rkent(100, mu = Line(120, 50), k = 5, b = 1)
kent_mle(x)
vmf_mle(x)
}
