% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabeta.R
\name{drillcore}
\alias{drillcore}
\alias{plane_from_drillcore}
\alias{line_from_drillcore}
\title{Orientation of structures from drill core orientation angles}
\usage{
plane_from_drillcore(alpha, beta, azi, inc)

line_from_drillcore(alpha, beta, gamma, azi, inc)
}
\arguments{
\item{alpha}{numeric vector. Alpha angle in degrees}

\item{beta}{numeric vector. Beta angle in degrees}

\item{azi}{numeric. Azimuth of drill core axis orientation in degrees (measured clockwise from North).}

\item{inc}{numeric. Inclination of drill core axis in degrees (Note: \strong{negative values for downward}).}

\item{gamma}{numeric. Gamma angle in degrees}
}
\value{
numeric two-column matrix with Dip direction and dip angle of plane or
azimuth and plunge of line (in degrees).
}
\description{
Calculates the orientation of a plane or line from internal core angles (
alpha, beta and gamma) of oriented drill cores
}
\examples{
azi <- 225
inc <- -45
plane_from_drillcore(60, 320, azi, inc)
plane_from_drillcore(45, 220, azi, inc)

# multiple alpha-beta measurements
stereoplot()
stereo_point(Line(azi, -inc), lab = "CA")
plane_from_drillcore(alpha = c(60, 45), beta = c(320, 220), azi, inc) |> 
stereo_point(lab = c("A", "B"))
}
