% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabeta.R
\encoding{UTF-8}
\name{drillcore}
\alias{drillcore}
\alias{drillcore_transformation}
\title{Orientation of structures from drill core orientation angles}
\usage{
drillcore_transformation(azi, inc, alpha, beta, gamma = NULL)
}
\arguments{
\item{azi}{numeric. Angle between North and the borehole trajectory projected
to the horizontal. The angle is measured clockwise from north and has a value
between 0° and 360°.}

\item{inc}{numeric. Acute angle between the horizontal plane and the
trajectory of the borehole. The value of the inclination can be between
−90° and 90°, where \strong{\code{inc}>0° corresponds to a borehole pointing downwards}.}

\item{alpha}{numeric. Acute dihedral angle between the geological plane and
the trajectory of the borehole. The angle is restricted to be between 0° and
90°, where 90° corresponds to a plane perpendicular to the borehole, i.e.
the trajectory of the borehole is parallel to the normal vector of the plane.}

\item{beta}{numeric. Angle from a reference line (line of the top of the
roof of the borehole profile) to the lower inflexion point of the fracture
trace on the borehole wall, i.e. where the perimeter of the borehole is the
tangent of the fracture trace. The angle is measured clockwise
looking in the direction of the borehole trajectory and can hence be between
0° and 360°.}

\item{gamma}{(optional) numeric. Linear feature on a plane measured in
\strong{clockwise} direction from ellipse long axis at \strong{DOWN hole end} (positive angle).
If measured clockwise on a plane facing UP hole, the angle is negative.}
}
\value{
object of class \code{"Plane"}. If gamma is specified, \code{"Line"} object is
returned.
}
\description{
Calculates the orientation of a plane or line from internal core angles
(alpha, beta, and gamma) of oriented drill cores
}
\examples{
# examples from Roger Marjoribanks (2016);
# http://rogermarjoribanks.info/wp-content/uploads/2016/03/Plotting-alpha-to-locate-P.jpg
azi <- 225
inc <- 45

# single alpha-beta measurement
drillcore_transformation(azi, inc, alpha = 60, beta = 320)
drillcore_transformation(azi, inc, 45, 220)

# example from Stigsson and Munier:
drillcore_transformation(120, 55, 50, 270)

# multiple alpha-beta measurements
my_alphas <- c(60, 45)
my_betas <- c(320, 220)
res <- drillcore_transformation(azi, inc, alpha = my_alphas, beta = my_betas)

# Plot core-axis, and planes in stereonet
plot(Line(azi, inc), lab = "core-axis")
points(res, col = 2:3)
lines(res, col = 2:3)
text(res, labels = c("A", "B"), col = 2:3, pos = 4)

# gamma measurements
my_gammas <- c(20, -10)
res2 <- drillcore_transformation(azi, inc, my_alphas, my_betas, my_gammas)
points(res2, col = 2:3)
text(res2, labels = c("lA", "lB"), col = 2:3, pos = 4)
}
\references{
Stigsson, M., & Munier, R. (2013). Orientation uncertainty goes
bananas: An algorithm to visualise the uncertainty sample space on stereonets
for oriented objects measured in boreholes. Computers and Geosciences, 56,
56–61. https://doi.org/10.1016/j.cageo.2013.03.001
}
