% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortensor.R
\name{ellipsoid-class}
\alias{ellipsoid-class}
\alias{is.ellipsoid}
\alias{as.ellipsoid}
\title{Ellipsoid class}
\usage{
is.ellipsoid(object)

as.ellipsoid(object)
}
\arguments{
\item{object}{3x3 matrix}
}
\value{
\code{is.ellipsoid} returns \code{TRUE} if \code{x} is an \code{"ellipsoid"} object, and \code{FALSE} otherwise.

\code{as.ellipsoid} coerces a 3x3 matrix into an \code{"ellipsoid"} object.
}
\description{
In deformation analysis, the quadratic forms of the three-dimensional stretches
are represented by the \code{ellipsoid} class.
It can be used to represents either ellipsoid objects or finite strain ellipsoids.
}
\examples{
test <- as.ellipsoid(diag(3))
is.ellipsoid(test)
}
\seealso{
\code{\link[=ortensor]{ortensor()}}
}
\concept{ellipsoid}
