% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid.R
\name{ellipsoid-class}
\alias{ellipsoid-class}
\alias{is.ellipsoid}
\alias{as.ellipsoid}
\alias{ellipsoid}
\alias{ellipsoid.default}
\alias{ellipsoid.defgrad}
\title{Ellipsoid class}
\usage{
is.ellipsoid(x)

as.ellipsoid(x)

ellipsoid(x, left, ...)

\method{ellipsoid}{default}(x, left = NULL, ...)

\method{ellipsoid}{defgrad}(x, left = TRUE)
}
\arguments{
\item{x}{Either a matrix or a \code{"defgrad"} object}
}
\value{
\code{is.ellipsoid} returns \code{TRUE} if \code{x} is an \code{"ellipsoid"} object, and \code{FALSE} otherwise.

\code{as.ellipsoid} coerces a 3x3 matrix into an \code{"ellipsoid"} object.
}
\description{
In deformation analysis, the quadratic forms of the three-dimensional stretches
are represented by the \code{ellipsoid} class.
It can be used to represents either ellipsoid objects or finite strain ellipsoids.
}
\examples{
test <- as.ellipsoid(diag(3))
is.ellipsoid(test)

R <- defgrad_from_ratio(2, 3)
ellipsoid(R)
}
\seealso{
\code{\link[=ortensor]{ortensor()}}
}
\concept{ellipsoid}
