% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid.R
\name{ellipsoid-params}
\alias{ellipsoid-params}
\alias{volume}
\alias{volume.default}
\alias{volume.ellipsoid}
\alias{volume.ortensor}
\alias{lode}
\alias{lode.default}
\alias{lode.ellipsoid}
\alias{lode.ortensor}
\alias{nadai}
\alias{nadai.default}
\alias{nadai.ellipsoid}
\alias{nadai.ortensor}
\alias{jelinek}
\alias{jelinek.default}
\alias{jelinek.ellipsoid}
\alias{jelinek.ortensor}
\alias{flinn}
\alias{flinn.default}
\alias{flinn.ortensor}
\alias{flinn.ellipsoid}
\alias{size_invariant}
\alias{size_invariant.default}
\alias{size_invariant.ortensor}
\alias{size_invariant.ellipsoid}
\alias{strain_invariant}
\alias{strain_invariant.default}
\alias{strain_invariant.ortensor}
\alias{strain_invariant.ellipsoid}
\alias{shape_invariant}
\alias{shape_invariant.default}
\alias{shape_invariant.ortensor}
\alias{shape_invariant.ellipsoid}
\alias{kind}
\alias{kind.default}
\alias{kind.ortensor}
\alias{kind.ellipsoid}
\title{Ellipsoid shape parameters}
\usage{
volume(x)

\method{volume}{default}(x)

\method{volume}{ellipsoid}(x)

\method{volume}{ortensor}(x)

lode(x)

\method{lode}{default}(x)

\method{lode}{ellipsoid}(x)

\method{lode}{ortensor}(x)

nadai(x)

\method{nadai}{default}(x)

\method{nadai}{ellipsoid}(x)

\method{nadai}{ortensor}(x)

jelinek(x)

\method{jelinek}{default}(x)

\method{jelinek}{ellipsoid}(x)

\method{jelinek}{ortensor}(x)

flinn(x)

\method{flinn}{default}(x)

\method{flinn}{ortensor}(x)

\method{flinn}{ellipsoid}()

size_invariant(x)

\method{size_invariant}{default}(x)

\method{size_invariant}{ortensor}(x)

\method{size_invariant}{ellipsoid}(x)

strain_invariant(x)

\method{strain_invariant}{default}(x)

\method{strain_invariant}{ortensor}(x)

\method{strain_invariant}{ellipsoid}(x)

shape_invariant(x)

\method{shape_invariant}{default}(x)

\method{shape_invariant}{ortensor}(x)

\method{shape_invariant}{ellipsoid}()

kind(x)

\method{kind}{default}(x)

\method{kind}{ortensor}(x)

\method{kind}{ellipsoid}(x)
}
\arguments{
\item{x}{numeric. either a 3-element vector giving the ellipsoid's semi-axis
lengths (in any order), an object of class \code{"ellipsoid"}, or an object of class \code{"ortensor"}.}
}
\value{
positive numeric
}
\description{
Ellipsoid shape parameters
}
\details{
\deqn{e_i = \log s_i} with \eqn{s1 \geq s2 \geq s3} the semi-axis lengths of the ellipsoid.

Lode's shape parameter:
\deqn{\nu = \frac{2 e_2 - e_1 - e_3}{e_1-e_3}}
with \eqn{e_1 \geq e_2 \geq e_3}. Note that \eqn{\nu}  is undefined for spheres,
but we arbitrarily declare \eqn{\nu=0} for them.
Otherwise \eqn{-1 \geq \nu \geq 1}. \eqn{\nu=-1} for prolate spheroids and \eqn{\nu=1} for oblate spheroids.

Octahedral shear strain \eqn{e_s} (Nadai 1963):
\deqn{e_s = \sqrt{\frac{(e_1 - e_2)^2 + (e_2 - e_3)^2 + (e_1 - e_3)^2 }{3}}}

Strain symmetry (Flinn 1963):
\deqn{k = \frac{s_1/s_2 - 1}{s_2/s_3 - 1}}

and strain intensity (Flinn 1963):
\deqn{d = \sqrt{(s_1/s_2 - 1)^2 + (s_2/s_3 - 1)^2}}

Jelinek (1981)'s \eqn{P_j} parameter:
\deqn{P_j = e^{\sqrt{2 \vec{v}\cdot \vec{v}}}}
with \eqn{\vec{v} = e_i - \frac{\sum e_i}{3}}
}
\examples{
# Generate some random data
set.seed(20250411)
dat <- rvmf(100, k = 20)
s <- principal_stretch(dat)

# Volume of ellipsoid
volume(s)

#  Size-related tensor invariant of ellipsoids
size_invariant(s)

# Strain-related tensor invariant of ellipsoids
strain_invariant(s)

# Shape-related tensor invariant of ellipsoids
shape_invariant(s)

# Lode's shape parameter
lode(s)

# Nadai's octahedral shear strain
nadai(s)

# Jelinek Pj parameter
jelinek(s)

# Flinn's intensity and symmetry parameters
flinn(s)

kind(s)
}
\references{
Flinn, Derek.(1963): "On the statistical analysis of fabric diagrams."
Geological Journal 3.2: 247-253.

Lode, Walter (1926): "Versuche über den Einfluß der mittleren Hauptspannung
auf das Fließen der Metalle Eisen, Kupfer und Nickel“
(\emph{"Experiments on the influence of the mean principal stress on the flow of
the metals iron, copper and nickel"}], Zeitschrift für Physik, vol. 36 (November),
pp. 913–939, \doi{10.1007/BF01400222}

Nadai, A., and Hodge, P. G., Jr. (1963): "Theory of Flow and Fracture of Solids,
vol. II." ASME. J. Appl. Mech. December 1963; 30(4): 640. \doi{10.1115/1.3636654}

Jelinek, Vit. "Characterization of the magnetic fabric of rocks."
Tectonophysics 79.3-4 (1981): T63-T67.
}
\seealso{
\code{\link[=shape_params]{shape_params()}}, \code{\link[=ot_eigen]{ot_eigen()}}
}
