% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortensor.R
\name{ellipsoid-params}
\alias{ellipsoid-params}
\alias{ell_volume}
\alias{ell_lode}
\alias{ell_nadai}
\alias{ell_jelinek}
\alias{ell_flinn}
\alias{ell_size_invariant}
\alias{ell_strain_invariant}
\alias{ell_shape_invariant}
\title{Ellipsoid shape parameters}
\usage{
ell_volume(s)

ell_lode(s)

ell_nadai(s)

ell_jelinek(s)

ell_flinn(s)

ell_size_invariant(s)

ell_strain_invariant(s)

ell_shape_invariant(s)
}
\arguments{
\item{s}{numeric. 3-element vector giving the ellipsoid's semi-axis
lengths, in any order (can be the square-roots of eigenvalues).}
}
\value{
positive numeric
}
\description{
Ellipsoid shape parameters
}
\details{
\deqn{e_i = \log s_i} with \eqn{s1 \geq s2 \geq s3} the semi-axis lengths of the ellipsoid.

Lode's shape parameter:
\deqn{\nu = \frac{2 e_2 - e_1 - e_3}{e_1-e_3}}
with \eqn{e_1 \geq e_2 \geq e_3}. Note that \eqn{\nu}  is undefined for spheres,
but we arbitrarily declare \eqn{\nu=0} for them.
Otherwise \eqn{-1 \geq \nu \geq 1}. \eqn{\nu=-1} for prolate spheroids and \eqn{\nu=1} for oblate spheroids.

Octahedral shear strain \eqn{e_s} (Nadai 1963):
\deqn{e_s = \sqrt{\frac{(e_1 - e_2)^2 + (e_2 - e_3)^2 + (e_1 - e_3)^2 }{3}}}

Strain symmetry (Flinn 1963):
\deqn{k = \frac{s_1/s_2 - 1}{s_2/s_3 - 1}}

and strain intensity (Flinn 1963):
\deqn{d = \sqrt{(s_1/s_2 - 1)^2 + (s_2/s_3 - 1)^2}}

Jelinek (1981)'s \eqn{P_j} parameter:
\deqn{P_j = e^{\sqrt{2 \vec{v}\cdot \vec{v}}}}
with \eqn{\vec{v} = e_i - \frac{\sum e_i}{3}}
}
\examples{
# Generate some random data
set.seed(1)
dat <- rvmf(100, k = 20)
s <- principal_stretch(dat)

# Volume of ellipsoid
ell_volume(s)

#  Size-related tensor invariant of ellipsoids
ell_size_invariant(s)

# Strain-related tensor invariant of ellipsoids
ell_strain_invariant(s)

# Shape-related tensor invariant of ellipsoids
ell_shape_invariant(s)

# Lode's shape parameter
ell_lode(s)

# Nadai's octahedral shear strain
ell_nadai(s)

# Jelinek Pj parameter
ell_jelinek(s)

# Flinn's intensity and symmetry parameters
ell_flinn(s)
}
\references{
Flinn, Derek.(1963): "On the statistical analysis of fabric diagrams." Geological Journal 3.2: 247-253.

Lode, Walter (1926): "Versuche über den Einfluß der mittleren Hauptspannung auf das Fließen der Metalle Eisen, Kupfer und Nickel“
(\emph{"Experiments on the influence of the mean principal stress on the flow of the metals iron, copper and nickel"}], Zeitschrift für Physik, vol. 36 (November), pp. 913–939, DOI: 10.1007/BF01400222

Nadai, A., and Hodge, P. G., Jr. (1963): "Theory of Flow and Fracture of Solids, vol. II." ASME. J. Appl. Mech. December 1963; 30(4): 640. https://doi.org/10.1115/1.3636654

Jelinek, Vit. "Characterization of the magnetic fabric of rocks." Tectonophysics 79.3-4 (1981): T63-T67.
}
\seealso{
\code{\link[=or_shape_params]{or_shape_params()}}, \code{\link[=ot_eigen]{ot_eigen()}}
}
