% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid.R
\name{ellipsoid_from_stretch}
\alias{ellipsoid_from_stretch}
\title{Ellipsoid tensor from principal stretches}
\usage{
ellipsoid_from_stretch(x = 1, y = 1, z = 1)
}
\arguments{
\item{x, y, z}{numeric. Magnitudes of principal stretches}
}
\value{
object of class \code{"ellipsoid"}
}
\description{
Return diagonal tensor defined by magnitudes of principal stretches
}
\examples{
el <- ellipsoid_from_stretch(4, 3, 1)
principal_stretch(el)
}
