% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid.R
\name{ellipsoid_from_stretch}
\alias{ellipsoid_from_stretch}
\title{Ellipsoid tensor from principal stretches}
\usage{
ellipsoid_from_stretch(x = 1, y = 1, z = 1)
}
\arguments{
\item{x, y, z}{numeric. Magnitudes of principal stretches}
}
\value{
object of class \code{"ellipsoid"}
}
\description{
Return diagonal tensor defined by magnitudes of principal stretches
}
\details{
The eigenvalues \eqn{\lambda} of the deformation matrix are the quadratic
forms of the principal stretches \eqn{s} (\eqn{s = 1 + \epsilon = l/l_0}).
}
\examples{
el <- ellipsoid_from_stretch(4, 3, 1)
principal_stretch(el)

lapply(seq.int(nrow(hossack1968)), function(i) {
  ellipsoid_from_stretch(hossack1968[i, 3], hossack1968[i, 2], hossack1968[i, 1])
})
}
\seealso{
Other ellipsoid: 
\code{\link{ellipsoid-class}},
\code{\link{ellipsoid-params}}
}
\concept{ellipsoid}
