% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_lines}
\alias{example_lines}
\title{Example lines}
\format{
An object of class \code{data.frame}
}
\usage{
data('example_lines')
}
\description{
example data set
}
\examples{
data("example_lines")
head(example_lines)
}
\keyword{datasets}
