% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_lines_df}
\alias{example_lines_df}
\title{Example lines}
\format{
An object of class \code{data.frame}
}
\usage{
data('example_lines_df')
}
\description{
example data set
}
\examples{
data("example_lines_df")
head(example_lines_df)
}
\keyword{datasets}
