% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_planes}
\alias{example_planes}
\title{Example planes}
\format{
An object of class \code{data.frame}
}
\usage{
data('example_planes')
}
\description{
example data set
}
\examples{
data("example_planes")
head(example_planes)
}
\keyword{datasets}
