% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_plots.R
\name{fabric_indexes}
\alias{fabric_indexes}
\title{Fabric intensity and shape indices}
\usage{
fabric_indexes(x)
}
\arguments{
\item{x}{object of class \code{"ortensor"}, \code{"ellipsoid"}, or
\code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"} where the rows are the
observations and the columns are the coordinates.}
}
\value{
numeric vector containing the fabric shape and intensity indices:
\describe{
\item{\code{P}}{Point (Vollmer 1990). Range: (0, 1)}
\item{\code{G}}{Girdle (Vollmer 1990). Range: (0, 1)}
\item{\code{R}}{Random (Vollmer 1990). Range: (0, 1)}
\item{\code{B}}{Cylindricity (Vollmer 1990). Range: (0, 1)}
\item{\code{C}}{Cylindricity or Fabric strength (Woodcock 1977). Range: (0, \code{Inf})}
\item{\code{I}}{Cylindricity or Fabric intensity (Lisle 1985). Range: (0, 5)}
\item{\code{D}}{"Distance" from uniformity, linear from R to P, and R to G
(Vollmer 2020). Range: (0, 1). End members are: uniform D = 0, girdle D = 0.5,
cluster D = 1. The 99\% level for a test against uniformity for a sample size
of 300 is D = 0.1.}
}
}
\description{
Fabric intensity and shape indices
}
\examples{
set.seed(20250411)
mu <- Line(120, 50)
x <- rvmf(100, mu = mu, k = 1)
fabric_indexes(x)

data(holst)
holst_R <- lapply(seq_along(holst[, 1]), function(i){
ellipsoid(defgrad_from_ratio(holst[i, 1], holst[i, 2])) 
})
sapply(holst_R, fabric_indexes) |> t()
}
\seealso{
\code{\link[=shape_params]{shape_params()}}
}
