% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_plots.R
\name{fabric_indexes}
\alias{fabric_indexes}
\title{Fabric intensity and shape indices}
\usage{
fabric_indexes(x)
}
\arguments{
\item{x}{numeric. Can be three element vector, three column array, or an object of class \code{"line"} or \code{"plane"}}
}
\value{
numeric vector containing the fabric shape and intensity indices:
\describe{
\item{\code{P}}{Point (Vollmer 1990), range: (0, 1)}
\item{\code{G}}{Girdle (Vollmer 1990), range: (0, 1)}
\item{\code{R}}{Random (Vollmer 1990), range: (0, 1)}
\item{\code{B}}{cylindricity (Vollmer 1990), range: (0, 1)}
\item{\code{C}}{cylindricity or Fabric strength (Woodcock 1977), range: (0, Inf)}
\item{\code{I}}{cylindricity or Fabric intensity (Lisle 1985), range: (0, 5)}
}
}
\description{
Fabric intensity and shape indices
}
\examples{
set.seed(1)
mu <- Line(120, 50)
x <- rvmf(100, mu = mu, k = 1)
fabric_indexes(x)
}
