% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{fault-plot}
\alias{fault-plot}
\alias{fault_plot}
\alias{hoeppner}
\alias{angelier}
\title{Add fault data to existing plot}
\usage{
fault_plot(x, type = c("angelier", "hoeppner"), ...)

hoeppner(x, ...)

angelier(x, ...)
}
\arguments{
\item{x}{\code{"Fault"} object}

\item{type}{character. One of \code{"angelier"} (for "Angelier plot") or \code{"hoeppner"} (for "Hoeppner plot"). See details.}

\item{...}{arguments passed to \code{\link[=stereo_arrows]{stereo_arrows()}}}
}
\value{
Plot
}
\description{
Add fault data to existing plot
}
\details{
\strong{Angelier plot} shows all planes as \emph{great circles} and lineations as points. Fault striae are plotted as vectors on top of the lineation pointing in the movement direction of the hangingwall. Easy to read in case of homogeneous or small datasets.

\strong{Hoeppner plot} shows all planes as \emph{poles} while lineations are not shown. Fault striae are plotted as vectors on top of poles pointing in the movement direction of the hangingwall. Useful in case of large or heterogeneous datasets.
}
\examples{
f <- Fault(c("a" = 120, "b" = 125, "c" = 100), c(60, 62, 50), c(110, 25, 30), c(58, 9, 23), c(1, -1, 1))

stereoplot(title = 'Angelier plot')
angelier(f, col = 1:nrow(f))


stereoplot(title = 'Hoeppner plot')
hoeppner(f, col = 1:nrow(f))
}
\references{
Angelier, J. Tectonic analysis of fault slip data sets, J. Geophys. Res. 89 (B7), 5835-5848 (1984)

Hoeppener, R. Tektonik im Schiefergebirge. Geol Rundsch 44, 26-58 (1955). https://doi.org/10.1007/BF01802903
}
\seealso{
\code{\link[=stereo_arrows]{stereo_arrows()}}
}
