% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faults.R
\name{fault_analysis}
\alias{fault_analysis}
\title{Simple fault analysis}
\usage{
fault_analysis(x, ptangle = 90)
}
\arguments{
\item{x}{fault object}

\item{ptangle}{angle between P and T axes in degrees (90 by default).}
}
\value{
list
}
\description{
Calculates PT-axes, kinematic plane (M), and dihedra separation plane (d)
}
\examples{
f <- Fault(c(120, 120, 100), c(60, 60, 50), c(110, 25, 30), c(58, 9, 23), c(1, -1, 1))
fault_analysis(f)
}
