% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faults.R
\name{Fault_components}
\alias{Fault_components}
\alias{Fault_rake}
\alias{Fault_slip}
\alias{Fault_plane}
\title{Extract components of fault object}
\usage{
Fault_rake(x)

Fault_slip(x)

Fault_plane(x)
}
\arguments{
\item{x}{object of class \code{"Fault"}}
}
\value{
numeric. plane, line or angle in degrees, respectively
}
\description{
\code{\link[=Fault_plane]{Fault_plane()}} extracts the orientation of the fault plane,
\code{\link[=Fault_slip]{Fault_slip()}} extracts the orientation of the slip vector, and
\code{\link[=Fault_rake]{Fault_rake()}} extracts the rake of the fault. i.e. the angle between fault slip vector and fault strike.
}
\examples{
f <- Fault(c(120, 120, 100), c(60, 60, 50), c(110, 25, 30), c(58, 9, 23), c(1, -1, 1))
Fault_plane(f)
Fault_slip(f)
Fault_rake(f)
}
