% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fault_displacement.R
\name{fault_displacements}
\alias{fault_displacements}
\title{Fault displacement components}
\usage{
fault_displacements(
  dip = NULL,
  delta = NULL,
  rake = NULL,
  verticalthrow = NULL,
  dipslip = NULL,
  heave = NULL,
  netslip = NULL,
  horizontalthrow = NULL,
  strikeslip = NULL
)
}
\arguments{
\item{dip}{fault's dip angle (in degrees)}

\item{delta}{angle between horizontal displacement vector and fault's strike (in degrees)}

\item{rake}{(in degrees)}

\item{verticalthrow}{vertical throw}

\item{dipslip}{dip slip component}

\item{heave}{apparent horizontal offset perpendicular to strike}

\item{netslip}{offset on fault plane parallel to the fault's motion direction}

\item{horizontalthrow}{apparent horizontal offset parallel to fault's motion direction}

\item{strikeslip}{strike-slip component}
}
\value{
array
}
\description{
Calculate fault displacement from (at least three) components of of the
fault displacement system
}
\details{
see vignette for description of fault displacement components
}
\examples{
\dontrun{
fault_displacements(strikeslip = 2, verticalthrow = -5, heave = 3)
}
}
