% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faults.R
\name{fault_from_rake}
\alias{fault_from_rake}
\title{Fault from plane and rake}
\usage{
fault_from_rake(p, rake, sense = NULL)
}
\arguments{
\item{p}{plane object}

\item{rake}{Angle in degrees}

\item{sense}{(optional) integer. Sense of the line on a fault plane. Either 1or -1 for normal or thrust offset, respectively.}
}
\value{
returns a \code{"Pair"} object if \code{sense=NULL}, otherwise a \code{"Fault"} object
}
\description{
Fault from plane and rake
}
\examples{
fault_from_rake(Plane(c(120, 120, 100), c(60, 60, 50)), c(84.7202, -10, 30), sense = c(1, -1, 1))
}
