% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faults.R
\name{Fault_from_rake}
\alias{Fault_from_rake}
\title{Fault from plane and rake}
\usage{
Fault_from_rake(p, rake, ...)
}
\arguments{
\item{p}{object of class \code{"Plane"}}

\item{rake}{Angle in degrees in the range  −180° and 180°}

\item{...}{optional arguments passed to \code{\link[=Fault]{Fault()}}}
}
\value{
\code{"Fault"} object
}
\description{
Fault from plane and rake
}
\details{
Rake is used to describe the direction of fault motion with respect
to the strike (measured anticlockwise from the horizontal, up is positive; values between −180° and 180°):
\itemize{
\item left-lateral strike slip: rake near 0°
\item right-lateral strike slip: rake near 180°
\item normal: rake near −90°
\item reverse/thrust: rake near +90°
}
}
\examples{
fr <- Fault_from_rake(Plane(c(120, 120, 100), c(60, 60, 50)), c(84.7202, -10, 30))
plot(fr, col = 1:3)

fr2 <- Fault_from_rake(Plane(c(90, 90, 90), c(80, 40, 10)), c(10, 20, 90))
plot(fr2, col = 1:3)
}
