% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faults.R
\name{Fault_from_rake_quadrant}
\alias{Fault_from_rake_quadrant}
\title{Fault from rake and quadrant notation}
\usage{
Fault_from_rake_quadrant(
  p,
  rake,
  quadrant,
  type = c("plunge", "rake"),
  sense = NULL
)
}
\arguments{
\item{p}{\code{"Plane"} object}

\item{rake}{numeric. Rake angle in degrees}

\item{quadrant}{character. Quadrant of plunge or rake direction}

\item{type}{character. Either \code{"plunge"} or \code{"rake"} for specifying which quadrant convention is used}

\item{sense}{Either 1 (for normal fault movement) or -1 (reverse fault movement). Only used when \code{type=="rake"}}
}
\value{
\code{"Fault"} object
}
\description{
Fault from rake and quadrant notation
}
\details{
\code{type=="plunge"} This is the angle measured in the fault plane between the
strike given by either right or left-hand rule and the lineation.
The angle is recorded in a clockwise sense (looking down upon the fault plane) and has a range
from 0 to 180\%deg;. The quadrant of plunge indicates the direction
of the strike from which the rake angle is measured. The \code{sense} argument is
ignored, as it it implied by the sign of the rake.

\code{type=="rake"} Rake is the \strong{acute} angle measured in the fault plane between the strike of the fault and the
lineation. Starting from the strike line, the angle is measured in a sense
which is down the dip of the plane. Quadrant of rake indicate the direction of
the strike  from which the rake angle is measured, i.e. whether right-hand or left-handrule is followed.
Angle ranges from 0 to 90 °. Use \code{sense} argument to specify the sense of motion.
}
\examples{
dip <- c(5, 10, 15, 30, 40, 55, 65, 75, 90)
dip_dir <- c(180, 225, 270, 315, 360, 0, 45, 90, 135)
rake1 <- c(0, 45, 90, 135, 180, 45, 90, 135, 180)
plunge_quadrant <- c("E", "S", "W", "N", "E", "W", "E", "S", "W")
Fault_from_rake_quadrant(Plane(dip_dir, dip), rake1, plunge_quadrant, type = "plunge")

rake2 <- c(0, 45, 90, 45, 0, 45, 90, 45, 0)
rake_quadrant <- c("E", "S", "S", "E", "E", "W", "N", "S", "W")
Fault_from_rake_quadrant(Plane(dip_dir, dip), rake2, rake_quadrant, type = "rake")
}
\seealso{
Other parse-orientations: 
\code{\link{azimuth_to_cardinal}()},
\code{\link{quadrant2dd}()},
\code{\link{sense_from_strikeslip}()},
\code{\link{split}()},
\code{\link{strikeslip_kinematics}()}
}
\concept{parse-orientations}
