% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faults.R
\name{Fault_PT}
\alias{Fault_PT}
\title{Simple fault analysis}
\usage{
Fault_PT(x, ptangle = 90)
}
\arguments{
\item{x}{object of class \code{"Fault"}}

\item{ptangle}{numeric. angle between P and T axes in degrees (90° by default).}
}
\value{
list
}
\description{
Calculates PT-axes, kinematic plane (M), and dihedra separation plane (d)
}
\examples{
f <- Fault(c(120, 120, 100), c(60, 60, 50), c(110, 25, 30), c(58, 9, 23), c(1, -1, 1))
Fault_PT(f)
}
\seealso{
\code{\link[=stress_inversion]{stress_inversion()}}
}
