% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faults.R
\name{fault_rake}
\alias{fault_rake}
\title{Rake of a fault}
\usage{
fault_rake(x)
}
\arguments{
\item{x}{fault object}
}
\value{
numeric. Angle in degrees
}
\description{
Angle between fault slip vector and fault strike.
}
\examples{
f <- Fault(120, 60, 110, 58, 1)
fault_rake(f)
}
