% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fault_displacement.R
\name{fault_tensor}
\alias{fault_tensor}
\title{Fault deformation gradient tensor}
\usage{
fault_tensor(displacements, dip_direction = NULL)
}
\arguments{
\item{displacements}{data.frame containing the strike slip, the
heave, and the vertical throw of the fault's displacement components.}

\item{dip_direction}{(optional) dip direction in degrees.
If \code{NULL}, the fault tensor will be given in the fault displacement
coordinates. Otherwise, the tensor will be in the geographic reference frame.}
}
\description{
Creates the fault deformation gradient tensor from displacement components.
If the dip direction is know, the tensor will be rotated into the
geographic reference frame.
}
\details{
x axis of tensor = heave, y = strike slip, z = vertical throw (positive for thrusting, negative for normal faulting)
}
\examples{

fault_tensor(displacements = data.frame(strikeslip = 2, verticalthrow = -5, heave = 3), dip_direction = 0)
}
