% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{fisher-inference}
\alias{fisher-inference}
\alias{fisher_inference}
\alias{fisher_inference.Vec3}
\alias{fisher_inference.Ray}
\alias{fisher_inference.Line}
\alias{fisher_inference.Plane}
\title{Confidence Region for the Fisher Distribution Mean.}
\source{
modified after \code{geologyGeometry} by Davis, J.R.
}
\usage{
fisher_inference(x, alpha)

\method{fisher_inference}{Vec3}(x, alpha = 0.05)

\method{fisher_inference}{Ray}(x, alpha = 0.05)

\method{fisher_inference}{Line}(x, alpha = 0.05)

\method{fisher_inference}{Plane}(x, alpha = 0.05)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Ray"}, or \code{"Plane"}, where the rows are the observations and the columns are the coordinates.}

\item{alpha}{A real number, between 0 and 1. The significance level for the confidence region.}
}
\value{
A list with members
\describe{
\item{\code{muHat}}{a ray, identical to \code{\link[=sph_mean]{sph_mean()}}). The mean vector of the distribution.}
\item{\code{kappaHat}}{a non-negative real number). The concentration parameter.}
\item{\code{angle}}{a real number in \verb{[0, pi]}). \code{angle} is the radius of the
confidence region, measured along the surface of the sphere. In radians if
\code{x} is a \code{Vec3} class, in degrees otherwise.}
}#'
}
\description{
Confidence Region for the Fisher Distribution Mean.
}
\details{
Experiments with Fisher-distributed data sets suggest that the sample size n doesn't affect the accuracy much. kappa == 1 is too dispersed, but kappa == 3 is fine.
}
\examples{
set.seed(20250411)
x <- rvmf(100, mu = Ray(120, 50), k = 5)
r <- fisher_inference(x)
print(r)

plot(x)
points(r$muHat, col = 'red')
lines(r$muHat, ang = r$angle, col = 'red')
}
\references{
Tauxe (2010, p. 214).
L. Tauxe 2010. Essentials of Paleomagnetism. xvi + 489 pp. Berkeley: University of California Press.
}
\seealso{
\code{\link[=rvmf]{rvmf()}} for simulating a von Mises-Fisher distribution, and \code{\link[=fisher_MLE]{fisher_MLE()}} to
estimate distribution parameters.

Other distribution-inference: 
\code{\link{bingham-inference}},
\code{\link{watson-inference}}
}
\concept{distribution-inference}
