% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{fisher-mle}
\alias{fisher-mle}
\alias{fisher_MLE}
\alias{fisher_MLE.Vec3}
\alias{fisher_MLE.Ray}
\alias{fisher_MLE.Plane}
\title{Maximum likelihood estimation of the Fisher parameters.}
\usage{
fisher_MLE(x)

\method{fisher_MLE}{Vec3}(x)

\method{fisher_MLE}{Ray}(x)

\method{fisher_MLE}{Plane}(x)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Ray"}, or \code{"Plane"}, where the rows are the observations and the columns are the coordinates.}
}
\value{
A list with members
\describe{
\item{\code{muHat}}{the mean ray, identical to \code{\link[=sph_mean]{sph_mean()}}}
\item{\code{rBar}}{non-negative real number. Mean resultant length.}
\item{\code{kappaHat}}{a positive real number. Concentration parameter \eqn{\kappa} of the Fisher distribution}
}
}
\description{
MLE parameters describing a von Mises-Fisher distribution for isotropic directional vectors.
Based on Mardia and Jupp (2000, p. 198).
}
\examples{
set.seed(20250411)
x <- rvmf(100, mu = Ray(120, 50), k = 5)

fisher_MLE(x)
}
\seealso{
\code{\link[=fisher_inference]{fisher_inference()}} for confidence regions, and \code{\link[=rvmf]{rvmf()}} to
simulate a distribution. \code{\link[=vmf_MLE]{vmf_MLE()}} is an alternative MLE function.

Other distribution-MLE: 
\code{\link{bingham-mle}},
\code{\link{dist.mle}},
\code{\link{watson-mle}}
}
\concept{distribution-MLE}
