% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stats.R
\name{fisher_ftest}
\alias{fisher_ftest}
\title{Test of mean orientations}
\usage{
fisher_ftest(x, y, alpha = 0.05, na.rm = TRUE)
}
\arguments{
\item{x, y}{objects of class \code{"Vec3"}, \code{"Line"}, or \code{"Plane"}, , where the
rows are the observations and the columns are the coordinates.}

\item{alpha}{numeric. Significance level for the confidence angle (default is 0.05 for a 95\% confidence angle).}

\item{na.rm}{logical. Whether \code{NA} values should be removed before the computation proceeds.}
}
\value{
list indicating the F-statistic and the p-value.
}
\description{
Test against the null-hypothesis that the samples are drawn from the same Fisher population.
}
\examples{
set.seed(20250411)
x <- rvmf(100, mu = Line(120, 50), k = 20)
y <- rvmf(100, mu = Line(180, 45), k = 20)

stereoplot()
stereo_point(x, col = 1)
stereo_point(y, col = 2)

fisher_ftest(x, y)
}
