% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stats.R
\name{fisher_statistics}
\alias{fisher_statistics}
\title{Fisher's statistics}
\usage{
fisher_statistics(x, w = NULL, alpha = 0.05, na.rm = TRUE)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, or \code{"Plane"}.}

\item{w}{numeric. Optional weights for each observation.}

\item{alpha}{numeric. Significance level for the confidence angle (default is 0.05 for a 95\% confidence angle).}

\item{na.rm}{logical. Whether \code{NA} values should be removed before the computation proceeds.}
}
\value{
list, with
\describe{
\item{\code{"k"}}{estimated concentration parameter \eqn{\kappa} for the von Mises-Fisher
distribution}
\item{\code{"csd"}}{estimated angular standard deviation enclosing 63\% of the orientation data. Angle is in degrees if \code{x} is a spherical object, and raidan if otherwise.}
\item{\code{"a95"}}{Confidence limit for given \code{p}. Angle is in degrees if \code{x} is a spherical object, and raidan if otherwise.}
}
}
\description{
Estimates concentration parameter, angular standard deviation, and
confidence limit.
}
\examples{
set.seed(20250411)
x <- rvmf(100, mu = Line(120, 50), k = 5)
fisher_statistics(x)
}
