% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{fisher_statistics}
\alias{fisher_statistics}
\title{Fisher's statistics}
\usage{
fisher_statistics(x)
}
\arguments{
\item{x}{numeric. Can be three element vector, three column array, or an
object of class \code{"line"} or \code{"plane"}}
}
\value{
list, with
\describe{
\item{\code{"k"}}{estimated concentration parameter kappa for the von Mises-Fisher
distribution}
\item{\code{"csd"}}{estimated angular standard deviation}
\item{\code{"a95"}}{confidence limit}
}
}
\description{
Estimates concentration parameter, angular standard deviation, and
confidence limit.
}
\examples{
x <- rvmf(100, mu = Line(120, 50), k = 5)
fisher_statistics(x)
}
