% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_analysis.R
\name{flinn_plot}
\alias{flinn_plot}
\title{Flinn diagram}
\usage{
flinn_plot(R_XY, R_YZ, main = "Flinn diagram", R.max = NULL, log = FALSE, ...)
}
\arguments{
\item{R_XY}{ratio of maximum strain and intermediate strain}

\item{R_YZ}{ratio of intermediate strain and minimum strain}

\item{main}{character. The main title (on top).}

\item{R.max}{numeric}

\item{log}{logical}

\item{...}{plotting arguments passed to \code{\link[graphics:points]{graphics::points()}}}
}
\value{
plot
}
\description{
Flinn diagram
}
\examples{
R_XY <- holst[, 'R_XY']
R_YZ <- holst[, 'R_YZ']
flinn_plot(R_XY, R_YZ, log = FALSE)
flinn_plot(R_XY, R_YZ, log = TRUE)
}
