% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_plots.R
\name{flinn_plot}
\alias{flinn_plot}
\alias{flinn_plot.default}
\alias{flinn_plot.ortensor}
\alias{flinn_plot.ellipsoid}
\alias{flinn_plot.spherical}
\title{Flinn diagram}
\usage{
flinn_plot(
  x,
  main = "Flinn diagram",
  R.max = NULL,
  log = FALSE,
  add = FALSE,
  ...
)

\method{flinn_plot}{default}(
  x,
  main = "Flinn diagram",
  R.max = NULL,
  log = FALSE,
  add = FALSE,
  ...
)

\method{flinn_plot}{ortensor}(x, ...)

\method{flinn_plot}{ellipsoid}(x, ...)

\method{flinn_plot}{spherical}(x, ...)
}
\arguments{
\item{x}{accepts the following objects: a two-column matrix where first column is the ratio of maximum strain and
intermediate strain (X/Y) and second column is the the ratio of intermediate strain and minimum strain (Y/Z);
objects of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}; or \code{"ortensor"} objects.}

\item{main}{character. The main title (on top).}

\item{R.max}{numeric. Maximum aspect ratio for scaling.}

\item{log}{logical. Whether the axes should be in logarithmic scale.}

\item{add}{logical. Should data be plotted to an existing plot?}

\item{...}{plotting arguments passed to \code{\link[graphics:points]{graphics::points()}}}
}
\value{
plot and when stored as an object, the multiplication factors for X, Y and Z.
}
\description{
Flinn diagram
}
\examples{
data(holst)
R_XY <- holst[, "R_XY"]
R_YZ <- holst[, "R_YZ"]
flinn_plot(cbind(R_XY, R_YZ), log = FALSE, col = "#B63679", pch = 16)
flinn_plot(cbind(R_XY, R_YZ), log = TRUE, col = "#B63679", pch = 16, type = 'b')

set.seed(20250411)
mu <- Line(120, 50)
x <- rvmf(100, mu = mu, k = 1)
flinn_plot(x, R.max = 2)

set.seed(20250411)
y <- rvmf(100, mu = mu, k = 20)
flinn_plot(ortensor(y), col = "red", R.max = 2, add = TRUE)
}
\references{
Flinn, D. (1965). On the Symmetry Principle and the Deformation
Ellipsoid. Geological Magazine, 102(1), 36–45. \doi{10.1017/S0016756800053851}
}
\seealso{
Other fabric-plot: 
\code{\link{hsu_plot}()},
\code{\link{vollmer_plot}()},
\code{\link{woodcock_plot}()}
}
\concept{fabric-plot}
