% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_analysis.R
\name{flinn_plot}
\alias{flinn_plot}
\title{Flinn diagram}
\usage{
flinn_plot(R_XY, R_YZ, main = "Flinn diagram", R.max = NULL, log = FALSE, ...)
}
\arguments{
\item{R_XY}{ratio of maximum strain and intermediate strain}

\item{R_YZ}{ratio of intermediate strain and minimum strain}

\item{main}{character. The main title (on top).}

\item{R.max}{numeric. Maximum aspect ratio for scaling.}

\item{log}{logical. Whether the axes should be in logarithmic scale.}

\item{...}{plotting arguments passed to \code{\link[graphics:points]{graphics::points()}}}
}
\value{
plot and when stored as an object, the multiplication factors for X, Y and Z.
}
\description{
Flinn diagram
}
\examples{
data(holst)
R_XY <- holst[, "R_XY"]
R_YZ <- holst[, "R_YZ"]
flinn_plot(R_XY, R_YZ, log = FALSE, col = "#B63679", pch = 16)
flinn_plot(R_XY, R_YZ, log = TRUE, col = "#B63679", pch = 16)
}
\references{
Flinn, D. (1965). On the Symmetry Principle and the Deformation
Ellipsoid. Geological Magazine, 102(1), 36–45. \doi{10.1017/S0016756800053851}
}
\seealso{
Other fabric-plot: 
\code{\link{hsu_fabric_plot}()},
\code{\link{hsu_plot}()},
\code{\link{vollmer_plot}()},
\code{\link{woodcock_plot}()}
}
\concept{fabric-plot}
