% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_mean.R
\name{geodesic-mean}
\alias{geodesic-mean}
\alias{geodesic_mean}
\alias{geodesic_mean.Vec3}
\alias{geodesic_mean.Line}
\alias{geodesic_mean.Ray}
\alias{geodesic_mean.Plane}
\alias{geodesic_mean.Pair}
\title{The Fréchet (geodesic \eqn{L^2}) mean}
\source{
\code{geologyGeometry} (J.R. Davis): http://www.joshuadavis.us/software/

\code{geologyGeometry} (J.R. Davis): http://www.joshuadavis.us/software/
}
\usage{
geodesic_mean(x, ...)

\method{geodesic_mean}{Vec3}(x, ...)

\method{geodesic_mean}{Line}(x, ...)

\method{geodesic_mean}{Ray}(x, ...)

\method{geodesic_mean}{Plane}(x, ...)

\method{geodesic_mean}{Pair}(x, ...)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"},
where the rows are the observations and the columns are the coordinates.}

\item{...}{parameters passed to \code{\link[=geodesic_meanvariance_ray]{geodesic_meanvariance_ray()}} (if \code{x} is a Ray), \code{\link[=geodesic_meanvariance_line]{geodesic_meanvariance_line()}} (if \code{x} is a Vec3, Line or Plane)
or \code{\link[=geodesic_mean_pair]{geodesic_mean_pair()}} (if \code{x} is a Pair or a Fault).}
}
\value{
the Fréchet mean vector as an object of class \code{x}.
}
\description{
An iterative algorithm for computing the Fréchet mean, i.e. the vector that
minimizes the Fréchet variance.
}
\examples{
set.seed(20250411)
geodesic_mean(example_planes)
}
\references{
Davis, J. R., & Titus, S. J. (2017). Modern methods of analysis
for three-dimensional orientational data. Journal of Structural Geology,
96, 65–89. \doi{10.1016/j.jsg.2017.01.002}
}
\seealso{
\code{\link[=geodesic_var]{geodesic_var()}} for Fréchet variance, \code{\link[=sph_mean]{sph_mean()}} for the arithmetic mean, \code{\link[=projected_mean]{projected_mean()}} for projected mean
}
