% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohr.R
\name{ggMohr}
\alias{ggMohr}
\title{Plot Mohr Circle in ggplots}
\usage{
ggMohr(
  sigma1,
  sigma2,
  sigma3,
  coulomb = c(70, 0.6),
  sliding = 0.81,
  units = "MPa",
  fill = "gray",
  alpha = 0.5
)
}
\arguments{
\item{sigma1, sigma2, sigma3}{numeric. Magnitudes of major, intermediate, and minor principal stresses}

\item{coulomb}{numeric 2 element vector. Coulomb criterion containing the cohesion and the coefficient of sliding: (\code{c(70, 0.6)})}

\item{sliding}{Sliding criteria (\code{0.81} by default)}

\item{units}{units of \code{sigma1}, \code{sigma2}, and \code{sigma3} (\code{"MPa"} by default).}

\item{fill}{fill color of Mohr circle}

\item{alpha}{opacity of Mohr circle}
}
\description{
Plot Mohr Circle in ggplots
}
\examples{
ggMohr(1025, 400, 250)
}
