% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohR.R
\name{ggMohr}
\alias{ggMohr}
\title{Plot Mohr Circle in ggplots}
\usage{
ggMohr(s1, s2, s3, coulomb = c(70, 0.6), sliding = 0.81, units = "MPa")
}
\description{
Plot Mohr Circle in ggplots
}
\examples{
ggMohr(1025, 400, 250)
}
