% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_gg.R
\name{ggstereo}
\alias{ggstereo}
\title{Stereonet using ggplot}
\usage{
ggstereo(
  data = NULL,
  mapping = aes(),
  earea = TRUE,
  centercross = TRUE,
  grid = FALSE,
  grid.spacing = 10,
  grid.rot = 0,
  ...
)
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link[ggplot2:fortify]{ggplot2::fortify()}}. If not specified, must be
supplied in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot. If not
specified, must be supplied in each layer added to the plot.}

\item{earea}{logical. Whether the projection is equal-area ("Schmidt net")
(\code{TRUE}, the default), or equal-angle ("Wulff net") (\code{FALSE}).}

\item{centercross}{logical. Whether a center cross should be added.}

\item{grid}{logical. Whether a gid should be added.}

\item{grid.spacing}{numeric. Grid spacing in degree}

\item{grid.rot}{numeric. Angle (in degrees) to rotate the grid.}

\item{...}{argument passed to \code{\link[ggplot2:geom_polygon]{ggplot2::geom_polygon()}}}
}
\value{
ggplot
}
\description{
Stereonet using ggplot
}
\examples{
if (require("mapproj")) {
  test_data <- rbind(
    rvmf(100, mu = Line(90, 45), k = 10),
    rvmf(50, mu = Line(0, 0), k = 20)
  )

  ggstereo(grid = TRUE) +
    ggplot2::geom_point(data = gg(test_data), ggplot2::aes(x = x, y = y))

  ggstereo(earea = FALSE, centercross = TRUE) +
    ggplot2::geom_point(data = gg(test_data), ggplot2::aes(x = x, y = y))
}
}
