% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defgrad.R
\name{gradient}
\alias{gradient}
\alias{velgrad_from_defgrad}
\alias{defgrad_from_velgrad}
\title{Velocity gradient and Deformation gradient tensors}
\usage{
velgrad_from_defgrad(R, time = 1)

defgrad_from_velgrad(V, time = 1, steps = 1)
}
\arguments{
\item{R}{3x3 matrix. Deformation gradient tensor.}

\item{time}{numeric. Total time (default is 1)}

\item{V}{3x3 matrix. Velocity gradient tensor.}

\item{steps}{numeric. Time increments (default is 1)}
}
\value{
3x3 matrix. If steps is > 1, then a list
of matrices is returned.
}
\description{
Calculates the velocity gradient tensor as the matrix logarithm  of the
deformation gradient tensor divided by given time, and
the deformation gradient tensor accumulated after some time.
}
\examples{
D <- defgrad_from_comp(xx = 2, xy = 1, zz = 0.5)
L <- velgrad_from_defgrad(D, time = 10)
L
defgrad_from_velgrad(L, time = 10, steps = 2)
}
