% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deftensor.R
\name{gradient}
\alias{gradient}
\alias{is.velgrad}
\alias{as.velgrad}
\alias{velgrad}
\alias{velgrad.default}
\alias{velgrad.defgrad}
\title{Velocity gradient gradient tensors}
\usage{
is.velgrad(x)

as.velgrad(object)

velgrad(x, time, ...)

\method{velgrad}{default}(x, ...)

\method{velgrad}{defgrad}(x, time = 1, ...)
}
\arguments{
\item{x}{3x3 matrix. Deformation gradient tensor.}

\item{object}{3x3 \code{"matrix"}}

\item{time}{numeric. Total time (default is 1)}

\item{...}{parameters passed to function call}
}
\value{
3x3 matrix. If steps is > 1, then a list
of matrices is returned.
}
\description{
The velocity gradient tensor describes the velocity of particles at any
instant during the deformation. Velocity gradient tensor from deformation
gradient tensor.
}
\details{
\code{velgrad()} calculates the velocity gradient tensor as the matrix logarithm of the
deformation gradient tensor divided by given time, and
the deformation gradient tensor accumulated after some time.
}
\examples{
d <- defgrad_from_generalshear(k = 2.5, gamma = 0.9)
l <- velgrad(d, time = 10)
d_steps <- defgrad(l, time = 10, steps = 2)

# apply on orientation data
set.seed(20250411)
v <- rvmf(100, mu = Line(0, 90), k= 100)
v_trans <- lapply(d_steps, function(i){transform_linear(v, i)})

# plot in stereonet
axes <- Vec3(c(1, 0, 0), c(0, 1, 0), c(0, 0, 1))
stereo_path(v_trans, type = "l", add = FALSE)
stereo_path(v_trans, type = "p", col = assign_col(seq_along(v_trans)), pch = 16, cex = .4)
points(axes, pch = 15); text(axes, labels = c('x', 'y', 'z'), pos = 1) 
}
\seealso{
\code{\link[=defgrad]{defgrad()}}, \code{\link[=stereo_path]{stereo_path()}} for plotting
}
