% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deftensor.R
\name{gradient}
\alias{gradient}
\alias{is.velgrad}
\alias{as.velgrad}
\alias{velgrad}
\alias{velgrad.default}
\alias{velgrad.defgrad}
\title{Velocity gradient and Deformation gradient tensors}
\usage{
is.velgrad(x)

as.velgrad(object)

velgrad(x, ...)

\method{velgrad}{default}(x, ...)

\method{velgrad}{defgrad}(x, time = 1, ...)
}
\arguments{
\item{x}{3x3 matrix. Deformation gradient tensor.}

\item{object}{3x3 \code{"matrix"}}

\item{...}{parameters passed to function call}

\item{time}{numeric. Total time (default is 1)}
}
\value{
3x3 matrix. If steps is > 1, then a list
of matrices is returned.
}
\description{
Calculates the velocity gradient tensor as the matrix logarithm  of the
deformation gradient tensor divided by given time, and
the deformation gradient tensor accumulated after some time.
}
\examples{
D <- defgrad_from_comp(xx = 2, xy = 1, zz = 0.5)
L <- velgrad(D, time = 10)
print(L)

defgrad(L, time = 10, steps = 2)
}
