% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gray_example}
\alias{gray_example}
\title{Bedding and cleavage data}
\format{
An object of class \code{data.frame}
}
\usage{
data('gray_example')
}
\description{
16 bedding and cleavage orientations from Gray et al. (1980) to find best small and great circle
}
\examples{
data("gray_example")
head(gray_example)
}
\keyword{datasets}
