% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{gray_example}
\alias{gray_example}
\title{Example data from Gray et al. 1980 to find best small and great circle}
\format{
An object of class \code{data.frame}
}
\usage{
data('gray_example')
}
\description{
example data set
}
\examples{
data("gray_example")
head(gray_example)
}
\keyword{datasets}
