% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holst}
\alias{holst}
\title{3D Strain Data}
\format{
An object of class \code{matrix}
}
\usage{
data('holst')
}
\description{
Example data from Holst and Fossen (1987) containing 3D strain data from 17 localities in the West Norwegian Caledonides.
The data gives the  difference of the  \eqn{\epsilon_1} and \eqn{\epsilon_2} (\code{e1e2}) and \eqn{\epsilon_2} and \eqn{\epsilon_3} (\code{e2e3})
}
\examples{
data("holst")
head(holst)
}
\keyword{datasets}
