% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hossack1968}
\alias{hossack1968}
\title{3D Strain Data II}
\format{
An object of class \code{matrix}
}
\usage{
data('hossack1968')
}
\description{
Example data from J.R. Hossack (1968) containing 3D strain data from 81
localities from the Bygdin Conglomerate by the Upper Jotun Nappe. The data
gives the mean axial ratios (X, Y, Z) of the deformed pebbles.
Different to Hossack, here X is the longest axes, and Z the shortest.
}
\examples{
data("hossack1968")
head(hossack1968)
}
\seealso{
Other datasets: 
\code{\link{angelier1990}},
\code{\link{example_lines}},
\code{\link{example_lines_df}},
\code{\link{example_planes}},
\code{\link{example_planes_df}},
\code{\link{gray_example}},
\code{\link{holst}},
\code{\link{ramsay}},
\code{\link{shebandowan}},
\code{\link{simongomez}}
}
\concept{datasets}
\keyword{datasets}
