% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_analysis.R
\encoding{UTF-8}
\name{hsu_plot}
\alias{hsu_plot}
\title{Hsü plot}
\usage{
hsu_plot(R_XY, R_YZ, main = "Hsü diagram", es.max = NULL, ...)
}
\arguments{
\item{R_XY}{ratio of maximum strain and intermediate strain}

\item{R_YZ}{ratio of intermediate strain and minimum strain}

\item{main}{character. The main title (on top).}

\item{es.max}{maximum strain for scaling.}

\item{...}{plotting arguments passed to \code{\link[graphics:points]{graphics::points()}}}
}
\value{
plot and when stored as an object, the Lode parameter \code{Lode} and the natural octahedral strain \code{es}.
}
\description{
3D strain diagram using the Hsü (1965) method to display the natural octahedral strain and the Lode parameter.
}
\examples{
R_XY <- holst[, "R_XY"]
R_YZ <- holst[, "R_YZ"]
hsu_plot(R_XY, R_YZ, col = "#B63679", pch = 16)
}
