% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_plots.R
\name{hsu_plot}
\alias{hsu_plot}
\alias{hsu_plot.ortensor}
\alias{hsu_plot.spherical}
\alias{hsu_plot.ellispoid}
\alias{hsu_plot.default}
\alias{hsu_plot.list}
\title{Fabric plot of Hsu (1965)}
\usage{
hsu_plot(x, ...)

\method{hsu_plot}{ortensor}(x, labels = NULL, add = FALSE, es.max = 3, main = "Hsu diagram", ...)

\method{hsu_plot}{spherical}(x, ...)

\method{hsu_plot}{ellispoid}(x, ...)

\method{hsu_plot}{default}(x, labels = NULL, add = FALSE, es.max = 3, main = "Hsu diagram", ...)

\method{hsu_plot}{list}(x, labels = NULL, add = FALSE, es.max = 3, main = "Hsu diagram", ...)
}
\arguments{
\item{x}{accepts the following objects: a two-column matrix where first column is the ratio of maximum strain and
intermediate strain (X/Y) and second column is the the ratio of intermediate strain and minimum strain (Y/Z);
objects of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}; or \code{"ortensor"} objects.}

\item{...}{plotting arguments passed to \code{\link[graphics:points]{graphics::points()}}}

\item{labels}{character. text labels}

\item{add}{logical. Should data be plotted to an existing plot?}

\item{es.max}{maximum strain for scaling.}

\item{main}{character. The main title (on top).}
}
\value{
plot and when stored as object, a list containing the Lode parameter \code{lode} and the natural octahedral strain \code{es}.
}
\description{
3D strain diagram using the Hsu (1965) method to display the natural
octahedral strain (Nádai, 1950) and Lode's parameter (Lode, 1926).
}
\examples{
R_XY <- holst[, "R_XY"]
R_YZ <- holst[, "R_YZ"]
hsu_plot(cbind(R_XY, R_YZ), col = "#B63679", pch = 16, type = "b")

set.seed(20250411)
mu <- Line(120, 50)
x <- rvmf(100, mu = mu, k = 1)
hsu_plot(x, labels = "x")

set.seed(20250411)
y <- rvmf(100, mu = mu, k = 20)
hsu_plot(ortensor(y), labels = "y", col = "red", add = TRUE)
}
\references{
Lode, W. (1926). Versuche über den Einfluß der mittleren Hauptspannung auf
das Fließen der Metalle Eisen. Kupfer und Nickel. Zeitschrift Für Physik,
36(11–12), 913–939. \doi{10.1007/BF01400222}

Nádai, A. (1950). Theory of flow and fracture of solids. McGraw-Hill.

Hsu, T. C. (1966). The characteristics of coaxial and non-coaxial strain
paths. Journal of Strain Analysis, 1(3), 216–222. \doi{10.1243/03093247V013216}

Hossack, J. R. (1968). Pebble deformation and thrusting in the Bygdin area
(Southern Norway). Tectonophysics, 5(4), 315–339. \doi{10.1016/0040-1951(68)90035-8}
}
\seealso{
\code{\link[=lode]{lode()}} for Lode parameter, and \link{nadai} for natural octahedral strain.

Other fabric-plot: 
\code{\link{flinn_plot}()},
\code{\link{vollmer-plot}},
\code{\link{woodcock_plot}()}
}
\concept{fabric-plot}
