% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_analysis.R
\name{hypercontour}
\alias{hypercontour}
\title{Calculates densities for fabric and strain data}
\usage{
hypercontour(
  r,
  phi,
  proj = c("eqd", "eqa", "stg", "ort", "gno", "lin", "rdl", "rfp"),
  normalize = TRUE,
  rmax = NULL,
  kappa = 40,
  nnodes = 50L
)
}
\arguments{
\item{r}{numeric. Aspect ratio of deformed object (long axis / short axis)}

\item{phi}{numeric. Orientation of long axis of deformed object (in degrees)}

\item{proj}{character. Projection,  \code{'eqd'} for equidistant (Elliot plot),
\code{'eqa'} for equal-area, \code{'stg'} for stereographic, \code{'ort'} for orthographic,
\code{'gno'} for gnomonic, \code{'lin'} for exponential (linear R), \code{'rdl'} for radial,
\code{'rfp'} for Rf/phi (cylindrical instead of polar).}

\item{normalize}{logical.}

\item{rmax}{maximum R value (if \code{NULL}, computed automatically)}

\item{kappa}{smoothing parameter}

\item{nnodes}{grid resolution. higher is more accurate but slower, 30 is good, but 50 is recommended for final plots, default = 50.}
}
\value{
list with:
\itemize{
\item x, y: vectors of the projection coordinates for the density grid
\item z: density matrix
\item points: projected data points
\item frame: plot frame (circle or square)
}
}
\description{
Densities in hyperboloidal projections of geological fabric and
finite strain data density calculations done on the unit hyperboloid
(Vollmer, 2018). Options are given for equidistant (Elliott), equal-area,
stereographic, orthographic, exponential, and radial projections, as polar
azimuthal or cylindrical (cartesian, RfPhi-type) plots.
}
\details{
The data must be in a comma delimited csv text file with one (R, phi) pair
per line, where R = strain ratio (max/min), phi = orientation of long (max)
axis from x. Contours are equally spaced over the probability density
distribution.
}
\examples{
data(ramsay)
out <- hypercontour(ramsay[, "R"], ramsay[, "phi"], proj = "rfp")
head(out$z)
}
\references{
Vollmer, F.W., 2018. Automatic contouring of geological fabric and finite
strain data on the unit hyperboloid. Computers & Geosciences,
https://doi.org/10.1016/j.cageo.2018.03.006
}
\seealso{
\code{\link[=Rphi_plot]{Rphi_plot()}}, \code{\link[=Rphi_polar_plot]{Rphi_polar_plot()}}
}
