% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{lines.spherical}
\alias{lines.spherical}
\title{Add Lines to a Plot}
\usage{
\method{lines}{spherical}(x, ang = 90, ...)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"},
where the rows are the observations and the columns are the coordinates.}

\item{ang}{numeric. Conical angle in degrees.}

\item{...}{arguments passed to \code{\link[graphics:lines]{graphics::lines()}}}
}
\description{
Add Lines to a Plot
}
\examples{
set.seed(20250411)
stereoplot()
lines(rvmf(n = 5), ang = runif(5, 0, 90), col = 1:5)
}
\seealso{
Other stereo-plot: 
\code{\link{fault-plot}},
\code{\link{plot-spherical}},
\code{\link{points.spherical}()},
\code{\link{stereo_arrows}()},
\code{\link{stereo_cones}},
\code{\link{stereo_confidence}()},
\code{\link{stereo_contour}},
\code{\link{stereo_point}()},
\code{\link{stereo_segment}()},
\code{\link{stereoplot}()},
\code{\link{stereoplot_guides}()},
\code{\link{stereoplot_ticks}()},
\code{\link{text.spherical}()}
}
\concept{stereo-plot}
