% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO3.R
\name{mean-pair}
\alias{mean-pair}
\alias{sph_mean_pair}
\alias{sph_var_pair}
\title{Mean orientation of a set of pairs or faults}
\usage{
sph_mean_pair(x, group = NULL)

sph_var_pair(x, group = NULL)
}
\arguments{
\item{x}{object of class \code{"Pair"} or \code{"Fault"}}

\item{group}{character. Symmetry group of \code{x}. One of
\code{"orthorhombic"} (line-in-plane symmetry, e.g. foliation-lineations, cylindrical fold
orientations, triaxial ellipsoid orientations, and earthquake
focal mechanisms, and olivine),
\code{"triclinic"} (ray-in-plane symmetry, e.g. faults with slip directions),
\code{"trigonal"} (e.g. alpha-quartz),
\code{"hexagonal"} (e.g. beta-quartz), or
\code{"trivial"} (rotations).
If \code{NULL}, the group will be
automatically picked based on the class of \code{x}.}
}
\value{
object of class \code{"Pair"} or \code{"Fault"}, respectively
}
\description{
The Frechet (geodesic \eqn{L^2}) mean and variance of a pair of foliations and lineations
}
\examples{
my_fault <- Fault(
c("a" = 120, "b" = 120, "c" = 100), 
c(60, 60, 50), 
c(110, 25, 30), 
c(58, 9, 23), 
c(1, -1, 1)
)
sph_mean_pair(my_fault)
sph_var_pair(my_fault)
}
