\name{merge.list}
\alias{merge.list}
\title{Function that takes 2 lists and merges them fairly effeciently}
\usage{
  merge.list(x, y = NULL, mergeUnnamed = TRUE, ...)
}
\arguments{
  \item{x}{a list}

  \item{y}{a second list}
}
\description{
  Function that takes 2 lists and merges them fairly
  effeciently
}

