% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohR.R
\name{Mohr_plot}
\alias{Mohr_plot}
\title{Mohr Circle plot}
\usage{
Mohr_plot(
  sigma_x = NA,
  sigma_z = NA,
  tau_xz = NA,
  sigma1 = NA,
  sigma2 = NA,
  sigma3 = NA,
  unit = "MPa",
  col = "black",
  n = 512,
  full.circle = FALSE,
  include.zero = TRUE,
  ...
)
}
\arguments{
\item{sigma_x}{numeric. Magnitude of normal stress acting in the horizontal direction}

\item{sigma_z}{numeric. Magnitude of normal stress acting in the vertical direction}

\item{tau_xz}{numeric. Magnitude of shear stress acting on the same plane as \code{"sigma_x"}}

\item{sigma1}{numeric. Magnitude of major principal stress}

\item{sigma2}{numeric. (optional) Magnitude of intermediate principal stress.
This will add the additional Mohr circles between the principal stresses.}

\item{sigma3}{numeric. Magnitude of minor principal stress}

\item{unit}{character. The unit used for magnitude of stress (\code{"MPa"} by default). \code{NULL} if unitless.}

\item{col}{color for Mohr circle.}

\item{n}{integer. Resolution given amount of points along the generated path
representing the full Mohr circle (\code{512} by default).}

\item{full.circle}{logical. Should the complete Mohr circle be shown, or only
the upper (positive shear stress) part of the circle?}

\item{include.zero}{logical. the plot range be extended to include normal_stress = 0?}

\item{...}{optional graphical parameters.}
}
\description{
plots the Mohr Circle
}
\note{
One of the following two sets of data must be entered
\enumerate{
\item{\code{"sigma_x"}, \code{"sigma_z"}, \code{"tau_xz"}}
\item{\code{"sigma1"}, \code{"sigma3"}}
}
}
\examples{
Mohr_plot(sigma_x = 80, sigma_z = 120, unit = "kPa", tau_xz = 20, col = "#B63679", lwd = 2)

Mohr_plot(sigma1 = 1025, sigma2 = 450, sigma3 = 250)

## Deviatoric stress from stress ratio:
R <- 0.4 # Stress ratio after Gephart & Forsyth (1984)
sigma1 <- 1; sigma3 <- 0
sigma2 <- sigma1 - R * (sigma1 - sigma3)

## unitless Mohr diagram
Mohr_plot(
  sigma1 = sigma1, sigma2 = sigma2, sigma3 = sigma3,
  col = "#B63679", lwd = 2, unit = NULL, include.zero = FALSE
)
}
\seealso{
\code{\link[=Mohr_calc]{Mohr_calc()}}, \code{\link[=ggMohr]{ggMohr()}}
}
