% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_quantitites.R
\name{Mohr_strain}
\alias{Mohr_strain}
\title{Mohr Circle Diagram for Strain}
\usage{
Mohr_strain(
  lambda1,
  lambda2 = NA,
  lambda3,
  phi = NULL,
  col = "black",
  n = 512,
  full.circle = FALSE,
  include.zero = TRUE,
  xlim = NULL,
  ylim = NULL,
  round = 1,
  ...
)
}
\arguments{
\item{lambda1, lambda2, lambda3}{numeric. Magnitude of quadratic elongation \eqn{\lambda}}

\item{phi}{numeric. (optional) Angle (in degrees) for a specific strain}

\item{col}{color for Mohr circle.}

\item{n}{integer. Resolution given amount of points along the generated path
representing the full Mohr circle (\code{512} by default).}

\item{full.circle}{logical. Should the complete Mohr circle be shown, or only
the upper (positive shear stress) part of the circle?}

\item{include.zero}{logical. the plot range be extended to include \code{lambda = 0}?}

\item{xlim, ylim}{range of plot}

\item{round}{integer indicating the number of decimal places to be used for rounding.}

\item{...}{optional graphical parameters.}
}
\description{
Plots the Mohr Circle for Strain
}
\examples{
Mohr_strain(lambda1 = 4, lambda3 = 0.25, phi = 25)
}
\seealso{
\code{\link[=Mohr_plot]{Mohr_plot()}} for Stress. \link{strain} for converting strain quantities
}
