% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohR.R
\name{MohrCircle_plot}
\alias{MohrCircle_plot}
\title{Mohr Circle plot}
\usage{
MohrCircle_plot(
  sigmaX = NA,
  sigmaZ = NA,
  tauXZ = NA,
  sigma1 = NA,
  sigma3 = NA,
  unit = "MPa",
  col = "black",
  ...
)
}
\arguments{
\item{sigmaX}{numeric. Magnitude of normal stress acting in the horizontal direction}

\item{sigmaZ}{numeric. Magnitude of normal stress acting in the vertical direction}

\item{tauXZ}{numeric. Magnitude of shear stress acting on the same plane as \code{"sigmaX"}}

\item{sigma1}{numeric. Magnitude of major principal stress}

\item{sigma3}{numeric. Magnitude of minor principal stress}

\item{unit}{character. The unit used for magnitude of stress (\code{"MPa"} by default).}

\item{col}{color for Mohr circle.}

\item{...}{optional graphical parameters.}
}
\description{
plots the Mohr Circle
}
\note{
One of the following two sets of data must be entered
\enumerate{
\item{\code{"sigmaX"}, \code{"sigmaZ"}, \code{"tauXZ"}}
\item{\code{"sigma1"}, \code{"sigma3"}}
}
}
\examples{
MohrCircle_plot(sigmaX = 80, sigmaZ = 120, unit = "kPa", tauXZ = 20, col = '#B63679', lwd = 2)
MohrCircle_plot(sigma1 = 1025, sigma3 = 250, col = '#B63679', lwd = 2)
}
\seealso{
\code{\link[=MohrCircle_calc]{MohrCircle_calc()}}, \code{\link[=ggMohr]{ggMohr()}}
}
\author{
Kyle Elmy and Jim Kaklamanos
}
