% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{operations}
\alias{operations}
\alias{vlength}
\alias{vsum}
\alias{vnorm}
\alias{vcross}
\alias{vdot}
\alias{vrotate}
\alias{vangle}
\alias{vproject}
\alias{vreject}
\title{Vector operations}
\usage{
vlength(x)

vsum(x)

vnorm(x)

vcross(x, y)

vdot(x, y)

vrotate(x, rotaxis, rotangle)

vangle(x, y)

vproject(x, y)

vreject(x, y)
}
\arguments{
\item{x, y, rotaxis}{numeric vector, array, or object of class \code{"line"} or \code{"plane"}}

\item{rotangle}{numeric. Angle in radians or degrees if \code{"rotaxis"} is an object of \code{"line"} or \code{"plane"}}
}
\description{
Vector operations
}
\details{
\code{vdot(x, y)} is equivalent to \code{x \%*\% t(y)}
}
\examples{
vec1 <- cbind(1, 0, 0)
vec2 <- cbind(0, 0, 1)

vlength(vec1)
vsum(vec1)
vnorm(vec1)
vcross(vec1, vec2)
vdot(vec1, vec2)
vrotate(vec1, vec2, pi / 2)
vangle(vec1, vec2)
vproject(vec1, vec2)
vreject(vec1, vec2)
}
