% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{operations}
\alias{operations}
\alias{vlength}
\alias{vsum}
\alias{vnorm}
\alias{vscale}
\alias{vcross}
\alias{vdot}
\alias{vrotate}
\alias{vangle}
\alias{vproject}
\alias{vreject}
\title{Vector operations}
\usage{
vlength(x)

vsum(x, w = NULL)

vnorm(x)

vscale(x, l)

vcross(x, y)

vdot(x, y)

vrotate(x, rotaxis, rotangle)

vangle(x, y)

vproject(x, y)

vreject(x, y)
}
\arguments{
\item{x, y, rotaxis}{numeric vector, array, or object of class \code{"line"} or \code{"plane"}}

\item{w}{weights.}

\item{l}{numeric. length to scale \code{x}}

\item{rotangle}{numeric. Angle in radians or degrees if \code{"rotaxis"} is an object of \code{"line"} or \code{"plane"}}
}
\description{
Vector operations
}
\details{
\code{vlength(x)} returns the length of a vector \code{x}.

\code{vsum(x)} returns the vector sum of \code{x}. If \code{w} is specified.

\code{vscale(x)} returns a vector x with length \code{l}.

\code{vnorm(x)} returns the normalized vector \code{x} (i.e. \code{vlength(x) = 1})

\code{vcross(x, y)} returns the cross product of vectors \code{x} and \code{y} which is a new vector perpendicular to both.

\code{vdot(x, y)} returns the dot product (or scalar product) of x and y. \code{vdot(x, y)} is equivalent to \code{x \%*\% t(y)}

\code{vrotate(x, rotaxis, rotangle)} rotates x about a vector by a specified angle.

\code{vproject(x,y)} projects x on vector y, i.e. returns a vector that has same orientation as y but length of x.

\code{vreject(x,y)} Is the vector between \code{x} and the projected vector of \code{x} onto \code{y} (also called the vector resolute of x perpendicular to y)
}
\examples{
vec1 <- cbind(1, 0, 0)
vec2 <- cbind(0, 0, 1)

vlength(vec1)
vsum(vec1)
vnorm(vec1)
vcross(vec1, vec2)
vdot(vec1, vec2)
vrotate(vec1, vec2, pi / 2)
vangle(vec1, vec2)
vproject(vec1, vec2)
vreject(vec1, vec2)
}
