% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortensor.R
\name{ortensor}
\alias{ortensor}
\alias{ortensor.spherical}
\title{Orientation tensor}
\usage{
\method{ortensor}{spherical}(x, norm = TRUE, w = NULL)

ortensor(x, norm = TRUE, w = NULL)
}
\arguments{
\item{x}{Object of class \code{"Vec3"}, \code{"Line"} or \code{"Plane"}}

\item{norm}{logical. Whether the tensor should be normalized or not.}

\item{w}{numeric. weightings}
}
\value{
matrix
}
\description{
3D orientation tensor, which characterize data distribution using
eigenvalue method. See (Watson 1966, Scheidegger 1965).
}
\details{
The normalized orientation tensor is given as \deqn{D = \frac{1}{n} (x_i, y_i, z_i) (x_i, y_i, z_i)^T}
n = 1
}
\examples{
set.seed(20250411)
x <- rfb(100, mu = Line(120, 50), k = 1, A = diag(c(10, 0, 0)))
ortensor(x)
}
\seealso{
\code{\link[=eigen.spherical]{eigen.spherical()}}, \code{\link[=inertia_tensor.spherical]{inertia_tensor.spherical()}}
}
