% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{ortensor}
\alias{ortensor}
\title{Orientation tensor}
\usage{
ortensor(x, norm = TRUE)
}
\arguments{
\item{x}{numeric. Can be three element vector, three column array, or an
object of class \code{"line"} or \code{"plane"}}

\item{norm}{logical. Whether the tensor should be normalized or not.}
}
\value{
matrix
}
\description{
3D orientation tensor, which characterize data distribution using
eigenvalue method. See (Watson 1966, Scheidegger 1965).
}
\examples{
set.seed(1)
x <- rfb(100, mu = Line(120, 50), k = 1, A = diag(c(10, 0, 0)))
ortensor(x)
}
\seealso{
\code{\link[=or_eigen]{or_eigen()}}
}
