% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortensor.R
\name{ortensor}
\alias{ortensor}
\alias{is.ortensor}
\alias{as.ortensor}
\alias{ortensor.Vec3}
\alias{ortensor.Line}
\alias{ortensor.Ray}
\alias{ortensor.Plane}
\alias{ortensor.Pair}
\title{Orientation tensor}
\usage{
is.ortensor(object)

as.ortensor(object)

ortensor(x, norm, w, shift)

\method{ortensor}{Vec3}(x, norm = TRUE, w = NULL, shift = NULL)

\method{ortensor}{Line}(x, norm = TRUE, w = NULL, shift = NULL)

\method{ortensor}{Ray}(x, norm = TRUE, w = NULL, shift = NULL)

\method{ortensor}{Plane}(x, norm = TRUE, w = NULL, shift = NULL)

\method{ortensor}{Pair}(x, norm = TRUE, w = NULL, shift = FALSE)
}
\arguments{
\item{object}{3x3 matrix}

\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"},
where the rows are the observations and the columns are the coordinates.}

\item{norm}{logical. Whether the tensor should be normalized or not.}

\item{w}{numeric. weightings}

\item{shift.}{logical. Only for \code{"Pair"} objects: Tensor is by default shifted towards positive
eigenvalues, so it could be used as Scheidegger orientation tensor for
plotting. When original Lisle tensor is needed, set shift to \code{FALSE}.}
}
\value{
\code{ortensor()} returns an object of class \code{"ortensor"}

\code{is.ortensor} returns \code{TRUE} if \code{x} is an \code{"ortensor"} object, and \code{FALSE} otherwise.

\code{as.ortensor} coerces a 3x3 matrix into an \code{"ortensor"} object.
}
\description{
3D orientation tensor, which characterize data distribution using
eigenvalue method. See (Watson 1966, Scheidegger 1965).
}
\details{
The normalized orientation tensor is given as \deqn{D = \frac{1}{n} (x_i, y_i, z_i) (x_i, y_i, z_i)^T}
}
\examples{
set.seed(20250411)
x <- rfb(100, mu = Line(120, 50), k = 1, A = diag(c(10, 0, 0)))
ortensor(x, w = runif(nrow(x)))

test <- as.ortensor(diag(3))
is.ortensor(test)

# Orientation tensor for Pairs
ortensor(angelier1990$TYM)
}
\references{
Watson, G. S. (1966). The Statistics of Orientation Data. The Journal of
Geology, 74(5), 786–797.

Scheidegger, A. E. (1964). The tectonic stress and tectonic motion direction
in Europe and Western Asia as calculated from earthquake fault plane solutions.
Bulletin of the Seismological Society of America, 54(5A), 1519–1528.
\doi{10.1785/BSSA05405A1519}

Lisle, R. (1989). The Statistical Analysis of Orthogonal Orientation Data.
The Journal of Geology, 97(3), 360-364.
}
\seealso{
\code{\link[=inertia_tensor]{inertia_tensor()}}

Other ortensor: 
\code{\link{ot_eigen}()},
\code{\link{strain_shape}}
}
\concept{ortensor}
