% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{osmundsen2010}
\alias{osmundsen2010}
\title{Fault-Slip Data from Northern Norway}
\format{
An object of class \code{Pair}
}
\usage{
data('osmundsen2010')
}
\description{
198 measurements fault-slip measurements from a quarry in Røsand on the
Lofoten-Vesterålen archipelago in Northern Norway
}
\examples{
data("osmundsen2010")
head(osmundsen2010)
}
\references{
Osmundsen, P.T., Redfield, T.F., Hendriks, B.H.W., Bergh, S.G., Hansen, J.-A.,
Henderson, I.H.C., Dehls, J., Lauknes, T.R., Larsen, Y., Anda, E., Davidsen,
B., 2010. Fault-controlled Alpine topography in Norway. Journal of the
Geological Society 167, 83-98.
}
\seealso{
Other datasets: 
\code{\link{angelier1990}},
\code{\link{example_lines}},
\code{\link{example_lines_df}},
\code{\link{example_planes}},
\code{\link{example_planes_df}},
\code{\link{gray_example}},
\code{\link{holst}},
\code{\link{hossack1968}},
\code{\link{ramsay}},
\code{\link{shebandowan}},
\code{\link{simongomez}},
\code{\link{strabo_prj}}
}
\concept{datasets}
\keyword{datasets}
