% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortensor.R
\name{ot_eigen}
\alias{ot_eigen}
\title{Eigenvalues and Eigenvectors of a Set of Vectors}
\usage{
ot_eigen(x, scaled = FALSE)
}
\arguments{
\item{x}{Object of class \code{"Vec3"}, \code{"Line"} or \code{"Plane"}}

\item{scaled}{logical. Whether the Eigenvectors should be scaled by the
Eigenvalues (only effective if \code{x} is in Cartesian coordinates).}
}
\value{
list containing
\describe{
\item{\code{values}}{Eigenvalues}
\item{\code{vectors}}{Eigenvectors in coordinate system of \code{x}}
}
}
\description{
Decomposition of Orientation Tensor Eigenvectors and Eigenvalues
}
\examples{
set.seed(20250411)
mu <- rvmf(n = 1)
x <- rfb(100, mu = mu, k = 1, A = diag(c(10, 0, 0)))
x_eigen <- ot_eigen(x)
x_eigen
plot(x, col = "grey")
points(mu, col = 4)
text(mu, labels = "Mean", col = 4, pos = 4)
points(x_eigen$vectors, col = c(1, 2, 3))
text(x_eigen$vectors, col = c(1, 2, 3), labels = c("E1", "E2", "E3"), pos = 4)
}
\seealso{
\code{\link[=ortensor]{ortensor()}}
}
