% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_pole.R
\name{perm-rsq}
\alias{perm-rsq}
\alias{perm_rsq}
\alias{perm_rsq_pvalue}
\title{Permutation Tests of Regressions}
\usage{
perm_rsq(n_perm, FUN, x, ...)

perm_rsq_pvalue(r_perm, r)
}
\arguments{
\item{n_perm}{A real number (positive integer). The number of permutations to try.}

\item{FUN}{An R function. Must return a result list with fields \code{convergence},
\code{min_eigenvalue}, \code{r_squared} Typically \code{\link[=regression_greatcircle]{regression_greatcircle()}} or a
similar regression function.}

\item{x}{A real vector. The values of the independent variable. Assumed to be
the first argument passed to \code{FUN}.}

\item{...}{Other arguments passed to \code{FUN}, after \code{x}.}

\item{r_perm}{vector of permutated \eqn{R^2} values}

\item{r}{original \eqn{R^2} value}
}
\value{
\code{perm_rsq} returns a a real vector. The maximum length is \code{n_perms}.
Often the length is less than \code{n_perms}, because the regression failed (as
signaled by error != 0 or \code{min_eigenvalue} <= 0).

\code{perm_rsq_pvalue} returns the p-value.
}
\description{
\code{perm_rsq} retrieves the \eqn{R^2} values from successful \code{n_perm} permutations
of a regression function.
\code{perm_rsq_pvalue} post-processes this vector to count how many of its entries
exceed the original \eqn{R^2} value. That fraction is a p-value for the null
hypothesis that the regression result is meaningless.
}
\examples{
set.seed(20250411)
data("gray_example")

# original regression
bestgc_clea <- regression_greatcircle(gray_example[1:8, ])

# permutation test
pr <- perm_rsq(100, FUN = regression_greatcircle, x = gray_example[1:8, ])

# p-value
perm_rsq_pvalue(pr, bestgc_clea)
}
\seealso{
Other geodesic-regression: 
\code{\link{best_fit}}
}
\concept{geodesic-regression}
