% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{plot-spherical}
\alias{plot-spherical}
\alias{plot.Line}
\alias{plot.Vec3}
\alias{plot.Ray}
\alias{plot.Plane}
\alias{plot.Pair}
\alias{plot.Fault}
\title{Plot spherical objects}
\usage{
\method{plot}{Line}(x, upper.hem = FALSE, earea = TRUE, grid.params = list(), ...)

\method{plot}{Vec3}(x, upper.hem = FALSE, earea = TRUE, grid.params = list(), ...)

\method{plot}{Ray}(x, upper.hem = FALSE, earea = TRUE, grid.params = list(), ...)

\method{plot}{Plane}(x, upper.hem = FALSE, earea = TRUE, grid.params = list(), ...)

\method{plot}{Pair}(x, upper.hem = FALSE, earea = TRUE, grid.params = list(), ...)

\method{plot}{Fault}(x, upper.hem = FALSE, earea = TRUE, grid.params = list(), ...)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"},
where the rows are the observations and the columns are the coordinates.}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}, the default).}

\item{earea}{logical \code{TRUE} for Lambert equal-area projection (also "Schmidt net"; the default), or
\code{FALSE} for meridional stereographic projection (also "Wulff net" or "Stereonet").}

\item{grid.params}{list.}

\item{...}{parameters passed to \code{\link[=stereo_point]{stereo_point()}}, \code{\link[=stereo_smallcircle]{stereo_smallcircle()}}, \code{\link[=stereo_greatcircle]{stereo_greatcircle()}}, or \code{\link[=fault_plot]{fault_plot()}}}
}
\description{
Plot spherical objects
}
\details{
If \code{x} is a Ray, than solid symbols show rays pointing in the lower hemisphere,
and open symbols point into the upper hemisphere
}
\examples{
plot(rvmf(10, mu = Vec3(1, 0, 0))) # Vec
plot(Line(c(90, 80), c(10, 75)), lab = c("L1", "L2"))
plot(Ray(c(90, 80), c(10, 75), sense = c(1, -1)), lab = c("L1", "L2"))
plot(Plane(120, 30), col = "red")
plot(Pair(120, 50, 36, 8))
plot(Fault(120, 50, 36, 8, -1))
}
