% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{points.spherical}
\alias{points.spherical}
\title{Add Points to a Plot}
\usage{
\method{points}{spherical}(x, upper.hem = FALSE, earea = TRUE, ...)
}
\arguments{
\item{x}{objects of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"}.}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}, the default).}

\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection (the default), or \code{FALSE} for meridional stereographic projection.}

\item{...}{arguments passed to \code{\link[graphics:points]{graphics::points()}}}
}
\description{
Add Points to a Plot
}
\examples{
stereoplot()
points(rvmf(n = 100))
points(Plane(120, 30), col = "red", pch = 19)
}
