% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faults.R
\name{principal_fault}
\alias{principal_fault}
\title{Calculate Principal Fault planes from stress vectors and friction}
\usage{
principal_fault(s1, s3, friction = 0.6)
}
\arguments{
\item{s1, s3}{Principal stress vectors as spherical objects}

\item{friction}{numeric. Coefficient of friction}
}
\value{
\code{"Fault"} object
}
\description{
Calculate Principal Fault planes from stress vectors and friction
}
\examples{
res_TYM <- slip_inversion(angelier1990$TYM, n_iter = 10)
pr_TYM <- principal_fault(res_TYM$principal_axes[1, ], res_TYM$principal_axes[3, ])

stereoplot()
fault_plot(angelier1990$TYM, col = "grey")
fault_plot(pr_TYM, col = "red")
points(res_TYM$principal_axes, pch = 16)
}
