% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortensor.R
\name{projected_mean}
\alias{projected_mean}
\title{Projected Mean}
\usage{
projected_mean(x, ...)
}
\arguments{
\item{x}{either an object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}, where the
rows are the observations and the columns are the coordinates, or an \code{"ortensor"} object.}

\item{...}{additional arguments passed to \code{\link[=ortensor]{ortensor()}} (ignored if \code{x} is \code{"ortensor"} object).}
}
\value{
Vector in coordinate system of \code{x}
}
\description{
Eigenvector with largest eigenvalue of the orientation tensor
}
\examples{
example_lines_df$quality
projected_mean(example_lines, w = runif(nrow(example_lines)))
}
\references{
Bachmann, F., Hielscher, R., Jupp, P. E., Pantleon, W., Schaeben, H., &
Wegert, E. (2010). Inferential statistics of electron backscatter diffraction
data from within individual crystalline grains. Journal of Applied Crystallography, 43(6), 1338–1355. https://doi.org/10.1107/S002188981003027X
}
\seealso{
\code{\link[=ot_eigen]{ot_eigen()}} for eigenvalues of orientation tensor,
\code{\link[=sph_mean]{sph_mean()}} for arithmetic mean, \code{\link[=geodesic_mean]{geodesic_mean()}} for geodesic mean.
}
