% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ramsay}
\alias{ramsay}
\title{R/phi data}
\format{
An object of class \code{matrix}
}
\usage{
data('ramsay')
}
\description{
Example strain data from Ramsay and Huber 1983
}
\examples{
data("ramsay")
head(ramsay)
}
\seealso{
Other datasets: 
\code{\link{angelier1990}},
\code{\link{example_lines}},
\code{\link{example_lines_df}},
\code{\link{example_planes}},
\code{\link{example_planes_df}},
\code{\link{gray_example}},
\code{\link{holst}},
\code{\link{shebandowan}},
\code{\link{simongomez}}
}
\concept{datasets}
\keyword{datasets}
