% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ramsay}
\alias{ramsay}
\title{R/phi data}
\format{
An object of class \code{matrix}
}
\usage{
data('ramsay')
}
\description{
example data set from Ramsay and Huber 1983
}
\examples{
data("ramsay")
head(ramsay)
}
\keyword{datasets}
