% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sph_distributions.R
\name{rbing}
\alias{rbing}
\alias{rbingham}
\alias{rbingham_eig}
\title{Bingham Distribution}
\usage{
rbingham(n, A, .class = c("Vec3", "Line", "Ray", "Plane"))

rbingham_eig(n, eigenvalues, .class = c("Vec3", "Line", "Ray", "Plane"))
}
\arguments{
\item{n}{integer. number of random samples to be generated}

\item{A}{symmetric matrix}

\item{.class}{character.}

\item{eigenvalues}{numeric, three-element vector. Eigenvalues of the diagonal symmetric matrix of the Bingham distribution.}
}
\value{
A spherical object of class \code{.class} and length \code{n}
}
\description{
Simulation of random values from a Bingham distribution with any given symmetric matrix A.
}
\examples{
a <- cov(iris[, 1:3])
r <- rbingham(100, a, "Line")
contour(r)
points(r)

re <- rbingham_eig(100, c(100, 1, 0), "Line")
contour(re)
points(re)
}
\references{
Fallaize C. J. and Kypraios T. (2016). Exact bayesian inference for the Bingham distribution. Statistics and Computing, 26(1): 349–360. \url{http://arxiv.org/pdf/1401.2894v1.pdf}
}
\seealso{
Other random: 
\code{\link{rfb}()},
\code{\link{rkent}()},
\code{\link{rrot}()},
\code{\link{runif.spherical}()},
\code{\link{rwatson}()},
\code{\link{vonmises-fisher}}
}
\concept{random}
