% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion.R
\name{reduced_stress}
\alias{reduced_stress}
\title{Reduced stress tensor}
\usage{
reduced_stress(fault, method = c("michael", "angelier"))
}
\arguments{
\item{fault}{\code{"Fault"} object where the rows are the observations, and the columns the coordinates.}

\item{method}{character. The inversion algorithm. One of \code{"michael"} (the default) for a bootstrapped linear inversion,
or \code{"angelier"} for a iterative direct inversion.}
}
\value{
\code{"ellipsoid"} object. 3x3 matrix (i.e. 2nd rank tensor).
}
\description{
Calculates the reduced stress tensor (without optimization).
}
\examples{
f <- angelier1990$TYM
reduced_stress(f)
}
\seealso{
\code{\link[=slip_inversion]{slip_inversion()}}

Other stress-tensor: 
\code{\link{fault_instability_criterion}()},
\code{\link{stress_shape}()},
\code{\link{tau-comp}},
\code{\link{tau2rup}()},
\code{\link{tau2stress}()}
}
\concept{stress-tensor}
