% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_analysis.R
\name{RGN_plot}
\alias{RGN_plot}
\title{Rigid Grain Net (RGN)}
\usage{
RGN_plot(
  r,
  theta,
  angle_error = 3,
  boot = 100L,
  probs = 0.972,
  grid = 0.05,
  main = "Rigid-Grain-Net",
  ...
)
}
\arguments{
\item{r}{numeric. The porphyroclast aspect ratio (long axis/short axis)}

\item{theta}{numeric. Angle between long axis and foliation (in degrees)}

\item{angle_error}{numeric. Uncertainty of angle measurement. \code{3} by default.}

\item{boot}{integer. Number of bootstrap resamples}

\item{probs}{integer. Probability with values in \eqn{[0, 1]} to estimate
critical shape factor, i.e. the largest shape factor of measurements outside
of critical hyperbole.}

\item{grid}{numeric. Spacing of hyperboles.}

\item{main}{character. The main title (on top).}

\item{...}{plotting arguments passed to \code{\link[graphics:points]{graphics::points()}}}
}
\value{
a plot or a list of the calculated \code{B} (shape factor) and \code{theta} values,
and the bootstrapped confidence interval of the critical B value (\code{Rc_CI}).
}
\description{
The rigid grain net after (Jessup et al. 2007) plots the distribution the
strain ratio (\code{R}) of orientation (\code{phi}) of
porphyroclast over the theoretical distribution of tailless clasts. The plot estimates
the critical shape factor \code{Rc} marking the transition between the stable-sink
position and infinitely rotating porphyroclasts.
This critical shape factor can be interpreted as the the \strong{mean kinmatic vorticity number}.
Here the \code{Rc} is estimated using the bootstrap method described in Stephan et al. (2025).
}
\examples{
data(ramsay)
set.seed(20250411)

# assuming the mean orientation is the foliation:
theta <- tectonicr::circular_mean(ramsay[, 2]) - ramsay[, 2]

RGN_plot(ramsay[, "R"], theta, col = "darkred")
}
\references{
Jessup, Micah J., Richard D. Law, and Chiara Frassi.
"The rigid grain net (RGN): an alternative method for estimating mean
kinematic vorticity number (Wm)." Journal of Structural Geology 29.3 (2007):
411-421. doi: 10.1016/j.jsg.2006.11.003

Stephan, Tobias, et al. "Going with the flow—Changes of vorticity control
gold enrichment in Archean shear zones (Shebandowan Greenstone Belt,
Superior Province, Canada)." Journal of Structural Geology (2025): 105542.
doi: 10.1016/j.jsg.2025.105542
}
