% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sph_distribution.R
\name{rkent}
\alias{rkent}
\title{Kent distribution}
\source{
Adapted from \code{\link[Directional:rkent]{Directional::rkent()}}
}
\usage{
rkent(n = 100, mu = c(1, 0, 0), k = 5, b)
}
\arguments{
\item{n}{integer. number of random samples to be generated}

\item{mu}{numeric. Can be three element vector, three column array, or an
object of class \code{"line"} or \code{"plane"}}

\item{k}{numeric. The concentration parameter (\eqn{\kappa})}

\item{b}{numeric. \eqn{\beta} (ellipticity): \eqn{0 \leq \beta < \kappa}}
}
\description{
Simulation of random values from a spherical Kent distribution.
}
\examples{
set.seed(20250411)
x <- rkent(100, mu = Line(120, 50), k = 5, b = 1)
stereoplot()
stereo_point(x)
}
\seealso{
\code{\link[=rvmf]{rvmf()}} for von Mises-Fisher distribution
}
