% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sph_distributions.R
\name{rkent}
\alias{rkent}
\title{Kent distribution}
\source{
Adapted from \code{\link[Directional:rkent]{Directional::rkent()}}
}
\usage{
rkent(n = 100, mu = Vec3(1, 0, 0), k = 5, b)
}
\arguments{
\item{n}{integer. number of random samples to be generated}

\item{mu}{Mean vector. object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}, where the
rows are the observations and the columns are the coordinates.}

\item{k}{numeric. The concentration parameter (\eqn{\kappa}) of the von
Mises-Fisher distribution}

\item{b}{numeric. \eqn{\beta} (ellipticity): \eqn{0 \leq \beta < \kappa}}
}
\description{
Simulation of random values from a spherical Kent distribution.
}
\examples{
set.seed(20250411)
x <- rkent(100, mu = Line(120, 50), k = 5, b = 1)
plot(x)
}
\seealso{
\code{\link[=rvmf]{rvmf()}} to draw samples from the von Mises Fisher distribution
around a specified mean vector. \code{\link[=runif.spherical]{runif.spherical()}} to draw from a a spherical uniform distribution.
\code{\link[=rfb]{rfb()}} to draw from a Fisher-Bingham distribution.
}
