% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_math.R
\name{rotate}
\alias{rotate}
\alias{rotate.Vec3}
\alias{rotate.Ray}
\alias{rotate.Line}
\alias{rotate.Plane}
\alias{rotate.Pair}
\title{Vector Rotation}
\usage{
rotate(x, rotaxis, rotangle)

\method{rotate}{Vec3}(x, rotaxis, rotangle)

\method{rotate}{Ray}(x, rotaxis, rotangle)

\method{rotate}{Line}(x, rotaxis, rotangle)

\method{rotate}{Plane}(x, rotaxis, rotangle)

\method{rotate}{Pair}(x, rotaxis, rotangle)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"},
where the rows are the observations and the columns are the coordinates.}

\item{rotaxis}{Axis of rotation given as object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}.}

\item{rotangle}{Angle of rotation in radians for \code{"Vec3"} objects and in degrees for \code{"Line"}, \code{"Ray"} and \code{"Plane"} objects.}
}
\value{
objects of same class as \code{x}
}
\description{
Vector Rotation
}
\examples{
vec1 <- Vec3(1, 0, 0)
vec2 <- Vec3(0, 0, 1)
rotate(vec1, vec2, pi / 2)

# rotate Fault data (sense of motion changes!)
rotate(simongomez[1:5, ], Ray(90, 10), 80)
}
\seealso{
\code{\link[=defgrad_from_axisangle]{defgrad_from_axisangle()}}; \code{\link[=transform_linear]{transform_linear()}}
}
