% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{rotate_stereogrid}
\alias{rotate_stereogrid}
\title{Center gridlines on a given point}
\usage{
rotate_stereogrid(x, d = 10, col = "gray90", lwd = 0.5, lty = 1, ...)
}
\arguments{
\item{x}{center position of grid lines.}

\item{d}{integer. Angle distance between guides. Default: 10}

\item{col}{color of guide lines}

\item{lwd}{linewidth of guide lines}

\item{lty}{linetype of guide lines}

\item{...}{arguments passed to \code{\link[graphics:lines]{graphics::lines()}}}
}
\description{
Center gridlines on a given point
}
\examples{
stereoplot(guide = FALSE)
rotate_stereogrid(Plane(120, 50), earea = FALSE)
}
