% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sph_distributions.R
\name{runif.spherical}
\alias{runif.spherical}
\title{Uniformly Distributed Vectors on the Sphere}
\source{
Adapted fom \code{\link[rotasym:unif]{rotasym::r_unif_sphere()}} and \code{rayUniform()} from
\code{geologyGeometry} by Davis, J.R.
}
\usage{
runif.spherical(
  n = 100,
  class = c("Vec3", "Ray", "Line", "Plane"),
  method = c("geologyGeometry", "gss", "sfs", "rotasym")
)
}
\arguments{
\item{n}{integer. number of random samples to be generated}

\item{class}{character. Coordinate class of the output vectors.}

\item{method}{character. The algorithm for generating uniformly distributed
vectors. Either \code{"geologyGeometry"} (the default) for generating random points in Cartesian coordinates
(as in the \code{geologyGeometry} code compilation),
\code{"sfs"} for the "Spherical Fibonacci Spiral points on a sphere",
\code{"gss"} for "Golden Section Spiral points on a sphere", or the algorithm
\code{\link[rotasym:unif]{rotasym::r_unif_sphere()}} from the rotasym package.}
}
\value{
object of class specified by \code{"class"} argument
}
\description{
Create uniformly distributed vectors on the sphere
}
\details{
\code{"sfs"} algorithm is from on John Burkardt (http://people.sc.fsu.edu/~jburkardt/),
\verb{"gss} is from  http://www.softimageblog.com/archives/115
}
\examples{
set.seed(20250411)
x1 <- runif.spherical(n = 100, "Ray", method = "sfs")
contour(x1)

x2 <- runif.spherical(n = 100, "Ray", method = "gss")
contour(x2)

x3 <- runif.spherical(n = 100, "Ray", method = "rotasym")
contour(x3)

x4 <- runif.spherical(n = 100, "Ray", method = "geologyGeometry")
contour(x4)
}
\seealso{
Other random: 
\code{\link{rbing}},
\code{\link{rfb}()},
\code{\link{rkent}()},
\code{\link{rrot}()},
\code{\link{rwatson}()},
\code{\link{vonmises-fisher}}
}
\concept{random}
