% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sph_distributions.R
\name{rwatson}
\alias{rwatson}
\title{Watson Distribution}
\source{
\code{geologyGeometry} by Davis, J.R.
}
\usage{
rwatson(n, mu, k)
}
\arguments{
\item{n}{integer. number of random samples to be generated}

\item{mu}{Mean vector. object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}, where the
rows are the observations and the columns are the coordinates.}

\item{k}{numeric. The concentration parameter (\eqn{\kappa}) of the von
Mises-Fisher distribution}
}
\value{
vector of class \code{mu} of length \code{n}
}
\description{
A naive acceptance-rejection sampling algorithm, based on bounding the
density (with respect to the distance from \code{mu}) with a constant. For large
\code{kappa}, this method grows inefficient. For \code{kappa} == 100, about 13 tries
are needed per success. For \code{kappa} == -100, about 18 tries are needed.
}
\examples{
set.seed(20250411)
r <- rwatson(100, mu = Ray(120, 50), k = 10)

contour(r)
points(r)
}
\seealso{
Other random: 
\code{\link{rbing}},
\code{\link{rfb}()},
\code{\link{rkent}()},
\code{\link{rrot}()},
\code{\link{runif.spherical}()},
\code{\link{vonmises-fisher}}
}
\concept{random}
