% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scale_weights}
\alias{scale_weights}
\title{Scaling weightings}
\usage{
scale_weights(
  e,
  error_type = c("rank", "angle", "rup"),
  scaling = c("lin", "inv_lin", "inv_square", "exp"),
  replace_na = TRUE,
  norm = FALSE
)
}
\arguments{
\item{e}{numeric. Weights}

\item{error_type}{character. One of \code{"rank"} (a numeric value ranking
measurement quality), \code{"angle"} (a reading error expressed as angles in
degrees), and \code{"rup"} (RUP values from a previous fault-slip inversion)}

\item{scaling}{character. Scaling function to use. One of \code{"lin"} (leaves
weights \eqn{e} as is), \code{"inv_lin"} (\eqn{1/e}), \code{"inv_square"} (\eqn{1/e^2}),
and \code{"exp"} (\eqn{\exp{-(x-1)}})}

\item{replace_na}{logical. Imputation? Whether \code{NA} should be replaced by the
mean of the weights? (\code{TRUE} by default)}

\item{norm}{logical. Whether the scaled weights should be normalized by
their mean? (\code{FALSE} by default)}
}
\value{
numeric.
}
\description{
Scales quantitative errors or qualities of orientation measurements. Useful for some
statistical summaries or fault-slip inversion.
}
\examples{
set.seed(20250411)
# Generate some random weights from 1 (poor) to 5 (good)
err <- sample(1:5, size = 10, replace = TRUE)

# Introduce 3 random NAs
err[sample(length(err), 3)] <- NA

scale_weights(err, error_type = 'rank', scaling = 'inv_square', norm = TRUE)
}
\seealso{
\code{\link[=slip_inversion_angelier]{slip_inversion_angelier()}}
}
