% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shebandowan}
\alias{shebandowan}
\title{Vorticity from Rotated Porphyroclasts}
\format{
An object of class \code{data.frame}
}
\usage{
data('shebandowan')
}
\description{
194 measurements of aspect ratio and orientation (wrt. foliation) of rotated
porphyroclasts in a mylonitic, greenschist-grade metavolcanic rock from the
Shebandowan Greenstone Belt (Superior Province, Canada).
Sample HT-17-XZ from Stephan et al. (2025)
}
\examples{
data("shebandowan")
head(shebandowan)
}
\references{
Stephan, T., Phillips, N., Tiitto, H., Perez, A., Nwakanma, M.,
Creaser, R., & Hollings, P. (2025). Going with the flow - Changes of
vorticity control gold enrichment in Archean shear zones
(Shebandowan Greenstone Belt, Superior Province, Canada).
Journal of Structural Geology, 201, 105542. https://doi.org/10.1016/j.jsg.2025.105542
}
\keyword{datasets}
