% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion.R
\name{sigma-comp}
\alias{sigma-comp}
\alias{sigma2shearnorm}
\alias{sigma2tendency}
\title{Resolved Shear and Normal Stress}
\usage{
sigma2shearnorm(sigma, fault, friction = 0.6)

sigma2tendency(sigma, fault, friction = 0.6)
}
\arguments{
\item{sigma}{symmetric 3x3 matrix. The (reduced) stress tensor.}

\item{fault}{\code{"Fault"} object where the rows are the observations, and the columns the coordinates.}

\item{friction}{numeric. Coefficient of friction (0.6 by default)}
}
\value{
2-column numeric array giving the relative normal and shear stress components
for each fault in \code{fault}.
}
\description{
\code{sigma2stress()} calculate normal and shear stress components, while
\code{sigma2tendency()} computes the tendency for slip and dilatency for a given set
of faults and a given stress tensor.
}
\examples{
f <- angelier1990$TYM
s <- reduced_stress(f)
sigma2shearnorm(s, f)

sigma2tendency(s, f)
}
\seealso{
\code{\link[=slip_inversion]{slip_inversion()}}

Other stress-tensor: 
\code{\link{fault_instability_criterion}()},
\code{\link{reduced_stress}()},
\code{\link{sigma2rup}()},
\code{\link{sigma2stress}()},
\code{\link{stress_shape}()}
}
\concept{stress-tensor}
