% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohR.R
\name{sigma13}
\alias{sigma13}
\title{Principle stresses}
\usage{
sigma13(sigmaX, sigmaZ, tauXZ)
}
\arguments{
\item{sigmaX}{numeric. Magnitude of normal stress acting in the horizontal direction}

\item{sigmaZ}{numeric. Magnitude of normal stress acting in the vertical direction}

\item{tauXZ}{numeric. Magnitude of shear stress acting on the same plane as \code{"sigmaX"}}
}
\value{
A four-element list containing
\describe{
\item{\code{"sigma1"}}{magnitude of major principal stress}
\item{\code{"sigma3"}}{magnitude of minor principal stress}
\item{\code{"theta1"}}{direction of major principal stress (degrees)}
\item{\code{"theta3"}}{direction of minor principal stress (degrees)}
}
}
\description{
calculates the magnitudes and directions of the principal stresses S1 and S2
}
\examples{
sigma13(sigmaX = 80, sigmaZ = 120, tauXZ = 20)
}
\author{
Kyle Elmy and Jim Kaklamanos
}
