% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion.R
\name{sigma2stress}
\alias{sigma2stress}
\title{Principal Stresses from Stress Tensor}
\usage{
sigma2stress(sigma)
}
\arguments{
\item{sigma}{symmetric 3x3 matrix. The (reduced) stress tensor.}
}
\value{
list with the following components
\describe{
\item{\code{"sigma_vals"}}{numeric. The relative magnitudes of the principal stress axes.}
\item{\code{"sigma_axes"}}{The principal stress axes as \code{Line} objects.}
}
}
\description{
The eigenvector and eigenvalues of a stress tensor give the orientations and
relative magnitudes of the principal stress axes.
}
\examples{
f <- angelier1990$TYM
s <- reduced_stress(f)
sigma2stress(s)
}
\seealso{
\code{\link[=slip_inversion]{slip_inversion()}}

Other stress-tensor: 
\code{\link{fault_instability_criterion}()},
\code{\link{reduced_stress}()},
\code{\link{sigma-comp}},
\code{\link{sigma2rup}()},
\code{\link{stress_shape}()}
}
\concept{stress-tensor}
