% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simongomez}
\alias{simongomez}
\title{Example data set}
\format{
An object of class \verb{"Fault}
}
\usage{
data('simongomez')
}
\description{
80 Fault-slip orientations from Simon-Gomez ???
}
\examples{
data("simongomez")
head(simongomez)
}
\seealso{
Other datasets: 
\code{\link{angelier1990}},
\code{\link{example_lines}},
\code{\link{example_lines_df}},
\code{\link{example_planes}},
\code{\link{example_planes_df}},
\code{\link{gray_example}},
\code{\link{holst}},
\code{\link{ramsay}},
\code{\link{shebandowan}}
}
\concept{datasets}
\keyword{datasets}
