% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simongomez}
\alias{simongomez}
\title{Example data set}
\format{
An object of class \code{data.frame}
}
\usage{
data('simongomez')
}
\description{
80 Fault-slip orientations from Simon-Gomez
}
\examples{
data("simongomez")
head(simongomez)
}
\keyword{datasets}
