% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion.R
\name{slip_inversion}
\alias{slip_inversion}
\title{Stress Inversion for Fault-Slip Data}
\usage{
slip_inversion(x, method = c("michael", "angelier"), ...)
}
\arguments{
\item{x}{\code{"Fault"} object where the rows are the observations, and the columns the coordinates.}

\item{method}{character. The inversion algorithm. One of \code{"michael"} (the default) for a bootstrapped linear inversion,
or \code{"angelier"} for a iterative direct inversion.}

\item{...}{arguments passed to \code{\link[=slip_inversion_angelier]{slip_inversion_angelier()}} or \code{\link[=slip_inversion_michael]{slip_inversion_michael()}}}
}
\value{
list. See \code{\link[=slip_inversion_angelier]{slip_inversion_angelier()}} and \code{\link[=slip_inversion_michael]{slip_inversion_michael()}}
}
\description{
Linear and direction inversion of fault-slip data to derive the reduced stress tensor.
}
\examples{
set.seed(20250411)
# Use Angelier examples:
dat <- angelier1990$TYM

# Linear inversion
TYM_michael <- slip_inversion(dat, method = "michael")

# Direct inversion
TYM_angelier <- slip_inversion(dat, method = "angelier")

stereoplot(guides = FALSE)
fault_plot(dat, col = "gray80")
points(TYM_michael$principal_axes, pch = 3, col = 1:3)
points(TYM_angelier$principal_axes, pch = 2, col = 1:3)
legend("topleft", col = 1, legend = c("Michael (1984)", "Angelier (1990)"), pch = c(3, 2))
}
\references{
Angelier, J. (1990). Inversion of field data in fault tectonics to obtain
the regional stress—III. A new rapid direct inversion method by analytical
means. Geophys. J. Int, 103, 363–376. \url{https://doi.org/10.1111/j.1365-246X.1990.tb01777.x}

Michael, A. J. (1984). Determination of stress from slip data: Faults and
folds. Journal of Geophysical Research: Solid Earth, 89(B13), 11517–11526.
\url{https://doi.org/10.1029/JB089iB13p11517}
}
\seealso{
Other stress-inversion: 
\code{\link{Fault_PT}()},
\code{\link{slip_inversion_angelier}()},
\code{\link{slip_inversion_michael}()},
\code{\link{slip_inversion_simple}()}
}
\concept{stress-inversion}
