% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion.R
\name{slip_inversion}
\alias{slip_inversion}
\title{Stress Inversion for Fault-Slip Data}
\usage{
slip_inversion(x, method = c("michael", "angelier", "hansen"), ...)
}
\arguments{
\item{x}{\code{"Fault"} object where the rows are the observations, and the columns the coordinates.}

\item{method}{character. The inversion algorithm, one of \code{"michael"} (the default)
for a bootstrapped linear inversion after Micheal (1984), \code{"angelier"}
for an iterative direct inversion after Angelier (1990) and Mostafa (2005), and
\code{"hansen"} for direct inversion using the 9d parameter space after Hansen (2013).}

\item{...}{arguments passed to \code{\link[=slip_inversion_angelier]{slip_inversion_angelier()}} or \code{\link[=slip_inversion_michael]{slip_inversion_michael()}}}
}
\value{
a named list with the following components:
\describe{
\item{\code{stress_tensor}}{\code{"ellipsoid"} object. Best-fit devitoric stress tensor in input coordinate frame}
\item{\code{principal_axes}}{\code{"Line"} objects. Orientation of the principal stress axes as unit vectors (max to min)}
\item{\code{tensor_params}}{the four tensor parameters (Eq. 4.87)}
\item{\code{principal_vals}}{eigenvalues of the stress tensor (\eqn{\sigma_1 >= \sigma_2 >= \sigma_3})}
\item{\code{stress_shape}}{list Stress shape ratio. See \code{\link[=stress_shape]{stress_shape()}}.}
\item{\code{misfit}}{list. Misfit parameters. See \code{\link[=slip_inversion_misfit]{slip_inversion_misfit()}}.}
\item{\code{SHmax}}{numeric. Direction of maximum horizontal stress (in degrees)}
\item{\code{tau_mean}}{numeric. Average resolved shear stress on each plane. Should be close to 1.}
\item{\code{stress_components}}{matrix. The resolved shear and normal stresses, the slip and dilation tendency on each plane. See \code{\link[=tau2shearnorm]{tau2shearnorm()}} and \code{\link[=tau2tendency]{tau2tendency()}}.}
\item{\code{n_iter}}{number of Mostafa iterations performed}
}
}
\description{
Convenient function for linear and direction inversion of fault-slip data to
derive the reduced stress tensor.
}
\examples{
set.seed(20250411)
# Use Angelier examples
par(mfrow = c(1, length(angelier1990)))
invisible(lapply(angelier1990, function(x){

# Inversion after Michael (1984)
res_michael <- slip_inversion(x, method = "michael", n_iter = 100, n = 100, res = 100)

# Inversion after Angelier (1990)
res_angelier <- slip_inversion(x, method = "angelier")

#res_hansen <- slip_inversion(x, method = "hansen")

stereoplot(guides = FALSE)
fault_plot(x, col = "gray80")
points(res_michael$principal_axes, pch = 3, col = 2:4)
points(res_angelier$principal_axes, pch = 2, col = 2:4)
#points(res_hansen$nine$principal_axes, pch = 1, col = 2:4)
legend("topleft", col = 1, legend = c("Michael (1984)", "Angelier (1990)"), pch = c(3, 2, 1))
}))
}
\references{
Angelier, J. (1990). Inversion of field data in fault tectonics to obtain
the regional stress—III. A new rapid direct inversion method by analytical
means. Geophys. J. Int, 103, 363–376. \url{https://doi.org/10.1111/j.1365-246X.1990.tb01777.x}

Hansen, J. A. (2013). Direct inversion of stress, strain or strain rate
including vorticity: A linear method of homogenous fault-slip data inversion
independent of adopted hypothesis. Journal of Structural Geology, 51, 3–13.
https://doi.org/10.1016/j.jsg.2013.03.014

Michael, A. J. (1984). Determination of stress from slip data: Faults and
folds. Journal of Geophysical Research: Solid Earth, 89(B13), 11517–11526.
\url{https://doi.org/10.1029/JB089iB13p11517}
}
\seealso{
Other stress-inversion: 
\code{\link{Fault_PT}()},
\code{\link{slip_inversion_angelier}()},
\code{\link{slip_inversion_hansen}()},
\code{\link{slip_inversion_michael}()},
\code{\link{slip_inversion_simple}()}
}
\concept{stress-inversion}
