% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion_hansen.R
\name{slip_inversion_hansen}
\alias{slip_inversion_hansen}
\title{9D Direct Inversion of Fault Slip}
\usage{
slip_inversion_hansen(x, flip = FALSE)
}
\arguments{
\item{x}{object of class \code{"Pair"} or \code{"Fault"}}

\item{flip}{logical. Flip if you want to have the negative stress tensor, i.e. sigma 1 and 3 will be flipped.}
}
\value{
list
}
\description{
Direct inversion of stress, strain or strain rate including vorticity using
9D parameter space using the method by Hansen (2013)
}
\examples{
par(mfrow = c(1, 2))
invisible(lapply(angelier1990, function(x){
res <- slip_inversion_hansen(x, TRUE)
plot(x, col = 'lightgrey')
title(sub = paste0("R = ", round(res$phi, 2)))
points(res$principal_axes, col = 2:4, pch = 16, cex = 2)
text(res$principal_axes, labels = 1:3, col = 2:4, adj = -1)
points(res$vorticity_axis, col = 5, pch = 17, cex = 2)
}))
}
\references{
Hansen, J. A. (2013). Direct inversion of stress, strain or strain rate
including vorticity: A linear method of homogenous fault-slip data inversion
independent of adopted hypothesis. Journal of Structural Geology, 51, 3–13.
https://doi.org/10.1016/j.jsg.2013.03.014
}
\seealso{
Other stress-inversion: 
\code{\link{Fault_PT}()},
\code{\link{slip_inversion}()},
\code{\link{slip_inversion_angelier}()},
\code{\link{slip_inversion_michael}()},
\code{\link{slip_inversion_simple}()}
}
\concept{stress-inversion}
