% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion_hansen.R
\name{slip_inversion_hansen}
\alias{slip_inversion_hansen}
\title{9D Direct Inversion for Fault Slip Including Vorticity}
\usage{
slip_inversion_hansen(x, flip = FALSE, friction = 0.6)
}
\arguments{
\item{x}{object of class \code{"Pair"} or \code{"Fault"} with at least 7 rows.}

\item{flip}{logical. Flip if you want to have the negative stress tensor, i.e.
sigma 1 and 3 will be flipped.}

\item{friction}{numeric. Coefficient of friction (0.6 by default)}
}
\value{
list
}
\description{
Direct inversion of stress, strain or strain rate including vorticity using
9D parameter space using the method by Hansen (2013). It can be applied
regardless whether the dynamic or the kinematic hypothesis is adopted;
it can handle datasets representing two to seven degrees of freedom; and it
is not dependent on the correct assessment of slip sense.
}
\examples{
# Use Angelier examples:
nx <- length(angelier1990)
par(mfrow = c(1, length(angelier1990)))

invisible(lapply(seq_len(nx), function(i){

# inversion
x <- angelier1990[[i]]
res <- slip_inversion_hansen(x, TRUE)

phi_val <- round(res$stress_shape$phi, 2)
rup_val <-  round(res$misfit$rup, 2)
w_val <- round(res$vorticity_mag, 2)

stereoplot(title = names(angelier1990)[i], guides = FALSE)
stereo_shmax(res$SHmax)
fault_plot(x, col = assign_col(res$misfit$rup))
points(res$principal_axes, col = 2:4, pch = 16, cex = 2)
text(res$principal_axes, labels = rownames(res$principal_axes), col = 2:4, adj = -.5)
points(res$vorticity_axis, col = 5, pch = 17, cex = 2)
text(res$vorticity_axis, labels = bquote(omega), col = 5, adj = -.5)

title(sub = bquote(Phi == .(phi_val) ~ "|" ~ bar("RUP") == .(rup_val) * '\%'~"|"~omega == .(w_val)))
}))
}
\references{
Hansen, J. A. (2013). Direct inversion of stress, strain or strain rate
including vorticity: A linear method of homogenous fault-slip data inversion
independent of adopted hypothesis. Journal of Structural Geology, 51, 3–13.
https://doi.org/10.1016/j.jsg.2013.03.014
}
\seealso{
Other stress-inversion: 
\code{\link{Fault_PT}()},
\code{\link{slip_inversion}()},
\code{\link{slip_inversion_angelier}()},
\code{\link{slip_inversion_michael}()},
\code{\link{slip_inversion_simple}()}
}
\concept{stress-inversion}
