% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion_hansen.R
\name{slip_inversion_hansen_boot}
\alias{slip_inversion_hansen_boot}
\title{9D Direct Inversion for Fault Slip Including Vorticity with Confidence Intervals}
\usage{
slip_inversion_hansen_boot(
  x,
  flip = FALSE,
  type = c("9d", "6d"),
  n_iter = 100L,
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{x}{object of class \code{"Pair"} or \code{"Fault"} with at least 7 rows.}

\item{flip}{logical. Flip if you want to have the negative stress tensor, i.e.
sigma 1 and 3 will be flipped.}

\item{type}{character. Inversion method, either \code{"9d"} (the default) for using the
9-dimensional or \code{"6d"} for the 6-dimensional parameter space.}

\item{n_iter}{integer. Number of bootstrap replicates (100 by default)}

\item{conf.level}{numeric. Confidence level of the interval (0.95 by default)}

\item{...}{optional parameters passed to \code{\link[=confidence_ellipse]{confidence_ellipse()}}}
}
\value{
See \code{\link[=slip_inversion_hansen]{slip_inversion_hansen()}} and \code{\link[=slip_inversion_michael]{slip_inversion_michael()}}
}
\description{
Fault-slip inversion method after Hansen (2013) with bootstrapped confidence intervals
}
\examples{
set.seed(20250411)
res <- slip_inversion_hansen_boot(osmundsen2010, n_iter = 100, n = 1000, res = 100)

# some stress shape
phi_val <- round(res$phi_CI, 2)

# vorticity
w_val <- round(res$vorticity_mag_CI, 2)

# Plot the faults
plot(osmundsen2010, col = "grey", lwd = 0.1, cex = 0.5)
stereo_confidence(res$principal_axes_CI$sigma1, col = 2)
stereo_confidence(res$principal_axes_CI$sigma2, col = 3)
stereo_confidence(res$principal_axes_CI$sigma3, col = 4)
stereo_confidence(res$vorticity_axis_CI, col = 5)
text(res$principal_axes, label = rownames(res$principal_axes), col = 2:4, adj = -.25)
text(res$vorticity_axis, labels = bquote(omega), col = 5, adj = -.5)
title(
  main = "Lofoten / Northern Norway\n(Osmundsen et al. 2010)",
  sub = bquote(atop(
    varphi ~ "(95\% CI)" == "[" * .(phi_val[1]) * "," ~ .(phi_val[2]) * "]",
    ~omega ~ "(95\%)" == "[" * .(w_val[1]) * "," ~ .(w_val[2]) * "]"
  ))
)

dev.off()
}
\seealso{
Other hansen: 
\code{\link{slip_inversion_hansen}()}
}
\concept{hansen}
