% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion.R
\name{slip_inversion_michael}
\alias{slip_inversion_michael}
\title{Stress Inversion for Fault-Slip Data after Michael (1984)}
\usage{
slip_inversion_michael(x, n_iter = 100L, conf.level = 0.95, flip = FALSE, ...)
}
\arguments{
\item{x}{\code{"Fault"} object where the rows are the observations, and the
columns the coordinates. Object must be complete, i.e. no \code{NA} values.
For Michael's, Angelier's, and Yamaji-Sato's methods, at least 4 rows
of fault measurements are required, while Hansen's method requires at least 7.}

\item{n_iter}{integer. Number of bootstrap replicates (100 by default)}

\item{conf.level}{numeric. Confidence level of the interval (0.95 by default)}

\item{flip}{logical. Flip if you want to have the negative stress tensor, i.e. sigma 1 and 3 will be flipped.}

\item{...}{optional parameters passed to \code{\link[=confidence_ellipse]{confidence_ellipse()}}}
}
\value{
Additionally, this child functions appends the following list components:
\describe{
\item{\code{principal_axes_CI}}{list containing the confidesnce ellipses for the 3 principal stress vectors. See \code{\link[=confidence_ellipse]{confidence_ellipse()}} for details.}
\item{\code{principal_vals_CI}}{3-column vector containing the lower and upper margins of the confidence interval of the principal vals}
\item{\code{SHmax_CI}}{numeric. Confidence interval of \code{SHmax} angle}
\item{\code{R_CI},\code{phi_CI},\code{bott_CI}}{Confidence interval for \code{R}}
\item{\code{alpha_CI},\code{beta_CI},\code{theta_CI}}{numeric. Confidence intervals of  \code{alpha}, \code{beta}, and \code{theta} angles}
}
}
\description{
Direct stress inversion (based on Michael, 1984) determines the orientation
of the principal stresses from fault slip data.
Confidence intervals are estimated by bootstrapping.
This inversion is simplified by the assumption that the magnitude of the
tangential traction on the various fault planes, at the time of rupture, is similar.
}
\details{
The goal of slip inversion is to find the single uniform stress tensor that
most likely caused the faulting events. With only slip data to constrain the
stress tensor the isotropic component can not be determined, unless
assumptions about the fracture criterion are made. Hence inversion will be
for the deviatoric stress tensor only.
A single fault can not completely constrain the deviatoric stress tensor a,
therefore it is necessary to simultaneously solve for a number of faults, so
that a single a that best satisfies all of the faults is found.
}
\examples{
set.seed(20250411)

# Use Angelier examples:
nx <- length(angelier1990)
par(mfrow = c(2, nx/2))

invisible(lapply(seq_len(nx), function(i) {
  # inversion
  x <- angelier1990[[i]]
  res <- slip_inversion_michael(x, n_iter = 100, n = 1000, res = 100)

  # some stress shape
  phi_val <- round(res$phi_CI, 2)

  # misfit
  rup_val <- round(res$rup_CI, 2)

  # Plot the faults (color-coded by RUP\%) and show the principal stress axes
  stereoplot(guides = FALSE)
  stereo_shmax(res$SHmax)
  fault_plot(x, col = assign_col(res$misfit$rup))
  stereo_confidence(res$principal_axes_CI$sigma1, col = 2)
  stereo_confidence(res$principal_axes_CI$sigma2, col = 3)
  stereo_confidence(res$principal_axes_CI$sigma3, col = 4)
  text(res$principal_axes, label = rownames(res$principal_axes), col = 2:4, adj = -.25)
  legend("topleft", col = 2:4, legend = rownames(res$principal_axes), pch = 16)
  title(
  main = names(angelier1990)[i],
  sub = bquote(atop(varphi ~ "(95\% CI)" == "[" * .(phi_val[1]) * "," ~ .(phi_val[2]) * "]",
  ~ bar("RUP") ~ "(95\% CI)" == "[" * .(rup_val[1]) * "," ~ .(rup_val[2]) * "] \%")
  ))
}))
dev.off()
}
\references{
Michael, A. J. (1984). Determination of stress from slip data:
Faults and folds. Journal of Geophysical Research: Solid Earth, 89(B13), 11517–11526.
\doi{10.1029/JB089iB13p11517}
}
\seealso{
\code{\link[=Fault_PT]{Fault_PT()}} for a simple P-T stress analysis,
\code{\link[=SH]{SH()}} and \code{\link[=SH_from_tensor]{SH_from_tensor()}} to calculate the azimuth of the maximum horizontal stress;
\code{\link[=Mohr_plot]{Mohr_plot()}} for graphical representation of the deviatoric stress tensor.

Other stress-inversion: 
\code{\link{Fault_PT}()},
\code{\link{slip_inversion}()},
\code{\link{slip_inversion_angelier}()},
\code{\link{slip_inversion_hansen}()},
\code{\link{slip_inversion_simple}()},
\code{\link{slip_inversion_wissi}()},
\code{\link{slip_inversion_yamaji_sato}()}
}
\concept{stress-inversion}
