% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion.R
\name{slip_inversion_misfit}
\alias{slip_inversion_misfit}
\title{Misfit parameters of slip inversion}
\usage{
slip_inversion_misfit(tau, fault)
}
\arguments{
\item{tau}{symmetric 3x3 matrix. The (reduced) stress tensor.}

\item{fault}{\code{"Fault"} object where the rows are the observations, and the columns the coordinates.}
}
\value{
list. \describe{
\item{\code{"alpha"}}{numeric. Angelier (1990)'s angles between the shear stress
vector and the slip vector, ranging from 0&degree; (perfect fit) to 180° (inconsistent fit).}
\item{\code{"beta"}}{numeric. Michael (1984)'s angles between the tangential
traction predicted by the best stress tensor and the slip vector on each plane, ranging from 0 to 90°.}
\item{\code{"theta"}}{numeric. Angle between slip planes and \eqn{\sigma_1} ranging from 0 to 180°.}
\item{\code{"rup"}}{numeric. "Ratio Upsilon" (RUP) parameter after Angelier (1990), ranging frm 0 (perfect fit) to 200\% (misfit). See \code{\link[=tau2rup]{tau2rup()}}.}
\item{\code{"quality"}}{factor. Ranked misfit classification based on RUP values. See \code{\link[=tau2rup]{tau2rup()}}.}
\item{\code{"quality_summary"}}{integer. Counts of faults in the RUP-based quality ranks.}
\item{\code{"misfit_means"}}{Mean values of \code{alpha}, \code{beta}, \code{theta}, and \code{rup.}}
}
}
\description{
Misfit parameters of slip inversion
}
\examples{
f <- angelier1990$TYM
tau <- reduced_stress(f)
slip_inversion_misfit(tau, f) 
}
