% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion.R
\name{slip_inversion_misfit}
\alias{slip_inversion_misfit}
\title{Misfit parameters of slip inversion}
\usage{
slip_inversion_misfit(tau, fault)
}
\arguments{
\item{tau}{symmetric 3x3 matrix. The (reduced) stress tensor.}

\item{fault}{\code{"Fault"} object where the rows are the observations, and the columns the coordinates.}
}
\value{
list. \describe{
\item{\code{alpha}}{numeric. Deviation angle between slickenline and shear traction as a line
(0-90&degree). This is the standard misfit used in Angelier (1990) because
field slickenlines are geometrically lines, not vectors. Ranging from 0&degree;
(perfect fit) to 90° (inconsistent fit).}
\item{\code{alpha_mean}}{numeric. The mean of \code{alpha}, the ie. the mean deviation
of predicted from observed slip.}
\item{\code{rup}}{numeric. "Ratio Upsilon" (RUP) parameter after Angelier (1990), ranging frm 0 (perfect fit) to 200\% (misfit). See \code{\link[=tau2rup]{tau2rup()}}.}
\item{\code{quality}}{factor. Ranked misfit classification based on RUP values. See \code{\link[=tau2rup]{tau2rup()}}.}
\item{\code{rup_mean}}{numeric. The mean RUP.}
\item{\code{quality_summary}}{integer. Counts of faults in the RUP-based quality ranks.}
\item{\code{flipped}}{logical. Are the signs of slip vectors flipped, i.e.,
dot product of the slip ray and the predicted shear traction is negative?}
}
}
\description{
Misfit parameters of slip inversion
}
\examples{
f <- angelier1990$TYM
tau <- reduced_stress(f)
slip_inversion_misfit(tau, f)
}
