% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion.R
\name{slip_inversion_simple}
\alias{slip_inversion_simple}
\title{Simple Statistical Fault-Slip Inversion}
\usage{
slip_inversion_simple(x, cluster_fun = stats::kmeans, n_grid = 1000L)
}
\arguments{
\item{x}{object of class \code{"Fault"}}

\item{cluster_fun}{function for cluster, must have number of desired cluster
as second input outputs and vector \code{cluster}. The default is \code{\link[stats:kmeans]{stats::kmeans()}}}

\item{n_grid}{integer. Number to optimize grid search for stress shape parameter R}
}
\value{
list.
}
\description{
Measurements of fault-slip data are often scattered due to measurement
errors and the wavy nature of fault planes and fault striations/slickenlines.
The fault scatter is large due to noise, rather than representing the actual
geometry of
the fault set. The idea of this algorithm is to cluster the fault data set to
identify the conjugate set of
faults and their the mean orientation Using the Wallace-Bott Hypothesis
and Anderson's theory, it then calculates the orientation of the principal stresses,
and uses the angles to the fault planes to derive the a best-fit stress shape parameter R.
}
\examples{
set.seed(20250411)

nx <- length(angelier1990)
par(mfrow = c(2, nx/2))

invisible(lapply(seq_len(nx), function(i) {
  x <- angelier1990[[i]]
  xres <- slip_inversion_simple(x)
  
  stereoplot(title = names(angelier1990)[i], guides = FALSE,
  sub = paste0(
    "beta: ", round(xres$beta, 2),
    " deg | R: ", round(xres$R, 2)
  ))
  hoeppener(x, col = assign_col(xres$beta_angles))
  angelier(xres$mean_planes, pch = 16, col = viridis::magma(2, end = 0.8), cex = 1)
  points(xres$principal_axes, pch = 16, col = viridis::rocket(3, end = 0.8), cex = 1)
  text(xres$principal_axes,
    labels = rownames(xres$principal_axes),
    col = viridis::rocket(3, end = 0.8), cex = 1, adj = c(-.25, -.25)
  )
}))

dev.off()
}
\seealso{
Other stress-inversion: 
\code{\link{Fault_PT}()},
\code{\link{slip_inversion}()},
\code{\link{slip_inversion_angelier}()},
\code{\link{slip_inversion_hansen}()},
\code{\link{slip_inversion_michael}()},
\code{\link{slip_inversion_wissi}()},
\code{\link{slip_inversion_yamaji_sato}()}
}
\concept{stress-inversion}
