% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stats.R
\name{sph-dist}
\alias{sph-dist}
\alias{dist.Vec3}
\alias{dist.Line}
\alias{dist.Ray}
\alias{dist.Plane}
\alias{dist.Pair}
\title{Angular distance matrix for orientation vectors}
\usage{
\method{dist}{Vec3}(x, ...)

\method{dist}{Line}(x, ...)

\method{dist}{Ray}(x, ...)

\method{dist}{Plane}(x, ...)

\method{dist}{Pair}(x, ...)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"},
where the rows are the observations and the columns are the coordinates.}

\item{...}{optional parameters passed to \code{\link[stats:dist]{stats::as.dist()}}}
}
\value{
distance matrix
}
\description{
This function computes and returns the distance matrix computed by using the
Cosine similarity to compute the distances between the rows of a data matrix.
}
\examples{
set.seed(20250411)
dist(rvmf(100, mu = Line(90, 0), k = 20))
}
