% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{split_trailing_letters}
\alias{split_trailing_letters}
\title{Parse measurement and direction strings}
\usage{
split_trailing_letters(x)
}
\arguments{
\item{x}{character or number}
}
\value{
list
}
\description{
Parse measurement and direction strings
}
\examples{
test <- c("45NW", "4SE")
split_trailing_letters(test)
}
\seealso{
Other parse-orientations: 
\code{\link{Fault_from_rake_quadrant}()},
\code{\link{azimuth_to_cardinal}()},
\code{\link{quadrant2dd}()},
\code{\link{sense_from_strikeslip}()},
\code{\link{strikeslip_kinematics}()}
}
\concept{parse-orientations}
