% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stats.R
\name{stats}
\alias{stats}
\alias{mean.spherical}
\alias{sd}
\alias{sd.spherical}
\alias{var}
\alias{var.spherical}
\alias{confidence_angle}
\alias{rdegree}
\alias{sd_error}
\alias{delta}
\alias{estimate_k}
\title{Statistical estimators of the distribution of a set of vectors}
\usage{
\method{mean}{spherical}(x, w = NULL, na.rm = TRUE)

sd(x, na.rm = TRUE, ...)

\method{sd}{spherical}(x, w = NULL, na.rm = TRUE)

var(x, na.rm = TRUE, ...)

\method{var}{spherical}(x, w = NULL, na.rm = TRUE)

confidence_angle(x, w = NULL, alpha = 0.05, na.rm = TRUE)

rdegree(x, w = NULL, na.rm = FALSE)

sd_error(x, w = NULL, na.rm = FALSE)

delta(x, w = NULL, na.rm = TRUE)

estimate_k(x, w = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, or \code{"Plane"}.}

\item{w}{numeric. Optional weights for each observation.}

\item{na.rm}{logical. Whether \code{NA} values should be removed before the computation proceeds.}

\item{...}{arguments passed to function call}

\item{alpha}{numeric. Significance level for the confidence angle (default is 0.05 for a 95\% confidence angle).}
}
\description{
Statistical estimators of the distribution of a set of vectors
}
\details{
\code{mean} returns the spherical mean of a set of vectors
(object of class of \code{x}).

\code{var} returns the spherical variance (numeric), based on resultant length
(Mardia 1972).

\code{delta} returns the cone angle containing ~63\% of the data (in degree if \code{x} is a \code{"Plane"} or \code{"Line"}, or in radians
if otherwise). For enough large sample it approaches the angular standard
deviation (\code{"csd"}) of the Fisher statistics.

\code{rdegree} returns the degree of preferred orientation of vectors, range: (0, 1).

\code{sd_error} returns the spherical standard error (numeric). If the number of
data is less than 25, if will print a additional message, that the output
value might not be a good estimator.

\code{confidence_angle} returns the semi-vertical angle \eqn{q} about the
mean \eqn{\mu} (in degree if \code{x} is a \code{"Plane"} or \code{"Line"}, or in radians
if otherwise). The \eqn{100(1-\alpha)\%} confidence interval is than given by \eqn{\mu \pm q}.

\code{estimate_k} returns the estimated concentration of the von Mises-Fisher distribution \eqn{\kappa} (after Sra, 2011).
}
\examples{
set.seed(20250411)
x <- rvmf(100, mu = Line(120, 50), k = 5)
mean(x)
var(x)
delta(x)
rdegree(x)
sd_error(x)
confidence_angle(x)
estimate_k(x)

#' weights:
x2 <- Line(c(0, 0), c(0, 90))
mean(x2)
mean(x2, w = c(1, 2))
var(x2)
var(x2, w = c(1, 2))
}
