% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stats.R
\name{stats}
\alias{stats}
\alias{sph_mean}
\alias{sph_sd}
\alias{sph_var}
\alias{sph_confidence_angle}
\alias{rdegree}
\alias{sd_error}
\alias{delta}
\alias{estimate_k}
\title{Statistical estimators of the distribution of a set of vectors}
\usage{
sph_mean(x, na.rm = TRUE, ...)

sph_sd(x, ...)

sph_var(x, ...)

sph_confidence_angle(x, w = NULL, alpha = 0.05, na.rm = TRUE)

rdegree(x, w = NULL, na.rm = FALSE)

sd_error(x, w = NULL, na.rm = FALSE)

delta(x, w = NULL, na.rm = TRUE)

estimate_k(x, w = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}.}

\item{na.rm}{logical. Whether \code{NA} values should be removed before the computation proceeds.}

\item{...}{arguments passed to function call}

\item{w}{numeric. Optional weights for each observation.}

\item{alpha}{numeric. Significance level for the confidence angle (default is 0.05 for a 95\% confidence angle).}
}
\description{
Statistical estimators of the distribution of a set of vectors
}
\details{
\code{sph_mean} returns the spherical arithmetic mean of a set of vectors
(object of class of \code{x}).

\code{sph_var} returns the spherical variance (numeric), based on resultant length
(Mardia 1972).

\code{sph_sd} returns the spherical standard deviation (numeric) given as the half
apical angle of a cone about the mean vector. In degrees if \code{x} is a
\code{"Plane"} or \code{"Line"}, or in radians if otherwise.

\code{delta} returns the half apical angle of the cone containing ~63\% of the data
(in degrees if \code{x} is a \code{"Plane"} or \code{"Line"}, or in radians
if otherwise). For enough large sample it approaches the angular standard
deviation (\code{"csd"}) of the Fisher statistics.

\code{rdegree} returns the degree of preferred orientation of vectors, range: (0, 1).

\code{sd_error} returns the spherical standard error (numeric). If the number of
data is less than 25, if will print a additional message, that the output
value might not be a good estimator.

\code{sph_confidence_angle} returns the half-apical angle \eqn{q} of a cone about the
mean \eqn{\mu} (in degrees if \code{x} is a \code{"Plane"} or \code{"Line"}, or in radians
if otherwise). The \eqn{100(1-\alpha)\%} confidence interval is than given by \eqn{\mu \pm q}.

\code{estimate_k} returns the estimated concentration of the von Mises-Fisher distribution \eqn{\kappa} (after Sra, 2011).
}
\examples{
set.seed(20250411)
x <- rvmf(100, mu = Line(120, 50), k = 5)
sph_mean(x)
sph_sd(x)
sph_var(x)
delta(x)
rdegree(x)
sd_error(x)
sph_confidence_angle(x)
estimate_k(x)

#' weights:
x2 <- Line(c(0, 0), c(0, 90))
sph_mean(x2)
sph_mean(x2, w = c(1, 2))
sph_var(x2)
sph_var(x2, w = c(1, 2))
}
\seealso{
\code{\link[=projected_mean]{projected_mean()}} for projected mean, \code{\link[=geodesic_mean_line]{geodesic_mean_line()}} and \code{\link[=geodesic_mean_pair]{geodesic_mean_pair()}} for the geodesic mean of lines and pairs, respectively.
or \code{\link[=geodesic_mean]{geodesic_mean()}} and \code{\link[=geodesic_var]{geodesic_var()}} as a convenience wrapper for all spherical data types.
}
