% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{stats}
\alias{stats}
\alias{v_mean}
\alias{v_var}
\alias{v_delta}
\alias{v_rdegree}
\alias{v_sde}
\alias{v_confidence_angle}
\alias{estimate_k}
\title{Statistical estimators of the distribution of a set of vectors}
\usage{
v_mean(x, w = NULL)

v_var(x, w = NULL)

v_delta(x, w = NULL)

v_rdegree(x)

v_sde(x)

v_confidence_angle(x, alpha = 0.05)

estimate_k(x)
}
\arguments{
\item{x}{numeric. Can be three element vector, three column array, or an
object of class \code{"line"} or \code{"plane"}}

\item{w}{weightings}

\item{alpha}{significance level (0.05 by default)}
}
\description{
Statistical estimators of the distribution of a set of vectors
}
\details{
\code{v_mean} returns the spherical mean of a set of vectors
(object of class of \code{x}). The function is shortcut for
\code{vresultant(x, mean = TRUE)} when the argument is a \code{"plane"} or \code{"line"} object.

\code{v_var} returns the spherical variance (numeric), based on resultant length
(Mardia 1972).

\code{v_delta} returns the cone angle containing ~63\% of the data (in degree if \code{x} is a \code{"plane"} or \code{"line"}, or in radians
if otherwise). For enough large sample it approaches the angular standard
deviation (\code{"csd"}) of the Fisher statistics.

\code{v_rdegree} returns the degree of preferred orientation of vectors (in \%).

\code{v_sde} returns the spherical standard error (numeric). If the number of
data is less than 25, if will print a additional message, that the output
value might not be a good estimator.

\code{v_confidence_angle} returns the semi-vertical angle \eqn{q} about the
mean \eqn{\mu} (in degree if \code{x} is a \code{"plane"} or \code{"line"}, or in radians
if otherwise). The \eqn{100(1-\alpha)\%} confidence interval is than given by \eqn{\mu \pm q}.
\code{estimate_k} returns the estimated concentration of the von Mises-Fisher distribution \eqn{\kappa} (after Sra, 2011).
}
\examples{
x <- rvmf(100, mu = Line(120, 50), k = 5)
v_mean(x)
v_var(x)
v_delta(x)
v_rdegree(x)
v_sde(x)
v_confidence_angle(x)
estimate_k(x)
}
\seealso{
\code{\link[=vresultant]{vresultant()}}, \code{\link[=fisher_statistics]{fisher_statistics()}}
}
