% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereo_arrows}
\alias{stereo_arrows}
\title{Add Arrows to a Stereoplot}
\usage{
stereo_arrows(
  x,
  sense,
  scale = 0.05,
  angle = 10,
  length = 0.1,
  upper.hem = FALSE,
  earea = TRUE,
  ...
)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, or \code{"Plane"}.}

\item{sense}{numeric. Sense of the line on a fault plane. Either
\code{1}or \code{-1} for normal or thrust offset, respectively. The "sense" is the sign
of the fault's rake (see \code{\link[=Fault_from_rake]{Fault_from_rake()}} for details).}

\item{scale}{numeric. Scales the length of the vector. \code{0.05} by default}

\item{angle}{numeric. Angle from the shaft of the arrow to the edge of the arrow head.}

\item{length}{numeric. Length of the edges of the arrow head (in inches).}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}, the default).}

\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection (the default), or \code{FALSE} for meridional stereographic projection.}

\item{...}{arguments passed to \code{\link[graphics:arrows]{graphics::arrows()}}}
}
\description{
A quiver plot displays displacement vectors into pointing into the direction of movement.
}
\examples{
set.seed(20250411)
stereoplot()
p <- rvmf(n = 100)
points(p, pch = 16, cex = .5)
stereo_arrows(p, sense = 1, col = "red")
}
\seealso{
\code{\link[=hoeppner]{hoeppner()}}, \code{\link[=angelier]{angelier()}}
}
