% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereo_cones}
\alias{stereo_cones}
\alias{stereo_smallcircle}
\alias{stereo_greatcircle}
\title{Stereographic Projection of Cones}
\usage{
stereo_smallcircle(
  x,
  d = 90,
  col = 1,
  N = 1000,
  upper.hem = FALSE,
  earea = TRUE,
  lty = 1,
  lwd = 1,
  BALL.radius = 1,
  ...
)

stereo_greatcircle(x, ...)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"},
where the rows are the observations and the columns are the coordinates.}

\item{d}{numeric. conical angle in degrees.}

\item{col, lty, lwd}{color, line type, and line width parameters}

\item{N}{integer. number of points to calculate}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}, the default).}

\item{earea}{logical \code{TRUE} for Lambert equal-area projection (also "Schmidt net"; the default), or
\code{FALSE} for meridional stereographic projection (also "Wulff net" or "Stereonet").}

\item{BALL.radius}{numeric size of sphere}

\item{...}{optional graphical parameters passed to \code{\link[graphics:lines]{graphics::lines()}}}
}
\description{
Visualization of smallcircles and greatcircles in a stereographic projection.
}
\examples{
stereoplot()
stereo_point(Line(90, 5), lab = "L")
stereo_smallcircle(Line(90, 5), d = 10)
stereo_point(Plane(120, 30), lab = "P", col = "red")
stereo_greatcircle(Plane(120, 30), col = "red")

stereoplot()
stereo_point(Line(c(129, 90), c(30, 5)), lab = c("L1", "L2"))
stereo_smallcircle(Line(c(129, 90), c(30, 5)), d = c(10, 5), col = 1:2, lty = 1:2, lwd = 1:2)
}
\seealso{
\code{\link[=lines.spherical]{lines.spherical()}}, \code{\link[=stereo_segment]{stereo_segment()}}, \code{\link[=stereo_lines]{stereo_lines()}}

Other stereo-plot: 
\code{\link{fault-plot}},
\code{\link{lines.spherical}()},
\code{\link{plot-spherical}},
\code{\link{points.spherical}()},
\code{\link{stereo_arrows}()},
\code{\link{stereo_confidence}()},
\code{\link{stereo_contour}},
\code{\link{stereo_lines}()},
\code{\link{stereo_point}()},
\code{\link{stereo_segment}()},
\code{\link{stereoplot}()},
\code{\link{stereoplot_guides}()},
\code{\link{stereoplot_ticks}()},
\code{\link{text.spherical}()}
}
\concept{stereo-plot}
