% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereo_confidence}
\alias{stereo_confidence}
\title{Plot Bootstrapped Confidence Ellipse}
\usage{
stereo_confidence(
  x,
  params = list(),
  .center = TRUE,
  col = par("col"),
  cex = par("cex"),
  pch = 16,
  upper.hem = NULL,
  earea = NULL,
  radius = NULL,
  ...
)
}
\arguments{
\item{x}{Spherical object or a list containing the output of an earlier call of \code{\link[=confidence_ellipse]{confidence_ellipse()}}}

\item{params}{list. Parameters passed to \code{\link[=confidence_ellipse]{confidence_ellipse()}}}

\item{.center}{logical. Whether the ellipse's center should be plotted?}

\item{col}{Color of the ellipse and its center}

\item{pch, cex}{Plotting symbol and size of the ellipse center. Ignored if \code{.center} is \code{FALSE}.}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}). Defaults to
\code{getOption("structr.upper.hem")}.}

\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection, or \code{FALSE} for meridional stereographic projection.
Defaults to \code{getOption("structr.earea")}.}

\item{radius}{numeric. Radius of circle. Defaults to \code{getOption("structr.radius")}.}

\item{...}{graphical parameters passed to \code{\link[graphics:lines]{graphics::lines()}}}
}
\value{
output of \code{\link[=confidence_ellipse]{confidence_ellipse()}}
}
\description{
Adds an ellipse marking the bootstrapped confidence interval of the arithmetic mean to an existing plot
}
\examples{
set.seed(20250411)
plot(example_lines, col = "grey")
stereo_confidence(example_lines, params = list(n = 100, res = 100), col = "red")
}
\seealso{
\code{\link[=confidence_ellipse]{confidence_ellipse()}}

Other stereo-plot: 
\code{\link{arrows}()},
\code{\link{fault-plot}},
\code{\link{lines}()},
\code{\link{plot-spherical}},
\code{\link{points.spherical}()},
\code{\link{stereo_cones}},
\code{\link{stereo_contour}},
\code{\link{stereo_lines}()},
\code{\link{stereo_point}()},
\code{\link{stereo_segment}()},
\code{\link{stereoplot}()},
\code{\link{stereoplot_guides}()},
\code{\link{stereoplot_ticks}()},
\code{\link{text.spherical}()}
}
\concept{stereo-plot}
