% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereo_confidence}
\alias{stereo_confidence}
\title{Plot the bootstrapped confidence ellipse}
\usage{
stereo_confidence(
  x,
  params = list(),
  col = par("col"),
  cex = par("cex"),
  pch = 16,
  center = TRUE,
  ...
)
}
\arguments{
\item{x}{Spherical object or a list containing the output of an earlier call of \code{\link[=confidence_ellipse]{confidence_ellipse()}}}

\item{params}{list. Parameters passed to \code{\link[=confidence_ellipse]{confidence_ellipse()}}}

\item{col}{Color of the ellipse and its center}

\item{pch, cex}{Plotting symbol and size of the ellipse center. Ignored if \code{center} is \code{FALSE}.}

\item{center}{logical. Whether the ellipse's center should be plotted?}

\item{...}{graphical parameters passed to \code{\link[graphics:lines]{graphics::lines()}}}
}
\value{
output of \code{\link[=confidence_ellipse]{confidence_ellipse()}}
}
\description{
Adds an ellipse marking the bootstrapped confidence interval of the arithmetic mean to an existing plot
}
\examples{
set.seed(20250411)
plot(example_lines, col = "grey")
stereo_confidence(example_lines, params = list(n = 100, res = 100), col = "red")
}
\seealso{
\code{\link[=confidence_ellipse]{confidence_ellipse()}}
}
