% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_contours.R
\name{stereo_contour}
\alias{stereo_contour}
\alias{contour.spherical}
\alias{filled.contour.spherical}
\alias{image.spherical}
\title{Plot spherical densities in a stereonet}
\usage{
\method{contour}{spherical}(
  x,
  add = FALSE,
  density.params = list(),
  nlevels = 10L,
  col.palette = viridis::viridis,
  col = NULL,
  col.params = list(),
  ...
)

\method{filled.contour}{spherical}(
  x,
  add = FALSE,
  density.params = list(),
  nlevels = 10L,
  col.palette = viridis::viridis,
  col.params = list()
)

\method{image}{spherical}(
  x,
  add = FALSE,
  density.params = list(),
  nlevels = 10L,
  col.palette = viridis::viridis,
  col.params = list(),
  ...
)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Plane"} or \code{'spherical.density'} (for plotting only).}

\item{add}{logical. Whether the contours should be added to an existing plot.}

\item{nlevels}{integer. Number of contour levels for plotting}

\item{col.palette}{a color palette function to be used to assign colors in the plot.}

\item{col}{colour(s) for the contour lines drawn. If \code{NULL}, lines are color based on \code{col.palette}.}

\item{col.params}{list. Arguments passed to \code{col.palette}}

\item{...}{optional parameters passed to \code{\link[=image.default]{image.default()}} or \code{\link[=contour.default]{contour.default()}}.}

\item{dens.params}{list of parameters passed to \link{density.spherical}}

\item{type}{character. Type of plot: \code{'contour'} for contour lines,
\code{'contour_filled'} for filled contours, or \code{'image'} for a raster image.}
}
\value{
list containing the stereographic x and coordinates of of the grid,
the counts, and the density.
}
\description{
Kamb counts and densities on the sphere
}
\examples{
set.seed(20250411)
x <- rfb(100, mu = Line(120, 10), k = 5, A = diag(c(-1, 0, 1)))

contour(x)

filled.contour(x,
  col.params = list(direction = -1, begin = .05, end = .95, alpha = .75)
)
stereo_point(x, col = "black", pch = 21)

image(x)
stereo_point(x, col = "lightgrey", pch = 21)


# complete example:
par(mfrow = c(2, 1))
wp <- 6 / ifelse(is.na(example_planes$quality), 6, example_planes$quality)
my_planes <- Plane(example_planes$dipdir, example_planes$dip)
fabric_p <- or_shape_params(my_planes)$Vollmer["D"]
my_planes_eig <- or_eigen(my_planes)

stereoplot(guides = TRUE, col = "grey96")
stereo_point(my_planes, col = "grey", pch = 16, cex = .5)
stereo_density(my_planes, type = "contour", add = TRUE, weights = wp)
stereo_point(as.plane(my_planes_eig$vectors[3, ]), col = "black", pch = 16)
stereo_greatcircle(as.plane(my_planes_eig$vectors[3, ]), col = "black", pch = 16)
title(
  main = "Planes",
  sub = paste0(
    "N: ", nrow(my_planes), " | Fabric strength: ", round(fabric_p, 2),
    "\nLambert equal area, lower hemisphere projection"
  )
)

my_lines <- Line(example_lines$trend, example_lines$plunge)
wl <- 6 / ifelse(is.na(example_lines$quality), 6, example_lines$quality)
fabric_l <- or_shape_params(my_lines)$Vollmer["D"]

stereoplot(guides = TRUE, col = "grey96")
stereo_point(my_lines, col = "grey", pch = 16, cex = .5)
stereo_density(my_lines, type = "contour", add = TRUE, weights = wl)
stereo_point(v_mean(my_lines, w = wl), col = "black", pch = 16)
title(
  main = "Lines",
  sub = paste0(
    "N: ", nrow(my_lines), " | Fabric strength: ", round(fabric_l, 2),
    "\nLambert equal area, lower hemisphere projection"
  )
)
}
\seealso{
\code{\link[=count_points]{count_points()}}
}
