% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet.R
\name{stereo_coords}
\alias{stereo_coords}
\title{Stereographic projection}
\usage{
stereo_coords(az, inc, upper.hem = FALSE)
}
\description{
Transformation of spherical coordinates into the stereographic projection
}
