% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contouring.R
\name{stereo_density}
\alias{stereo_density}
\alias{spherical_density}
\alias{projected_density}
\title{Calculate and plot densities in a stereonet}
\usage{
spherical_density(
  x,
  FUN = exponential_kamb,
  n = 128L,
  sigma = 3,
  weights = NULL,
  upper.hem = FALSE,
  r = 1
)

projected_density(
  x,
  n = 128L,
  sigma = 3,
  weights = NULL,
  upper.hem = FALSE,
  r = 1
)

stereo_density(
  x,
  ...,
  type = c("contour", "contour_filled", "image"),
  nlevels = 10L,
  col.palette = viridis,
  col = NULL,
  add = TRUE,
  col.params = list()
)
}
\arguments{
\item{x}{Object of class \code{"line"} or \code{"plane"} or \code{'spherical.density'} (for plotting only).}

\item{FUN}{density estimation function; one of \code{\link[=exponential_kamb]{exponential_kamb()}} (the default),
\link{kamb_count}, and \code{\link[=schmidt_count]{schmidt_count()}}.}

\item{n}{integer. Gridzise. 128 by default.}

\item{sigma}{numeric. Radius for Kamb circle used for counting. 3 by default.}

\item{weights}{(optional) numeric vector of length of \code{azi}.
The relative weight to be applied to each input measurement. The array
will be normalized to sum to 1, so absolute value of the \code{weights} do not
affect the result. Defaults to \code{NULL}}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}, the default).}

\item{r}{numeric. radius of stereonet circle}

\item{...}{optional parameters passed to \code{\link[=spherical_density]{spherical_density()}}.}

\item{type}{character. Type of plot: \code{'contour'} for contour lines,
\code{'contour_filled'} for filled contours, or \code{'image'} for a raster image.}

\item{nlevels}{integer. Number of contour levels for plotting}

\item{col.palette}{a color palette function to be used to assign colors in the plot.}

\item{col}{colour(s) for the contour lines drawn. If \code{NULL}, lines are color based on \code{col.palette}.}

\item{add}{logical. Whether the contours should be added to an existing plot.}

\item{col.params}{list. Arguments passed to \code{col.palette}}
}
\value{
list containing the stereographic x and coordinates of of the grid,
the counts, and the density.
}
\description{
Linear Kamb counts and densities on the sphere
}
\examples{
set.seed(20250411)
test <- rfb(100, mu = Line(120, 10), k = 5, A = diag(c(-1, 0, 1)))
test_densities <- spherical_density(x = test, n = 100, sigma = 3, weights = runif(100))

stereo_density(test_densities, type = "image", add = FALSE)
stereo_point(test, col = "lightgrey", pch = 21)

stereoplot(guides = FALSE)
stereo_point(test, col = "lightgrey", pch = 19)
stereo_density(test_densities, type = "contour", add = TRUE)

stereo_density(test_densities, type = "contour_filled", add = FALSE)
stereo_point(test, col = "lightgrey", pch = 21)
}
\seealso{
\code{\link[=count_points]{count_points()}}
}
