% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet.R
\name{stereo_fault}
\alias{stereo_fault}
\title{Stereographic projection of faults}
\usage{
stereo_fault(
  x,
  hoeppner = FALSE,
  greatcirles = TRUE,
  pch = 21,
  col = 1,
  lwd = 1,
  lty = 1,
  lab = NULL,
  cex = 0.75,
  text.pos = 4,
  upper.hem = FALSE,
  ...
)
}
\arguments{
\item{x}{Object of class \code{"fault"}}

\item{hoeppner}{logical. \code{TRUE} for Hoeppner plot}

\item{greatcirles}{logical. Whether greatcircles are displayed (\code{TRUE}, the default) or poles to planes (\code{FALSE})}

\item{pch, col, lwd, lty}{lotting parameters for planes and lines}

\item{lab}{character. text labels}

\item{cex}{character expansion of labels}

\item{text.pos}{position for labels}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}, the default).}

\item{...}{optional graphical parameters}
}
\description{
Visualization of faults (planes and lines) in a stereographic projection.
}
\note{
\code{"plane"} objects will be displayed as pole to the plane.
}
\examples{
\dontrun{
faults <- Fault(
  c(0, 90, 180, 270),
  c(80, 45, 80, 45),
  c(0, 180, 180, 315),
  c(80, 5, 80, 36),
  c(1, -1, 1, 0)
)
stereoplot()
stereo_fault(faults, col = 1:4)
}
}
