% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereo_lines}
\alias{stereo_lines}
\title{Lines in a Stereoplot}
\usage{
stereo_lines(x, upper.hem = NULL, earea = NULL, radius = NULL, ...)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"},
where the rows are the observations and the columns are the coordinates.}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}). Defaults to
\code{getOption("structr.upper.hem")}.}

\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection, or \code{FALSE} for meridional stereographic projection.
Defaults to \code{getOption("structr.earea")}.}

\item{radius}{numeric. Radius of circle. Defaults to \code{getOption("structr.radius")}.}

\item{...}{optional graphical parameters passed to \code{\link[graphics:lines]{graphics::lines()}} and (if \code{fill=TRUE}) \code{\link[graphics:polygon]{graphics::polygon()}}}
}
\value{
two-column matrix of the stereographic or equal-area coordinates
}
\description{
Draws simple lines between vector points in stereographic or equal-area projection
}
\examples{
plot(example_lines, col = "grey")
stereo_lines(example_lines[1:2, ], col = "red")
}
\seealso{
\code{\link[=slerp]{slerp()}}, \link{stereo_greatcircle}, \code{\link[=stereo_lines]{stereo_lines()}}, \code{\link[=stereo_segment]{stereo_segment()}}

Other stereo-plot:
\code{\link{arrows}},
\code{\link{fault-plot}},
\code{\link{lines}},
\code{\link{plot-spherical}},
\code{\link[=points.spherical]{points.spherical()}},
\code{\link{stereo_cones}},
\code{\link[=stereo_confidence]{stereo_confidence()}},
\code{\link{stereo_contour}},
\code{\link[=stereo_point]{stereo_point()}},
\code{\link[=stereo_segment]{stereo_segment()}},
\code{\link[=stereoplot]{stereoplot()}},
\code{\link[=stereoplot_guides]{stereoplot_guides()}},
\code{\link[=stereoplot_ticks]{stereoplot_ticks()}},
\code{\link[=text.spherical]{text.spherical()}}
}
\concept{stereo-plot}
