% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereo_lines}
\alias{stereo_lines}
\title{Lines in a Stereoplot}
\usage{
stereo_lines(x, upper.hem = FALSE, earea = TRUE, BALL.radius = 1, ...)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"},
where the rows are the observations and the columns are the coordinates.}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}, the default).}

\item{earea}{logical \code{TRUE} for Lambert equal-area projection (also "Schmidt net"; the default), or
\code{FALSE} for meridional stereographic projection (also "Wulff net" or "Stereonet").}

\item{BALL.radius}{numeric size of sphere}

\item{...}{optional graphical parameters passed to \code{\link[graphics:lines]{graphics::lines()}}}
}
\value{
two-column matrix of the stereographic or equal-area coordinates
}
\description{
Draws simple lines between vector points in stereographic or equal-area projection
}
\examples{
plot(example_lines, col = "grey")
stereo_lines(example_lines[1:2, ], col = "red")
}
\seealso{
\code{\link[=slerp]{slerp()}}, \link{stereo_greatcircle}, \code{\link[=stereo_lines]{stereo_lines()}}, \code{\link[=stereo_segment]{stereo_segment()}}

Other stereo-plot: 
\code{\link{fault-plot}},
\code{\link{lines.spherical}()},
\code{\link{plot-spherical}},
\code{\link{points.spherical}()},
\code{\link{stereo_arrows}()},
\code{\link{stereo_cones}},
\code{\link{stereo_confidence}()},
\code{\link{stereo_contour}},
\code{\link{stereo_point}()},
\code{\link{stereo_segment}()},
\code{\link{stereoplot}()},
\code{\link{stereoplot_guides}()},
\code{\link{stereoplot_ticks}()},
\code{\link{text.spherical}()}
}
\concept{stereo-plot}
