% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet.R
\name{stereo_point}
\alias{stereo_point}
\title{Plot lines and planes}
\usage{
stereo_point(
  azi,
  inc,
  plane = FALSE,
  upper.hem = FALSE,
  col = 1,
  pch = 20,
  lab = NULL,
  text.pos = 4,
  cex = 0.75,
  ...
)
}
\arguments{
\item{azi, inc}{numeric. Angles of azimuth (or dip direction) and inclination (dip) in degrees.}

\item{plane}{logical. Whether the object is a linear (`FALSE`) of planar feature (`TRUE`).}

\item{upper.hem}{logical. Whether the projection is shown for upper hemisphere (`TRUE`) or lower hemisphere (`FALSE`, the default).}

\item{col}{color}

\item{pch}{plotting character}

\item{lab}{character. text labels}

\item{text.pos}{position for labels}

\item{cex}{character expansion of labels}

\item{...}{optional graphical parameters}
}
\description{
Visualization of lines, planes in a stereographic projection.
}
\examples{
stereoplot()
stereo_point(90, 80, lab = "L")
stereo_point(120, 30, plane = TRUE, lab = "P", col = "red")
}
