% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereo_point}
\alias{stereo_point}
\title{Stereographic Projection of Lines and Planes}
\usage{
stereo_point(
  x,
  col = 1,
  pch = 20,
  lab = NULL,
  text.pos = 4,
  cex = 1,
  upper.hem = FALSE,
  earea = TRUE,
  ...
)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"},
where the rows are the observations and the columns are the coordinates.}

\item{col}{color}

\item{pch}{plotting character}

\item{lab}{character. text labels}

\item{text.pos}{position for labels}

\item{cex}{character expansion of labels}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}, the default).}

\item{earea}{logical \code{TRUE} for Lambert equal-area projection (also "Schmidt net"; the default), or
\code{FALSE} for meridional stereographic projection (also "Wulff net" or "Stereonet").}

\item{...}{optional graphical parameters}
}
\value{
two-column matrix of the stereographic or equal-area coordinates
}
\description{
Visualization of lines, planes in a stereographic projection.
}
\note{
\code{"Plane"} and \code{"Fault"} objects will be displayed as pole to the plane (only).
}
\examples{
stereoplot()
stereo_point(Line(c(90, 80), c(10, 75)), lab = c("L1", "L2"))
stereo_point(Plane(120, 30), lab = "P", col = "red")
}
\seealso{
Other stereo-plot: 
\code{\link{fault-plot}},
\code{\link{lines.spherical}()},
\code{\link{plot-spherical}},
\code{\link{points.spherical}()},
\code{\link{stereo_arrows}()},
\code{\link{stereo_cones}},
\code{\link{stereo_confidence}()},
\code{\link{stereo_contour}},
\code{\link{stereo_lines}()},
\code{\link{stereo_segment}()},
\code{\link{stereoplot}()},
\code{\link{stereoplot_guides}()},
\code{\link{stereoplot_ticks}()},
\code{\link{text.spherical}()}
}
\concept{stereo-plot}
