% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet.R
\name{stereo_point}
\alias{stereo_point}
\title{Stereographic projection of lines and planes}
\usage{
stereo_point(
  x,
  col = 1,
  pch = 20,
  lab = NULL,
  text.pos = 4,
  cex = 0.75,
  upper.hem = FALSE,
  ...
)
}
\arguments{
\item{x}{Object of class \code{"line"} or \code{"plane"}}

\item{col}{color}

\item{pch}{plotting character}

\item{lab}{character. text labels}

\item{text.pos}{position for labels}

\item{cex}{character expansion of labels}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}, the default).}

\item{...}{optional graphical parameters}
}
\description{
Visualization of lines, planes in a stereographic projection.
}
\note{
\code{"plane"} and \code{"fault"} objects will be displayed as pole to the plane (only).
}
\examples{
stereoplot()
stereo_point(Line(azi = c(90, 80), inc = c(10, 75)), lab = c("L", "L"))
stereo_point(Plane(120, 30), lab = "P", col = "red")
}
