% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereo_segment}
\alias{stereo_segment}
\title{Great-circle segment between two vectors}
\usage{
stereo_segment(
  x,
  y,
  upper.hem = FALSE,
  earea = TRUE,
  n = 100L,
  BALL.radius = 1,
  ...
)
}
\arguments{
\item{x, y}{objects of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}, the default).}

\item{earea}{logical \code{TRUE} for Lambert equal-area projection (also "Schmidt net"; the default), or
\code{FALSE} for meridional stereographic projection (also "Wulff net" or "Stereonet").}

\item{n}{integer. number of points along greatcircle (100 by default)}

\item{BALL.radius}{numeric size of sphere}

\item{...}{graphical parameters passed to \code{\link[graphics:lines]{graphics::lines()}}}
}
\description{
Plots the great-circle segment between two vectors
}
\examples{
x <- Line(120, 7)
y <- Line(10, 13)
plot(rbind(x, y))
stereo_segment(x, y, col = "red")

# For multiple segments use lapply():
set.seed(20250411)
mu <- Line(45, 10)
x <- rvmf(100, mu = mu)
plot(x)
lapply(seq_len(nrow(x)), FUN = function(i) {
  stereo_segment(x[i, ], mu, col = i)
})
points(mu, pch = 16, col = "white")
}
\seealso{
\code{\link[=slerp]{slerp()}}, \link{stereo_greatcircle}
}
