% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet.R
\name{stereo_smallcircle}
\alias{stereo_smallcircle}
\title{Stereographic projection of cones}
\usage{
stereo_smallcircle(
  x,
  d = 90,
  col = 1,
  N = 100,
  BALL.radius = 0.25,
  upper.hem = FALSE,
  lty = 1,
  lwd = 1,
  ...
)
}
\arguments{
\item{x}{Object of class \code{"line"} or \code{"plane"}}

\item{d}{numeric. conical angle in degrees.}

\item{col, lty, lwd}{color, line type, and line width parameters}

\item{N}{integer. number of points to calculate}

\item{BALL.radius}{numeric size of sphere}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}, the default).}

\item{...}{optional graphical parameters}
}
\description{
Visualization of smallcircles and greatcircles in a stereographic projection.
}
\examples{
stereoplot()
stereo_point(Line(90, 5), lab = "L")
stereo_smallcircle(Line(90, 5), d = 10)
stereo_point(Plane(120, 30), lab = "P", col = "red")
stereo_greatcircle(Plane(120, 30), col = "red", N = 1000)
}
