% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet.R
\name{stereo_smallcircle}
\alias{stereo_smallcircle}
\title{Plot cones}
\usage{
stereo_smallcircle(azi, inc, d = 90, col = 1, N = 100, BALL.radius = 0.25, ...)
}
\arguments{
\item{azi, inc}{numeric. Angles of azimuth (or dip direction) and inclination (dip) in degrees.}

\item{d}{numeric. conical angle in degrees.}

\item{col}{color}

\item{N}{integer. number of points to calculate}

\item{BALL.radius}{numeric size of sphere}

\item{...}{optional graphical parameters}
}
\description{
Visualization of smallcircles and greatcircles in a stereographic projection.
}
\examples{
stereoplot()
stereo_point(90, 80, lab = "L")
stereo_smallcircle(90, 80, d = 10)
stereo_point(120, 30, plane = TRUE, lab = "P", col = "red")
stereo_greatcircle(120, 30, col = "red", N = 1000)
}
