% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereoplot}
\alias{stereoplot}
\title{Stereographic and Equal Area Projection}
\source{
Adapted from the \code{RFOC} package
}
\usage{
stereoplot(
  earea = TRUE,
  guides = TRUE,
  d = 10,
  col = "gray90",
  lwd = 0.5,
  lty = 1,
  border.col = "black",
  title = NULL,
  sub = NULL,
  origin.text = "N",
  labels = FALSE,
  ladj = 0.05,
  centercross = TRUE,
  ticks = 90,
  radius = 1,
  center = NULL
)
}
\arguments{
\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection (the default), or \code{FALSE} for meridional stereographic projection.}

\item{guides}{logical. Whether guides should be added to the plot (\code{TRUE} by default)}

\item{d}{integer. Angle distance between guides. Default: 10}

\item{col}{Color of guide lines}

\item{lwd}{Width of guide lines}

\item{lty}{Type of guide lines}

\item{border.col}{color of primitive circle (frame), center-cross and ticks of the stereo plot}

\item{title, sub}{character. Title and subtitle of plot}

\item{origin.text}{character. Text at origin of plot}

\item{labels}{this can either be a logical value specifying whether (numerical)
annotations are to be made next to the tick marks, or a character or expression
vector of labels to be placed next to the tick points.}

\item{ladj}{adjustment for all labels away from origin of projection circle.
This essentially an amount that is added to \code{radius} and the length of the ticks.}

\item{centercross}{logical. Whether a center cross should be added (\code{TRUE} by default)}

\item{ticks}{integer. Angle between ticks. if \code{NULL} (the default), no ticks are drawn.}

\item{radius}{numeric. Radius of circle}

\item{center}{An object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}
specifying the center of the projection If \code{NULL} (the default), the center
is at the origin of the plot.}
}
\description{
Initialize the plot for equal-angle stereographic projections (Wulff) or Lambert
equal-area projections (Schmidt).
}
\examples{
stereoplot()

stereoplot(ticks = 30, title = "title", sub = "subtitle", border.col = "purple", labels = TRUE)

stereoplot(center = Line(120, 50))
}
\seealso{
Other stereo-plot: 
\code{\link{arrows}()},
\code{\link{fault-plot}},
\code{\link{lines}()},
\code{\link{plot-spherical}},
\code{\link{points.spherical}()},
\code{\link{stereo_cones}},
\code{\link{stereo_confidence}()},
\code{\link{stereo_contour}},
\code{\link{stereo_lines}()},
\code{\link{stereo_point}()},
\code{\link{stereo_segment}()},
\code{\link{stereoplot_guides}()},
\code{\link{stereoplot_ticks}()},
\code{\link{text.spherical}()}
}
\concept{stereo-plot}
