% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet.R
\name{stereoplot}
\alias{stereoplot}
\title{Initialize the Stereo plot}
\source{
[RFOC::net()]
}
\usage{
stereoplot(add = FALSE, col = gray(0.9), border = "black", lwd = 1, main = "N")
}
\arguments{
\item{add}{logical, `TRUE` adds to existing plot}

\item{col}{color of lines}

\item{border}{color of outer rim of stereo plot}

\item{lwd}{linewidth of lines description}
}
\description{
Plot equal-area stereographic projection, i.e. Lambert azimuthal Equal-Area projection (Schmidt).
}
\examples{
stereoplot()
}
