% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet.R
\name{stereoplot}
\alias{stereoplot}
\title{Stereographic projection}
\source{
\code{\link[RFOC:net]{RFOC::net()}}
}
\usage{
stereoplot(col = grDevices::gray(0.9), border = "black", lwd = 1, main = "N")
}
\arguments{
\item{col}{color of lines}

\item{border}{color of outer rim of stereo plot}

\item{lwd}{linewidth of lines description}

\item{main}{title of the plot (\code{"N"} by default)}
}
\description{
Initialize the plot for equal-area stereographic projections, i.e. Lambert
azimuthal Equal-Area projections (Schmidt).
}
\examples{
stereoplot()
}
