% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereoplot}
\alias{stereoplot}
\title{Stereographic projection}
\source{
Adapted from the \code{RFOC} package
}
\usage{
stereoplot(
  earea = TRUE,
  guides = TRUE,
  d = 10,
  col = "lightgray",
  lwd = 1,
  lty = 1,
  border.col = "black",
  title = NULL,
  sub = NULL,
  origin.text = "N",
  labels = FALSE,
  ladj = 0.04,
  centercross = TRUE,
  ticks = NULL,
  radius = 1
)
}
\arguments{
\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection (the default), or \code{FALSE} for meridional stereographic projection.}

\item{guides}{logical. Whether guides should be added to the plot (\code{TRUE} by default)}

\item{d}{integer. Angle distance between guides. Default: 10}

\item{col}{color of guide lines}

\item{lwd}{linewidth of guide lines}

\item{lty}{linetype of guide lines}

\item{border.col}{color of primitive circle (frame), center-cross and ticks of the stereo plot}

\item{title, sub}{character. Title and subtitle of plot}

\item{origin.text}{character. Text at origin of stereoplot.}

\item{labels}{this can either be a logical value specifying whether (numerical)
annotations are to be made next to the tickmarks, or a character or expression
vector of labels to be placed next to the tickpoints.}

\item{ladj}{adjustment for all labels away from origin of stereoplot circle.
This essentially an amount that is added to \code{radius} and the length of the ticks.}

\item{centercross}{logical. Whether a center cross should be added (\code{TRUE} by default)}

\item{ticks}{integer. Angle between ticks. if \code{NULL} (the default), no ticks are drawn.}

\item{radius}{numeric. Radius of circle}
}
\description{
Initialize the plot for equal-area stereographic projections (Wulff) or Lambert
Equal-Area projections (Schmidt).
}
\examples{
stereoplot(ticks = 30, title = "title", sub = "subtitle", border.col = "purple", labels = TRUE)
}
