% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet.R
\name{stereoplot}
\alias{stereoplot}
\title{Stereographic projection}
\source{
Adapted from the \code{RFOC} package
}
\usage{
stereoplot(
  earea = TRUE,
  guides = TRUE,
  d = 10,
  col = grDevices::gray(0.9),
  lwd = 1,
  lty = 1,
  border = "black",
  title = NULL,
  sub = NULL,
  centercross = TRUE,
  ticks = NULL
)
}
\arguments{
\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection (the default), or \code{FALSE} for meridional stereographic projection.}

\item{guides}{logical. Whether guides should be added to the plot (\code{TRUE} by default)}

\item{d}{integer. Angle distance between guides. Default: 10}

\item{col}{color of guide lines}

\item{lwd}{linewidth of guide lines}

\item{lty}{linetype of guide lines}

\item{border}{color of outer rim of stereo plot}

\item{title, sub}{character. Title and subtitle of the plot}

\item{centercross}{logical. Whether a center cross should be added (\code{TRUE} by default)}

\item{ticks}{integer. Angle between ticks. if \code{NULL} (the default), no ticks are drawn.}
}
\description{
Initialize the plot for equal-area stereographic projections (Wulff) or Lambert
Equal-Area projections (Schmidt).
}
\examples{
stereoplot(ticks = 45, title = "title", sub = "subtitle")
}
