% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereoplot_guides}
\alias{stereoplot_guides}
\title{Stereoplot Gridlines}
\usage{
stereoplot_guides(d = 10, earea = TRUE, radius = 1, center = NULL, ...)
}
\arguments{
\item{d}{angle between grid lines}

\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection (the default), or \code{FALSE} for meridional stereographic projection.}

\item{radius}{numeric. Radius of circle}

\item{center}{Center position of gridlines. If \code{NULL} (the default), gridlines are centered on the origin of the stereoplot. Otherwise, this should is an spherical object.}

\item{...}{optional arguments passed to \code{\link[graphics:lines]{graphics::lines()}}}
}
\description{
Adds equal-area or equal-angle projection gridlines to an existing stereoplot.
}
\examples{
plot(c(-1, 1), c(-1, 1), type = "n", asp = 1)
stereoplot_guides(d = 5, earea = FALSE, col = "green", rotation = 20)

plot(c(-1, 1), c(-1, 1), type = "n", asp = 1)
stereoplot_guides(d = 15, earea = TRUE, col = "orange", rotation = 90)

plot(c(-1, 1), c(-1, 1), type = "n", asp = 1)
stereoplot_guides(d = 15, earea = FALSE, center = Line(120, 50))
}
\seealso{
Other stereo-plot: 
\code{\link{arrows}()},
\code{\link{fault-plot}},
\code{\link{lines}()},
\code{\link{plot-spherical}},
\code{\link{points.spherical}()},
\code{\link{stereo_cones}},
\code{\link{stereo_confidence}()},
\code{\link{stereo_contour}},
\code{\link{stereo_lines}()},
\code{\link{stereo_point}()},
\code{\link{stereo_segment}()},
\code{\link{stereoplot}()},
\code{\link{stereoplot_ticks}()},
\code{\link{text.spherical}()}
}
\concept{stereo-plot}
