% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereoplot_guides}
\alias{stereoplot_guides}
\title{Stereoplot gridlines}
\usage{
stereoplot_guides(d = 10, earea = TRUE, radius = 1, ...)
}
\arguments{
\item{d}{angle between grid lines}

\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection (the default), or \code{FALSE} for meridional stereographic projection.}

\item{radius}{numeric. Radius of circle}

\item{...}{optional arguments passed to \code{\link[graphics:lines]{graphics::lines()}}}
}
\description{
Adds equal-area or equal-angle projection gridlines to an existing stereoplot.
}
\examples{
plot(c(-1, 1), c(-1, 1), type = "n", asp = 1)
stereoplot_guides(d = 5, earea = FALSE, col = "green", rotation = 20)

plot(c(-1, 1), c(-1, 1), type = "n", asp = 1)
stereoplot_guides(d = 15, earea = TRUE, col = "orange", rotation = 90)
}
\seealso{
\code{\link[=stereoplot]{stereoplot()}}, \code{\link[=stereoplot_frame]{stereoplot_frame()}}, \code{\link[=stereoplot_ticks]{stereoplot_ticks()}}
}
