% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{stereoplot_ticks}
\alias{stereoplot_ticks}
\title{Stereoplot Tick Marks}
\usage{
stereoplot_ticks(
  length = 0.02,
  angle = 10,
  labels = FALSE,
  ladj = 2 * length,
  radius = NULL,
  rotation = 0,
  cex = 0.8,
  ...
)
}
\arguments{
\item{length}{numeric. Length of ticks as fraction of \code{radius}}

\item{angle}{numeric. Division angle in degrees}

\item{labels}{this can either be a logical value specifying whether (numerical)
annotations are to be made next to the tick marks, or a character or expression
vector of labels to be placed next to the tick points.}

\item{ladj}{adjustment for all labels away from origin of projection circle.
This essentially an amount that is added to \code{radius} and the length of the ticks.}

\item{radius}{numeric. Radius of circle. Defaults to \code{getOption("structr.radius")}.}

\item{rotation}{numeric. Rotation (positive for counter-clockwise) of tick marks and labels}

\item{cex}{numeric. Character expansion controls the font size of thee labels.}

\item{...}{optional arguments passed to \code{\link[graphics:segments]{graphics::segments()}} and \code{\link[graphics:text]{graphics::text()}}}
}
\description{
Adds stereoplot trick marks to an existing plot
}
\examples{
plot(c(-1, 1), c(-1, 1), type = "n", asp = 1)
stereoplot_frame()
stereoplot_ticks(length = 0.05, angle = 45, col = "blue", lwd = 2, labels = TRUE)
}
\seealso{
Other stereo-plot: 
\code{\link{arrows}()},
\code{\link{fault-plot}},
\code{\link{lines}()},
\code{\link{plot-spherical}},
\code{\link{points.spherical}()},
\code{\link{stereo_cones}},
\code{\link{stereo_confidence}()},
\code{\link{stereo_contour}},
\code{\link{stereo_lines}()},
\code{\link{stereo_point}()},
\code{\link{stereo_segment}()},
\code{\link{stereoplot}()},
\code{\link{stereoplot_guides}()},
\code{\link{text.spherical}()}
}
\concept{stereo-plot}
