% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{strabo}
\alias{strabo}
\alias{read_strabo_xls}
\alias{read_strabo_mobile}
\alias{read_strabo_JSON}
\title{Import orientation data from \emph{StraboSpot}}
\usage{
read_strabo_xls(file, tag_cols = FALSE, sf = TRUE)

read_strabo_mobile(file, sf = TRUE)

read_strabo_JSON(file, sf = TRUE)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.}

\item{tag_cols}{logical. Whether the Tag columns should be summarized in a
single column (may lead to duplicate rows).}

\item{sf}{logical. Whether the output should be a spatial \code{"sf"} object
using the Longitude and Latitude columns.}
}
\value{
\code{list} containing the following objects:
\describe{
\item{\code{data}}{\code{"tbl_df"} object. Metadata.}
\item{\code{spots}}{\code{"tbl_df"} object or \code{"sf"} if \code{sf == TRUE}. Locations and spot descriptions.}
\item{\code{tags}}{\code{"tbl_df"} object. Tags and their descriptions.}
\item{\code{planar}}{Plane elements. Same row IDs as in \code{data}.}
\item{\code{linear}}{Line elements. Same row IDs as in \code{data}.}
}
}
\description{
Reads the \code{XLS} format export of field book data, the \code{JSON} project file, or
the \code{txt} export of StraboMobile data from \url{strabospot.org/my_data} and
creates a list with the metadata, and the line or plane orientations.
}
\examples{
\dontrun{
# import from excel file
read_strabo_xls("path/to/my/file.xlsx")

# import from text file
read_strabo_mobile("path/to/my/file.txt")

# import from .json file
read_strabo_JSON("path/to/my/file.json")
}
}
