% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{strabo}
\alias{strabo}
\alias{read_strabo_xls}
\alias{read_strabo_mobile}
\alias{read_strabo_JSON}
\title{Import orientation data from \emph{StraboSpot}}
\usage{
read_strabo_xls(file, tag_cols = FALSE, sf = TRUE)

read_strabo_mobile(file, sf = TRUE)

read_strabo_JSON(file, sf = TRUE)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.}

\item{tag_cols}{logical. Whether the Tag columns should be summarized in a
single column (may lead to duplicate rows).}

\item{sf}{logical. Whether the output should be a spatial \code{"sf"} object
using the Longitude and Latitude columns.}

\item{dataset}{character. name of the dataset extracted from \code{JSON} project
file.}
}
\value{
\code{list} containing the following objects:
\describe{
\item{\code{data}}{\code{"tbl_df"} object or \code{"sf"} if \code{sf == TRUE}. Metadata.}
\item{\code{tags}}{\code{"tbl_df"} object. Tags and their descriptsions.}
\item{\code{planar}}{Plane elements. Same row IDs as in \code{data}.}
\item{\code{linear}}{Line elements. Same row IDs as in \code{data}.}
}
}
\description{
Reads the \code{XLS} format export of field book data, the \code{JSON} project file, or
the \code{txt} export of StraboMobile data from \url{strabospot.org/my_data} and
creates a list with the metadata, and the line or plane orientations.
}
\examples{
\dontrun{
file <- "E:/Lakehead/Field Work/StraboSpot_Output_11_01_2022_TS.xlsx"
dt <- read_strabo_xls(file)
stereoplot()
stereo_greatcircle(dt$planar, col = "lightgrey")
stereo_point(dt$linear)

read_strabo_mobile("C:/Users/tobis/Downloads/StraboSpot_Search_06_16_2023.txt")

read_strabo_JSON("G:/My Drive/Moss_Lake/data.json")
}
}
