% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strabo_prj}
\alias{strabo_prj}
\title{Example of StraboSpot Project}
\format{
An object of class \code{strabo}
}
\usage{
data('strabo_prj')
}
\description{
Strabospot project containing 347 foliation and stretching lineation
measurements from the Shebandowan Greenstone Belt and
Quetico Subprovince (Stephan et al., 2025)
}
\examples{
data("strabo_prj")
head(strabo_prj)
}
\references{
Stephan, T., Phillips, N., Tiitto, H., Perez, A., Nwakanma, M.,
Creaser, R., & Hollings, P. (2025). Going with the flow - Changes of
vorticity control gold enrichment in Archean shear zones
(Shebandowan Greenstone Belt, Superior Province, Canada).
Journal of Structural Geology, 201, 105542. \doi{10.1016/j.jsg.2025.105542}
}
\seealso{
Other datasets:
\code{\link{angelier1990}},
\code{\link{example_lines}},
\code{\link{example_lines_df}},
\code{\link{example_planes}},
\code{\link{example_planes_df}},
\code{\link{gray_example}},
\code{\link{holst}},
\code{\link{hossack1968}},
\code{\link{osmundsen2010}},
\code{\link{ramsay}},
\code{\link{shebandowan}},
\code{\link{simongomez}}
}
\concept{datasets}
\keyword{datasets}
