% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortensor.R
\name{strain_shape}
\alias{strain_shape}
\alias{principal_stretch}
\alias{principal_stretch.default}
\alias{principal_stretch.Vec3}
\alias{principal_stretch.Line}
\alias{principal_stretch.Ray}
\alias{principal_stretch.Plane}
\alias{principal_stretch.ortensor}
\alias{principal_stretch.ellipsoid}
\alias{principal_strain}
\alias{shape_params}
\alias{shape_params.default}
\alias{shape_params.spherical}
\alias{shape_params.ortensor}
\alias{shape_params.ellipsoid}
\title{Principal Stretches, Strain and Shape Parameters based on the Orientation Tensor.}
\usage{
principal_stretch(x, ...)

\method{principal_stretch}{default}(x, ...)

\method{principal_stretch}{Vec3}(x, ...)

\method{principal_stretch}{Line}(x, ...)

\method{principal_stretch}{Ray}(x, ...)

\method{principal_stretch}{Plane}(x, ...)

\method{principal_stretch}{ortensor}(x, ...)

\method{principal_stretch}{ellipsoid}(x, ...)

principal_strain(x)

shape_params(x, ...)

\method{shape_params}{default}(x, ...)

\method{shape_params}{spherical}(x, ...)

\method{shape_params}{ortensor}(x, ...)

\method{shape_params}{ellipsoid}(x, ...)
}
\arguments{
\item{x}{object of class \code{"ortensor"}, \code{"ellipsoid"}, or
\code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"} where the rows are the
observations and the columns are the coordinates.}

\item{...}{optional parameters passed to \code{\link[=ortensor]{ortensor()}}}
}
\value{
list
}
\description{
Principal Stretches, Strain and Shape Parameters based on the Orientation Tensor.
}
\details{
\describe{
\item{\code{stretch_ratios}}{Sqrt of eigenvalue ratios}
\item{\code{strain_ratios}}{Log of stretch ratios}
\item{\code{Ramsay}}{strain symmetry (Ramsay, 1983)}
\item{\code{Woodcock}}{Woodcock shape}
\item{\code{Flinn}}{Flinn strain intensity}
\item{\code{Nadai}}{natural octahedral unit strain and shear (Nadai, 1963)}
\item{\code{Lisle_intensity}}{Intensity index (Lisle, 1985)}
\item{\code{Waterson_intensity}}{strain intensity (Watterson, 1968)}
\item{\code{lode}}{Lode parameter (Lode, 1926)}
\item{\code{kind}}{Descriptive type of ellipsoid: \code{"O"} - isotrope, \code{"L"} - L-tectonite, \code{"LLS"} - oblate L-tectonite, \code{"S"} - S-tectonite, \code{"SSL"} - prolate S-tectonite}
\item{\code{MAD}}{maximum angular deviation (Kirschvink, 1980)}
}
}
\examples{
set.seed(1)
mu <- Line(120, 50)
x <- rvmf(100, mu = mu, k = 20)
principal_stretch(x)
principal_strain(x)
shape_params(x)
}
\references{
Flinn, Derek.(1963): "On the statistical analysis of fabric diagrams." Geological Journal 3.2: 247-253.

Jelinek, Vit. "Characterization of the magnetic fabric of rocks." Tectonophysics 79.3-4 (1981): T63-T67.

Kirschvink, J. (1980): The least-squares line and plane and the analysis of palaeomagnetic data. Geophysical Journal International, 62(3), 699-718.

Lisle, Richard J.  (1985): "The use of the orientation tensor for the description and statistical testing of fabrics." Journal of Structural Geology 7.1: 115-117.

Lode, Walter (1926): "Versuche über den Einfluß der mittleren Hauptspannung auf das Fließen der Metalle Eisen, Kupfer und Nickel“
(\emph{"Experiments on the influence of the mean principal stress on the flow of the metals iron, copper and nickel"}], Zeitschrift für Physik, vol. 36 (November), pp. 913–939, DOI: 10.1007/BF01400222

Mardia, Kantilal Varichand. (1975): "Statistics of directional data." Journal of the Royal Statistical Society Series B: Statistical Methodology 37.3: 349-371.

Nadai, A., and Hodge, P. G., Jr. (1963): "Theory of Flow and Fracture of Solids, vol. II." ASME. J. Appl. Mech. December 1963; 30(4): 640. https://doi.org/10.1115/1.3636654

Ramsay, John G. (1967): "Folding and fracturing of rocks." Mc Graw Hill Book Company 568.

Vollmer, Frederick W. (1990): "An application of eigenvalue methods to structural domain analysis." Geological Society of America Bulletin 102.6: 786-791.

Vollmer, Frederick W. (2020): "Representing Progressive Fabric Paths on a Triangular Plot Using a Fabric Density Index and Crystal Axes Eigenvector Barycenters." Geological Society of America Abstracts. Vol. 52.

Watterson, Juan. (1968): "Homogeneous deformation of the gneisses of Vesterland, south-west Greenland". No. 78. CA Reitzel.

Woodcock, N. H.  (1977): "Specification of fabric shapes using an eigenvalue method." Geological Society of America Bulletin 88.9: 1231-1236.
}
\seealso{
\code{\link[=vollmer]{vollmer()}} for Vollmer 1990 shape parameters of the orientation tensor.
\code{\link[=ellipsoid]{ellipsoid()}} class

More details on shape parameters: \code{\link[=lode]{lode()}}, \code{\link[=nadai]{nadai()}},
\code{\link[=jelinek]{jelinek()}}, \code{\link[=flinn]{flinn()}}. Details on eigenvectors from orientation
tensors: \code{\link[=ot_eigen]{ot_eigen()}}.

Other ortensor: 
\code{\link{ortensor}()},
\code{\link{ot_eigen}()}
}
\concept{ortensor}
