% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{strain_shape}
\alias{strain_shape}
\alias{principal_stretch}
\alias{principal_strain}
\alias{or_shape_params}
\title{Principal stretches, strain and shape parameters based on the orientation tensor.}
\usage{
principal_stretch(x)

principal_strain(x)

or_shape_params(x)
}
\arguments{
\item{x}{numeric. Can be three element vector, three column array, or an
object of class \code{"line"} or \code{"plane"}}
}
\description{
Principal stretches, strain and shape parameters based on the orientation tensor.
}
\examples{
set.seed(1)
mu <- Line(120, 50)
x <- rvmf(100, mu = mu, k = 20)
principal_stretch(x)
principal_strain(x)
or_shape_params(x)
}
