% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohR.R
\name{stress-comp}
\alias{stress-comp}
\alias{diff_stress}
\alias{mean_stress}
\alias{shear_stress}
\alias{normal_stress}
\alias{fracture_angle}
\alias{slip_tendency}
\alias{dilatation_tendency}
\title{Stress components}
\usage{
diff_stress(sigma1, sigma3)

mean_stress(sigma1, sigma3)

shear_stress(sigma1, sigma3, theta)

normal_stress(sigma1, sigma3, theta)

fracture_angle(mu)

slip_tendency(sigma_s, sigma_n)

dilatation_tendency(sigma1, sigma3, sigma_n)
}
\arguments{
\item{sigma1, sigma3}{numeric. Magnitudes of maximum and minimum principal stress (\eqn{\sigma_1} and \eqn{\sigma_3}), respectively.}

\item{theta}{numeric. Angle \eqn{\theta} between fracture and \eqn{\sigma_1}.}

\item{mu}{numeric. Coefficient of internal friction \eqn{\mu}.}

\item{sigma_s, sigma_n}{numeric. Magnitudes of shear and normal stress (\eqn{\sigma_s} and \eqn{\sigma_n}), respectively.}
}
\value{
numeric
}
\description{
Calculates some stress components
}
\examples{
s1 <- 1025
s3 <- 250

diff_stress(s1, s3)
mean_stress(s1, s3)
ss <- shear_stress(s1, s3, theta = 35); print(ss)
sn <- normal_stress(s1, s3, theta = 35); print(sn)
fracture_angle(mu = 0.6)

slip_tendency(ss, sn)
dilatation_tendency(s1, s3, sn)
}
