% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_inversion.R
\name{stress_inversion}
\alias{stress_inversion}
\title{Fault slip stress inversion}
\usage{
stress_inversion(x, friction = 0.6)
}
\arguments{
\item{x}{\code{"Fault"} object}

\item{friction}{numeric. Value(s) for coefficient of friction}
}
\value{
list
}
\description{
Michael (1984) method
}
\examples{
stress_inversion(angelier1990$TYM)
stress_inversion(angelier1990$AVB)
}
\references{
Michael, A. J. (1984). Determination of stress from slip data: Faults and folds. Journal of Geophysical Research: Solid Earth, 89(B13), 11517–11526. https://doi.org/10.1029/JB089iB13p11517
}
