% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/structr.R
\docType{package}
\name{structr}
\alias{structr}
\alias{structr-package}
\title{library(structr)}
\description{
Free and open-source R package for analyzing and visualizing
orientation and stress data for structural geology.

Free and open-source R package for analyzing and visualizing orientation and
stress data for structural geology.
}
\details{
A list of documented functions may be viewed by typing
\code{help(package='structr')}.

A list of documented functions may be viewed by typing
\code{help(package="structr")}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tobiste.github.io/structr/}
}


Useful links:
\itemize{
  \item \url{https://tobiste.github.io/structr/}
}

}
\author{
Tobias Stephan
}
