% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stats.R
\name{summary.spherical}
\alias{summary.spherical}
\title{Summary statistics}
\usage{
\method{summary}{spherical}(x, ...)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, or \code{"Plane"}.}

\item{...}{parameters passed to \code{\link[=mean]{mean()}}, \code{\link[=sd]{sd()}},}
}
\description{
Lists the mean, vartiance, 68\% cone, and the confidence cone around the mean.
}
