% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stats.R
\name{summary.spherical}
\alias{summary.spherical}
\title{Summary statistics}
\usage{
\method{summary}{spherical}(object, ...)
}
\arguments{
\item{...}{parameters passed to \code{\link[=sph_mean]{sph_mean()}}, \code{\link[=sph_var]{sph_var()}}, \code{\link[=delta]{delta()}}, and \code{\link[=sph_confidence_angle]{sph_confidence_angle()}}}
}
\value{
named vector
}
\description{
Calculates the arithmetic mean, variance, 68\% cone, and the confidence cone around the mean.
}
\examples{
set.seed(20250411)
summary(rvmf(100, mu = Line(90, 20), k = 20))
}
