% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohr.R
\name{tauMax}
\alias{tauMax}
\title{Maximum in-plane shear stress}
\usage{
tauMax(sigmaX, sigmaZ, tauXZ)
}
\arguments{
\item{sigmaX}{numeric. Magnitude of normal stress acting in the horizontal direction}

\item{sigmaZ}{numeric. Magnitude of normal stress acting in the vertical direction}

\item{tauXZ}{numeric. Magnitude of shear stress acting on the same plane as \code{"sigmaX"}}
}
\value{
A two-element list containing
\describe{
\item{\code{"tauMax"}}{maximum in-plane shear stress}
\item{\code{"theta"}}{angle of maximum in-plane shear stress (in degrees)}
}
}
\description{
calculates the magnitude and direction of the maximum in-plane shear stress
}
\examples{
tauMax(sigmaX = 80, sigmaZ = 120, tauXZ = 20)
}
\author{
Kyle Elmy and Jim Kaklamanos
}
