% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet_base.R
\name{text.spherical}
\alias{text.spherical}
\title{Add Points to a Plot}
\usage{
\method{text}{spherical}(x, labels = seq_along(x[, 1]), upper.hem = FALSE, earea = TRUE, ...)
}
\arguments{
\item{x}{objects of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"}, \code{"Pair"}, or \code{"Fault"}.}

\item{labels}{a character vector or \link{expression} specifying
    the \emph{text} to be written.  An attempt is made to coerce other
    language objects (names and calls) to expressions, and vectors and
    other classed objects to character vectors by \code{\link{as.character}}.
    If \code{labels} is longer than \code{x} and
    \code{y}, the coordinates are recycled to the length of \code{labels}.}

\item{upper.hem}{logical. Whether the projection is shown for upper
hemisphere (\code{TRUE}) or lower hemisphere (\code{FALSE}, the default).}

\item{earea}{logical. Projection, either \code{TRUE} for Lambert equal-area
projection (the default), or \code{FALSE} for meridional stereographic projection.}

\item{...}{arguments passed to \code{\link[graphics:text]{graphics::text()}}}
}
\description{
Add Points to a Plot
}
\examples{
stereoplot()
points(Line(c(90, 80), c(10, 75)), col = 1:2)
text(Line(c(90, 80), c(10, 75)), labels = c("L1", "L2"), col = 1:2, pos = 3)
}
