% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vec_list}
\alias{vec_list}
\title{List of vectors}
\usage{
vec_list(x)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, or \code{"Plane"}.}
}
\value{
list
}
\description{
Creates a list of Cartesian vectors  from an spherical objects. This is a convenience
function to link with the package \code{geologyGeometry} by J. R. Davis
}
\examples{
vec_list(example_lines) 
}
