% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_math.R
\name{vecmath}
\alias{vecmath}
\alias{crossprod.spherical}
\alias{\%*\%.spherical}
\alias{rotate.spherical}
\alias{angle.spherical}
\alias{project.spherical}
\alias{reject.spherical}
\alias{transform.spherical}
\title{Vector math operations}
\usage{
\method{crossprod}{spherical}(x, y)

\method{\%*\%}{spherical}(x, y)

\method{rotate}{spherical}(x, rotaxis, rotangle)

\method{angle}{spherical}(x, y)

\method{project}{spherical}(x, y)

\method{reject}{spherical}(x, y)

\method{transform}{spherical}(x, A, norm = FALSE)
}
\arguments{
\item{x, y}{objects of class \code{"Vec3"}, \code{"Line"}, or \code{"Plane"}.}

\item{rotaxis}{Axis of rotation given as object of class \code{"Vec3"}, \code{"Line"}, or \code{"Plane"}.}

\item{rotangle}{Angle of rotation in radians for \code{"Vec3"} objects and in degrees for \code{"Line"} and \code{"Plane"} objects.}

\item{A}{numeric matrix. Transformation matrix.}

\item{norm}{logical. If \code{TRUE}, the transformed vectors are normalized to unit length.}
}
\value{
objects of class \code{"Vec3"}, \code{"Line"}, or \code{"Plane"}
}
\description{
Vector math operations
}
\examples{
vec1 <- Vec3(1, 0, 0)
vec2 <- Vec3(0, 0, 1)

vlength(vec1)
vsum(vec1)
vnorm(vec1)
crossprod(vec1, vec2)
vdot(vec1, vec2)
vec1 \%*\% vec2
rotate(vec1, vec2, pi / 2)
angle(vec1, vec2)
project(vec1, vec2)
reject(vec1, vec2)
}
