% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_math.R
\name{vecmath}
\alias{vecmath}
\alias{vector_length}
\alias{vector_norm}
\alias{crossprod.Vec3}
\alias{crossprod.Line}
\alias{crossprod.Ray}
\alias{crossprod.Plane}
\alias{dotprod}
\alias{rotate}
\alias{rotate.spherical}
\alias{angle}
\alias{angle.spherical}
\alias{project}
\alias{project.spherical}
\alias{reject}
\alias{reject.default}
\alias{transform_linear}
\title{Vector math operations}
\usage{
vector_length(x)

vector_norm(x)

\method{crossprod}{Vec3}(x, y = NULL, ...)

\method{crossprod}{Line}(x, y = NULL, ...)

\method{crossprod}{Ray}(x, y = NULL, ...)

\method{crossprod}{Plane}(x, y = NULL, ...)

dotprod(x, y)

rotate(x, rotaxis, rotangle)

\method{rotate}{spherical}(x, rotaxis, rotangle)

angle(x, y)

\method{angle}{spherical}(x, y)

project(x, y)

\method{project}{spherical}(x, y)

reject(x, y)

\method{reject}{default}(x, y)

transform_linear(x, A, norm = FALSE)
}
\arguments{
\item{x, y}{objects of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}.}

\item{...}{arguments passed to function call}

\item{rotaxis}{Axis of rotation given as object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}.}

\item{rotangle}{Angle of rotation in radians for \code{"Vec3"} objects and in degrees for \code{"Line"}, \code{"Ray"} and \code{"Plane"} objects.}

\item{A}{numeric 3x3 matrix. Transformation matrix.}

\item{norm}{logical. If \code{TRUE}, the transformed vectors are normalized to unit length.}
}
\value{
objects of same class as \code{x}, i.e. one of \code{"Vec3"}, \code{"Line"}, \code{"Ray"} or
\code{"Plane"}. \code{vector_length()} and \code{\%*\%} return a real number. \code{angle()}
returns a numeric angle (in degrees, unless \code{x} is class \code{"Vec3"}).
}
\description{
Vector math operations
}
\details{
\describe{
\item{\code{vector_length()}}{the length of a vector: \eqn{||x|| = \sqrt{x_1^2 + x_2^2 + x_3^2}}}
\item{\code{vector_norm()}}{the normalized vector: \eqn{\hat{x} = \frac{x}{||x||}}}
\item{\code{crossprod()}}{the cross-product of two vectors, i.e. the vector perpendicular to the 2 vectors. If \code{y = NULL} is taken to be the sam,e vector as \code{x}: \deqn{x \times y = (x_2 y_3 - x_3 y_2, x_3 y_1 - x_1 y_3, x_1 y_2 - x_2 y_1)}}.
\item{\code{dotprod()}}{the dot product of two vectors: \eqn{x \cdot y = x_1 y_1 + x_2 y_2 + x_3 y_3}}
\item{\code{angle()}}{angle between two vectors:  \eqn{\theta = \arccos{\frac{x \cdot y}{||x|| \, ||y||}}}}
\item{\code{project()}}{projection of one vector onto the other (changes the vector length of second vector, unless their are unit vectors):
\deqn{proj_y(x) = \frac{x \cdot y}{||y||^2} y}}
\item{\code{transform_linear()}}{Linear transformation of a vector by a 3x3 matrix: \eqn{x' = A x}}
}
}
\examples{
vec1 <- Vec3(1, 0, 0)
vec2 <- Vec3(0, 0, 1)

vector_length(vec1) # length of a vector
crossprod(vec1, vec2) # cross product
dotprod(vec1, vec2) # dot product
rotate(vec1, vec2, pi / 2) # rotation
angle(vec1, vec2) # angle between vectors
project(vec1, vec2) # projection of a vector
transform_linear(vec1, matrix(runif(9), 3, 3)) # linear transformation
}
