% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deftensor.R
\name{vel_rate}
\alias{vel_rate}
\alias{velgrad_rate}
\alias{velgrad_spin}
\title{Rate and spin of velocity gradient tensor}
\usage{
velgrad_rate(x)

velgrad_spin(x)
}
\arguments{
\item{x}{3x3 matrix. Velocity gradient tensor.}
}
\value{
3x3 matrix
}
\description{
Rate and spin of velocity gradient tensor
}
\examples{
R <- defgrad_from_comp(xx = 2, xy = 1, zz = 0.5)
L <- velgrad_from_defgrad(R, time = 10)
velgrad_rate(L)
velgrad_spin(L)
}
