% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_plots.R
\name{VollmerPlot}
\alias{VollmerPlot}
\title{Fabric plot of Vollmer (1990)}
\usage{
VollmerPlot(x, lab = NULL, add = FALSE, grid = TRUE, increment = 20, ...)
}
\arguments{
\item{x}{numeric. Can be three element vector, three column array, or an object of class \code{"line"} or \code{"plane"}}

\item{lab}{character. text labels}

\item{add}{logical. Should data be plotted to an existing plot?}

\item{grid}{logical. Should the ternary diagram have grid lines?}

\item{increment}{Spacing of grid lines, if \code{grid=TRUE}. Ignored otherwise}

\item{...}{optional graphical parameters}
}
\description{
Creates a fabric plot using the eigenvalue method
}
\examples{
set.seed(1)
mu <- Line(120, 50)
x <- rvmf(100, mu = mu, k = 1)
VollmerPlot(x, lab = "x")
y <- rvmf(100, mu = mu, k = 20)
VollmerPlot(y, lab = "y", add = TRUE, col = "red")
title("Fabric plot of Vollmer (1990)")
}
\references{
Vollmer, F. W. (1990). An application of eigenvalue methods to structural domain analysis. Geological Society of America Bulletin, 102, 786–791.
}
\seealso{
\code{\link[=WoodcockPlot]{WoodcockPlot()}}, \code{\link[=fabric_indexes]{fabric_indexes()}}
}
