% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sph_distributions.R
\name{vonmises-fisher}
\alias{vonmises-fisher}
\alias{rvmf}
\alias{dvmf}
\title{von Mises-Fisher Distribution}
\source{
Adapted fom \code{\link[rotasym:vMF]{rotasym::r_vMF()}} and \code{\link[rotasym:vMF]{rotasym::d_vMF()}}
}
\usage{
rvmf(n = 100, mu = Vec3(1, 0, 0), k = 5)

dvmf(x, mu, k = 5)
}
\arguments{
\item{n}{integer. number of random samples to be generated}

\item{mu}{Mean vector. object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}, where the
rows are the observations and the columns are the coordinates.}

\item{k}{numeric. The concentration parameter (\eqn{\kappa}) of the von
Mises-Fisher distribution}

\item{x}{object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}, where the
rows are the observations and the columns are the coordinates.}
}
\description{
Density and random generation for the spherical normal distribution with mean
and concentration parameter (\eqn{\kappa}) .
}
\examples{
set.seed(20250411)
x <- rvmf(100, mu = Ray(120, 50), k = 5)
contour(x)

dx <- dvmf(x, mu = Ray(120, 50))
head(dx)

plot(x, col = assign_col(dx))
}
\seealso{
Other random: 
\code{\link{rbing}},
\code{\link{rfb}()},
\code{\link{rkent}()},
\code{\link{rrot}()},
\code{\link{runif.spherical}()}
}
\concept{random}
