% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{vonmises-fisher}
\alias{vonmises-fisher}
\alias{rvmf}
\alias{dvmf}
\title{The von Mises-Fisher Distribution}
\usage{
rvmf(n = 100, mu = c(1, 0, 0), k = 5)

dvmf(x, mu, k = 5)
}
\arguments{
\item{n}{integer. number of random samples to be generated}

\item{k}{numeric.  the concentration parameter (kappa) of the the von Mises-Fisher distributiuon}

\item{x, mu}{numeric. Can be three element vector, three column matrix, or an object of class \code{"line"} or \code{"plane"}}
}
\description{
Density and random generation for the spherical normal distribution with mean and kappa.
}
\examples{
# example code
x <- rvmf(100, mu = Line(120, 50), k = 5)
stereoplot()
stereo_point(x)
}
