% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sph_distribution.R
\name{vonmises-fisher}
\alias{vonmises-fisher}
\alias{rvmf}
\alias{dvmf}
\title{von Mises-Fisher Distribution}
\source{
Adapted fom \code{\link[rotasym:vMF]{rotasym::r_vMF()}} and \code{\link[rotasym:vMF]{rotasym::d_vMF()}}
}
\usage{
rvmf(n = 100, mu = c(1, 0, 0), k = 5)

dvmf(x, mu, k = 5)
}
\arguments{
\item{n}{integer. number of random samples to be generated}

\item{k}{numeric. The concentration parameter (\eqn{\kappa}) of the the von
Mises-Fisher distribution}

\item{x, mu}{numeric. Can be three element vector, three column array, or an
object of class \code{"line"} or \code{"plane"}}
}
\description{
Density and random generation for the spherical normal distribution with mean
and kappa.
}
\examples{
set.seed(20250411)
x <- rvmf(100, mu = Line(120, 50), k = 5)
stereoplot()
stereo_point(x)
}
\seealso{
\code{\link[=v_unif]{v_unif()}} for alternative algorithms to generate uniform
distributed samples on a sphere, \code{\link[=rkent]{rkent()}} for Kent distribution.
}
