% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_math.R
\name{vslerp}
\alias{vslerp}
\title{Spherical Linear Interpolation (Slerp)}
\usage{
vslerp(x, y, t, na.rm = TRUE)
}
\arguments{
\item{x, y}{objects of class \code{"Vec3"}, \code{"Line"}, or \code{"Plane"}.}

\item{t}{numeric. interpolation factor (\verb{t = [0, 1]}).}

\item{na.rm}{logical. Whether \code{NA} values should be removed before the computation proceeds.}
}
\description{
Returns the spherical linear interpolation of points between two vectors
}
\details{
A Slerp path is the spherical geometry equivalent of a path along a line
segment in the plane; a great circle is a spherical geodesic.
}
\note{
For non-unit vectors the interpolation is not uniform.
}
\examples{
t <- seq(0, 1, .05)
res <- sapply(t, function(i){
vslerp(Line(120, 55), Line(0, 90), t = i) |> unclass()
}) |> t() |> rbind() |> as.Line()

plot(res, col = assign_col(t))
points(rbind(Line(120, 55),  Line(10, 80)))
}
