% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{vslerp}
\alias{vslerp}
\title{Spherical Linear Interpolation (Slerp)}
\usage{
vslerp(x, y, t)
}
\arguments{
\item{x, y}{numeric. Can be three element vector, three column array, or an
object of class \code{"line"} or \code{"plane"}}

\item{t}{numeric. interpolation factor (\verb{t = [0, 1]}).}
}
\description{
Returns the spherical linear interpolation of points between two vectors
}
\details{
A Slerp path is the spherical geometry equivalent of a path along a line
segment in the plane; a great circle is a spherical geodesic.
}
\note{
For non-unit vectors the interpolation is not uniform
}
