% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{vtransform}
\alias{vtransform}
\title{Affine transformation of vector by matrix}
\usage{
vtransform(x, A, norm = TRUE)
}
\description{
Affine transformation of vector by matrix
}
\examples{
mat <- cbind(V1 = c(1, 0, 0), V2 = c(0, 1, 0), V3 = c(0, 0, -1))
vec <- c(1, 1, 1)
vtransform(vec, mat)
}
