% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{watson-inference}
\alias{watson-inference}
\alias{watson_inference}
\alias{watson_inference.Vec3}
\alias{watson_inference.Line}
\alias{watson_inference.Plane}
\title{One-sample inference about the mean of the Watson distribution.}
\usage{
watson_inference(x, alpha, shape)

\method{watson_inference}{Vec3}(x, alpha = 0.05, shape = NULL)

\method{watson_inference}{Line}(x, alpha = 0.05, shape = NULL)

\method{watson_inference}{Plane}(x, alpha = 0.05, shape = NULL)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, or \code{"Plane"}, where the rows are the observations and the columns are the coordinates.}

\item{alpha}{A real number, between 0 and 1. The significance level for the confidence region.}

\item{shape}{\code{NULL} or character, either 'bipolar' or 'girdle'. If \code{NULL}, then
this function chooses automatically.}
}
\value{
A list with members \verb{$shape}, \verb{$tBar}, \verb{$rhs}, \verb{$pvalue}.
\describe{
\item{\code{shape}}{is either 'bipolar' or 'girdle'. If 'bipolar', then the confidence region consists of
all lines u such that \verb{u^T \%*\% $tBar \%*\% u > $rhs}. If 'girdle', then the
confidence region consists of all lines u such that \verb{u^T \%*\% $tBar \%*\% u < $rhs}.}
\item{\code{tBar}}{orientation tesor}
\item{\code{rhs}}{}
\item{\code{pvalue}}{is an R function that takes as input a line u0 and produces as output
a real number in \verb{[0, 1]} --- the p-value for the null hypothesis that the
Watson mean is u0.}
}
}
\description{
Assumes large concentration --- either kappa >> 0 or kappa << 0. From Mardia and Jupp (2000, Section 10.7.3).
}
\examples{
r <- watson_inference(example_lines)
print(r)

r$pvalue(Line(60, 10))
}
