% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{watson-mle}
\alias{watson-mle}
\alias{watson_MLE}
\alias{watson_MLE.Vec3}
\alias{watson_MLE.Line}
\alias{watson_MLE.Plane}
\title{Maximum Likelihood Estimation of the Watson Distribution Parameters.}
\usage{
watson_MLE(x, shape)

\method{watson_MLE}{Vec3}(x, shape = NULL)

\method{watson_MLE}{Line}(x, shape = NULL)

\method{watson_MLE}{Plane}(x, shape = NULL)
}
\arguments{
\item{x}{object of class \code{"Vec3"}, \code{"Line"}, or \code{"Plane"}, where the rows are the observations and the columns are the coordinates.}

\item{shape}{\code{NULL} or character, either 'bipolar' or 'girdle'. If \code{NULL}, then
this function chooses automatically.}
}
\value{
A list with members
\describe{
\item{\code{muHat}}{a line, the MLE of the mean (identical to \code{\link[=projected_mean]{projected_mean()}}
if x has a bipolar shape),}
\item{\code{kappaHat}}{a real number, the MLE of the concentration,}
\item{\code{shape}}{character, either \code{'bipolar'} or \code{'girdle'},}
\item{\code{d3}}{a positive real number, the D3 from which \code{kappaHat} was
computed, and}
\item{\code{eigenvalues}}{(the eigenvalues of the \eqn{\bar{T}} matrix
(orientation tensor), in descending order (see \code{\link[=ot_eigen]{ot_eigen()}}).}
}
}
\description{
MLE parameters describing a Watson distribution for isotropic axial vectors.
From Mardia and Jupp (2000, Section 10.3.2).
}
\examples{
r <- watson_MLE(example_lines)
print(r)

plot(example_lines)
points(r$muHat, col = 'red', pch = 16, cex = 1.5)
}
\seealso{
\code{\link[=watson_inference]{watson_inference()}} for confidence regions, and \code{\link[=rwatson]{rwatson()}} to
simulate a distribution.

Other distribution-MLE: 
\code{\link{bingham-mle}},
\code{\link{dist.mle}},
\code{\link{fisher-mle}}
}
\concept{distribution-MLE}
