% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_plots.R
\name{woodcock_plot}
\alias{woodcock_plot}
\title{Fabric Plot of Woodcock (1977)}
\usage{
woodcock_plot(
  x,
  labels = NULL,
  add = FALSE,
  max = 7,
  main = "Woodcock diagram",
  weights = NULL,
  guides = NULL,
  ...
)
}
\arguments{
\item{x}{either an object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, \code{"Plane"},
\code{"Pair"}, or \code{"Fault"} where the
rows are the observations and the columns are the coordinates, or an \code{"ortensor"} object.}

\item{labels}{character. text labels}

\item{add}{logical. Should data be plotted to an existing plot?}

\item{max}{numeric. Maximum value for x and y axes. If \code{NULL}, it is calculated from the data.}

\item{main}{character. The main title for the plot.}

\item{weights}{numeric. Weightings}

\item{guides}{logical. Whether guides of the projection grid should be added
to the plot. Defaults to \code{getOption("structr.guides")}.}

\item{...}{optional graphical parameters}
}
\value{
A plot and when stored as an object, the orientation tensor's
eigenvalues and eigenvectors as a list.
}
\description{
Creates a fabric plot using the eigenvalue method
}
\examples{
set.seed(20250411)
mu <- Line(120, 50)
x <- rvmf(100, mu = mu, k = 1)
woodcock_plot(x, lab = "x")
y <- rvmf(100, mu = mu, k = 20)
woodcock_plot(y, lab = "y", add = TRUE, col = "red")
}
\references{
Woodcock, N. H. (1977). Specification of fabric shapes using an eigenvalue method. Geological Society of America Bulletin88, 1231<U+2013>1236.
}
\seealso{
\code{\link[=vollmer]{vollmer()}}, \code{\link[=ot_eigen]{ot_eigen()}}

Other fabric-plot:
\code{\link[=flinn_plot]{flinn_plot()}},
\code{\link[=hsu_plot]{hsu_plot()}},
\code{\link{vollmer-plot}}
}
\concept{fabric-plot}
