% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_plots.R
\name{woodcock_plot}
\alias{woodcock_plot}
\title{Fabric plot of Woodcock (1977)}
\usage{
woodcock_plot(x, labels = NULL, add = FALSE, max = 7, ...)
}
\arguments{
\item{x}{either an object of class \code{"Vec3"}, \code{"Line"}, \code{"Ray"}, or \code{"Plane"}, where the
rows are the observations and the columns are the coordinates, or an \code{"ortensor"} object.}

\item{labels}{character. text labels}

\item{add}{logical. Should data be plotted to an existing plot?}

\item{max}{numeric. Maximum value for x and y axes. If \code{NULL}, it is calculated from the data.}

\item{...}{optional graphical parameters}
}
\value{
A plot and when stored as an object, the orientation tensor's
eigenvalues and eigenvectors as a list.
}
\description{
Creates a fabric plot using the eigenvalue method
}
\examples{
set.seed(20250411)
mu <- Line(120, 50)
x <- rvmf(100, mu = mu, k = 1)
woodcock_plot(x, lab = "x", main = "Fabric plot of Woodcock (1977)")
y <- rvmf(100, mu = mu, k = 20)
woodcock_plot(y, lab = "y", add = TRUE, col = "red")
}
\references{
Woodcock, N. H. (1977). Specification of fabric shapes using an eigenvalue method. Geological Society of America Bulletin88, 1231<U+2013>1236. http://pubs.geoscienceworld.org/gsa/gsabulletin/article-pdf/88/9/1231/3418366/i0016-7606-88-9-1231.pdf
}
\seealso{
\code{\link[=fabric_indexes]{fabric_indexes()}}, \code{\link[=ot_eigen]{ot_eigen()}}

Other fabric-plot: 
\code{\link{flinn_plot}()},
\code{\link{hsu_plot}()},
\code{\link{vollmer_plot}()}
}
\concept{fabric-plot}
